Cardano.Ledger.Shelley

Documentation

type Shelley = ShelleyEra

Deprecated: In favor of ShelleyEra

data ShelleyEra

data ShelleyTx era

data ShelleyTxOut era

data family TxBody era

The body of a transaction.

data family Tx era

data ShelleyTxAuxData era

nativeMultiSigTag ∷ ByteString

Magic number "memorialized" in the ValidateScript class under the method: scriptPrefixTag:: Core.Script era -> Bs.ByteString, for the Shelley Era.

hardforkAllegraAggregatedRewards ∷ ProtVer → Bool

hardforkAlonzoAllowMIRTransfer ∷ ProtVer → Bool

Starting with protocol version 5, the MIR certs will also be able to transfer funds between the reserves and the treasury. Additionally, the semantics for the pervious functionality will change a bit. Before version 5 redundancies in the instantaneous reward mapping were handled by overriding. Now they are handled by adding the values and allowing for negatives updates, provided the sum for each key remains positive.

hardforkAlonzoValidatePoolRewardAccountNetID ∷ ProtVer → Bool

Starting with protocol version 5, we will validate the network ID for the reward account listed in stake pool registration certificates.

hardforkBabbageForgoRewardPrefilter ∷ ProtVer → Bool

Starting with protocol version 7, the reward calculation no longer filters out unregistered stake addresses at the moment the calculation begins. See the Shelley Ledger Errata 17.2.