Cardano.Ledger.Shelley.API.ByronTranslation

Documentation

translateToShelleyLedgerState :: forall c. (Crypto c, ADDRHASH c ~ Blake2b_224) => ShelleyGenesis (ShelleyEra c) -> EpochNo -> ChainValidationState -> NewEpochState (ShelleyEra c)

Exported for testing purposes

translateCompactTxOutByronToShelley :: CompactTxOut -> TxOut (ShelleyEra c)

translateTxIdByronToShelley :: (Crypto c, ADDRHASH c ~ Blake2b_224) => TxId -> TxId c

We use the same hashing algorithm so we can unwrap and rewrap the bytes. We don't care about the type that is hashed, which will differ going from Byron to Shelley, we just use the hashes as IDs.