Cardano.Ledger.Shelley.API.Genesis

Documentation

class (EraTxOut era, Default (StashedAVVMAddresses era), EraGov era) ⇒ CanStartFromGenesis era where

Deprecated: Use EraTransition instead

Indicates that this era may be bootstrapped from ShelleyGenesis.

Minimal complete definition

fromShelleyPParams

Associated Types

type AdditionalGenesisConfig era ∷ Type

Additional genesis configuration necessary for this era.

type AdditionalGenesisConfig era = ()

Methods

fromShelleyPParams ∷ AdditionalGenesisConfig era → PParams (ShelleyEra (EraCrypto era)) → PParams era

Deprecated: Use tcInitialPParamsG instead

Upgrade PParams from ShelleyEra all the way to the current one.

initialState ∷ ShelleyGenesis (EraCrypto era) → AdditionalGenesisConfig era → NewEpochState era

Deprecated: Use createInitialState instead

Construct an initial state given a ShelleyGenesis and any appropriate AdditionalGenesisConfig for the era.

initialStateFromGenesis

Arguments

∷ ∀ era. CanStartFromGenesis era

⇒ ShelleyGenesis (EraCrypto era)
Genesis type

→ AdditionalGenesisConfig era

→ NewEpochState era

Helper function for constructing the initial state for any era