Cardano.Ledger.Shelley.API.Types

Documentation

data Network

Constructors

Testnet

Mainnet

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo32 !TxIx !CertIx

data Addr

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network PaymentCredential StakeReference

AddrBootstrap BootstrapAddress

data Block h era

Constructors

Block

Fields

blockHeader ∷ !h

blockBody ∷ !(BlockBody era)

type Hash = Hash HASH

type CertifiedVRF c = CertifiedVRF (VRF c)

type VerKeyVRF c = VerKeyVRF (VRF c)

type SignKeyVRF c = SignKeyVRF (VRF c)

type SignedDSIGN = SignedDSIGN DSIGN

type SignKeyDSIGN = SignKeyDSIGN DSIGN

type SignedKES c = SignedKES (KES c)

type VerKeyKES c = VerKeyKES (KES c)

type SignKeyKES c = SignKeyKES (KES c)

data StrictMaybe a

Constructors

SNothing

SJust !a

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

newtype KeyHash (r ∷ KeyRole)

Discriminated hash of public Key

Constructors

KeyHash

Fields

unKeyHash ∷ Hash ADDRHASH (VerKeyDSIGN DSIGN)

data TxIn

The input of a UTxO.

Constructors

TxIn !TxId !TxIx

newtype TxId

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

unTxId ∷ SafeHash EraIndependentTxBody

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map TxIn (TxOut era)

data Metadatum

A generic metadatum type.

Constructors

Map ![(Metadatum, Metadatum)]

List ![Metadatum]

I !Integer

B !ByteString

S !Text

newtype VKey (kd ∷ KeyRole)

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey ∷ VerKeyDSIGN DSIGN

data KeyRole

The role of a key.

All key roles are fixed and unique, except for the Witness role. In particular, keys can be cast to a Witness role with the help of asWitness, because same witness can be valid for many roles.

In fact, it is perfectly allowable for a key to be used in many roles by the end user; there is nothing prohibiting somebody using the same underlying key or a script as their payment and staking credential, as well as the key for their stake pool. However, in the ledger code mixing up keys with different roles could be catastrophic, that is why we have this separation.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

Guard

newtype GenDelegs

Constructors

GenDelegs

Fields

unGenDelegs ∷ Map (KeyHash 'Genesis) GenDelegPair

data GenDelegPair

Constructors

GenDelegPair

Fields

genDelegKeyHash ∷ !(KeyHash 'GenesisDelegate)

genDelegVrfHash ∷ !(VRFVerKeyHash 'GenDelegVRF)

newtype ScriptHash

Constructors

ScriptHash (Hash ADDRHASH EraIndependentScript)

data WitVKey (kr ∷ KeyRole) where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ VKey kr → SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody) → WitVKey kr

data BootstrapWitness

Constructors

BootstrapWitness

Fields

bwKey ∷ !(VKey 'Witness)

bwSignature ∷ !(SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody))

bwChainCode ∷ !ChainCode

bwAttributes ∷ !ByteString

type KESignable c = Signable (KES c)

data CertIx

Certificate index. There is mkCertIxPartial that can be used for testing when constructing from other integral types that are larger than Word16

data TxIx

Transaction index.

data Globals

Constructors

Globals

Fields

epochInfo ∷ !(EpochInfo (Either Text))

slotsPerKESPeriod ∷ !Word64

stabilityWindow ∷ !Word64

The window size in which our chosen chain growth property guarantees at least k blocks. From the paper "Ouroboros praos: An adaptively-secure, semi-synchronous proof-of-stake protocol". The stabilityWindow constant is used in a number of places; for example, protocol updates must be submitted at least twice this many slots before an epoch boundary.

randomnessStabilisationWindow ∷ !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

securityParameter ∷ !(NonZero Word64)

Maximum number of blocks we are allowed to roll back

maxKESEvo ∷ !Word64

Maximum number of KES iterations

quorum ∷ !Word64

Quorum for update system votes and MIR certificates

maxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system

activeSlotCoeff ∷ !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

networkId ∷ !Network

The network ID

systemStart ∷ !SystemStart

System start time

newtype Port

Constructors

Port

Fields

portToWord16 ∷ Word16

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

data ProtVer

Constructors

ProtVer

Fields

pvMajor ∷ !Version

pvMinor ∷ !Natural

data StakeReference

Constructors

StakeRefBase !StakeCredential

StakeRefPtr !Ptr

StakeRefNull

data Credential (kr ∷ KeyRole)

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !ScriptHash

KeyHashObj !(KeyHash kr)

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin ∷ Integer

newtype Withdrawals

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map RewardAccount Coin

data RewardAccount

An account based address for rewards

Constructors

RewardAccount

Fields

raNetwork ∷ !Network

raCredential ∷ !(Credential 'Staking)

data ChainAccountState where

Bundled Patterns

pattern AccountState ∷ Coin → Coin → ChainAccountState

type family AccountState era = (r ∷ Type) | r → era

type family AccountState era = (r ∷ Type) | r → era

data PoolDistr

A map of stake pool IDs (the hash of the stake pool operator's verification key) to IndividualPoolStake. Also holds absolute values necessary for the calculations in the computeDRepDistr.

Constructors

PoolDistr

Fields

unPoolDistr ∷ !(Map (KeyHash 'StakePool) IndividualPoolStake)

pdTotalActiveStake ∷ !(CompactForm Coin)

Total stake delegated to registered stake pools. In addition to the stake considered for the individualPoolStake Rational, we add proposal-deposits to this field.

data PoolParams

A stake pool.

Constructors

PoolParams

Fields

ppId ∷ !(KeyHash 'StakePool)

ppVrf ∷ !(VRFVerKeyHash 'StakePoolVRF)

ppPledge ∷ !Coin

ppCost ∷ !Coin

ppMargin ∷ !UnitInterval

ppRewardAccount ∷ !RewardAccount

ppOwners ∷ !(Set (KeyHash 'Staking))

ppRelays ∷ !(StrictSeq StakePoolRelay)

ppMetadata ∷ !(StrictMaybe PoolMetadata)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl ∷ !Url

pmHash ∷ !ByteString

data PoolCert

Constructors

RegPool !PoolParams
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool) !EpochNo
A stake pool retirement certificate.

data family TxBody era

The body of a transaction.

data SnapShots

Snapshots of the stake distribution.

Note that ssStakeMark and ssStakeMarkPoolDistr are lazy on purpose since we only want to force the thunk after one stability window when we know that they are stable (so that we do not compute them if we do not have to). See more info in the Optimize TICKF ADR

Constructors

SnapShots

Fields

ssStakeMark ∷ SnapShot

ssStakeMarkPoolDistr ∷ PoolDistr

ssStakeSet ∷ !SnapShot

ssStakeGo ∷ !SnapShot

ssFee ∷ !Coin

data SnapShot

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

ssStake ∷ !Stake

ssDelegations ∷ !(VMap VB VB (Credential 'Staking) (KeyHash 'StakePool))

ssPoolParams ∷ !(VMap VB VB (KeyHash 'StakePool) PoolParams)

newtype Stake

Type of stake as map from staking credential to coins associated. Any staking credential that has no stake will not appear in this Map, even if it is registered. For this reason, this data type should not be used for infering whether credential is registered or not.

Constructors

Stake

Fields

unStake ∷ VMap VB VP (Credential 'Staking) (CompactForm Coin)

class (EraAccounts era, ToJSON (CertState era), EncCBOR (CertState era), DecShareCBOR (CertState era), Share (CertState era) ~ (Interns (Credential 'Staking), Interns (KeyHash 'StakePool), Interns (Credential 'DRepRole), Interns (Credential 'HotCommitteeRole)), Default (CertState era), NoThunks (CertState era), NFData (CertState era), Show (CertState era), Eq (CertState era)) ⇒ EraCertState era where

The state associated with the DELPL rule, which combines the DELEG rule and the POOL rule.

Associated Types

type CertState era = (r ∷ Type) | r → era

Methods

certDStateL ∷ Lens' (CertState era) (DState era)

certPStateL ∷ Lens' (CertState era) (PState era)

obligationCertState ∷ CertState era → Obligations

Calculate total possible refunds in the system that are related to certificates

There is an invariant that the sum of all the fields should be the same as the utxosDeposited field of the UTxOState. Note that this does not depend upon the current values of the Key and Pool deposits of the PParams.

certsTotalDepositsTxBody ∷ PParams era → CertState era → TxBody era → Coin

Compute the total deposits from the Certs of a TxBody.

This is the contribution of a TxBody towards the deposit pot (utxosDeposit field of the UTxOState) of the system

certsTotalRefundsTxBody ∷ PParams era → CertState era → TxBody era → Coin

Compute the total refunds from the Certs of a TxBody.

This is the contribution of a TxBody towards the total Obligations of the system See Obligations and obligationCertState for more information.

type family CertState era = (r ∷ Type) | r → era

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psVRFKeyHashes ∷ !(Map (VRFVerKeyHash 'StakePoolVRF) (NonZero Word64))

VRF key hashes that have been registered via PoolParams

psStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

The state of current stake pools.

psFutureStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

The state of future stake pools. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring ∷ !(Map (KeyHash 'StakePool) EpochNo)

A map of retiring stake pools to the epoch when they retire.

data DState era

The state used by the DELEG rule, which roughly tracks stake delegation and some governance features.

Constructors

DState

Fields

dsAccounts ∷ !(Accounts era)

Keep track of the account state (eg. balance, deposit, stake-pool delegation, etc.) for all registered stake credentials.

dsFutureGenDelegs ∷ !(Map FutureGenDeleg GenDelegPair)

Future genesis key delegations

dsGenDelegs ∷ !GenDelegs

Genesis key delegations

dsIRewards ∷ !InstantaneousRewards

Instantaneous Rewards

data InstantaneousRewards

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking) Coin)

iRTreasury ∷ !(Map (Credential 'Staking) Coin)

deltaReserves ∷ !DeltaCoin

deltaTreasury ∷ !DeltaCoin

data ShelleyTx era

Constructors

ShelleyTx

Fields

stBody ∷ !(TxBody era)

stWits ∷ !(TxWits era)

stAuxData ∷ !(StrictMaybe (TxAuxData era))

data ShelleyTxOut era

Constructors

TxOutCompact !CompactAddr !(CompactForm (Value era))

Bundled Patterns

pattern ShelleyTxOut ∷ (HasCallStack, Era era, Val (Value era)) ⇒ Addr → Value era → ShelleyTxOut era

newtype ShelleyTxAuxData era

Constructors

MkShelleyTxAuxData (MemoBytes (ShelleyTxAuxDataRaw era))

Bundled Patterns

pattern ShelleyTxAuxData ∷ ∀ era. Era era ⇒ Map Word64 Metadatum → ShelleyTxAuxData era

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState ∷ !(UTxOState era)

The current unspent transaction outputs.

lsCertState ∷ !(CertState era)

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Number of the epoch when this NewEpochState was modified last. With respect to block and transactions validation this will always be the current epoch number. However, when it comes to the TICK rule, it will be the epoch number of the previous epoch whenever we are crossing the epoch boundary.

nesBprev ∷ !BlocksMade

Blocks made before current epoch

nesBcur ∷ !BlocksMade

Blocks made in current epoch

nesEs ∷ !(EpochState era)

Epoch state

nesRu ∷ !(StrictMaybe PulsingRewUpdate)

Possible reward update

nesPd ∷ !PoolDistr

Stake distribution within the stake pool

stashedAVVMAddresses ∷ !(StashedAVVMAddresses era)

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo ∷ SlotNo

ledgerEpochNo ∷ Maybe EpochNo

ledgerIx ∷ TxIx

ledgerPp ∷ PParams era

ledgerAccount ∷ ChainAccountState

data EpochState era

Constructors

EpochState

Fields

esChainAccountState ∷ !ChainAccountState

esLState ∷ !(LedgerState era)

esSnapshots ∷ !SnapShots

esNonMyopic ∷ !NonMyopic

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://github.com/intersectmbo/cardano-ledger/releases/latest/download/pool-ranking.pdf

data RewardUpdate

The ultimate goal of a reward update computation. Aggregating rewards for each staking credential.

Constructors

RewardUpdate

Fields

deltaT ∷ !DeltaCoin

deltaR ∷ !DeltaCoin

rs ∷ !(Map (Credential 'Staking) (Set Reward))

deltaF ∷ !DeltaCoin

nonMyopic ∷ !NonMyopic

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

sgsFuturePParams ∷ !(FuturePParams era)

Prediction of parameter changes that might happen on the epoch boundary.

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very careful, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

utxosUtxo ∷ !(UTxO era)

utxosDeposited ∷ !Coin

utxosFees ∷ !Coin

utxosGovState ∷ !(GovState era)

utxosInstantStake ∷ !(InstantStake era)

utxosDonation ∷ !Coin

newtype ProposedPPUpdates era

Update operation for protocol parameters structure PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis) (PParamsUpdate era))

data NonMyopic

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo ∷ SlotNo

deCurEpochNo ∷ EpochNo

Lazy on purpose, because not all certificates need to know the current EpochNo

ptr_ ∷ Ptr

deChainAccountState ∷ ChainAccountState

ppDE ∷ PParams era

data ShelleyDELEG era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo ∷ SlotNo

delegsEpochNo ∷ EpochNo

Lazy on purpose, because not all certificates need to know the current EpochNo

delegsIx ∷ TxIx

delegspp ∷ PParams era

delegsTx ∷ Tx era

delegsAccount ∷ ChainAccountState

data ShelleyDELEGS era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo ∷ SlotNo

delplEpochNo ∷ EpochNo

delPlPtr ∷ Ptr

delPlPp ∷ PParams era

delPlAccount ∷ ChainAccountState

data ShelleyDELPL era

data ShelleyLEDGER era

data ShelleyLEDGERS era

data ShelleyLedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo ∷ SlotNo

ledgersEpochNo ∷ EpochNo

ledgersPp ∷ PParams era

ledgersAccount ∷ ChainAccountState

data ShelleyNEWEPOCH era

data PoolEnv era

Constructors

PoolEnv EpochNo (PParams era)

data ShelleyPOOL era

data ShelleyPOOLREAP era

data PpupEnv era

Constructors

PPUPEnv SlotNo (PParams era) GenDelegs

data ShelleyPPUP era

data SnapEnv era

Constructors

SnapEnv (LedgerState era) (PParams era)

data ShelleyTICK era

data ShelleyTICKF era

data ShelleyUTXO era

data UtxoEnv era

Constructors

UtxoEnv

Fields

ueSlot ∷ SlotNo

uePParams ∷ PParams era

ueCertState ∷ CertState era

data ShelleyUTXOW era

data MultiSig era

data GenesisDelegCert

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis) !(KeyHash 'GenesisDelegate) !(VRFVerKeyHash 'GenDelegVRF)

data MIRCert

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ !MIRPot

mirRewards ∷ !MIRTarget

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking) DeltaCoin)

SendToOppositePotMIR !Coin

data ShelleyDelegCert

Constructors

ShelleyRegCert !StakeCredential
A stake credential registration certificate.

ShelleyUnRegCert !StakeCredential
A stake credential deregistration certificate.

ShelleyDelegCert !StakeCredential !(KeyHash 'StakePool)
A stake delegation certificate.

data ShelleyTxWits era

data ShelleyBlockBody era where

Bundled Patterns

pattern ShelleyBlockBody ∷ ∀ era. (EraTx era, SafeToHash (TxWits era)) ⇒ StrictSeq (Tx era) → ShelleyBlockBody era
Constuct a BlockBody (with all it bytes) from just Tx's

data ShelleyGenesisStaking

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools ∷ ListMap (KeyHash 'StakePool) PoolParams

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake ∷ ListMap (KeyHash 'Staking) (KeyHash 'StakePool)

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

data ShelleyGenesis

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !PositiveUnitInterval

sgSecurityParam ∷ !(NonZero Word64)

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTimeMicro

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(PParams ShelleyEra)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis) GenDelegPair)

sgInitialFunds ∷ ListMap Addr Coin

sgInitialFunds is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

sgStaking ∷ ShelleyGenesisStaking

sgStaking is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

data ValidationErr

Constructors

EpochNotLongEnough EpochSize Word64 Rational EpochSize

MaxKESEvolutionsUnsupported Word64 Word

QuorumTooSmall Word64 Word64 Word64

newtype NominalDiffTimeMicro

Unlike NominalDiffTime that supports Pico precision, this type only supports Micro precision.

Constructors

NominalDiffTimeMicro Micro

hashVerKeyVRF ∷ ∀ c (r ∷ KeyRoleVRF). Crypto c ⇒ VerKeyVRF (VRF c) → VRFVerKeyHash r

hashKey ∷ ∀ (kd ∷ KeyRole). VKey kd → KeyHash kd

Hash a given public key

balance ∷ EraTxOut era ⇒ UTxO era → Value era

coerceKeyRole ∷ ∀ (r ∷ KeyRole) (r' ∷ KeyRole). HasKeyRole a ⇒ a r → a r'

General coercion of key roles.

The presence of this function is mostly to help the user realise where they are converting key roles.

txIxToInt ∷ TxIx → Int

txIxFromIntegral ∷ (Integral a, MonadFail m) ⇒ a → m TxIx

certIxToInt ∷ CertIx → Int

certIxFromIntegral ∷ (Integral a, MonadFail m) ⇒ a → m CertIx

word64ToCoin ∷ Word64 → Coin

isOverlaySlot

Arguments

∷ SlotNo
The first slot of the given epoch.

→ UnitInterval
The decentralization parameter.

→ SlotNo
The slot to check.

→ Bool

Determine if the given slot is reserved for the overlay schedule.

individualPoolStake ∷ IndividualPoolStake → Rational

Pool stake distribution. This is a ratio of individualTotalPoolStake/pdTotalActiveStake

bBodySize ∷ EraBlockBody era ⇒ ProtVer → BlockBody era → Int

calculatePoolDistr ∷ SnapShot → PoolDistr

calculatePoolDistr' ∷ (KeyHash 'StakePool → Bool) → SnapShot → PoolDistr

bheader ∷ Block h era → h

bbody ∷ Block h era → BlockBody era

mkShelleyCertState ∷ EraCertState era ⇒ PState era → DState era → CertState era

computeRandomnessStabilisationWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the randomness stabilisation window from the security param and the active slot coefficient.

The value 4k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

computeStabilityWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the stability window (e.g. the number of slots needed for a block to become stable) from the security param and the active slot coefficient.

The value 3k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

txSeqBlockBodyShelleyL ∷ (EraBlockBody era, SafeToHash (TxWits era), BlockBody era ~ ShelleyBlockBody era) ⇒ Lens' (BlockBody era) (StrictSeq (Tx era))

mkBasicBlockBodyShelley ∷ (EraBlockBody era, SafeToHash (TxWits era), BlockBody era ~ ShelleyBlockBody era) ⇒ BlockBody era

shelleyBlockBodyHash ∷ ShelleyBlockBody era → Hash HASH EraIndependentBlockBody

shelleyBlockBodyTxs ∷ ShelleyBlockBody era → StrictSeq (Tx era)

auxDataSeqDecoder ∷ Int → IntMap a → Decoder s (Seq (Maybe a))

hashShelleySegWits

Arguments

∷ ByteString

→ ByteString
Bytes for transaction bodies

→ ByteString
Bytes for transaction witnesses

→ Hash HASH EraIndependentBlockBody
Bytes for transaction auxiliary datas

slotToNonce ∷ SlotNo → Nonce

incrBlocks ∷ Bool → KeyHash 'StakePool → BlocksMade → BlocksMade

coreAuxDataBytes ∷ EraTx era ⇒ Tx era → StrictMaybe ByteString

toShelleyGenesisPairs ∷ KeyValue e a ⇒ ShelleyGenesis → [a]

Deprecated: In favor of toKeyValuePairs

emptyGenesisStaking ∷ ShelleyGenesisStaking

Empty genesis staking

sgActiveSlotCoeff ∷ ShelleyGenesis → ActiveSlotCoeff

genesisUTxO ∷ ∀ era. EraTxOut era ⇒ ShelleyGenesis → UTxO era

initialFundsPseudoTxIn ∷ Addr → TxIn

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

validateGenesis ∷ ShelleyGenesis → Either [ValidationErr] ()

Do some basic sanity checking on the Shelley genesis file.

describeValidationErr ∷ ValidationErr → Text

mkShelleyGlobals ∷ ShelleyGenesis → EpochInfo (Either Text) → Globals

nominalDiffTimeMicroToMicroseconds ∷ NominalDiffTimeMicro → Micro

nominalDiffTimeMicroToSeconds ∷ NominalDiffTimeMicro → Pico

toNominalDiffTimeMicro ∷ NominalDiffTime → Maybe NominalDiffTimeMicro

toNominalDiffTimeMicroWithRounding ∷ NominalDiffTime → NominalDiffTimeMicro

fromNominalDiffTimeMicro ∷ NominalDiffTimeMicro → NominalDiffTime

secondsToNominalDiffTimeMicro ∷ Micro → NominalDiffTimeMicro

sgInitialFundsL ∷ Lens' ShelleyGenesis (ListMap Addr Coin)

sgStakingL ∷ Lens' ShelleyGenesis ShelleyGenesisStaking