Cardano.Ledger.Shelley.API.Validation

Description

Interface to the block validation and chain extension logic in the Shelley API.

Documentation

class (EraGov era, EraBlockBody era) ⇒ ApplyBlock era where

Minimal complete definition

Nothing

Methods

applyBlock ∷ SingEP ep → ValidationPolicy → Globals → NewEpochState era → Block BHeaderView era → (NewEpochState era, [PredicateFailure (EraRule "BBODY" era)], [Event (EraRule "BBODY" era)])

Run the BBODY rule with globalAssertionPolicy. This function always succeeds, but whenever validation is turned on it is necessary to check for presence of predicate failures before a call can be marked successful. Therefore it is recommended to call applyBlockEither instead.

default applyBlock ∷ (STS (EraRule "BBODY" era), BaseM (EraRule "BBODY" era) ~ ShelleyBase, Environment (EraRule "BBODY" era) ~ BbodyEnv era, State (EraRule "BBODY" era) ~ ShelleyBbodyState era, Signal (EraRule "BBODY" era) ~ Block BHeaderView era, State (EraRule "LEDGERS" era) ~ LedgerState era) ⇒ SingEP ep → ValidationPolicy → Globals → NewEpochState era → Block BHeaderView era → (NewEpochState era, [PredicateFailure (EraRule "BBODY" era)], [Event (EraRule "BBODY" era)])

applyTick ∷ SingEP ep → Globals → NewEpochState era → SlotNo → (NewEpochState era, [Event (EraRule "TICK" era)])

Run the TICK rule with globalAssertionPolicy and without any validation, since it can't fail anyways.

default applyTick ∷ (STS (EraRule "TICK" era), BaseM (EraRule "TICK" era) ~ ShelleyBase, Environment (EraRule "TICK" era) ~ (), State (EraRule "TICK" era) ~ NewEpochState era, Signal (EraRule "TICK" era) ~ SlotNo) ⇒ SingEP ep → Globals → NewEpochState era → SlotNo → (NewEpochState era, [Event (EraRule "TICK" era)])

applyBlockEither ∷ ApplyBlock era ⇒ SingEP ep → ValidationPolicy → Globals → NewEpochState era → Block BHeaderView era → Either (BlockTransitionError era) (NewEpochState era, [Event (EraRule "BBODY" era)])

Same as applyBlock, except it produces a Left when there are failures present and Right with result otherwise.

applyBlockEitherNoEvents ∷ ApplyBlock era ⇒ ValidationPolicy → Globals → NewEpochState era → Block BHeaderView era → Either (BlockTransitionError era) (NewEpochState era)

applyBlockNoValidaton ∷ ApplyBlock era ⇒ Globals → NewEpochState era → Block BHeaderView era → NewEpochState era

Re-apply a ledger block to the same state it has been applied to before.

This function does no validation of whether the block applies successfully; the caller implicitly guarantees that they have previously called applyBlockTransition on the same block and that this was successful.

applyTickNoEvents ∷ ApplyBlock era ⇒ Globals → NewEpochState era → SlotNo → NewEpochState era

Same as applyTick, but do not retain any ledger events

newtype BlockTransitionError era

Constructors

BlockTransitionError (NonEmpty (PredicateFailure (EraRule "BBODY" era)))

chainChecks

Arguments

∷ ∀ m. MonadError ChainPredicateFailure m

⇒ Version
Max major protocol version

→ ChainChecksPParams

→ BHeaderView

→ m ()