Cardano.Ledger.Shelley.API.Wallet

UTxOs

getUTxO :: NewEpochState era -> UTxO era

Get the full UTxO.

getUTxOSubset :: NewEpochState era -> Set (TxIn (Crypto era)) -> UTxO era

getFilteredUTxO :: Era era => NewEpochState era -> Set (Addr (Crypto era)) -> UTxO era

Get the UTxO filtered by address.

Stake Pools

getPools :: NewEpochState era -> Set (KeyHash 'StakePool (Crypto era))

Get the current registered stake pools.

getPoolParameters :: NewEpochState era -> Set (KeyHash 'StakePool (Crypto era)) -> Map (KeyHash 'StakePool (Crypto era)) (PoolParams (Crypto era))

Get the current registered stake pool parameters for a given set of stake pools. The result map will contain entries for all the given stake pools that are currently registered.

getTotalStake :: Globals -> NewEpochState era -> Coin

Calculate the current total stake.

poolsByTotalStakeFraction :: forall era. Globals -> NewEpochState era -> PoolDistr (Crypto era)

Get pool sizes, but in terms of total stake

The stake distribution uses active stake (so that the leader schedule is not affected by undelegated stake), but the wallet wants to display pool saturation for rewards purposes. For that, it needs the fraction of total stake.

This is not based on any snapshot, but uses the current ledger state.

data RewardInfoPool

Information about a stake pool

Constructors

RewardInfoPool

Fields

stake :: Coin

Absolute stake delegated to this pool

ownerPledge :: Coin

Pledge of pool owner(s)

ownerStake :: Coin

Absolute stake delegated by pool owner(s)

cost :: Coin

Pool cost

margin :: UnitInterval

Pool margin

performanceEstimate :: Double

Number of blocks produced divided by expected number of blocks. Can be larger than 1.0 for pool that gets lucky. (If some pools get unlucky, some pools must get lucky.)

data RewardParams

Global information that influences stake pool rewards

Constructors

RewardParams

Fields

nOpt :: Natural

Desired number of stake pools

a0 :: NonNegativeInterval

Influence of the pool owner's pledge on rewards

rPot :: Coin

Total rewards available for the given epoch

totalStake :: Coin

Maximum lovelace supply minus treasury

getRewardInfoPools :: (HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural) => Globals -> NewEpochState era -> (RewardParams, Map (KeyHash 'StakePool (Crypto era)) RewardInfoPool)

Retrieve the information necessary to calculate stake pool member rewards from the current stake distribution.

This information includes the current stake distribution aggregated by stake pools and pool owners, the current pool costs and margins, and performance estimates. Also included are global information such as the total stake or protocol parameters.

getRewardProvenance :: forall era. (HasField "_a0" (PParams era) NonNegativeInterval, HasField "_d" (PParams era) UnitInterval, HasField "_nOpt" (PParams era) Natural, HasField "_protocolVersion" (PParams era) ProtVer, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval) => Globals -> NewEpochState era -> (RewardUpdate (Crypto era), RewardProvenance (Crypto era))

Calculate stake pool rewards from the snapshot labeled go. Also includes information on how the rewards were calculated (RewardProvenance).

For a calculation of rewards based on the current stake distribution, see getRewardInfoPools.

TODO: Deprecate getRewardProvenance, because wallets are more likely to use getRewardInfoPools for up-to-date information on stake pool rewards.

getNonMyopicMemberRewards :: (HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural) => Globals -> NewEpochState era -> Set (Either Coin (Credential 'Staking (Crypto era))) -> Map (Either Coin (Credential 'Staking (Crypto era))) (Map (KeyHash 'StakePool (Crypto era)) Coin)

Calculate the Non-Myopic Pool Member Rewards for a set of credentials. For each given credential, this function returns a map from each stake pool (identified by the key hash of the pool operator) to the non-myopic pool member reward for that stake pool.

This is not based on any snapshot, but uses the current ledger state.

Transaction helpers

class (Era era, HasField "_minfeeA" (PParams era) Natural, HasField "_keyDeposit" (PParams era) Coin, HasField "_poolDeposit" (PParams era) Coin, HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era)))) => CLI era where

A collection of functons to help construction transactions from the cardano-cli.

Minimal complete definition

evaluateMinFee, evaluateConsumed, addKeyWitnesses, evaluateMinLovelaceOutput

Methods

evaluateMinFee :: PParams era -> Tx era -> Coin

The minimum fee calculation. Used for the default implentation of evaluateTransactionFee.

evaluateConsumed :: PParams era -> UTxO era -> TxBody era -> Value era

The consumed calculation. Used for the default implentation of evaluateTransactionBalance.

addKeyWitnesses :: Tx era -> Set (WitVKey 'Witness (Crypto era)) -> Tx era

evaluateTransactionBalance

Arguments

:: PParams era
The current protocol parameters.

-> UTxO era
The UTxO relevant to the transaction.

-> (KeyHash 'StakePool (Crypto era) -> Bool)
A predicate that a stake pool ID is new (i.e. unregistered). Typically this will be:

  (notMember stakepools)

-> TxBody era
The transaction being evaluated for balance.

-> Value era
The difference between what the transaction consumes and what it produces.

Evaluate the difference between the value currently being consumed by a transaction and the number of lovelace being produced. This value will be zero for a valid transaction.

evaluateTransactionFee

Arguments

:: PParams era
The current protocol parameters.

-> Tx era
The transaction.

-> Word
The number of key witnesses still to be added to the transaction.

-> Coin
The required fee.

Evaluate the fee for a given transaction.

evaluateMinLovelaceOutput :: PParams era -> TxOut era -> Coin

Evaluate the minimum lovelace that a given transaciton output must contain.

addShelleyKeyWitnesses :: (Era era, Witnesses era ~ WitnessSet era, AnnotatedData (Script era), ToCBOR (AuxiliaryData era), ToCBOR (TxBody era)) => Tx era -> Set (WitVKey 'Witness (Crypto era)) -> Tx era

data AdaPots

Constructors

AdaPots

Fields

treasuryAdaPot :: Coin

reservesAdaPot :: Coin

rewardsAdaPot :: Coin

utxoAdaPot :: Coin

depositsAdaPot :: Coin

feesAdaPot :: Coin

totalAdaES :: UsesValue era => EpochState era -> Coin

Calculate the total ada in the epoch state

totalAdaPotsES :: UsesValue era => EpochState era -> AdaPots

Calculate the total ada pots in the epoch state