Cardano.Ledger.Shelley.AdaPots

Documentation

data AdaPots

Constructors

AdaPots

Fields

treasuryAdaPot :: Coin

reservesAdaPot :: Coin

rewardsAdaPot :: Coin

utxoAdaPot :: Coin

keyDepositAdaPot :: Coin

poolDepositAdaPot :: Coin

depositsAdaPot :: Coin

feesAdaPot :: Coin

totalAdaES :: EraTxOut era => EpochState era -> Coin

Calculate the total ada in the epoch state

totalAdaPotsES :: EraTxOut era => EpochState era -> AdaPots

Calculate the total ada pots in the epoch state

data Produced

Itemizing what is Produced by a transaction

Constructors

Produced

Fields

proOutputs :: !Coin

proFees :: !Coin

proDeposits :: !Coin

data Consumed

Itemizing what is consumed by a transaction

Constructors

Consumed

Fields

conInputs :: !Coin

conRefunds :: !Coin

conWithdrawals :: !Coin

consumedTxBody :: ShelleyEraTxBody era => TxBody era -> PParams era -> CertState era -> UTxO era -> Consumed

Compute the Coin part of what is consumed by a TxBody, itemized as a Consume

producedTxBody :: ShelleyEraTxBody era => TxBody era -> PParams era -> CertState era -> Produced

Compute the Coin part of what is produced by a TxBody, itemized as a Produced