Cardano.Ledger.Shelley.BlockBody.Internal

Description

Provides BlockBody internals

Warning

This module is considered internal.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Documentation

data ShelleyBlockBody era

Constructors

ShelleyBlockBodyInternal

Fields

sbbTxs ∷ !(StrictSeq (Tx era))

sbbHash ∷ Hash HASH EraIndependentBlockBody

Memoized hash to avoid recomputation. Lazy on purpose.

sbbTxsBodyBytes ∷ ByteString

Bytes encoding Seq (TxBody era)

sbbTxsWitsBytes ∷ ByteString

Bytes encoding Seq (TxWits era)

sbbTxsAuxDataBytes ∷ ByteString

Bytes encoding a Seq (TxAuxData era). Missing indices have SNothing for metadata

Bundled Patterns

pattern ShelleyBlockBody ∷ ∀ era. (EraTx era, SafeToHash (TxWits era)) ⇒ StrictSeq (Tx era) → ShelleyBlockBody era
Constuct a BlockBody (with all it bytes) from just Tx's

mkBasicBlockBodyShelley ∷ (EraBlockBody era, SafeToHash (TxWits era), BlockBody era ~ ShelleyBlockBody era) ⇒ BlockBody era

txSeqBlockBodyShelleyL ∷ (EraBlockBody era, SafeToHash (TxWits era), BlockBody era ~ ShelleyBlockBody era) ⇒ Lens' (BlockBody era) (StrictSeq (Tx era))

auxDataSeqDecoder ∷ Int → IntMap a → Decoder s (Seq (Maybe a))

hashShelleySegWits

Arguments

∷ ByteString

→ ByteString
Bytes for transaction bodies

→ ByteString
Bytes for transaction witnesses

→ Hash HASH EraIndependentBlockBody
Bytes for transaction auxiliary datas

bBodySize ∷ EraBlockBody era ⇒ ProtVer → BlockBody era → Int

slotToNonce ∷ SlotNo → Nonce

incrBlocks ∷ Bool → KeyHash 'StakePool → BlocksMade → BlocksMade

coreAuxDataBytes ∷ EraTx era ⇒ Tx era → StrictMaybe ByteString

Orphan instances

EraBlockBody ShelleyEra

Instance details

Associated Types

type BlockBody ShelleyEra = (r ∷ Type)

Methods

mkBasicBlockBody ∷ BlockBody ShelleyEra

txSeqBlockBodyL ∷ Lens' (BlockBody ShelleyEra) (StrictSeq (Tx ShelleyEra))

fromTxSeq ∷ BlockBody ShelleyEra → StrictSeq (Tx ShelleyEra)

toTxSeq ∷ StrictSeq (Tx ShelleyEra) → BlockBody ShelleyEra

hashBlockBody ∷ BlockBody ShelleyEra → Hash HASH EraIndependentBlockBody

hashTxSeq ∷ BlockBody ShelleyEra → Hash HASH EraIndependentBlockBody

numSegComponents ∷ Word64