Cardano.Ledger.Shelley.BlockChain

Documentation

data ShelleyTxSeq era

Constructors

TxSeq' !(StrictSeq (ShelleyTx era)) ByteString ByteString ByteString

Bundled Patterns

pattern ShelleyTxSeq ∷ ∀ era. (EraTx era, Tx era ~ ShelleyTx era, SafeToHash (TxWits era)) ⇒ StrictSeq (Tx era) → ShelleyTxSeq era
Constuct a TxSeq (with all it bytes) from just Tx's

constructMetadata ∷ Int → Map Int (Annotator (TxAuxData era)) → Seq (Maybe (Annotator (TxAuxData era)))

Given a size and a mapping from indices to maybe metadata, return a sequence whose size is the size parameter and whose non-Nothing values correspond to the values in the mapping.

indexLookupSeq ∷ Int → Map Int a → Seq (Maybe a)

txSeqTxns ∷ ShelleyTxSeq era → StrictSeq (ShelleyTx era)

bbHash ∷ ShelleyTxSeq era → Hash HASH EraIndependentBlockBody

Hash a given block body

bBodySize ∷ EraSegWits era ⇒ ProtVer → TxSeq era → Int

slotToNonce ∷ SlotNo → Nonce

incrBlocks ∷ Bool → KeyHash 'StakePool → BlocksMade → BlocksMade

coreAuxDataBytes ∷ EraTx era ⇒ Tx era → StrictMaybe ByteString

txSeqDecoder ∷ ∀ era. EraTx era ⇒ Bool → ∀ s. Decoder s (Annotator (ShelleyTxSeq era))

The parts of the Tx in Blocks that have to have DecCBOR(Annotator x) instances. These are exactly the parts that are SafeToHash. | Decode a TxSeq, used in decoding a Block.

Orphan instances

EraSegWits ShelleyEra

Instance details

Associated Types

type TxSeq ShelleyEra = (r ∷ Type)

Methods

fromTxSeq ∷ TxSeq ShelleyEra → StrictSeq (Tx ShelleyEra)

toTxSeq ∷ StrictSeq (Tx ShelleyEra) → TxSeq ShelleyEra

hashTxSeq ∷ TxSeq ShelleyEra → Hash HASH EraIndependentBlockBody

numSegComponents ∷ Word64