Cardano.Ledger.Shelley.Core

Documentation

class (ShelleyEraTxCert era, EraTxBody era, AtMostEra "Babbage" era) ⇒ ShelleyEraTxBody era where

Methods

ttlTxBodyL ∷ ExactEra ShelleyEra era ⇒ Lens' (TxBody era) SlotNo

updateTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (Update era))

newtype Withdrawals

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map RewardAccount Coin

class EraTxCert era ⇒ ShelleyEraTxCert era where

Methods

mkRegTxCert ∷ StakeCredential → TxCert era

getRegTxCert ∷ TxCert era → Maybe StakeCredential

mkUnRegTxCert ∷ StakeCredential → TxCert era

getUnRegTxCert ∷ TxCert era → Maybe StakeCredential

mkDelegStakeTxCert ∷ StakeCredential → KeyHash 'StakePool → TxCert era

getDelegStakeTxCert ∷ TxCert era → Maybe (StakeCredential, KeyHash 'StakePool)

mkGenesisDelegTxCert ∷ AtMostEra "Babbage" era ⇒ GenesisDelegCert → TxCert era

getGenesisDelegTxCert ∷ AtMostEra "Babbage" era ⇒ TxCert era → Maybe GenesisDelegCert

mkMirTxCert ∷ AtMostEra "Babbage" era ⇒ MIRCert → TxCert era

getMirTxCert ∷ AtMostEra "Babbage" era ⇒ TxCert era → Maybe MIRCert

pattern MirTxCert ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ MIRCert → TxCert era

data MIRCert

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ !MIRPot

mirRewards ∷ !MIRTarget

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking) DeltaCoin)

SendToOppositePotMIR !Coin

pattern GenesisDelegTxCert ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ KeyHash 'Genesis → KeyHash 'GenesisDelegate → VRFVerKeyHash 'GenDelegVRF → TxCert era

pattern RegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential → TxCert era

pattern UnRegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential → TxCert era

pattern DelegStakeTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential → KeyHash 'StakePool → TxCert era

type family Value era

A value is something which quantifies a transaction output.

data Hash h a

class (KnownNat (SizeHash h), Typeable h) ⇒ HashAlgorithm h

Minimal complete definition

hashAlgorithmName, digest

newtype KeyHash (r ∷ KeyRole)

Discriminated hash of public Key

Constructors

KeyHash

Fields

unKeyHash ∷ Hash ADDRHASH (VerKeyDSIGN DSIGN)

class (Val (Value era), ToJSON (TxOut era), DecCBOR (Value era), DecCBOR (CompactForm (Value era)), MemPack (CompactForm (Value era)), EncCBOR (Value era), ToCBOR (TxOut era), EncCBOR (TxOut era), DecCBOR (TxOut era), DecShareCBOR (TxOut era), Share (TxOut era) ~ Interns (Credential 'Staking), NoThunks (TxOut era), NFData (TxOut era), Show (TxOut era), Eq (TxOut era), MemPack (TxOut era), EraPParams era) ⇒ EraTxOut era where

Abstract interface into specific fields of a TxOut

Minimal complete definition

mkBasicTxOut, upgradeTxOut, valueEitherTxOutL, addrEitherTxOutL, (getMinCoinSizedTxOut | getMinCoinTxOut)

Associated Types

type TxOut era = (r ∷ Type) | r → era

The output of a UTxO for a particular era

Methods

mkBasicTxOut ∷ Addr → Value era → TxOut era

upgradeTxOut ∷ TxOut (PreviousEra era) → TxOut era

Every era, except Shelley, must be able to upgrade a TxOut from a previous era.

valueTxOutL ∷ Lens' (TxOut era) (Value era)

compactValueTxOutL ∷ Lens' (TxOut era) (CompactForm (Value era))

valueEitherTxOutL ∷ Lens' (TxOut era) (Either (Value era) (CompactForm (Value era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

addrTxOutL ∷ Lens' (TxOut era) Addr

compactAddrTxOutL ∷ Lens' (TxOut era) CompactAddr

addrEitherTxOutL ∷ Lens' (TxOut era) (Either Addr CompactAddr)

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

The utility of this function comes from the fact that TxOut usually stores the address in either one of two forms: compacted or unpacked. In order to avoid extroneous conversions in getTxOutAddr and getTxOutCompactAddr we can define just this functionality. Also sometimes it is crucial to know at the callsite which form of address we have readily available without any conversions (eg. searching millions of TxOuts for a particular address)

getMinCoinSizedTxOut ∷ PParams era → Sized (TxOut era) → Coin

Produce the minimum lovelace that a given transaction output must contain. Information about the size of the TxOut is required in some eras. Use getMinCoinTxOut if you don't have the size readily available to you.

getMinCoinTxOut ∷ PParams era → TxOut era → Coin

Same as getMinCoinSizedTxOut, except information about the size of TxOut will be computed by serializing the TxOut. If the size turns out to be not needed, then serialization will have no overhead, since it is computed lazily.

type family TxOut era = (r ∷ Type) | r → era

The output of a UTxO for a particular era

class (EraTxBody era, EraTxWits era, EraTxAuxData era, EraPParams era, NFData (Tx era), NoThunks (Tx era), DecCBOR (Annotator (Tx era)), EncCBOR (Tx era), ToCBOR (Tx era), Show (Tx era), Eq (Tx era)) ⇒ EraTx era where

A transaction.

Minimal complete definition

mkBasicTx, bodyTxL, witsTxL, auxDataTxL, sizeTxF, validateNativeScript, getMinFeeTx

Associated Types

data Tx era

Methods

mkBasicTx ∷ TxBody era → Tx era

bodyTxL ∷ Lens' (Tx era) (TxBody era)

witsTxL ∷ Lens' (Tx era) (TxWits era)

auxDataTxL ∷ Lens' (Tx era) (StrictMaybe (TxAuxData era))

sizeTxF ∷ SimpleGetter (Tx era) Word32

For fee calculation and estimations of impact on block space

sizeTxForFeeCalculation ∷ Tx era → Word32

For fee calculation and estimations of impact on block space To replace sizeTxF after it has been proved equivalent to it .

validateNativeScript ∷ Tx era → NativeScript era → Bool

Using information from the transaction validate the supplied native script.

getMinFeeTx

Arguments

∷ PParams era

→ Tx era

→ Int
Size in bytes of reference scripts present in this transaction

→ Coin

Minimum fee calculation excluding witnesses

data family Tx era

data ByronEra

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

class (Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT) ⇒ Era era where

Minimal complete definition

Nothing

Associated Types

type EraName era ∷ Symbol

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra AllegraEra = ShelleyEra

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Methods

eraName ∷ String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @ByronEra
"Byron"

type family EraName era ∷ Symbol

type family PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra AllegraEra = ShelleyEra

type family ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type family ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type AtMostEra (eraMostEra ∷ Symbol) era = ProtVerAtMost era (ProtVerHigh (EraFromName eraMostEra))

Restrict the era to equal to eraName or come before it.

type AtLeastEra (atLeastEra ∷ Symbol) era = ProtVerAtLeast era (ProtVerLow (EraFromName atLeastEra))

Restrict the era to equal to atLeastEra or come after it

type ExactEra inEra era = ProtVerInBounds era (ProtVerLow inEra) (ProtVerHigh inEra)

Restrict an era to the specific era through the protocol version. This is equivalent to (inEra (Crypto era) ~ era)

type ProtVerInBounds era (l ∷ Nat) (h ∷ Nat) = (ProtVerAtLeast era l, ProtVerAtMost era h)

Restrict a lower and upper bounds of the protocol version for the particular era

type family ProtVerAtMost era (h ∷ Nat) where ...

Requirement for the era's lowest protocol version to be lower or equal to the supplied value

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type family ProtVerAtLeast era (l ∷ Nat) where ...

Requirement for the era's highest protocol version to be higher or equal to the supplied value

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

class EraRuleEvent rule era ~ Event (EraRule rule era) ⇒ InjectRuleEvent (rule ∷ Symbol) (t ∷ Type → Type) era where

Minimal complete definition

Nothing

Methods

injectEvent ∷ t era → EraRuleEvent rule era

class EraRuleFailure rule era ~ PredicateFailure (EraRule rule era) ⇒ InjectRuleFailure (rule ∷ Symbol) (t ∷ Type → Type) era where

Minimal complete definition

Nothing

Methods

injectFailure ∷ t era → EraRuleFailure rule era

data VoidEraRule (rule ∷ Symbol) era

This is a type with no inhabitans for the rules. It is used to indicate that a rule does not have a predicate failure as well as marking rules that have been disabled when comparing to prior eras.

type family EraRuleEvent (rule ∷ Symbol) era = (r ∷ Type) | r → rule era

type family EraRuleFailure (rule ∷ Symbol) era = (r ∷ Type) | r → rule era

EraRuleFailure type family is needed for injectivity, which STS' PredicateFailure does not provide for us unfortunately.

type family EraRule (rule ∷ Symbol) era = (r ∷ Type) | r → rule

Era STS map

data KeyRole

The role of a key.

All key roles are fixed and unique, except for the Witness role. In particular, keys can be cast to a Witness role with the help of asWitness, because same witness can be valid for many roles.

In fact, it is perfectly allowable for a key to be used in many roles by the end user; there is nothing prohibiting somebody using the same underlying key or a script as their payment and staking credential, as well as the key for their stake pool. However, in the ledger code mixing up keys with different roles could be catastrophic, that is why we have this separation.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

Guard

newtype HashHeader

Constructors

HashHeader

Fields

unHashHeader ∷ Hash HASH EraIndependentBlockHeader

class SafeToHash x ⇒ HashAnnotated x i | x → i where

Types that are SafeToHash AND have the type uniquely determines the index type tag of SafeHash index

The SafeToHash and the HashAnnotated classes are designed so that their instances can be easily derived (because their methods have default methods when the type is a newtype around a type that is SafeToHash). For example,

Minimal complete definition

Nothing

Methods

hashAnnotated ∷ x → SafeHash i

Create a (SafeHash i), given (HashAnnotated x i) instance.

class SafeToHash t where

Only Types that preserve their serialisation bytes are members of the class SafeToHash. There are only a limited number of primitive direct instances of SafeToHash, all but two of them are present in this file. Instead of making explicit instances, we almost always use a newtype (around a type S) where their is already an instance (SafeToHash S). In that case the newtype has its SafeToHash instance derived using newtype deriving. The prime example of s is MemoBytes. The only exceptions are the legacy Shelley types: Metadata and ShelleyTx, that preserve their serialization bytes using a different mechanism than the use of MemoBytes. SafeToHash is a superclass requirement of the classes HashAnnotated which provide more convenient ways to construct SafeHashes than using makeHashWithExplicitProxys.

Minimal complete definition

originalBytes

Methods

originalBytes ∷ t → ByteString

Extract the original bytes from t

originalBytesSize ∷ t → Int

makeHashWithExplicitProxys ∷ Proxy i → t → SafeHash i

data SafeHash i

A SafeHash is a hash of something that is safe to hash. Such types store their own serialisation bytes. The prime example is (MemoBytes t), but other examples are things that consist of only ByteStrings (i.e. they are their own serialization) or for some other reason store their original bytes.

We do NOT export the constructor SafeHash, but instead export other functions such as hashAnnotated and extractHash which have constraints that limit their application to types which preserve their original serialization bytes.

newtype TxAuxDataHash

Constructors

TxAuxDataHash

Fields

unTxAuxDataHash ∷ SafeHash EraIndependentTxAuxData

newtype VRFVerKeyHash (r ∷ KeyRoleVRF)

Discriminated hash of VRF Verification Key

Constructors

VRFVerKeyHash

Fields

unVRFVerKeyHash ∷ Hash HASH KeyRoleVRF

data KeyRoleVRF

Constructors

StakePoolVRF

GenDelegVRF

BlockIssuerVRF

newtype ScriptHash

Constructors

ScriptHash (Hash ADDRHASH EraIndependentScript)

data EraIndependentScriptIntegrity

data EraIndependentPParamView

data EraIndependentScriptData

type DataHash = SafeHash EraIndependentData

data EraIndependentData

data EraIndependentScript

data EraIndependentTxAuxData

data EraIndependentMetadata

data EraIndependentBlockBody

data EraIndependentBlockHeader

data EraIndependentTxBody

type ADDRHASH = Blake2b_224

Hashing algorithm used for hashing cryptographic keys and scripts. As the type synonym name alludes, this is the hashing algorithm used for addresses.

type HASH = Blake2b_256

Hashing algorithm used for hashing everything, except addresses, for which ADDRHASH is used.

class (Era era, Era (PreviousEra era)) ⇒ TranslateEra era (f ∷ Type → Type) where

Translation of types between eras, e.g., from Shelley to Allegra.

When era is just a phantom type parameter, an empty standalone deriving can be used:

newtype Foo era = Foo Int

instance TranslateEra AllegraEra Foo

Note that one could use DerivingAnyClass (deriving (TranslateEra (Allegra c))), but this would introduce an undesired coupling between the era-parametric type and (a) particular era(s). The intention is to have a module with orphan instances per era.

In most cases, the era parameter won't be phantom, and a manual instance will have to be written:

newtype Bar era = Bar (TxBody era)

instance TranslateEra AllegraEra Bar where
    translateEra ctxt = Bar <$> translateEra ctxt

-- With the following instance being in scope:
instance TranslatEra AllegraEra TxBody

Note: we use PreviousEra instead of NextEra as an era definitely knows its predecessor, but not necessarily its successor. Moreover, one could argue that it makes more sense to define the translation from era A to era B where era B is defined, than where era A is defined.

Minimal complete definition

Nothing

Associated Types

type TranslationError era (f ∷ Type → Type)

Most translations should be infallible (default instance), but we leave the door open for partial translations.

For a partial translation, override the default type to be () or a concrete error type.

type TranslationError era (f ∷ Type → Type) = Void

Methods

translateEra ∷ TranslationContext era → f (PreviousEra era) → Except (TranslationError era f) (f era)

Translate a type f parameterised by the era from an era to the era after it.

The translation is a given the translation context of era.

A default instance is provided for when the two types are Coercible.

type family TranslationError era (f ∷ Type → Type)

Most translations should be infallible (default instance), but we leave the door open for partial translations.

For a partial translation, override the default type to be () or a concrete error type.

type family TranslationContext era

Per-era context used for TranslateEra.

This context will be passed to the translation instances of all types of that particular era. In practice, most instances won't need the context, but this approach makes the translation composable (as opposed to having a separate context per type).

data PParamUpdate era t

Constructors

PParamUpdate

Fields

ppuTag ∷ Word

Used in CBOR and Plutus Data encoding of

ppuLens ∷ Lens' (PParamsUpdate era) (StrictMaybe t)

data PParam era where

Represents a single protocol parameter and the data required to serialize it.

Constructors

PParam

Fields

∷ ∀ t era. (DecCBOR t, EncCBOR t, FromJSON t, ToJSON t, ToPlutusData t)

⇒ { ppName ∷ Text

Used as JSON key

, ppLens ∷ Lens' (PParams era) t

, ppUpdate ∷ Maybe (PParamUpdate era t)

Not all protocol parameters have an update functionality in all eras

} → PParam era

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era)) ⇒ EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositCompactL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL, eraPParams

Associated Types

type PParamsHKD (f ∷ Type → Type) era = (r ∷ Type) | r → era

Protocol parameters where the fields are represented with a HKD

type UpgradePParams (f ∷ Type → Type) era

type DowngradePParams (f ∷ Type → Type) era

Methods

applyPPUpdates ∷ PParams era → PParamsUpdate era → PParams era

Applies a protocol parameters update

ppDG ∷ SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

ppProtocolVersionL ∷ Lens' (PParams era) ProtVer

ppuProtocolVersionL ∷ Lens' (PParamsUpdate era) (StrictMaybe ProtVer)

PParamsUpdate Protocol version

type family PParamsHKD (f ∷ Type → Type) era = (r ∷ Type) | r → era

Protocol parameters where the fields are represented with a HKD

type family UpgradePParams (f ∷ Type → Type) era

type family DowngradePParams (f ∷ Type → Type) era

newtype PParamsUpdate era

The type of updates to Protocol parameters

Constructors

PParamsUpdate (PParamsHKD StrictMaybe era)

newtype PParams era

Protocol parameters

Constructors

PParams (PParamsHKD Identity era)

data Reward

The Reward type captures:

if the reward is a member or leader reward

the stake pool ID associated with the reward

the number of Lovelace in the reward

Constructors

Reward

Fields

rewardType ∷ !RewardType

rewardPool ∷ !(KeyHash 'StakePool)

rewardAmount ∷ !Coin

data RewardType

The staking rewards in Cardano are all either:

member rewards - rewards given to a registered stake credential which has delegated to a stake pool, or

leader rewards - rewards given to a registered stake pool (in particular, given to the stake credential in the stake pool registration certificate).

See Figure 47, "Functions used in the Reward Splitting", of the formal specification for more details.

Constructors

MemberReward

LeaderReward

data PoolCert

Constructors

RegPool !PoolParams
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool) !EpochNo
A stake pool retirement certificate.

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Ord (TxCert era), Eq (TxCert era)) ⇒ EraTxCert era where

Associated Types

type TxCert era = (r ∷ Type) | r → era

type TxCertUpgradeError era

type TxCertUpgradeError era = Void

Methods

upgradeTxCert ∷ TxCert (PreviousEra era) → Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert ∷ TxCert era → Maybe (KeyHash 'Witness)

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert ∷ TxCert era → Maybe ScriptHash

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert ∷ PoolParams → TxCert era

getRegPoolTxCert ∷ TxCert era → Maybe PoolParams

mkRetirePoolTxCert ∷ KeyHash 'StakePool → EpochNo → TxCert era

getRetirePoolTxCert ∷ TxCert era → Maybe (KeyHash 'StakePool, EpochNo)

lookupRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking)

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking)

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether stake pool is registered or not

→ f (TxCert era)

→ Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (Credential 'Staking → Maybe Coin)
Lookup current deposit for Staking credential if one is registered

→ (Credential 'DRepRole → Maybe Coin)
Lookup current deposit for DRep credential if one is registered

→ f (TxCert era)

→ Coin

Compute the total refunds from a list of certificates.

type family TxCert era = (r ∷ Type) | r → era

type family TxCertUpgradeError era

class (EraTx era, Eq (BlockBody era), Show (BlockBody era), EncCBORGroup (BlockBody era), DecCBOR (Annotator (BlockBody era))) ⇒ EraBlockBody era where

Indicates that an era supports segregated witnessing.

This class embodies an isomorphism between 'BlockBody era' and 'StrictSeq (Tx era)', witnessed by the txSeqBlockBodyL lens.

Minimal complete definition

mkBasicBlockBody, txSeqBlockBodyL, hashBlockBody, numSegComponents

Associated Types

type BlockBody era = (r ∷ Type) | r → era

Methods

mkBasicBlockBody ∷ BlockBody era

txSeqBlockBodyL ∷ Lens' (BlockBody era) (StrictSeq (Tx era))

fromTxSeq ∷ BlockBody era → StrictSeq (Tx era)

toTxSeq ∷ StrictSeq (Tx era) → BlockBody era

hashBlockBody ∷ BlockBody era → Hash HASH EraIndependentBlockBody

Get the block body hash from the BlockBody. Note that this is not a regular "hash the stored bytes" function since the block body hash forms a small Merkle tree.

hashTxSeq ∷ BlockBody era → Hash HASH EraIndependentBlockBody

numSegComponents ∷ Word64

The number of segregated components

type family BlockBody era = (r ∷ Type) | r → era

class (Era era, Show (Script era), Eq (Script era), EqRaw (Script era), ToCBOR (Script era), EncCBOR (Script era), DecCBOR (Annotator (Script era)), NoThunks (Script era), SafeToHash (Script era), Eq (NativeScript era), Show (NativeScript era), NFData (NativeScript era), NoThunks (NativeScript era), EncCBOR (NativeScript era), DecCBOR (Annotator (NativeScript era))) ⇒ EraScript era where

Typeclass for script data types. Allows for script validation and hashing. You must understand the role of SafeToHash and scriptPrefixTag to make new instances. scriptPrefixTag is a magic number representing the tag of the script language. For each new script language defined, a new tag is chosen and the tag is included in the script hash for a script. The safeToHash constraint ensures that Scripts are never reserialised.

Associated Types

type Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

type NativeScript era = (r ∷ Type) | r → era

Methods

upgradeScript ∷ Script (PreviousEra era) → Script era

Every era, except Shelley, must be able to upgrade a Script from a previous era.

Warning - Important to note that any memoized binary representation will not be preserved. If you need to retain underlying bytes then you can use translateEraThroughCBOR

scriptPrefixTag ∷ Script era → ByteString

getNativeScript ∷ Script era → Maybe (NativeScript era)

fromNativeScript ∷ NativeScript era → Script era

type family Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

type family NativeScript era = (r ∷ Type) | r → era

class (EraScript era, Eq (TxWits era), EqRaw (TxWits era), Show (TxWits era), Monoid (TxWits era), NoThunks (TxWits era), ToCBOR (TxWits era), EncCBOR (TxWits era), DecCBOR (Annotator (TxWits era))) ⇒ EraTxWits era where

A collection of witnesses in a Tx

Minimal complete definition

addrTxWitsL, bootAddrTxWitsL, scriptTxWitsL

Associated Types

type TxWits era = (r ∷ Type) | r → era

Methods

mkBasicTxWits ∷ TxWits era

addrTxWitsL ∷ Lens' (TxWits era) (Set (WitVKey 'Witness))

bootAddrTxWitsL ∷ Lens' (TxWits era) (Set BootstrapWitness)

scriptTxWitsL ∷ Lens' (TxWits era) (Map ScriptHash (Script era))

type family TxWits era = (r ∷ Type) | r → era

class (Era era, Eq (TxAuxData era), EqRaw (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (Annotator (TxAuxData era)), HashAnnotated (TxAuxData era) EraIndependentTxAuxData) ⇒ EraTxAuxData era where

TxAuxData which may be attached to a transaction

Associated Types

type TxAuxData era = (r ∷ Type) | r → era

Methods

mkBasicTxAuxData ∷ TxAuxData era

metadataTxAuxDataL ∷ Lens' (TxAuxData era) (Map Word64 Metadatum)

validateTxAuxData ∷ ProtVer → TxAuxData era → Bool

type family TxAuxData era = (r ∷ Type) | r → era

class (EraTxOut era, EraTxCert era, EraPParams era, HashAnnotated (TxBody era) EraIndependentTxBody, DecCBOR (Annotator (TxBody era)), EncCBOR (TxBody era), ToCBOR (TxBody era), NoThunks (TxBody era), NFData (TxBody era), Show (TxBody era), Eq (TxBody era), EqRaw (TxBody era)) ⇒ EraTxBody era where

Minimal complete definition

mkBasicTxBody, inputsTxBodyL, outputsTxBodyL, feeTxBodyL, withdrawalsTxBodyL, auxDataHashTxBodyL, spendableInputsTxBodyF, allInputsTxBodyF, certsTxBodyL

Associated Types

data TxBody era

The body of a transaction.

Methods

mkBasicTxBody ∷ TxBody era

inputsTxBodyL ∷ Lens' (TxBody era) (Set TxIn)

outputsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (TxOut era))

feeTxBodyL ∷ Lens' (TxBody era) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody era) Withdrawals

auxDataHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe TxAuxDataHash)

spendableInputsTxBodyF ∷ SimpleGetter (TxBody era) (Set TxIn)

This getter will produce all inputs from the UTxO map that this transaction might spend, which ones will depend on the validity of the transaction itself. Starting in Alonzo this will include collateral inputs.

allInputsTxBodyF ∷ SimpleGetter (TxBody era) (Set TxIn)

This getter will produce all inputs from the UTxO map that this transaction is referencing, even if some of them cannot be spent by the transaction. For example starting with Babbage era it will also include reference inputs.

certsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (TxCert era))

getTotalDepositsTxBody

Arguments

∷ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether stake pool is registered or not

→ TxBody era

→ Coin

Compute the total deposits from the certificates in a TxBody.

This is the contribution of a TxBody towards the consumed amount by the transaction

getTotalRefundsTxBody

Arguments

∷ PParams era

→ (Credential 'Staking → Maybe Coin)
Lookup current deposit for Staking credential if one is registered

→ (Credential 'DRepRole → Maybe Coin)
Lookup current deposit for DRep credential if one is registered

→ TxBody era

→ Coin

Compute the total refunds from the Certs of a TxBody.

This is the contribution of a TxBody towards produced amount by the transaction

getGenesisKeyHashCountTxBody ∷ TxBody era → Int

This function is not used in the ledger rules. It is only used by the downstream tooling to figure out how many witnesses should be supplied for Genesis keys.

data family TxBody era

The body of a transaction.

pattern RetirePoolTxCert ∷ EraTxCert era ⇒ KeyHash 'StakePool → EpochNo → TxCert era

pattern RegPoolTxCert ∷ EraTxCert era ⇒ PoolParams → TxCert era

hashKey ∷ ∀ (kd ∷ KeyRole). VKey kd → KeyHash kd

Hash a given public key

hashScript ∷ EraScript era ⇒ Script era → ScriptHash

Compute ScriptHash of a Script for a particular era.

absurdEraRule ∷ ∀ (rule ∷ Symbol) era a. VoidEraRule rule era → a

eraProtVerLow ∷ Era era ⇒ Version

Get the value level Version of the lowest major protocol version for the supplied era.

eraProtVerHigh ∷ Era era ⇒ Version

Get the value level Version of the highest major protocol version for the supplied era.

eraProtVersions ∷ Era era ⇒ [Version]

List with all major versions that are used in the particular era.

atLeastEra ∷ ∀ (eraName ∷ Symbol) era. AtLeastEra eraName era ⇒ ()

Enforce era to be at least the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

atLeastEra @"Babbage" @ConwayEra
atLeastEra @"Babbage" @BabbageEra

However this will result in a type error

atLeastEra @"Babbage" @AlonzoEra

atMostEra ∷ ∀ (eraName ∷ Symbol) era. AtMostEra eraName era ⇒ ()

Enforce era to be at most the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

atMostEra @BabbageEra @ShelleyEra
atMostEra @AlonzoEra @MaryEra

However this will result in a type error

atMostEra @BabbageEra @ConwayEra

notSupportedInThisEra ∷ HasCallStack ⇒ a

notSupportedInThisEraL ∷ HasCallStack ⇒ Lens' a b

toEraCBOR ∷ (Era era, EncCBOR t) ⇒ t → Encoding

Convert a type that implements EncCBOR to plain Encoding using the lowest protocol version for the supplied era.

fromEraCBOR ∷ (Era era, DecCBOR t) ⇒ Decoder s t

Convert a type that implements DecCBOR to plain Decoder using the lowest protocol version for the supplied era

This action should not be used for decoders that require access to original bytes, use toPlainDecoder instead.

fromEraShareCBOR ∷ (Era era, DecShareCBOR t) ⇒ Decoder s t

Convert a type that implements DecShareCBOR to plain Decoder using the lowest protocol version for the supplied era

This action should not be used for decoders that require access to original bytes, use toPlainDecoder instead.

eraDecoder ∷ ∀ era t s. Era era ⇒ Decoder s t → Decoder s t

Convert a versioned Decoder to plain a Decoder using the lowest protocol version for the supplied era

This action should not be used for decoders that require access to original bytes, use eraDecoderWithBytes instead.

eraDecoderWithBytes ∷ ∀ era t s. Era era ⇒ ByteString → Decoder s t → Decoder s t

Just like eraDecoder, but for decoders that rely on access for underlying bytes

hashTxBodySignature ∷ SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody) → Hash HASH (SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody))

Hash a given signature

toVRFVerKeyHash ∷ ∀ v (r ∷ KeyRoleVRF). Hash HASH (VerKeyVRF v) → VRFVerKeyHash r

fromVRFVerKeyHash ∷ ∀ (r ∷ KeyRoleVRF) v. VRFVerKeyHash r → Hash HASH (VerKeyVRF v)

extractHash ∷ SafeHash i → Hash HASH i

Extract the hash out of a SafeHash

castSafeHash ∷ SafeHash i → SafeHash j

To change the index parameter of SafeHash (which is a phantom type) use castSafeHash

standardHashSize ∷ Int

standardAddrHashSize ∷ Int

translateEra' ∷ (TranslateEra era f, TranslationError era f ~ Void) ⇒ TranslationContext era → f (PreviousEra era) → f era

Variant of translateEra for when TranslationError is Void and the translation thus cannot fail.

translateEraMaybe ∷ (TranslateEra era f, TranslationError era f ~ ()) ⇒ TranslationContext era → f (PreviousEra era) → Maybe (f era)

Variant of translateEra for when TranslationError is (), converting the result to a Maybe.

translateEraThroughCBOR

Arguments

∷ (Era era, ToCBOR (ti (PreviousEra era)), DecCBOR (Annotator (to era)))

⇒ Text
Label for error reporting

→ ti (PreviousEra era)

→ Except DecoderError (to era)

Translate a type through its binary representation from previous era to the current one.

genericApplyPPUpdates ∷ (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) ⇒ PParams era → PParamsUpdate era → PParams era

emptyPParams ∷ EraPParams era ⇒ PParams era

emptyPParamsUpdate ∷ EraPParams era ⇒ PParamsUpdate era

ppLensHKD ∷ ∀ era f. Functor f ⇒ (PParamsHKD Identity era → f (PParamsHKD Identity era)) → PParams era → f (PParams era)

ppuLensHKD ∷ ∀ era f. Functor f ⇒ (PParamsHKD StrictMaybe era → f (PParamsHKD StrictMaybe era)) → PParamsUpdate era → f (PParamsUpdate era)

ppMinFeeAL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The linear factor for the minimum fee calculation

ppMinFeeBL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The constant factor for the minimum fee calculation

ppMaxBBSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word32

Maximal block body size

ppMaxTxSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word32

Maximal transaction size

ppMaxBHSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word16

Maximal block header size

ppKeyDepositL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The amount of a key registration deposit

ppPoolDepositL ∷ (EraPParams era, HasCallStack) ⇒ Lens' (PParams era) Coin

The amount of a pool registration deposit

ppPoolDepositCompactL ∷ EraPParams era ⇒ Lens' (PParams era) (CompactForm Coin)

The amount of a pool registration deposit in compacted form

ppEMaxL ∷ EraPParams era ⇒ Lens' (PParams era) EpochInterval

epoch bound on pool retirement

ppNOptL ∷ EraPParams era ⇒ Lens' (PParams era) Word16

Desired number of pools

ppA0L ∷ EraPParams era ⇒ Lens' (PParams era) NonNegativeInterval

Pool influence

ppRhoL ∷ EraPParams era ⇒ Lens' (PParams era) UnitInterval

Monetary expansion

ppTauL ∷ EraPParams era ⇒ Lens' (PParams era) UnitInterval

Treasury expansion

ppDL ∷ (EraPParams era, AtMostEra "Alonzo" era) ⇒ Lens' (PParams era) UnitInterval

Decentralization parameter

ppExtraEntropyL ∷ (EraPParams era, AtMostEra "Alonzo" era) ⇒ Lens' (PParams era) Nonce

Extra entropy

ppMinUTxOValueL ∷ (EraPParams era, AtMostEra "Mary" era) ⇒ Lens' (PParams era) Coin

Minimum UTxO value

ppMinPoolCostL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

Minimum Stake Pool Cost

ppuMinFeeAL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The linear factor for the minimum fee calculation

ppuMinFeeBL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The constant factor for the minimum fee calculation

ppuMaxBBSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal block body size

ppuMaxTxSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal transaction size

ppuMaxBHSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word16)

Maximal block header size

ppuKeyDepositL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a key registration deposit

ppuPoolDepositL ∷ (EraPParams era, HasCallStack) ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a pool registration deposit. The value must be small enough to fit into a Word64.

ppuPoolDepositCompactL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe (CompactForm Coin))

The amount of a pool registration deposit in compacted form

ppuEMaxL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

epoch bound on pool retirement

ppuNOptL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word16)

Desired number of pools

ppuA0L ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe NonNegativeInterval)

Pool influence

ppuRhoL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Monetary expansion

ppuTauL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Treasury expansion

ppuDL ∷ (EraPParams era, AtMostEra "Alonzo" era) ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Decentralization parameter

ppuExtraEntropyL ∷ (EraPParams era, AtMostEra "Alonzo" era) ⇒ Lens' (PParamsUpdate era) (StrictMaybe Nonce)

Extra entropy

ppuMinUTxOValueL ∷ (EraPParams era, AtMostEra "Mary" era) ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum UTxO value

ppuMinPoolCostL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum Stake Pool Cost

mapPParams ∷ (PParamsHKD Identity era1 → PParamsHKD Identity era2) → PParams era1 → PParams era2

mapPParamsUpdate ∷ (PParamsHKD StrictMaybe era1 → PParamsHKD StrictMaybe era2) → PParamsUpdate era1 → PParamsUpdate era2

upgradePParams ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ UpgradePParams Identity era → PParams (PreviousEra era) → PParams era

downgradePParams ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ DowngradePParams Identity era → PParams era → PParams (PreviousEra era)

upgradePParamsUpdate ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ UpgradePParams StrictMaybe era → PParamsUpdate (PreviousEra era) → PParamsUpdate era

downgradePParamsUpdate ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ DowngradePParams StrictMaybe era → PParamsUpdate era → PParamsUpdate (PreviousEra era)

getPoolCertTxCert ∷ EraTxCert era ⇒ TxCert era → Maybe PoolCert

poolCertKeyHashWitness ∷ PoolCert → KeyHash 'Witness

isRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake registering certificate

isUnRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake un-registering certificate

bootAddrTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) (Maybe BootstrapAddress)

coinTxOutL ∷ (HasCallStack, EraTxOut era) ⇒ Lens' (TxOut era) Coin

compactCoinTxOutL ∷ (HasCallStack, EraTxOut era) ⇒ Lens' (TxOut era) (CompactForm Coin)

isAdaOnlyTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) Bool

This is a getter that implements an efficient way to check whether TxOut contains ADA only.

mkCoinTxOut ∷ EraTxOut era ⇒ Addr → Coin → TxOut era

hashTxAuxData ∷ EraTxAuxData era ⇒ TxAuxData era → TxAuxDataHash

Compute a hash of TxAuxData

hashScriptTxWitsL ∷ EraTxWits era ⇒ Lens (TxWits era) (TxWits era) (Map ScriptHash (Script era)) [Script era]

This is a helper lens that will hash the scripts when adding as witnesses.

keyHashWitnessesTxWits ∷ EraTxWits era ⇒ TxWits era → Set (KeyHash 'Witness)

Extract all of the KeyHash witnesses provided in the TxWits

isNativeScript ∷ EraScript era ⇒ Script era → Bool

bBodySize ∷ EraBlockBody era ⇒ ProtVer → BlockBody era → Int

txIdTx ∷ EraTx era ⇒ Tx era → TxId

txIdTxBody ∷ EraTxBody era ⇒ TxBody era → TxId

wireSizeTxF ∷ EraTx era ⇒ SimpleGetter (Tx era) Word32

txsize computes the length of the serialised bytes (actual size)

binaryUpgradeTx ∷ (Era era, ToCBOR (Tx (PreviousEra era)), DecCBOR (Annotator (Tx era))) ⇒ Tx (PreviousEra era) → Except DecoderError (Tx era)

Translate a transaction through its binary representation from previous to current era.

binaryUpgradeTxBody ∷ (Era era, ToCBOR (TxBody (PreviousEra era)), DecCBOR (Annotator (TxBody era))) ⇒ TxBody (PreviousEra era) → Except DecoderError (TxBody era)

Translate a tx body through its binary representation from previous to current era.

binaryUpgradeTxWits ∷ (Era era, ToCBOR (TxWits (PreviousEra era)), DecCBOR (Annotator (TxWits era))) ⇒ TxWits (PreviousEra era) → Except DecoderError (TxWits era)

Translate tx witnesses through its binary representation from previous to current era.

binaryUpgradeTxAuxData ∷ (Era era, ToCBOR (TxAuxData (PreviousEra era)), DecCBOR (Annotator (TxAuxData era))) ⇒ TxAuxData (PreviousEra era) → Except DecoderError (TxAuxData era)

Translate tx auxData through its binary representation from previous to current era.

module Cardano.Ledger.Shelley.Governance