Cardano.Ledger.Shelley.Delegation

Documentation

class EraDCert era ⇒ ShelleyEraDCert era where

Methods

mkDCertMir ∷ ProtVerAtMost era 8 ⇒ MIRCert (EraCrypto era) → DCert era

getDCertMir ∷ DCert era → Maybe (MIRCert (EraCrypto era))

mkShelleyDCertDeleg ∷ ShelleyDelegCert (EraCrypto era) → DCert era

getShelleyDCertDeleg ∷ DCert era → Maybe (ShelleyDelegCert (EraCrypto era))

pattern DCertMir ∷ (ShelleyEraDCert era, ProtVerAtMost era 8) ⇒ MIRCert (EraCrypto era) → DCert era

pattern ShelleyDCertDeleg ∷ ShelleyEraDCert era ⇒ ShelleyDelegCert (EraCrypto era) → DCert era

data ShelleyDelegCert c

Constructors

RegKey !(StakeCredential c)
A stake key registration certificate.

DeRegKey !(StakeCredential c)
A stake key deregistration certificate.

Delegate !(Delegation c)
A stake delegation certificate.

delegCWitness ∷ ShelleyDelegCert c → Credential 'Staking c

Determine the certificate author

data ShelleyDCert era

A heavyweight certificate.

Constructors

ShelleyDCertDelegCert !(ShelleyDelegCert (EraCrypto era))

ShelleyDCertPool !(PoolCert (EraCrypto era))

ShelleyDCertGenesis !(ConstitutionalDelegCert (EraCrypto era))

ShelleyDCertMir !(MIRCert (EraCrypto era))

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ !MIRPot

mirRewards ∷ !(MIRTarget c)

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

isRegKey ∷ ShelleyEraDCert era ⇒ DCert era → Bool

Check for RegKey constructor

isDeRegKey ∷ ShelleyEraDCert era ⇒ DCert era → Bool

Check for DeRegKey constructor

isDelegation ∷ ShelleyEraDCert era ⇒ DCert era → Bool

Check for Delegation constructor

isRegPool ∷ EraDCert era ⇒ DCert era → Bool

Check for RegPool constructor

isRetirePool ∷ EraDCert era ⇒ DCert era → Bool

Check for RetirePool constructor

isGenesisDelegation ∷ EraDCert era ⇒ DCert era → Bool

Check for GenesisDelegate constructor

isInstantaneousRewards ∷ ShelleyEraDCert era ⇒ DCert era → Bool

isReservesMIRCert ∷ ShelleyEraDCert era ⇒ DCert era → Bool

isTreasuryMIRCert ∷ ShelleyEraDCert era ⇒ DCert era → Bool

requiresVKeyWitness ∷ ShelleyEraDCert era ⇒ DCert era → Bool

Returns True for delegation certificates that require at least one witness, and False otherwise. It is mainly used to ensure that calling a variant of cwitness is safe.

Serialization helpers

shelleyDCertDelegDecoder ∷ ShelleyEraDCert era ⇒ Word → Decoder s (Int, DCert era)

commonDCertDecoder ∷ EraDCert era ⇒ Word → Decoder s (Int, DCert era)

encodeShelleyDelegCert ∷ Crypto c ⇒ ShelleyDelegCert c → Encoding

encodePoolCert ∷ Crypto c ⇒ PoolCert c → Encoding

encodeConstitutionalCert ∷ Crypto c ⇒ ConstitutionalDelegCert c → Encoding

Re-exports

class (Era era, DecCBOR (DCert era), EncCBOR (DCert era), ToCBOR (DCert era), FromCBOR (DCert era), NoThunks (DCert era), NFData (DCert era), Show (DCert era), Eq (DCert era)) ⇒ EraDCert era where

Associated Types

type DCert era = (r ∷ Type) | r → era

Methods

mkDCertPool ∷ PoolCert (EraCrypto era) → DCert era

getDCertPool ∷ DCert era → Maybe (PoolCert (EraCrypto era))

mkDCertGenesis ∷ ConstitutionalDelegCert (EraCrypto era) → DCert era

getDCertGenesis ∷ DCert era → Maybe (ConstitutionalDelegCert (EraCrypto era))

pattern DCertPool ∷ EraDCert era ⇒ PoolCert (EraCrypto era) → DCert era

pattern DCertGenesis ∷ EraDCert era ⇒ ConstitutionalDelegCert (EraCrypto era) → DCert era

data Delegation c

The delegation of one stake key to another.

Constructors

Delegation

Fields

dDelegator ∷ !(StakeCredential c)

dDelegatee ∷ !(KeyHash 'StakePool c)

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

poolCWitness ∷ PoolCert c → Credential 'StakePool c

data ConstitutionalDelegCert c

Constitutional key delegation certificate

Constructors

ConstitutionalDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

genesisCWitness ∷ ConstitutionalDelegCert c → KeyHash 'Genesis c

Orphan instances

Crypto c ⇒ EraDCert (ShelleyEra c)

Instance details

Associated Types

type DCert (ShelleyEra c) = (r ∷ Type)

Methods

mkDCertPool ∷ PoolCert (EraCrypto (ShelleyEra c)) → DCert (ShelleyEra c)

getDCertPool ∷ DCert (ShelleyEra c) → Maybe (PoolCert (EraCrypto (ShelleyEra c)))

mkDCertGenesis ∷ ConstitutionalDelegCert (EraCrypto (ShelleyEra c)) → DCert (ShelleyEra c)

getDCertGenesis ∷ DCert (ShelleyEra c) → Maybe (ConstitutionalDelegCert (EraCrypto (ShelleyEra c)))