Cardano.Ledger.Shelley.Delegation.Certificates

Documentation

data DCert crypto

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert crypto)

DCertPool !(PoolCert crypto)

DCertGenesis !(GenesisDelegCert crypto)

DCertMir !(MIRCert crypto)

data DelegCert crypto

Constructors

RegKey !(StakeCredential crypto)
A stake key registration certificate.

DeRegKey !(StakeCredential crypto)
A stake key deregistration certificate.

Delegate !(Delegation crypto)
A stake delegation certificate.

data PoolCert crypto

Constructors

RegPool !(PoolParams crypto)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool crypto) !EpochNo
A stake pool retirement certificate.

data GenesisDelegCert crypto

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis crypto) !(KeyHash 'GenesisDelegate crypto) !(Hash crypto (VerKeyVRF crypto))

data MIRCert crypto

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot :: MIRPot

mirRewards :: MIRTarget crypto

newtype StakeCreds crypto

Constructors

StakeCreds

Fields

unStakeCreds :: Map (Credential 'Staking crypto) SlotNo

delegCWitness :: DelegCert crypto -> Credential 'Staking crypto

Determine the certificate author

poolCWitness :: PoolCert crypto -> Credential 'StakePool crypto

genesisCWitness :: GenesisDelegCert crypto -> KeyHash 'Genesis crypto

isRegKey :: DCert crypto -> Bool

Check for RegKey constructor

isDeRegKey :: DCert crypto -> Bool

Check for DeRegKey constructor

isDelegation :: DCert crypto -> Bool

Check for Delegation constructor

isGenesisDelegation :: DCert crypto -> Bool

Check for GenesisDelegate constructor

isRegPool :: DCert crypto -> Bool

Check for RegPool constructor

isRetirePool :: DCert crypto -> Bool

Check for RetirePool constructor

isInstantaneousRewards :: DCert crypto -> Bool

isReservesMIRCert :: DCert crypto -> Bool

isTreasuryMIRCert :: DCert crypto -> Bool

requiresVKeyWitness :: DCert crypto -> Bool

Returns True for delegation certificates that require at least one witness, and False otherwise. It is mainly used to ensure that calling a variant of cwitness is safe.