Cardano.Ledger.Shelley.EpochBoundary

Description

This modules implements the necessary functions for the changes that can happen at epoch boundaries.

Documentation

newtype Stake crypto

Type of stake as map from hash key to coins associated.

Constructors

Stake

Fields

unStake :: VMap VB VP (Credential 'Staking crypto) (CompactForm Coin)

sumAllStake :: Stake crypto -> Coin

sumStakePerPool :: VMap VB VB (Credential 'Staking crypto) (KeyHash 'StakePool crypto) -> Stake crypto -> Map (KeyHash 'StakePool crypto) Coin

data SnapShot crypto

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

_stake :: !(Stake crypto)

_delegations :: !(VMap VB VB (Credential 'Staking crypto) (KeyHash 'StakePool crypto))

_poolParams :: !(VMap VB VB (KeyHash 'StakePool crypto) (PoolParams crypto))

data SnapShots crypto

Snapshots of the stake distribution.

Constructors

SnapShots

Fields

_pstakeMark :: SnapShot crypto

_pstakeSet :: !(SnapShot crypto)

_pstakeGo :: !(SnapShot crypto)

_feeSS :: !Coin

emptySnapShot :: SnapShot crypto

emptySnapShots :: SnapShots crypto

poolStake :: KeyHash 'StakePool crypto -> VMap VB VB (Credential 'Staking crypto) (KeyHash 'StakePool crypto) -> Stake crypto -> Stake crypto

Get stake of one pool

obligation :: forall crypto pp anymap. (HasField "_keyDeposit" pp Coin, HasField "_poolDeposit" pp Coin, Foldable (anymap (Credential 'Staking crypto))) => pp -> anymap (Credential 'Staking crypto) Coin -> Map (KeyHash 'StakePool crypto) (PoolParams crypto) -> Coin

Calculate total possible refunds.

maxPool :: (HasField "_a0" pp NonNegativeInterval, HasField "_nOpt" pp Natural) => pp -> Coin -> Rational -> Rational -> Coin

Version of maxPool' that extracts a0 and nOpt from a PParam with the right HasField instances

maxPool' :: NonNegativeInterval -> Natural -> Coin -> Rational -> Rational -> Coin

Calculate maximal pool reward