Cardano.Ledger.Shelley.Governance

Documentation

class (EraPParams era, Eq (GovState era), Show (GovState era), NoThunks (GovState era), NFData (GovState era), EncCBOR (GovState era), DecCBOR (GovState era), DecShareCBOR (GovState era), Share (GovState era) ~ (Interns (Credential 'Staking), Interns (KeyHash 'StakePool), Interns (Credential 'DRepRole), Interns (Credential 'HotCommitteeRole)), ToCBOR (GovState era), FromCBOR (GovState era), Default (GovState era), ToJSON (GovState era)) ⇒ EraGov era where

Minimal complete definition

curPParamsGovStateL, prevPParamsGovStateL, futurePParamsGovStateL, obligationGovState

Associated Types

type GovState era = (r ∷ Type) | r → era

Methods

emptyGovState ∷ GovState era

Construct empty governance state

curPParamsGovStateL ∷ Lens' (GovState era) (PParams era)

Lens for accessing current protocol parameters

prevPParamsGovStateL ∷ Lens' (GovState era) (PParams era)

Lens for accessing the previous protocol parameters

futurePParamsGovStateL ∷ Lens' (GovState era) (FuturePParams era)

Lens for accessing the future protocol parameters.

This lens will produce DefinitePParamsUpdate whenever we are absolutely sure that the new PParams will be updated. Which means there will be no chance of a DefinitePParamsUpdate value until we are past the point of no return, which is 2 stability windows before the end of the epoch. This lens is mostly intended for ledger usage and nextEpochUpdatedPParams should be used instead whenever definite results are desired.

obligationGovState ∷ GovState era → Obligations

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

sgsFuturePParams ∷ !(FuturePParams era)

Prediction of parameter changes that might happen on the epoch boundary.

emptyShelleyGovState ∷ EraPParams era ⇒ ShelleyGovState era

data FuturePParams era

Constructors

NoPParamsUpdate
This indicates that there is definitely not going to be an update to PParams expected at the next epoch boundary.

DefinitePParamsUpdate !(PParams era)
This case specifies the PParams that will be adopted at the next epoch boundary.

PotentialPParamsUpdate (Maybe (PParams era))
With this case there is no guarantee that these will be the new PParams, users should not rely on this value to be computed efficiently and should use nextEpochPParams instead. The field is lazy on purpose, since we truly need to compute this field only towards the end of the epoch, which is done by solidifyFuturePParams two stability windows before the end of the epoch.

solidifyFuturePParams ∷ FuturePParams era → FuturePParams era

knownFuturePParams ∷ FuturePParams era → Maybe (PParams era)

Return new PParams only when it is known that there was an update proposed and it is guaranteed to be applied

nextEpochPParams ∷ EraGov era ⇒ GovState era → PParams era

This function is guaranteed to produce PParams that will be adopted at the next epoch boundary, whenever this function is applied to the GovState that was produced by ledger at any point that is two stability windows before the end of the epoch. If you need to know if there were actual changes to those PParams then use nextEpochUpdatedPParams instead.

nextEpochUpdatedPParams ∷ EraGov era ⇒ GovState era → StrictMaybe (PParams era)

This function is guaranteed to return updated PParams when it is called during the last two stability windows of the epoch and there were proposals to update PParams that all relevant parties reached consensus on. In other words whenever there is a definite update to PParams coming on the epoch boundary those PParams will be returned, otherwise it will return Nothing. This function is inexpensive and can be invoked at any time without danger of forcing some suspended computation.

curPParamsShelleyGovStateL ∷ Lens' (ShelleyGovState era) (PParams era)

prevPParamsShelleyGovStateL ∷ Lens' (ShelleyGovState era) (PParams era)

futurePParamsShelleyGovStateL ∷ Lens' (ShelleyGovState era) (FuturePParams era)

Orphan instances

EraGov ShelleyEra

Instance details

Associated Types

type GovState ShelleyEra = (r ∷ Type)

Methods

emptyGovState ∷ GovState ShelleyEra

curPParamsGovStateL ∷ Lens' (GovState ShelleyEra) (PParams ShelleyEra)

prevPParamsGovStateL ∷ Lens' (GovState ShelleyEra) (PParams ShelleyEra)

futurePParamsGovStateL ∷ Lens' (GovState ShelleyEra) (FuturePParams ShelleyEra)

obligationGovState ∷ GovState ShelleyEra → Obligations