Cardano.Ledger.Shelley.HardForks

Documentation

aggregatedRewards :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

allowMIRTransfer :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

Starting with protocol version 5, the MIR certs will also be able to transfer funds between the reserves and the treasury. Additionally, the semantics for the pervious functionality will change a bit. Before version 5 redundancies in the instantaneous reward mapping were handled by overriding. Now they are handled by adding the values and allowing for negatives updates, provided the sum for each key remains positive.

validatePoolRewardAccountNetID :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

Starting with protocol version 5, we will validate the network ID for the reward account listed in stake pool registration certificates.

allowScriptStakeCredsToEarnRewards :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

Starting with protocol version 5, Stake Credentials bound by scripts will be eligibile for staking rewards.

translateTimeForPlutusScripts :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

Starting with protocol version 6, we translate slots to time correctly for Plutus scripts.

missingScriptsSymmetricDifference :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

Starting with protocol version 7, the UTXO rule predicate failure MissingScriptWitnessesUTXOW will not be used for extraneous scripts

forgoRewardPrefilter :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

Starting with protocol version 7, the reward calculation no longer filters out unregistered stake addresses at the moment the calculation begins. See the Shelley Ledger Errata 17.2.

allowOutsideForecastTTL :: HasField "_protocolVersion" pp ProtVer => pp -> Bool

In versions 5 and 6, we allow the ttl field to lie outside the stability window.