Cardano.Ledger.Shelley.Internal

Description

Provide Tools for debugging Feel free to add new things as they are developed.

Documentation

trace ∷ String → a → a

The trace function outputs the trace message given as its first argument, before returning the second argument as its result.

For example, this returns the value of f x and outputs the message to stderr. Depending on your terminal (settings), they may or may not be mixed.

>>> let x = 123; f = show
>>> trace ("calling f with x = " ++ show x) (f x)
calling f with x = 123
"123"

The trace function should only be used for debugging, or for monitoring execution. The function is not referentially transparent: its type indicates that it is a pure function but it has the side effect of outputting the trace message.

totalAdaPotsES ∷ (EraTxOut era, EraGov era) ⇒ EpochState era → AdaPots

Calculate the total ada pots in the epoch state

compareAdaPots ∷ String → AdaPots → String → AdaPots → String

producedTxBody ∷ EraTxBody era ⇒ TxBody era → PParams era → CertState era → Produced

Compute the Coin part of what is produced by a TxBody, itemized as a Produced

consumedTxBody ∷ EraTxBody era ⇒ TxBody era → PParams era → CertState era → UTxO era → Consumed

Compute the Coin part of what is consumed by a TxBody, itemized as a Consume

showCred ∷ Credential x c → String

showIR ∷ InstantaneousRewards c → String

showKeyHash ∷ KeyHash c x → String

showListy ∷ Foldable t ⇒ (a → String) → t a → String

showMap ∷ (t1 → [Char]) → (t2 → [Char]) → Map t1 t2 → String

showWithdrawal ∷ Withdrawals c → String

showSafeHash ∷ SafeHash c i → String

synopsisCoinMap ∷ Maybe (Map k Coin) → String

Display a synopsis of a map to Coin

showTxCerts ∷ EraTxBody era ⇒ TxBody era → String