Cardano.Ledger.Shelley.LedgerState

Description

This module implements the operation rules for treating UTxO transactions (Tx) as state transformations on a ledger state (LedgerState), as specified in A Simplified Formal Specification of a UTxO Ledger.

UTxO

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map TxIn (TxOut era)

class CanGetUTxO (t ∷ Type → Type) where

Minimal complete definition

Nothing

Methods

utxoG ∷ SimpleGetter (t era) (UTxO era)

class CanGetUTxO t ⇒ CanSetUTxO (t ∷ Type → Type) where

Methods

utxoL ∷ Lens' (t era) (UTxO era)

Others to organize

type family AccountState era = (r ∷ Type) | r → era

data ChainAccountState where

Bundled Patterns

pattern AccountState ∷ Coin → Coin → ChainAccountState

class (EraAccounts era, ToJSON (CertState era), EncCBOR (CertState era), DecShareCBOR (CertState era), Share (CertState era) ~ (Interns (Credential 'Staking), Interns (KeyHash 'StakePool), Interns (Credential 'DRepRole), Interns (Credential 'HotCommitteeRole)), Default (CertState era), NoThunks (CertState era), NFData (CertState era), Show (CertState era), Eq (CertState era)) ⇒ EraCertState era where

The state associated with the DELPL rule, which combines the DELEG rule and the POOL rule.

Associated Types

type CertState era = (r ∷ Type) | r → era

Methods

certDStateL ∷ Lens' (CertState era) (DState era)

certPStateL ∷ Lens' (CertState era) (PState era)

obligationCertState ∷ CertState era → Obligations

Calculate total possible refunds in the system that are related to certificates

There is an invariant that the sum of all the fields should be the same as the utxosDeposited field of the UTxOState. Note that this does not depend upon the current values of the Key and Pool deposits of the PParams.

certsTotalDepositsTxBody ∷ PParams era → CertState era → TxBody era → Coin

Compute the total deposits from the Certs of a TxBody.

This is the contribution of a TxBody towards the deposit pot (utxosDeposit field of the UTxOState) of the system

certsTotalRefundsTxBody ∷ PParams era → CertState era → TxBody era → Coin

Compute the total refunds from the Certs of a TxBody.

This is the contribution of a TxBody towards the total Obligations of the system See Obligations and obligationCertState for more information.

data DState era

The state used by the DELEG rule, which roughly tracks stake delegation and some governance features.

Constructors

DState

Fields

dsAccounts ∷ !(Accounts era)

Keep track of the account state (eg. balance, deposit, stake-pool delegation, etc.) for all registered stake credentials.

dsFutureGenDelegs ∷ !(Map FutureGenDeleg GenDelegPair)

Future genesis key delegations

dsGenDelegs ∷ !GenDelegs

Genesis key delegations

dsIRewards ∷ !InstantaneousRewards

Instantaneous Rewards

data EpochState era

Constructors

EpochState

Fields

esChainAccountState ∷ !ChainAccountState

esLState ∷ !(LedgerState era)

esSnapshots ∷ !SnapShots

esNonMyopic ∷ !NonMyopic

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://github.com/intersectmbo/cardano-ledger/releases/latest/download/pool-ranking.pdf

data PulsingRewUpdate

The state used in the STS rules

Constructors

Pulsing !RewardSnapShot !Pulser

Complete !RewardUpdate

data FutureGenDeleg

Constructors

FutureGenDeleg

Fields

fGenDelegSlot ∷ !SlotNo

fGenDelegGenKeyHash ∷ !(KeyHash 'Genesis)

data InstantaneousRewards

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking) Coin)

iRTreasury ∷ !(Map (Credential 'Staking) Coin)

deltaReserves ∷ !DeltaCoin

deltaTreasury ∷ !DeltaCoin

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState ∷ !(UTxOState era)

The current unspent transaction outputs.

lsCertState ∷ !(CertState era)

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psVRFKeyHashes ∷ !(Map (VRFVerKeyHash 'StakePoolVRF) (NonZero Word64))

VRF key hashes that have been registered via PoolParams

psStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

The state of current stake pools.

psFutureStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

The state of future stake pools. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring ∷ !(Map (KeyHash 'StakePool) EpochNo)

A map of retiring stake pools to the epoch when they retire.

type RewardAccounts = Map (Credential 'Staking) Coin

Deprecated: In favor of Map (Credential Staking) Coin

data RewardUpdate

The ultimate goal of a reward update computation. Aggregating rewards for each staking credential.

Constructors

RewardUpdate

Fields

deltaT ∷ !DeltaCoin

deltaR ∷ !DeltaCoin

rs ∷ !(Map (Credential 'Staking) (Set Reward))

deltaF ∷ !DeltaCoin

nonMyopic ∷ !NonMyopic

data RewardSnapShot

To complete the reward update, we need a snap shot of the EpochState particular to this computation

Constructors

RewardSnapShot

Fields

rewFees ∷ !Coin

rewProtocolVersion ∷ !ProtVer

rewNonMyopic ∷ !NonMyopic

rewDeltaR1 ∷ !Coin

rewR ∷ !Coin

rewDeltaT1 ∷ !Coin

rewLikelihoods ∷ !(Map (KeyHash 'StakePool) Likelihood)

rewLeaders ∷ !(Map (Credential 'Staking) (Set Reward))

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very careful, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

utxosUtxo ∷ !(UTxO era)

utxosDeposited ∷ !Coin

utxosFees ∷ !Coin

utxosGovState ∷ !(GovState era)

utxosInstantStake ∷ !(InstantStake era)

utxosDonation ∷ !Coin

smartUTxOState ∷ EraStake era ⇒ PParams era → UTxO era → Coin → Coin → GovState era → Coin → UTxOState era

A valid (or self-consistent) UTxOState{utxosUtxo, utxosDeposited , utxosFees , utxosPpups , utxosStakeDistr} maintains an invariant between the utxosUtxo and utxosStakeDistr fields. the utxosStakeDistr field is the aggregation of Coin over the StakeReferences in the UTxO. It can be computed by a pure function from the _utxo field. In some situations, mostly unit or example tests, or when initializing a small UTxO, we want to create a UTxOState that computes the utxosStakeDistr from the utxosUtxo. This is aways safe to do, but if the utxosUtxo field is big, this can be very expensive, which defeats the purpose of memoizing the utxosStakeDistr field. So use of this function should be restricted to tests and initializations, where the invariant should be maintained.

TO IncrementalStake

mkShelleyCertState ∷ EraCertState era ⇒ PState era → DState era → CertState era

data ShelleyCertState era

Constructors

ShelleyCertState

Fields

shelleyCertPState ∷ !(PState era)

shelleyCertDState ∷ !(DState era)

Genesis State

genesisState ∷ ∀ era. (EraGov era, EraCertState era, EraStake era) ⇒ Map (KeyHash 'Genesis) GenDelegPair → UTxO era → LedgerState era

Creates the ledger state for an empty ledger which contains the specified transaction outputs.

Validation

consumed ∷ EraUTxO era ⇒ PParams era → CertState era → UTxO era → TxBody era → Value era

produced ∷ (EraUTxO era, EraCertState era) ⇒ PParams era → CertState era → TxBody era → Value era

Compute the lovelace which are created by the transaction For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

DelegationState

totalObligation ∷ (EraGov era, EraCertState era) ⇒ CertState era → GovState era → Coin

allObligations ∷ (EraGov era, EraCertState era) ⇒ CertState era → GovState era → Obligations

Epoch boundary

applyRUpd ∷ (EraGov era, EraCertState era) ⇒ RewardUpdate → EpochState era → EpochState era

Apply a RewardUpdate to the EpochState. Does several things 1) Adds reward coins to Rewards component of the UMap field of the DState, for actively delegated Stake 2) Adds to the Treasury of the ChainAccountState for non-actively delegated stake 3) Adds fees to the UTxOState

applyRUpdFiltered ∷ (EraGov era, EraCertState era) ⇒ RewardUpdate → EpochState era → (EpochState era, FilteredRewards era)

filterAllRewards ∷ (EraGov era, EraCertState era) ⇒ Map (Credential 'Staking) (Set Reward) → EpochState era → FilteredRewards era

data FilteredRewards era

Constructors

FilteredRewards

Fields

frRegistered ∷ !(Map (Credential 'Staking) (Set Reward))

These are registered, in the current Unified map of the CertState

frShelleyIgnored ∷ Map (Credential 'Staking) (Set Reward)

These are registered, but ignored in the ShelleyEra because of backward compatibility in non-Shelley Eras, this field will be Map.empty

frUnregistered ∷ Set (Credential 'Staking)

These are NOT registered in the current Unified map of the CertState

frTotalUnregistered ∷ Coin

Total Coin of the unregistered rewards. These will end up in the Treasury or Reserves.

createRUpd ∷ ∀ era. (EraGov era, EraCertState era) ⇒ EpochSize → BlocksMade → EpochState era → Coin → ActiveSlotCoeff → NonZero Word64 → ShelleyBase RewardUpdate

To create a reward update, run all 3 phases This function is not used in the rules, so it ignores RewardEvents

completeRupd ∷ PulsingRewUpdate → ShelleyBase (RewardUpdate, RewardEvent)

Phase 3 of reward update has several parts a) completeM the pulser (in case there are still computions to run) b) Combine the pulser provenance with the RewardProvenance c) Construct the final RewardUpdate d) Add the leader rewards to both the events and the computed Rewards

startStep ∷ ∀ era. (EraGov era, EraCertState era) ⇒ EpochSize → BlocksMade → EpochState era → Coin → ActiveSlotCoeff → NonZero Word64 → PulsingRewUpdate

pulseStep ∷ PulsingRewUpdate → ShelleyBase (PulsingRewUpdate, RewardEvent)

Run the pulser for a bit. If is has nothing left to do, complete it.

completeStep ∷ PulsingRewUpdate → ShelleyBase (PulsingRewUpdate, RewardEvent)

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState !EpochNo !BlocksMade !BlocksMade !(EpochState era) !(StrictMaybe PulsingRewUpdate) !PoolDistr !(StashedAVVMAddresses era)

type family StashedAVVMAddresses era where ...

Equations

StashedAVVMAddresses ShelleyEra = UTxO ShelleyEra

StashedAVVMAddresses _ = ()

stashedAVVMAddresses ∷ NewEpochState era → StashedAVVMAddresses era

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

getGKeys ∷ EraCertState era ⇒ NewEpochState era → Set (KeyHash 'Genesis)

updateNES ∷ EraGov era ⇒ NewEpochState era → BlocksMade → LedgerState era → NewEpochState era

Update new epoch state

circulation ∷ EpochState era → Coin → Coin

Calculate the current circulation

This is used in the rewards calculation, and for API endpoints for pool ranking.

Decay

decayFactor ∷ Float

Remove Bootstrap Redeem Addresses

returnRedeemAddrsToReserves ∷ ∀ era. EraTxOut era ⇒ EpochState era → EpochState era

updateNonMyopic ∷ NonMyopic → Coin → Map (KeyHash 'StakePool) Likelihood → NonMyopic

emptyRewardUpdate ∷ RewardUpdate

pvCanFollow

Arguments

∷ ProtVer
Previous protocol version

→ ProtVer
New protocol version

→ Bool

Check whether the new protocol version is a legitimate version bump with respect to the previous one.

availableAfterMIR ∷ MIRPot → ChainAccountState → InstantaneousRewards → Coin

This function returns the coin balance of a given pot, either the reserves or the treasury, after the instantaneous rewards and pot transfers are accounted for.

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

sgsFuturePParams ∷ !(FuturePParams era)

Prediction of parameter changes that might happen on the epoch boundary.

Lenses from Types

nesPdL ∷ Lens' (NewEpochState era) PoolDistr

nesEsL ∷ Lens' (NewEpochState era) (EpochState era)

nesELL ∷ Lens' (NewEpochState era) EpochNo

nesBprevL ∷ Lens' (NewEpochState era) (Map (KeyHash 'StakePool) Natural)

nesBcurL ∷ Lens' (NewEpochState era) (Map (KeyHash 'StakePool) Natural)

nesRuL ∷ Lens' (NewEpochState era) (StrictMaybe PulsingRewUpdate)

nesStashedAVVMAddressesL ∷ Lens' (NewEpochState era) (StashedAVVMAddresses era)

nesEpochStateL ∷ Lens' (NewEpochState era) (EpochState era)

esAccountStateL ∷ Lens' (EpochState era) ChainAccountState

Deprecated: In favor of chainAccountStateL

esSnapshotsL ∷ Lens' (EpochState era) SnapShots

esLStateL ∷ Lens' (EpochState era) (LedgerState era)

esNonMyopicL ∷ Lens' (EpochState era) NonMyopic

curPParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (PParams era)

prevPParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (PParams era)

futurePParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (FuturePParams era)

asTreasuryL ∷ Lens' ChainAccountState Coin

Deprecated: In favor of casTreasuryL

asReservesL ∷ Lens' ChainAccountState Coin

Deprecated: In favor of casReservesL

lsUTxOStateL ∷ Lens' (LedgerState era) (UTxOState era)

lsCertStateL ∷ Lens' (LedgerState era) (CertState era)

utxosUtxoL ∷ Lens' (UTxOState era) (UTxO era)

Deprecated: In favor of utxoL

utxosDepositedL ∷ Lens' (UTxOState era) Coin

utxosFeesL ∷ Lens' (UTxOState era) Coin

utxosGovStateL ∷ Lens' (UTxOState era) (GovState era)

utxosDonationL ∷ Lens' (UTxOState era) Coin

epochStateGovStateL ∷ Lens' (EpochState era) (GovState era)

epochStateStakeDistrL ∷ Lens' (EpochState era) (VMap VB VP (Credential 'Staking) (CompactForm Coin))

epochStateStakePoolsL ∷ EraCertState era ⇒ Lens' (EpochState era) (Map (KeyHash 'StakePool) StakePoolState)

epochStateDonationL ∷ Lens' (EpochState era) Coin

newEpochStateGovStateL ∷ Lens' (NewEpochState era) (GovState era)

epochStateTreasuryL ∷ Lens' (EpochState era) Coin

Deprecated: In favor of treasuryL

Lenses from CertState

dsGenDelegsL ∷ ∀ era f. Functor f ⇒ (GenDelegs → f GenDelegs) → DState era → f (DState era)

dsIRewardsL ∷ ∀ era f. Functor f ⇒ (InstantaneousRewards → f InstantaneousRewards) → DState era → f (DState era)

dsFutureGenDelegsL ∷ ∀ era f. Functor f ⇒ (Map FutureGenDeleg GenDelegPair → f (Map FutureGenDeleg GenDelegPair)) → DState era → f (DState era)

psStakePoolsL ∷ ∀ era f. Functor f ⇒ (Map (KeyHash 'StakePool) StakePoolState → f (Map (KeyHash 'StakePool) StakePoolState)) → PState era → f (PState era)

psFutureStakePoolsL ∷ ∀ era f. Functor f ⇒ (Map (KeyHash 'StakePool) StakePoolState → f (Map (KeyHash 'StakePool) StakePoolState)) → PState era → f (PState era)

psRetiringL ∷ ∀ era f. Functor f ⇒ (Map (KeyHash 'StakePool) EpochNo → f (Map (KeyHash 'StakePool) EpochNo)) → PState era → f (PState era)

psVRFKeyHashesL ∷ ∀ era f. Functor f ⇒ (Map (VRFVerKeyHash 'StakePoolVRF) (NonZero Word64) → f (Map (VRFVerKeyHash 'StakePoolVRF) (NonZero Word64))) → PState era → f (PState era)

Lenses from SnapShot(s)

ssStakeMarkL ∷ Lens' SnapShots SnapShot

ssStakeMarkPoolDistrL ∷ Lens' SnapShots PoolDistr

ssStakeSetL ∷ Lens' SnapShots SnapShot

ssStakeGoL ∷ Lens' SnapShots SnapShot

ssFeeL ∷ Lens' SnapShots Coin

ssStakeL ∷ Lens' SnapShot Stake

ssStakeDistrL ∷ Lens' SnapShot (VMap VB VP (Credential 'Staking) (CompactForm Coin))

ssDelegationsL ∷ Lens' SnapShot (VMap VB VB (Credential 'Staking) (KeyHash 'StakePool))

ssPoolParamsL ∷ Lens' SnapShot (VMap VB VB (KeyHash 'StakePool) PoolParams)