Cardano.Ledger.Shelley.LedgerState.NewEpochState

Documentation

availableAfterMIR ∷ MIRPot → AccountState → InstantaneousRewards c → Coin

This function returns the coin balance of a given pot, either the reserves or the treasury, after the instantaneous rewards and pot transfers are accounted for.

getGKeys ∷ NewEpochState era → Set (KeyHash 'Genesis (EraCrypto era))

genesisState ∷ ∀ era. EraGov era ⇒ Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)) → UTxO era → LedgerState era

Creates the ledger state for an empty ledger which contains the specified transaction outputs.

updateNES ∷ EraGov era ⇒ NewEpochState era → BlocksMade (EraCrypto era) → LedgerState era → NewEpochState era

Update new epoch state

returnRedeemAddrsToReserves ∷ ∀ era. EraTxOut era ⇒ EpochState era → EpochState era