Cardano.Ledger.Shelley.LedgerState.PulsingReward

Documentation

startStep ∷ ∀ era. EraGov era ⇒ EpochSize → BlocksMade (EraCrypto era) → EpochState era → Coin → ActiveSlotCoeff → Word64 → PulsingRewUpdate (EraCrypto era)

pulseStep ∷ PulsingRewUpdate c → ShelleyBase (PulsingRewUpdate c, RewardEvent c)

Run the pulser for a bit. If is has nothing left to do, complete it.

completeStep ∷ PulsingRewUpdate c → ShelleyBase (PulsingRewUpdate c, RewardEvent c)

createRUpd ∷ ∀ era. EraGov era ⇒ EpochSize → BlocksMade (EraCrypto era) → EpochState era → Coin → ActiveSlotCoeff → Word64 → ShelleyBase (RewardUpdate (EraCrypto era))

To create a reward update, run all 3 phases This function is not used in the rules, so it ignores RewardEvents

completeRupd ∷ PulsingRewUpdate c → ShelleyBase (RewardUpdate c, RewardEvent c)

Phase 3 of reward update has several parts a) completeM the pulser (in case there are still computions to run) b) Combine the pulser provenance with the RewardProvenance c) Construct the final RewardUpdate d) Add the leader rewards to both the events and the computed Rewards

circulation ∷ EpochState era → Coin → Coin

Calculate the current circulation

This is used in the rewards calculation, and for API endpoints for pool ranking.

updateNonMyopic ∷ NonMyopic c → Coin → Map (KeyHash 'StakePool c) Likelihood → NonMyopic c

decayFactor ∷ Float