Cardano.Ledger.Shelley.LedgerState.RefundsAndDeposits

Documentation

totalTxDeposits ∷ ShelleyEraTxBody era ⇒ PParams era → CertState era → TxBody era → Coin

Calculates the total amount of deposits needed for all pool registration and stake delegation certificates to be valid.

totalCertsDeposits

Arguments

∷ (EraPParams era, Foldable f)

⇒ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ f (DCert (EraCrypto era))

→ Coin

Determine the total deposit amount needed from a TxBody. The block may (legitimately) contain multiple registration certificates for the same pool, where the first will be treated as a registration and any subsequent ones as re-registration. As such, we must only take a deposit for the first such registration. It is even possible for a single transaction to have multiple pool registration for the same pool, so as we process pool registrations, we must keep track of those that are already registered, so we do not add a Deposit for the same pool twice.

Note that this is not an issue for key registrations since subsequent registration certificates would be invalid.

totalCertsDepositsCertState ∷ (EraPParams era, Foldable f) ⇒ PParams era → CertState era → f (DCert (EraCrypto era)) → Coin

keyTxRefunds ∷ ShelleyEraTxBody era ⇒ PParams era → CertState era → TxBody era → Coin

keyCertsRefunds

Arguments

∷ (EraPParams era, Foldable f)

⇒ PParams era

→ (StakeCredential (EraCrypto era) → Maybe Coin)
Function that can lookup current deposit, in case when the stake key is registered.

→ f (DCert (EraCrypto era))

→ Coin

Compute the key deregistration refunds in a transaction

keyCertsRefundsCertState ∷ (EraPParams era, Foldable f) ⇒ PParams era → CertState era → f (DCert (EraCrypto era)) → Coin

Compute the key deregistration refunds in a transaction