Cardano.Ledger.Shelley.LedgerState.Types

Documentation

type KeyPairs c = [(KeyPair 'Payment c, KeyPair 'Staking c)]

Deprecated: Use `Test.Cardano.Ledger.Core.KeyPair (KeyPairs)` instead

type RewardAccounts c = Map (Credential 'Staking c) Coin

data AccountState

Constructors

AccountState

Fields

asTreasury ∷ !Coin

asReserves ∷ !Coin

toAccountStatePairs ∷ KeyValue e a ⇒ AccountState → [a]

data EpochState era

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esLState ∷ !(LedgerState era)

esSnapshots ∷ !(SnapShots (EraCrypto era))

esNonMyopic ∷ !(NonMyopic (EraCrypto era))

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://github.com/intersectmbo/cardano-ledger/releases/latest/download/pool-ranking.pdf

toEpochStatePairs ∷ (EraTxOut era, EraGov era, KeyValue e a) ⇒ EpochState era → [a]

data IncrementalStake c

Incremental Stake, Stake along with possible missed coins from danging Ptrs. Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr does not have to point to anything until the epoch boundary, when we compute rewards and aggregate staking information for ranking. This is unusual but legal. In a non incremental system, we use whatever legal Ptrs exist at the epoch boundary. Here we are computing things incrementally, so we need to remember Ptrs that might point to something by the time the epoch boundary is reached. When the epoch boundary is reached we resolve these pointers, to see if any have become non-dangling since the time they were first used in the incremental computation.

Constructors

IStake

Fields

credMap ∷ !(Map (Credential 'Staking c) (CompactForm Coin))

ptrMap ∷ !(Map Ptr (CompactForm Coin))

toIncrementalStakePairs ∷ (KeyValue e a, Crypto c) ⇒ IncrementalStake c → [a]

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that istake == (updateStakeDistribution (UTxO Map.empty) (UTxO Map.empty) utxo) Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very careful, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

utxosUtxo ∷ !(UTxO era)

utxosDeposited ∷ Coin

This field is left lazy, because we only use it for assertions

utxosFees ∷ !Coin

utxosGovState ∷ !(GovState era)

utxosStakeDistr ∷ !(IncrementalStake (EraCrypto era))

utxosDonation ∷ !Coin

toUTxOStatePairs ∷ (EraTxOut era, EraGov era, KeyValue e a) ⇒ UTxOState era → [a]

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Number of the epoch when this NewEpochState was modified last. With respect to block and transactions validation this will always be the current epoch number. However, when it comes to the TICK rule, it will be the epoch number of the previous epoch whenever we are crossing the epoch boundary.

nesBprev ∷ !(BlocksMade (EraCrypto era))

Blocks made before current epoch

nesBcur ∷ !(BlocksMade (EraCrypto era))

Blocks made in current epoch

nesEs ∷ !(EpochState era)

Epoch state

nesRu ∷ !(StrictMaybe (PulsingRewUpdate (EraCrypto era)))

Possible reward update

nesPd ∷ !(PoolDistr (EraCrypto era))

Stake distribution within the stake pool

stashedAVVMAddresses ∷ !(StashedAVVMAddresses era)

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

type family StashedAVVMAddresses era where ...

Equations

StashedAVVMAddresses (ShelleyEra c) = UTxO (ShelleyEra c)

StashedAVVMAddresses _ = ()

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState ∷ !(UTxOState era)

The current unspent transaction outputs.

lsCertState ∷ !(CertState era)

toLedgerStatePairs ∷ (EraTxOut era, EraGov era, KeyValue e a) ⇒ LedgerState era → [a]

nesPdL ∷ Lens' (NewEpochState era) (PoolDistr (EraCrypto era))

nesEsL ∷ Lens' (NewEpochState era) (EpochState era)

unifiedL ∷ Lens' (NewEpochState era) (UMap (EraCrypto era))

nesELL ∷ Lens' (NewEpochState era) EpochNo

nesBprevL ∷ Lens' (NewEpochState era) (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

nesBcurL ∷ Lens' (NewEpochState era) (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

nesRuL ∷ Lens' (NewEpochState era) (StrictMaybe (PulsingRewUpdate (EraCrypto era)))

nesStashedAVVMAddressesL ∷ Lens' (NewEpochState era) (StashedAVVMAddresses era)

nesEpochStateL ∷ Lens' (NewEpochState era) (EpochState era)

esAccountStateL ∷ Lens' (EpochState era) AccountState

esSnapshotsL ∷ Lens' (EpochState era) (SnapShots (EraCrypto era))

esLStateL ∷ Lens' (EpochState era) (LedgerState era)

esNonMyopicL ∷ Lens' (EpochState era) (NonMyopic (EraCrypto era))

curPParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (PParams era)

prevPParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (PParams era)

futurePParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (FuturePParams era)

asTreasuryL ∷ Lens' AccountState Coin

asReservesL ∷ Lens' AccountState Coin

lsUTxOStateL ∷ Lens' (LedgerState era) (UTxOState era)

lsCertStateL ∷ Lens' (LedgerState era) (CertState era)

utxosUtxoL ∷ Lens' (UTxOState era) (UTxO era)

utxosDepositedL ∷ Lens' (UTxOState era) Coin

utxosFeesL ∷ Lens' (UTxOState era) Coin

utxosGovStateL ∷ Lens' (UTxOState era) (GovState era)

utxosStakeDistrL ∷ Lens' (UTxOState era) (IncrementalStake (EraCrypto era))

utxosDonationL ∷ Lens' (UTxOState era) Coin

credMapL ∷ Lens' (IncrementalStake c) (Map (Credential 'Staking c) (CompactForm Coin))

ptrMapL ∷ Lens' (IncrementalStake c) (Map Ptr (CompactForm Coin))

newEpochStateGovStateL ∷ Lens' (NewEpochState era) (GovState era)

epochStateGovStateL ∷ Lens' (EpochState era) (GovState era)

epochStateDonationL ∷ Lens' (EpochState era) Coin

epochStateTreasuryL ∷ Lens' (EpochState era) Coin

epochStateIncrStakeDistrL ∷ Lens' (EpochState era) (Map (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

epochStateRegDrepL ∷ Lens' (EpochState era) (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

epochStatePoolParamsL ∷ Lens' (EpochState era) (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

epochStateUMapL ∷ Lens' (EpochState era) (UMap (EraCrypto era))

epochStateStakeDistrL ∷ Lens' (EpochState era) (VMap VB VP (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

potEqualsObligation ∷ EraGov era ⇒ CertState era → UTxOState era → Bool

allObligations ∷ EraGov era ⇒ CertState era → GovState era → Obligations

totalObligation ∷ EraGov era ⇒ CertState era → GovState era → Coin