Cardano.Ledger.Shelley.PParams

Description

This module contains just the type of protocol parameters.

Documentation

emptyShelleyPParams ∷ ∀ era. Era era ⇒ ShelleyPParams Identity era

emptyShelleyPParamsUpdate ∷ ShelleyPParams StrictMaybe era

data ShelleyPParams f era

Protocol parameters.

Constructors

ShelleyPParams

Fields

sppMinFeeA ∷ !(HKD f Coin)

The linear factor for the minimum fee calculation

sppMinFeeB ∷ !(HKD f Coin)

The constant factor for the minimum fee calculation

sppMaxBBSize ∷ !(HKD f Word32)

Maximal block body size

sppMaxTxSize ∷ !(HKD f Word32)

Maximal transaction size

sppMaxBHSize ∷ !(HKD f Word16)

Maximal block header size

sppKeyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

sppPoolDeposit ∷ !(HKD f (CompactForm Coin))

The amount of a pool registration deposit

sppEMax ∷ !(HKD f EpochInterval)

epoch bound on pool retirement

sppNOpt ∷ !(HKD f Word16)

Desired number of pools

sppA0 ∷ !(HKD f NonNegativeInterval)

Pool influence

sppRho ∷ !(HKD f UnitInterval)

Monetary expansion

sppTau ∷ !(HKD f UnitInterval)

Treasury expansion

sppD ∷ !(HKD f UnitInterval)

Decentralization parameter

sppExtraEntropy ∷ !(HKD f Nonce)

Extra entropy

sppProtocolVersion ∷ !(HKD f ProtVer)

Protocol version

sppMinUTxOValue ∷ !(HKD f Coin)

Minimum UTxO value

sppMinPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

emptyPParams ∷ EraPParams era ⇒ PParams era

type family HKD (f ∷ Type → Type) a where ...

Higher Kinded Data

Equations

HKD Identity a = a

HKD f a = f a

data PPUpdateEnv

Deprecated: As unused

Constructors

PPUpdateEnv SlotNo GenDelegs

Deprecated: As unused

newtype ProposedPPUpdates era

Update operation for protocol parameters structure PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis) (PParamsUpdate era))

emptyPPPUpdates ∷ ProposedPPUpdates era

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

upgradeUpdate ∷ ∀ era. (EraPParams era, EraPParams (PreviousEra era)) ⇒ UpgradePParams StrictMaybe era → Update (PreviousEra era) → Update era

pvCanFollow

Arguments

∷ ProtVer
Previous protocol version

→ ProtVer
New protocol version

→ Bool

Check whether the new protocol version is a legitimate version bump with respect to the previous one.

hasLegalProtVerUpdate ∷ (AtMostEra "Babbage" era, EraPParams era) ⇒ PParams era → PParamsUpdate era → Bool

Check whether PParamsUpdate contains a valid ProtVer update. When a protocol version update is not included in PParamsUpdate it is considered a legal update.

shelleyPParams ∷ (EraPParams era, AtMostEra "Mary" era, AtMostEra "Alonzo" era, AtMostEra "Babbage" era) ⇒ [PParam era]

PParam

ppA0 ∷ EraPParams era ⇒ PParam era

ppD ∷ (EraPParams era, AtMostEra "Alonzo" era) ⇒ PParam era

ppEMax ∷ EraPParams era ⇒ PParam era

ppExtraEntropy ∷ (EraPParams era, AtMostEra "Alonzo" era) ⇒ PParam era

ppMaxBBSize ∷ EraPParams era ⇒ PParam era

ppKeyDeposit ∷ EraPParams era ⇒ PParam era

ppMinFeeA ∷ EraPParams era ⇒ PParam era

ppMinFeeB ∷ EraPParams era ⇒ PParam era

ppMinPoolCost ∷ EraPParams era ⇒ PParam era

ppMaxBHSize ∷ EraPParams era ⇒ PParam era

ppMaxTxSize ∷ EraPParams era ⇒ PParam era

ppNOpt ∷ EraPParams era ⇒ PParam era

ppProtocolVersion ∷ (EraPParams era, AtMostEra "Babbage" era) ⇒ PParam era

ppPoolDeposit ∷ EraPParams era ⇒ PParam era

ppRho ∷ EraPParams era ⇒ PParam era

ppTau ∷ EraPParams era ⇒ PParam era

Orphan instances

EraPParams ShelleyEra

Instance details

Associated Types

type PParamsHKD f ShelleyEra = (r ∷ Type)

type UpgradePParams f ShelleyEra

type DowngradePParams f ShelleyEra

Methods

applyPPUpdates ∷ PParams ShelleyEra → PParamsUpdate ShelleyEra → PParams ShelleyEra

emptyPParamsIdentity ∷ PParamsHKD Identity ShelleyEra

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe ShelleyEra

emptyUpgradePParamsUpdate ∷ UpgradePParams StrictMaybe ShelleyEra

upgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDApplicative f, EraPParams (PreviousEra ShelleyEra)) ⇒ UpgradePParams f ShelleyEra → PParamsHKD f (PreviousEra ShelleyEra) → PParamsHKD f ShelleyEra

downgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra ShelleyEra)) ⇒ DowngradePParams f ShelleyEra → PParamsHKD f ShelleyEra → PParamsHKD f (PreviousEra ShelleyEra)

hkdMinFeeAL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Coin)

hkdMinFeeBL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Coin)

hkdMaxBBSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Word32)

hkdMaxTxSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Word32)

hkdMaxBHSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Word16)

hkdKeyDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Coin)

hkdPoolDepositCompactL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f (CompactForm Coin))

hkdEMaxL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f EpochInterval)

hkdNOptL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Word16)

hkdA0L ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f NonNegativeInterval)

hkdRhoL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f UnitInterval)

hkdTauL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f UnitInterval)

hkdDL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Alonzo" ShelleyEra) ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f UnitInterval)

ppDG ∷ SimpleGetter (PParams ShelleyEra) UnitInterval

hkdExtraEntropyL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Alonzo" ShelleyEra) ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Nonce)

hkdProtocolVersionL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Babbage" ShelleyEra) ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f ProtVer)

ppProtocolVersionL ∷ Lens' (PParams ShelleyEra) ProtVer

ppuProtocolVersionL ∷ Lens' (PParamsUpdate ShelleyEra) (StrictMaybe ProtVer)

hkdMinUTxOValueL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Mary" ShelleyEra) ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Coin)

hkdMinPoolCostL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f ShelleyEra) (HKD f Coin)

eraPParams ∷ [PParam ShelleyEra]