Cardano.Ledger.Shelley.RewardUpdate

Description

How to compute the reward update compuation. Also, how to spread the compuation over many blocks, once the chain reaches a stability point.

Documentation

type RewardEvent = Map (Credential 'Staking) (Set Reward)

data RewardAns

The result of reward calculation is a pair of aggregate Maps. One for the accumulated answer, and one for the answer since the last pulse

Constructors

RewardAns

Fields

accumRewardAns ∷ !(Map (Credential 'Staking) Reward)

recentRewardAns ∷ !RewardEvent

type Pulser = RewardPulser ShelleyBase RewardAns

The type of RewardPulser we pulse on.

data RewardUpdate

The ultimate goal of a reward update computation. Aggregating rewards for each staking credential.

Constructors

RewardUpdate

Fields

deltaT ∷ !DeltaCoin

deltaR ∷ !DeltaCoin

rs ∷ !(Map (Credential 'Staking) (Set Reward))

deltaF ∷ !DeltaCoin

nonMyopic ∷ !NonMyopic

emptyRewardUpdate ∷ RewardUpdate

data RewardSnapShot

To complete the reward update, we need a snap shot of the EpochState particular to this computation

Constructors

RewardSnapShot

Fields

rewFees ∷ !Coin

rewProtocolVersion ∷ !ProtVer

rewNonMyopic ∷ !NonMyopic

rewDeltaR1 ∷ !Coin

rewR ∷ !Coin

rewDeltaT1 ∷ !Coin

rewLikelihoods ∷ !(Map (KeyHash 'StakePool) Likelihood)

rewLeaders ∷ !(Map (Credential 'Staking) (Set Reward))

data FreeVars

Constructors

FreeVars

Fields

fvDelegs ∷ !(VMap VB VB (Credential 'Staking) (KeyHash 'StakePool))

fvAddrsRew ∷ !(Set (Credential 'Staking))

fvTotalStake ∷ !Coin

fvProtVer ∷ !ProtVer

fvPoolRewardInfo ∷ !(Map (KeyHash 'StakePool) PoolRewardInfo)

rewardStakePoolMember ∷ FreeVars → RewardAns → Credential 'Staking → CompactForm Coin → RewardAns

The function to call on each reward update pulse. Called by the pulser.

data RewardPulser (m ∷ Type → Type) ans where

The type of a Pulser which uses rewardStakePoolMember as its underlying function. rewardStakePool will be partially applied to the component of type (FreeVars c) when pulsing. Note that we use two type equality (~) constraints to fix both the monad m and the ans type, to the context where we will use the type as a Pulser. The type must have m and ans as its last two parameters so we can make a Pulsable instance. RSLP = Reward Serializable Listbased Pulser

Constructors

RSLP ∷ (ans ~ RewardAns, m ~ ShelleyBase) ⇒ !Int → !FreeVars → !(VMap VB VP (Credential 'Staking) (CompactForm Coin)) → !ans → RewardPulser m ans

clearRecent ∷ RewardAns → RewardAns

data PulsingRewUpdate

The state used in the STS rules

Constructors

Pulsing !RewardSnapShot !Pulser

Complete !RewardUpdate