Cardano.Ledger.Shelley.Rewards

Documentation

newtype StakeShare

StakeShare type

Constructors

StakeShare

Fields

unStakeShare :: Rational

data PoolRewardInfo crypto

Stake Pool specific information needed to compute the rewards for its members.

Constructors

PoolRewardInfo

Fields

poolRelativeStake :: !StakeShare

The stake pool's stake divided by the total stake

poolPot :: !Coin

The maximum rewards available for the entire pool

poolPs :: !(PoolParams crypto)

The stake pool parameters

poolBlocks :: !Natural

The number of blocks the stake pool produced

poolLeaderReward :: !(LeaderOnlyReward crypto)

The leader reward

mkApparentPerformance :: UnitInterval -> Rational -> Natural -> Natural -> Rational

Calculate pool reward

data RewardType

Constructors

MemberReward

LeaderReward

data LeaderOnlyReward crypto

Constructors

LeaderOnlyReward

Fields

lRewardPool :: !(KeyHash 'StakePool crypto)

lRewardAmount :: !Coin

leaderRewardToGeneral :: LeaderOnlyReward c -> Reward c

data Reward crypto

Constructors

Reward

Fields

rewardType :: RewardType

rewardPool :: KeyHash 'StakePool crypto

rewardAmount :: Coin

leaderRew :: Coin -> PoolParams crypto -> StakeShare -> StakeShare -> Coin

Calculate pool leader reward

memberRew :: Coin -> PoolParams crypto -> StakeShare -> StakeShare -> Coin

Calculate pool member reward

aggregateRewards :: forall crypto pp. HasField "_protocolVersion" pp ProtVer => pp -> Map (Credential 'Staking crypto) (Set (Reward crypto)) -> Map (Credential 'Staking crypto) Coin

filterRewards :: forall crypto pp. HasField "_protocolVersion" pp ProtVer => pp -> Map (Credential 'Staking crypto) (Set (Reward crypto)) -> (Map (Credential 'Staking crypto) (Set (Reward crypto)), Map (Credential 'Staking crypto) (Set (Reward crypto)))

Filter the reward payments to those that will actually be delivered. This function exists since in Shelley, a stake credential earning rewards from multiple sources would only receive one reward.

sumRewards :: forall crypto pp. HasField "_protocolVersion" pp ProtVer => pp -> Map (Credential 'Staking crypto) (Set (Reward crypto)) -> Coin

rewardOnePoolMember

Arguments

:: HasField "_protocolVersion" pp ProtVer

=> pp
The protocol parameters

-> Coin
The total amount of stake in the system

-> Set (Credential 'Staking c)
The set of registered stake credentials

-> PoolRewardInfo c
Stake pool specific intermediate values needed to compute member rewards.

-> Credential 'Staking c
The stake credential whose reward is being calculated.

-> Coin
The stake controlled by the stake credential in the previous parameter above.

-> Maybe Coin
The reward for the given stake credential. This could be Nothing if the credential is no longer registered, if it is an owner, or if the reward is zero.

The stake pool member reward calculation

mkPoolRewardInfo :: (HasField "_d" (PParams era) UnitInterval, HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural) => PParams era -> Coin -> BlocksMade (Crypto era) -> Natural -> Stake (Crypto era) -> VMap VB VB (Credential 'Staking (Crypto era)) (KeyHash 'StakePool (Crypto era)) -> Map (KeyHash 'StakePool (Crypto era)) Coin -> Coin -> Coin -> PoolParams (Crypto era) -> Either StakeShare (PoolRewardInfo (Crypto era))

Calculate single stake pool specific values for the reward computation.

Note that if a stake pool has made no blocks in the given epoch, it will get no rewards, and so we do not need to return PoolRewardInfo. We do, however, need to return the relative stake of the pool in order to compute data for the stake pool ranking. Eventually we will remove the ranking information out of the ledger code and into a separate service, and at that point we can simplify this function to not care about ranking.