Cardano.Ledger.Shelley.Rules

Block

data ShelleyBBODY era

data ShelleyBbodyState era

Constructors

BbodyState !(State (EraRule "LEDGERS" era)) !BlocksMade

data BbodyEnv era

Constructors

BbodyEnv

Fields

bbodyPp ∷ PParams era

bbodyAccount ∷ ChainAccountState

data ShelleyBbodyPredFailure era

Constructors

WrongBlockBodySizeBBODY (Mismatch 'RelEQ Int)
mismatchSupplied ~ Actual body size. mismatchExpected ~ Claimed body size in the header.

InvalidBodyHashBBODY (Mismatch 'RelEQ (Hash HASH EraIndependentBlockBody))
mismatchSupplied ~ Actual hash. mismatchExpected ~ Claimed hash in the header.

LedgersFailure (PredicateFailure (EraRule "LEDGERS" era))

newtype ShelleyBbodyEvent era

Constructors

LedgersEvent (Event (EraRule "LEDGERS" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

type family State a

Type of the state which the system transitions between.

data ShelleyDELEG era

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo ∷ SlotNo

deCurEpochNo ∷ EpochNo

Lazy on purpose, because not all certificates need to know the current EpochNo

ptr_ ∷ Ptr

deChainAccountState ∷ ChainAccountState

ppDE ∷ PParams era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyDelegPredFailure era

Constructors

StakeKeyAlreadyRegisteredDELEG (Credential 'Staking)

StakeKeyNotRegisteredDELEG (Credential 'Staking)

StakeKeyNonZeroAccountBalanceDELEG Coin

StakeDelegationImpossibleDELEG (Credential 'Staking)

WrongCertificateTypeDELEG

GenesisKeyNotInMappingDELEG (KeyHash 'Genesis)

DuplicateGenesisDelegateDELEG (KeyHash 'GenesisDelegate)

InsufficientForInstantaneousRewardsDELEG MIRPot (Mismatch 'RelLTEQ Coin)

MIRCertificateTooLateinEpochDELEG (Mismatch 'RelLT SlotNo)

DuplicateGenesisVRFDELEG (VRFVerKeyHash 'GenDelegVRF)

MIRTransferNotCurrentlyAllowed

MIRNegativesNotCurrentlyAllowed

InsufficientForTransferDELEG MIRPot (Mismatch 'RelLTEQ Coin)

MIRProducesNegativeUpdate

MIRNegativeTransfer MIRPot Coin

newtype ShelleyDelegEvent era

Constructors

DelegNewEpoch EpochNo

data ShelleyDELEGS era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo ∷ SlotNo

delegsEpochNo ∷ EpochNo

Lazy on purpose, because not all certificates need to know the current EpochNo

delegsIx ∷ TxIx

delegspp ∷ PParams era

delegsTx ∷ Tx era

delegsAccount ∷ ChainAccountState

data ShelleyDelegsPredFailure era

Constructors

DelegateeNotRegisteredDELEG (KeyHash 'StakePool)
Target pool which is not registered

WithdrawalsNotInRewardsDELEGS Withdrawals
Withdrawals that are missing or do not withdrawal the entire amount

DelplFailure (PredicateFailure (EraRule "DELPL" era))
Subtransition Failures

newtype ShelleyDelegsEvent era

Constructors

DelplEvent (Event (EraRule "DELPL" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

validateStakePoolDelegateeRegistered ∷ PState era → KeyHash 'StakePool → Test (KeyHash 'StakePool)

data ShelleyDELPL era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo ∷ SlotNo

delplEpochNo ∷ EpochNo

delPlPtr ∷ Ptr

delPlPp ∷ PParams era

delPlAccount ∷ ChainAccountState

data ShelleyDelplPredFailure era

Constructors

PoolFailure (PredicateFailure (EraRule "POOL" era))

DelegFailure (PredicateFailure (EraRule "DELEG" era))

data ShelleyDelplEvent era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyLEDGER era

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo ∷ SlotNo

ledgerEpochNo ∷ Maybe EpochNo

ledgerIx ∷ TxIx

ledgerPp ∷ PParams era

ledgerAccount ∷ ChainAccountState

ledgerSlotNoL ∷ Lens' (LedgerEnv era) SlotNo

ledgerEpochNoL ∷ Lens' (LedgerEnv era) (Maybe EpochNo)

ledgerIxL ∷ Lens' (LedgerEnv era) TxIx

ledgerPpL ∷ Lens' (LedgerEnv era) (PParams era)

ledgerAccountL ∷ Lens' (LedgerEnv era) ChainAccountState

data ShelleyLedgerPredFailure era

Constructors

UtxowFailure (PredicateFailure (EraRule "UTXOW" era))

DelegsFailure (PredicateFailure (EraRule "DELEGS" era))

data ShelleyLedgerEvent era

Constructors

UtxowEvent (Event (EraRule "UTXOW" era))

DelegsEvent (Event (EraRule "DELEGS" era))

type family Event a

Event type.

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

epochFromSlot ∷ SlotNo → Reader Globals EpochNo

renderDepositEqualsObligationViolation ∷ (EraTx era, EraGov era, EraCertState era, Environment t ~ LedgerEnv era, Signal t ~ Tx era, State t ~ LedgerState era) ⇒ AssertionViolation t → String

shelleyLedgerAssertions ∷ (EraGov era, EraCertState era, State (rule era) ~ LedgerState era) ⇒ [Assertion (rule era)]

data ShelleyLEDGERS era

data ShelleyLedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo ∷ SlotNo

ledgersEpochNo ∷ EpochNo

ledgersPp ∷ PParams era

ledgersAccount ∷ ChainAccountState

newtype ShelleyLedgersPredFailure era

Constructors

LedgerFailure (PredicateFailure (EraRule "LEDGER" era))

newtype ShelleyLedgersEvent era

Constructors

LedgerEvent (Event (EraRule "LEDGER" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyPOOL era

data PoolEvent era

Constructors

RegisterPool (KeyHash 'StakePool)

ReregisterPool (KeyHash 'StakePool)

data PoolEnv era

Constructors

PoolEnv EpochNo (PParams era)

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyPoolPredFailure era

Constructors

StakePoolNotRegisteredOnKeyPOOL

Fields

(KeyHash 'StakePool)

KeyHash which cannot be retired since it is not registered

StakePoolRetirementWrongEpochPOOL (Mismatch 'RelGT EpochNo) (Mismatch 'RelLTEQ EpochNo)

StakePoolCostTooLowPOOL (Mismatch 'RelGTEQ Coin)

WrongNetworkPOOL

Fields

(Mismatch 'RelEQ Network)

(KeyHash 'StakePool)

Stake Pool ID

PoolMedataHashTooBig

Fields

(KeyHash 'StakePool)

Stake Pool ID

Int

Size of the metadata hash

VRFKeyHashAlreadyRegistered

Fields

(KeyHash 'StakePool)

Stake Pool ID

(VRFVerKeyHash 'StakePoolVRF)

VRF key attempted to use, that has already been registered

data ShelleyPPUP era

data PpupEnv era

Constructors

PPUPEnv SlotNo (PParams era) GenDelegs

data ShelleyPpupPredFailure era

Constructors

NonGenesisUpdatePPUP (Mismatch 'RelSubset (Set (KeyHash 'Genesis)))
An update was proposed by a key hash that is not one of the genesis keys. mismatchSupplied ~ key hashes which were a part of the update. mismatchExpected ~ key hashes of the genesis keys.

PPUpdateWrongEpoch EpochNo EpochNo VotingPeriod
An update was proposed for the wrong epoch. The first EpochNo is the current epoch. The second EpochNo is the epoch listed in the update. The last parameter indicates if the update was intended for the current or the next epoch.

PVCannotFollowPPUP ProtVer
An update was proposed which contains an invalid protocol version. New protocol versions must either increase the major number by exactly one and set the minor version to zero, or keep the major version the same and increase the minor version by exactly one.

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

sgsFuturePParams ∷ !(FuturePParams era)

Prediction of parameter changes that might happen on the epoch boundary.

newtype PpupEvent era

Constructors

PpupNewEpoch EpochNo

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data VotingPeriod

Constructors

VoteForThisEpoch

VoteForNextEpoch

votedFuturePParams

Arguments

∷ ∀ era. EraPParams era

⇒ ProposedPPUpdates era

→ PParams era
Protocol parameters to which the change will be applied.

→ Word64
Quorum needed to change the protocol parameters.

→ Maybe (PParams era)

If at least n nodes voted to change the same protocol parameters to the same values, return the given protocol parameters updated to these values. Here n is the quorum needed.

data ShelleyUTXO era

data UtxoEnv era

Constructors

UtxoEnv

Fields

ueSlot ∷ SlotNo

uePParams ∷ PParams era

ueCertState ∷ CertState era

data ShelleyUtxoPredFailure era

Constructors

BadInputsUTxO (Set TxIn)

ExpiredUTxO (Mismatch 'RelLTEQ SlotNo)

MaxTxSizeUTxO (Mismatch 'RelLTEQ Word32)

InputSetEmptyUTxO

FeeTooSmallUTxO (Mismatch 'RelGTEQ Coin)

ValueNotConservedUTxO (Mismatch 'RelEQ (Value era))

WrongNetwork Network (Set Addr)

WrongNetworkWithdrawal Network (Set RewardAccount)

OutputTooSmallUTxO [TxOut era]

UpdateFailure (EraRuleFailure "PPUP" era)

OutputBootAddrAttrsTooBig [TxOut era]

data UtxoEvent era

Constructors

TotalDeposits (SafeHash EraIndependentTxBody) Coin

UpdateEvent (Event (EraRule "PPUP" era))

TxUTxODiff
The UTxOs consumed and created by a signal tx

Fields

(UTxO era)

UTxO consumed

(UTxO era)

UTxO created

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

validSizeComputationCheck ∷ (EraTx era, SafeToHash (TxWits era), Signal (rule era) ~ Tx era) ⇒ Assertion (rule era)

updateUTxOState ∷ (EraTxBody era, EraStake era, EraCertState era, Monad m) ⇒ PParams era → UTxOState era → TxBody era → CertState era → GovState era → (Coin → m ()) → (UTxO era → UTxO era → m ()) → m (UTxOState era)

This monadic action captures the final stages of the UTXO(S) rule. In particular it applies all of the UTxO related aditions and removals, gathers all of the fees into the fee pot utxosFees and updates the utxosDeposited field. Continuation supplied will be called on the deposit - refund change, which is normally used to emit the TotalDeposits event.

Validations

validateInputSetEmptyUTxO ∷ EraTxBody era ⇒ TxBody era → Test (ShelleyUtxoPredFailure era)

Ensure that there is at least one input in the TxBody

txins txb ≠ ∅

validateFeeTooSmallUTxO ∷ EraUTxO era ⇒ PParams era → Tx era → UTxO era → Test (ShelleyUtxoPredFailure era)

Ensure that the fee is at least the amount specified by the minfee

minfee pp tx ≤ txfee txb

validateBadInputsUTxO ∷ UTxO era → Set TxIn → Test (ShelleyUtxoPredFailure era)

Ensure all transaction inputs are present in UTxO

inputs ⊆ dom utxo

validateWrongNetwork ∷ (EraTxOut era, Foldable f) ⇒ Network → f (TxOut era) → Test (ShelleyUtxoPredFailure era)

Make sure all addresses match the supplied NetworkId

∀(_ → (a, _)) ∈ txouts txb, netId a = NetworkId

validateWrongNetworkWithdrawal ∷ EraTxBody era ⇒ Network → TxBody era → Test (ShelleyUtxoPredFailure era)

Make sure all addresses match the supplied NetworkId

∀(a → ) ∈ txwdrls txb, netId a = NetworkId

validateOutputBootAddrAttrsTooBig ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Test (ShelleyUtxoPredFailure era)

Bootstrap (i.e. Byron) addresses have variable sized attributes in them. It is important to limit their overall size.

∀ ( _ ↦ (a,_)) ∈ txoutstxb,  a ∈ Addrbootstrap → bootstrapAttrsSize a ≤ 64

validateMaxTxSizeUTxO ∷ EraTx era ⇒ PParams era → Tx era → Test (ShelleyUtxoPredFailure era)

Ensure that the size of the transaction does not exceed the maxTxSize protocol parameter

txsize tx ≤ maxTxSize pp

validateValueNotConservedUTxO ∷ (EraUTxO era, EraCertState era) ⇒ PParams era → UTxO era → CertState era → TxBody era → Test (ShelleyUtxoPredFailure era)

Ensure that value consumed and produced matches up exactly

consumed pp utxo txb = produced pp poolParams txb

utxoEnvSlotL ∷ Lens' (UtxoEnv era) SlotNo

utxoEnvPParamsL ∷ Lens' (UtxoEnv era) (PParams era)

utxoEnvCertStateL ∷ Lens' (UtxoEnv era) (CertState era)

data ShelleyUTXOW era

data ShelleyUtxowPredFailure era

Constructors

InvalidWitnessesUTXOW [VKey 'Witness]

MissingVKeyWitnessesUTXOW (Set (KeyHash 'Witness))

MissingScriptWitnessesUTXOW (Set ScriptHash)

ScriptWitnessNotValidatingUTXOW (Set ScriptHash)

UtxoFailure (PredicateFailure (EraRule "UTXO" era))

MIRInsufficientGenesisSigsUTXOW (Set (KeyHash 'Witness))

MissingTxBodyMetadataHash TxAuxDataHash

MissingTxMetadata TxAuxDataHash

ConflictingMetadataHash (Mismatch 'RelEQ TxAuxDataHash)

InvalidMetadata

ExtraneousScriptWitnessesUTXOW (Set ScriptHash)

newtype ShelleyUtxowEvent era

Constructors

UtxoEvent (Event (EraRule "UTXO" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

transitionRulesUTXOW ∷ ∀ era. (EraUTxO era, ShelleyEraTxBody era, ScriptsNeeded era ~ ShelleyScriptsNeeded era, BaseM (EraRule "UTXOW" era) ~ ShelleyBase, Embed (EraRule "UTXO" era) (EraRule "UTXOW" era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Signal (EraRule "UTXOW" era) ~ Tx era, InjectRuleFailure "UTXOW" ShelleyUtxowPredFailure era, STS (EraRule "UTXOW" era), EraCertState era) ⇒ TransitionRule (EraRule "UTXOW" era)

A generic Utxow witnessing function designed to be used across many Eras. Note the embed argument lifts from the simple Shelley (ShelleyUtxowPredFailure) to the PredicateFailure (type family) of the context of where it is called.

Individual validation steps

validateFailedNativeScripts ∷ EraTx era ⇒ ScriptsProvided era → Tx era → Test (ShelleyUtxowPredFailure era)

validateMissingScripts ∷ ShelleyScriptsNeeded era → ScriptsProvided era → Test (ShelleyUtxowPredFailure era)

validateVerifiedWits ∷ EraTx era ⇒ Tx era → Test (ShelleyUtxowPredFailure era)

Determine if the UTxO witnesses in a given transaction are correct.

validateMetadata ∷ EraTx era ⇒ PParams era → Tx era → Test (ShelleyUtxowPredFailure era)

check metadata hash ((adh = ◇) ∧ (ad= ◇)) ∨ (adh = hashAD ad)

validateMIRInsufficientGenesisSigs ∷ (EraTx era, ShelleyEraTxBody era) ⇒ GenDelegs → Word64 → Set (KeyHash 'Witness) → Tx era → Test (ShelleyUtxowPredFailure era)

check genesis keys signatures for instantaneous rewards certificates

genSig := { hashKey gkey | gkey ∈ dom(genDelegs)} ∩ witsKeyHashes { c ∈ txcerts txb ∩ TxCert_mir} ≠ ∅ ⇒ |genSig| ≥ Quorum

validateNeededWitnesses

Arguments

∷ EraUTxO era

⇒ Set (KeyHash 'Witness)
Provided witness

→ CertState era

→ UTxO era

→ TxBody era

→ Test (ShelleyUtxowPredFailure era)

Verify that we provide at least all of the required witnesses

witsVKeyNeeded utxo tx ⊆ witsKeyHashes

Tick

data ShelleyEPOCH era

data ShelleyEpochEvent era

Constructors

PoolReapEvent (Event (EraRule "POOLREAP" era))

SnapEvent (Event (EraRule "SNAP" era))

UpecEvent (Event (EraRule "UPEC" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyNEWEPOCH era

data ShelleyNewEpochEvent era

Constructors

DeltaRewardEvent (Event (EraRule "RUPD" era))

RestrainedRewards EpochNo (Map (Credential 'Staking) (Set Reward)) (Set (Credential 'Staking))

TotalRewardEvent EpochNo (Map (Credential 'Staking) (Set Reward))

EpochEvent (Event (EraRule "EPOCH" era))

MirEvent (Event (EraRule "MIR" era))

TotalAdaPotsEvent AdaPots

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

updateRewards ∷ (EraGov era, EraCertState era) ⇒ EpochState era → EpochNo → RewardUpdate → Rule (ShelleyNEWEPOCH era) 'Transition (EpochState era)

calculatePoolDistr ∷ SnapShot → PoolDistr

calculatePoolDistr' ∷ (KeyHash 'StakePool → Bool) → SnapShot → PoolDistr

data ShelleyRUPD era

data RupdEnv era

Constructors

RupdEnv BlocksMade (EpochState era)

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

epochInfoRange ∷ Monad m ⇒ EpochInfo m → EpochNo → m (SlotNo, SlotNo)

data PulsingRewUpdate

The state used in the STS rules

Constructors

Pulsing !RewardSnapShot !Pulser

Complete !RewardUpdate

startStep ∷ ∀ era. (EraGov era, EraCertState era) ⇒ EpochSize → BlocksMade → EpochState era → Coin → ActiveSlotCoeff → NonZero Word64 → PulsingRewUpdate

pulseStep ∷ PulsingRewUpdate → ShelleyBase (PulsingRewUpdate, RewardEvent)

Run the pulser for a bit. If is has nothing left to do, complete it.

completeStep ∷ PulsingRewUpdate → ShelleyBase (PulsingRewUpdate, RewardEvent)

lift ∷ (MonadTrans t, Monad m) ⇒ m a → t m a

Lift a computation from the argument monad to the constructed monad.

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity ∷ a

data RupdEvent

Constructors

RupdEvent !EpochNo !(Map (Credential 'Staking) (Set Reward))

data ShelleySNAP era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

newtype SnapEvent era

Constructors

StakeDistEvent (Map (Credential 'Staking) (Coin, KeyHash 'StakePool))

data SnapEnv era

Constructors

SnapEnv (LedgerState era) (PParams era)

data ShelleyTICK era

type family State a

Type of the state which the system transitions between.

data ShelleyTickEvent era

Constructors

TickNewEpochEvent (Event (EraRule "NEWEPOCH" era))

TickRupdEvent (Event (EraRule "RUPD" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

adoptGenesisDelegs ∷ EraCertState era ⇒ EpochState era → SlotNo → EpochState era

data ShelleyTICKF era

validatingTickTransition ∷ ∀ tick era. (EraGov era, EraCertState era, Embed (EraRule "NEWEPOCH" era) (tick era), STS (tick era), State (tick era) ~ NewEpochState era, BaseM (tick era) ~ ShelleyBase, Environment (EraRule "NEWEPOCH" era) ~ (), State (EraRule "NEWEPOCH" era) ~ NewEpochState era, Signal (EraRule "NEWEPOCH" era) ~ EpochNo) ⇒ NewEpochState era → SlotNo → TransitionRule (tick era)

This is a limited version of bheadTransition which is suitable for the future ledger view.

validatingTickTransitionFORECAST ∷ ∀ tick era. (State (tick era) ~ NewEpochState era, BaseM (tick era) ~ ShelleyBase, State (EraRule "UPEC" era) ~ UpecState era, Signal (EraRule "UPEC" era) ~ (), Environment (EraRule "UPEC" era) ~ LedgerState era, Embed (EraRule "UPEC" era) (tick era), STS (tick era), GovState era ~ ShelleyGovState era, EraGov era, EraCertState era) ⇒ NewEpochState era → SlotNo → TransitionRule (tick era)

This is a limited version of validatingTickTransition which is only suitable for the future ledger view.

solidifyNextEpochPParams ∷ EraGov era ⇒ NewEpochState era → SlotNo → ShelleyBase (EpochNo, NewEpochState era)

This action ensures that once the current slot number is at the point of no return we mark the future PParams to be updated at the next epoch boundary. Also returns the current epoch number for convenience.

data ShelleyUPEC era

data UpecState era

Constructors

UpecState

Fields

usCurPParams ∷ !(PParams era)

Current protocol parameters.

usGovState ∷ !(ShelleyGovState era)

State of the protocol update transition system.

data ShelleyMIR era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyMirEvent era

Constructors

MirTransfer InstantaneousRewards

NoMirTransfer InstantaneousRewards Coin Coin
We were not able to perform an MIR transfer due to insufficient funds. This event gives the rewards we wanted to pay, plus the available reserves and treasury.

emptyInstantaneousRewards ∷ InstantaneousRewards

data ShelleyNEWPP era

data ShelleyNewppState era

Constructors

NewppState (PParams era) (ShelleyGovState era)

data NewppEnv era

Constructors

NewppEnv

Fields

neCertState ∷ CertState era

neUTxOState ∷ UTxOState era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyPOOLREAP era

data ShelleyPoolreapEvent era

Constructors

RetiredPools

Fields

refundPools ∷ Map (Credential 'Staking) (Map (KeyHash 'StakePool) (CompactForm Coin))

unclaimedPools ∷ Map (Credential 'Staking) (Map (KeyHash 'StakePool) (CompactForm Coin))

epochNo ∷ EpochNo

data ShelleyPoolreapState era

Constructors

PoolreapState

Fields

prUTxOSt ∷ UTxOState era

prChainAccountState ∷ ChainAccountState

prCertState ∷ CertState era

prCertStateL ∷ Lens' (ShelleyPoolreapState era) (CertState era)

prChainAccountStateL ∷ Lens' (ShelleyPoolreapState era) ChainAccountState

prUTxOStateL ∷ Lens' (ShelleyPoolreapState era) (UTxOState era)

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.