Cardano.Ledger.Shelley.Rules.Bbody

Documentation

data ShelleyBBODY era

data ShelleyBbodyState era

Constructors

BbodyState !(State (EraRule "LEDGERS" era)) !(BlocksMade (EraCrypto era))

data BbodyEnv era

Constructors

BbodyEnv

Fields

bbodyPp ∷ PParams era

bbodyAccount ∷ AccountState

data ShelleyBbodyPredFailure era

Constructors

WrongBlockBodySizeBBODY !Int !Int

InvalidBodyHashBBODY !(Hash (EraCrypto era) EraIndependentBlockBody) !(Hash (EraCrypto era) EraIndependentBlockBody)

LedgersFailure (PredicateFailure (EraRule "LEDGERS" era))

newtype ShelleyBbodyEvent era

Constructors

LedgersEvent (Event (EraRule "LEDGERS" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

type family State a

Type of the state which the system transitions between.

Orphan instances

InjectRuleFailure "BBODY" ShelleyDelegPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyDelegsPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelegsPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyDelplPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelplPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyLedgerPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyLedgerPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyLedgersPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyLedgersPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyPoolPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPoolPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyPpupPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyUtxoPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

InjectRuleFailure "BBODY" ShelleyUtxowPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxowPredFailure (ShelleyEra c) → EraRuleFailure "BBODY" (ShelleyEra c)

(Era era, BaseM ledgers ~ ShelleyBase, ledgers ~ EraRule "LEDGERS" era, STS ledgers, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), Era era) ⇒ Embed ledgers (ShelleyBBODY era)

Instance details

Methods

wrapFailed ∷ PredicateFailure ledgers → PredicateFailure (ShelleyBBODY era)

wrapEvent ∷ Event ledgers → Event (ShelleyBBODY era)

(EraSegWits era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), Embed (EraRule "LEDGERS" era) (ShelleyBBODY era), Environment (EraRule "LEDGERS" era) ~ ShelleyLedgersEnv era, Signal (EraRule "LEDGERS" era) ~ Seq (Tx era)) ⇒ STS (ShelleyBBODY era)

Instance details

Associated Types

type State (ShelleyBBODY era)

type Signal (ShelleyBBODY era)

type Environment (ShelleyBBODY era)

type BaseM (ShelleyBBODY era) ∷ Type → Type

type Event (ShelleyBBODY era)

type PredicateFailure (ShelleyBBODY era)

Methods

initialRules ∷ [InitialRule (ShelleyBBODY era)]

transitionRules ∷ [TransitionRule (ShelleyBBODY era)]

assertions ∷ [Assertion (ShelleyBBODY era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyBBODY era) → String