Cardano.Ledger.Shelley.Rules.Deleg

Documentation

data ShelleyDELEG era

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo ∷ !SlotNo

ptr_ ∷ !Ptr

acnt_ ∷ !AccountState

ppDE ∷ !(PParams era)

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyDelegPredFailure era

Constructors

StakeKeyAlreadyRegisteredDELEG !(Credential 'Staking (EraCrypto era))

StakeKeyInRewardsDELEG !(Credential 'Staking (EraCrypto era))
Indicates that the stake key is somehow already in the rewards map. This error is now redundant with StakeKeyAlreadyRegisteredDELEG. We should remove it and replace its one use with StakeKeyAlreadyRegisteredDELEG.

StakeKeyNotRegisteredDELEG !(Credential 'Staking (EraCrypto era))

StakeKeyNonZeroAccountBalanceDELEG !(Maybe Coin)

StakeDelegationImpossibleDELEG !(Credential 'Staking (EraCrypto era))

WrongCertificateTypeDELEG

GenesisKeyNotInMappingDELEG !(KeyHash 'Genesis (EraCrypto era))

DuplicateGenesisDelegateDELEG !(KeyHash 'GenesisDelegate (EraCrypto era))

InsufficientForInstantaneousRewardsDELEG !MIRPot !Coin !Coin

MIRCertificateTooLateinEpochDELEG !SlotNo !SlotNo

DuplicateGenesisVRFDELEG !(Hash (EraCrypto era) (VerKeyVRF (EraCrypto era)))

MIRTransferNotCurrentlyAllowed

MIRNegativesNotCurrentlyAllowed

InsufficientForTransferDELEG !MIRPot !Coin !Coin

MIRProducesNegativeUpdate

MIRNegativeTransfer !MIRPot !Coin

newtype ShelleyDelegEvent era

Constructors

DelegNewEpoch EpochNo

Orphan instances

(EraPParams era, ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ STS (ShelleyDELEG era)

Instance details

Associated Types

type State (ShelleyDELEG era)

type Signal (ShelleyDELEG era)

type Environment (ShelleyDELEG era)

type BaseM (ShelleyDELEG era) ∷ Type → Type

type Event (ShelleyDELEG era)

type PredicateFailure (ShelleyDELEG era)

Methods

initialRules ∷ [InitialRule (ShelleyDELEG era)]

transitionRules ∷ [TransitionRule (ShelleyDELEG era)]

assertions ∷ [Assertion (ShelleyDELEG era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyDELEG era) → String