Cardano.Ledger.Shelley.Rules.Delegs

Documentation

data ShelleyDELEGS era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo ∷ !SlotNo

delegsIx ∷ !TxIx

delegspp ∷ !(PParams era)

delegsTx ∷ !(Tx era)

delegsAccount ∷ !AccountState

data ShelleyDelegsPredFailure era

Constructors

DelegateeNotRegisteredDELEG !(KeyHash 'StakePool (EraCrypto era))
Target pool which is not registered

WithdrawalsNotInRewardsDELEGS !(Map (RewardAccount (EraCrypto era)) Coin)
Withdrawals that are missing or do not withdrawal the entire amount

DelplFailure !(PredicateFailure (EraRule "DELPL" era))
Subtransition Failures

newtype ShelleyDelegsEvent era

Constructors

DelplEvent (Event (EraRule "DELPL" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

validateZeroRewards ∷ ∀ era. DState era → Withdrawals (EraCrypto era) → Network → Test (Map (RewardAccount (EraCrypto era)) Coin)

validateStakePoolDelegateeRegistered ∷ PState era → KeyHash 'StakePool (EraCrypto era) → Test (KeyHash 'StakePool (EraCrypto era))

drainWithdrawals ∷ DState era → Withdrawals (EraCrypto era) → DState era

Orphan instances

InjectRuleFailure "DELEGS" ShelleyDelegPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure (ShelleyEra c) → EraRuleFailure "DELEGS" (ShelleyEra c)

InjectRuleFailure "DELEGS" ShelleyDelplPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelplPredFailure (ShelleyEra c) → EraRuleFailure "DELEGS" (ShelleyEra c)

InjectRuleFailure "DELEGS" ShelleyPoolPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPoolPredFailure (ShelleyEra c) → EraRuleFailure "DELEGS" (ShelleyEra c)

(EraTx era, ShelleyEraTxBody era, Embed (EraRule "DELPL" era) (ShelleyDELEGS era), Environment (EraRule "DELPL" era) ~ DelplEnv era, State (EraRule "DELPL" era) ~ CertState era, Signal (EraRule "DELPL" era) ~ TxCert era, EraRule "DELEGS" era ~ ShelleyDELEGS era) ⇒ STS (ShelleyDELEGS era)

Instance details

Associated Types

type State (ShelleyDELEGS era)

type Signal (ShelleyDELEGS era)

type Environment (ShelleyDELEGS era)

type BaseM (ShelleyDELEGS era) ∷ Type → Type

type Event (ShelleyDELEGS era)

type PredicateFailure (ShelleyDELEGS era)

Methods

initialRules ∷ [InitialRule (ShelleyDELEGS era)]

transitionRules ∷ [TransitionRule (ShelleyDELEGS era)]

assertions ∷ [Assertion (ShelleyDELEGS era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyDELEGS era) → String

(Era era, STS (ShelleyDELPL era), PredicateFailure (EraRule "DELPL" era) ~ ShelleyDelplPredFailure era, Event (EraRule "DELPL" era) ~ ShelleyDelplEvent era) ⇒ Embed (ShelleyDELPL era) (ShelleyDELEGS era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyDELPL era) → PredicateFailure (ShelleyDELEGS era)

wrapEvent ∷ Event (ShelleyDELPL era) → Event (ShelleyDELEGS era)