Cardano.Ledger.Shelley.Rules.Delpl

Documentation

data ShelleyDELPL era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo ∷ SlotNo

delPlPtr ∷ Ptr

delPlPp ∷ PParams era

delPlAccount ∷ AccountState

data ShelleyDelplPredFailure era

Constructors

PoolFailure (PredicateFailure (EraRule "POOL" era))

DelegFailure (PredicateFailure (EraRule "DELEG" era))

data ShelleyDelplEvent era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Deprecations

delPlAcnt ∷ DelplEnv era → AccountState

Deprecated: Use delPlAccount instead

Orphan instances

InjectRuleFailure "DELPL" ShelleyDelegPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure (ShelleyEra c) → EraRuleFailure "DELPL" (ShelleyEra c)

InjectRuleFailure "DELPL" ShelleyPoolPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPoolPredFailure (ShelleyEra c) → EraRuleFailure "DELPL" (ShelleyEra c)

(Era era, Embed (EraRule "DELEG" era) (ShelleyDELPL era), Environment (EraRule "DELEG" era) ~ DelegEnv era, State (EraRule "DELEG" era) ~ DState era, Embed (EraRule "POOL" era) (ShelleyDELPL era), Environment (EraRule "POOL" era) ~ PoolEnv era, State (EraRule "POOL" era) ~ PState era, Signal (EraRule "DELEG" era) ~ TxCert era, Embed (EraRule "POOL" era) (ShelleyDELPL era), Environment (EraRule "POOL" era) ~ PoolEnv era, Signal (EraRule "POOL" era) ~ PoolCert (EraCrypto era), TxCert era ~ ShelleyTxCert era) ⇒ STS (ShelleyDELPL era)

Instance details

Associated Types

type State (ShelleyDELPL era)

type Signal (ShelleyDELPL era)

type Environment (ShelleyDELPL era)

type BaseM (ShelleyDELPL era) ∷ Type → Type

type Event (ShelleyDELPL era)

type PredicateFailure (ShelleyDELPL era)

Methods

initialRules ∷ [InitialRule (ShelleyDELPL era)]

transitionRules ∷ [TransitionRule (ShelleyDELPL era)]

assertions ∷ [Assertion (ShelleyDELPL era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyDELPL era) → String

(ShelleyEraTxCert era, EraPParams era, ProtVerAtMost era 8, PredicateFailure (EraRule "DELEG" era) ~ ShelleyDelegPredFailure era, Event (EraRule "DELEG" era) ~ ShelleyDelegEvent era) ⇒ Embed (ShelleyDELEG era) (ShelleyDELPL era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyDELEG era) → PredicateFailure (ShelleyDELPL era)

wrapEvent ∷ Event (ShelleyDELEG era) → Event (ShelleyDELPL era)

(Era era, STS (ShelleyPOOL era), PredicateFailure (EraRule "POOL" era) ~ ShelleyPoolPredFailure era, Event (EraRule "POOL" era) ~ PoolEvent era) ⇒ Embed (ShelleyPOOL era) (ShelleyDELPL era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyPOOL era) → PredicateFailure (ShelleyDELPL era)

wrapEvent ∷ Event (ShelleyPOOL era) → Event (ShelleyDELPL era)