Cardano.Ledger.Shelley.Rules.Epoch

Documentation

data ShelleyEPOCH era

data ShelleyEpochPredFailure era

Constructors

PoolReapFailure (PredicateFailure (EraRule "POOLREAP" era))

SnapFailure (PredicateFailure (EraRule "SNAP" era))

UpecFailure (UpecPredFailure era)

data ShelleyEpochEvent era

Constructors

PoolReapEvent (Event (EraRule "POOLREAP" era))

SnapEvent (Event (EraRule "SNAP" era))

UpecEvent (Event (EraRule "UPEC" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

type UpecPredFailure era = UpecPredFailurePV (ProtVerLow era) era

Orphan instances

(EraTxOut era, EraGov era, GovState era ~ ShelleyGovState era, Embed (EraRule "SNAP" era) (ShelleyEPOCH era), Environment (EraRule "SNAP" era) ~ SnapEnv era, State (EraRule "SNAP" era) ~ SnapShots (EraCrypto era), Signal (EraRule "SNAP" era) ~ (), Embed (EraRule "POOLREAP" era) (ShelleyEPOCH era), Environment (EraRule "POOLREAP" era) ~ ShelleyPoolreapEnv era, State (EraRule "POOLREAP" era) ~ ShelleyPoolreapState era, Signal (EraRule "POOLREAP" era) ~ EpochNo, Embed (EraRule "UPEC" era) (ShelleyEPOCH era), Environment (EraRule "UPEC" era) ~ LedgerState era, State (EraRule "UPEC" era) ~ UpecState era, Signal (EraRule "UPEC" era) ~ (), Default (PParams era), Eq (UpecPredFailure era), Show (UpecPredFailure era)) ⇒ STS (ShelleyEPOCH era)

Instance details

Associated Types

type State (ShelleyEPOCH era)

type Signal (ShelleyEPOCH era)

type Environment (ShelleyEPOCH era)

type BaseM (ShelleyEPOCH era) ∷ Type → Type

type Event (ShelleyEPOCH era)

type PredicateFailure (ShelleyEPOCH era)

Methods

initialRules ∷ [InitialRule (ShelleyEPOCH era)]

transitionRules ∷ [TransitionRule (ShelleyEPOCH era)]

assertions ∷ [Assertion (ShelleyEPOCH era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyEPOCH era) → String

(Era era, STS (ShelleyPOOLREAP era), PredicateFailure (EraRule "POOLREAP" era) ~ ShelleyPoolreapPredFailure era, Event (EraRule "POOLREAP" era) ~ ShelleyPoolreapEvent era) ⇒ Embed (ShelleyPOOLREAP era) (ShelleyEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyPOOLREAP era) → PredicateFailure (ShelleyEPOCH era)

wrapEvent ∷ Event (ShelleyPOOLREAP era) → Event (ShelleyEPOCH era)

(EraTxOut era, PredicateFailure (EraRule "SNAP" era) ~ ShelleySnapPredFailure era, Event (EraRule "SNAP" era) ~ SnapEvent era) ⇒ Embed (ShelleySNAP era) (ShelleyEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleySNAP era) → PredicateFailure (ShelleyEPOCH era)

wrapEvent ∷ Event (ShelleySNAP era) → Event (ShelleyEPOCH era)

(Era era, STS (ShelleyUPEC era), UpecPredFailure era ~ ShelleyUpecPredFailure era, Event (EraRule "UPEC" era) ~ Void) ⇒ Embed (ShelleyUPEC era) (ShelleyEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyUPEC era) → PredicateFailure (ShelleyEPOCH era)

wrapEvent ∷ Event (ShelleyUPEC era) → Event (ShelleyEPOCH era)