Cardano.Ledger.Shelley.Rules.Ledgers

Documentation

data ShelleyLEDGERS era

data ShelleyLedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo ∷ SlotNo

ledgersPp ∷ PParams era

ledgersAccount ∷ AccountState

newtype ShelleyLedgersPredFailure era

Constructors

LedgerFailure (PredicateFailure (EraRule "LEDGER" era))

newtype ShelleyLedgersEvent era

Constructors

LedgerEvent (Event (EraRule "LEDGER" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Orphan instances

InjectRuleFailure "LEDGERS" ShelleyDelegPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

InjectRuleFailure "LEDGERS" ShelleyDelegsPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelegsPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

InjectRuleFailure "LEDGERS" ShelleyDelplPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyDelplPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

InjectRuleFailure "LEDGERS" ShelleyLedgerPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyLedgerPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

InjectRuleFailure "LEDGERS" ShelleyPoolPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPoolPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

InjectRuleFailure "LEDGERS" ShelleyPpupPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

InjectRuleFailure "LEDGERS" ShelleyUtxoPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

InjectRuleFailure "LEDGERS" ShelleyUtxowPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxowPredFailure (ShelleyEra c) → EraRuleFailure "LEDGERS" (ShelleyEra c)

(Era era, Embed (EraRule "LEDGER" era) (ShelleyLEDGERS era), Environment (EraRule "LEDGER" era) ~ LedgerEnv era, State (EraRule "LEDGER" era) ~ LedgerState era, Signal (EraRule "LEDGER" era) ~ Tx era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), Default (LedgerState era)) ⇒ STS (ShelleyLEDGERS era)

Instance details

Associated Types

type State (ShelleyLEDGERS era)

type Signal (ShelleyLEDGERS era)

type Environment (ShelleyLEDGERS era)

type BaseM (ShelleyLEDGERS era) ∷ Type → Type

type Event (ShelleyLEDGERS era)

type PredicateFailure (ShelleyLEDGERS era)

Methods

initialRules ∷ [InitialRule (ShelleyLEDGERS era)]

transitionRules ∷ [TransitionRule (ShelleyLEDGERS era)]

assertions ∷ [Assertion (ShelleyLEDGERS era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyLEDGERS era) → String

(Era era, STS (ShelleyLEDGER era), PredicateFailure (EraRule "LEDGER" era) ~ ShelleyLedgerPredFailure era, Event (EraRule "LEDGER" era) ~ ShelleyLedgerEvent era) ⇒ Embed (ShelleyLEDGER era) (ShelleyLEDGERS era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyLEDGER era) → PredicateFailure (ShelleyLEDGERS era)

wrapEvent ∷ Event (ShelleyLEDGER era) → Event (ShelleyLEDGERS era)