Cardano.Ledger.Shelley.Rules.NewEpoch

Documentation

data NEWEPOCH era

data NewEpochPredicateFailure era

Constructors

EpochFailure (PredicateFailure (EraRule "EPOCH" era))

CorruptRewardUpdate !(RewardUpdate (Crypto era))

MirFailure (PredicateFailure (EraRule "MIR" era))

data NewEpochEvent era

Constructors

DeltaRewardEvent (Event (EraRule "RUPD" era))

RestrainedRewards EpochNo (Map (Credential 'Staking (Crypto era)) (Set (Reward (Crypto era)))) (Set (Credential 'Staking (Crypto era)))

TotalRewardEvent EpochNo (Map (Credential 'Staking (Crypto era)) (Set (Reward (Crypto era))))

EpochEvent (Event (EraRule "EPOCH" era))

MirEvent (Event (EraRule "MIR" era))

TotalAdaPotsEvent AdaPots

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

calculatePoolDistr :: SnapShot crypto -> PoolDistr crypto