Cardano.Ledger.Shelley.Rules.Newpp

Documentation

data ShelleyNEWPP era

data ShelleyNewppState era

Constructors

NewppState (PParams era) (ShelleyGovState era)

data NewppEnv era

Constructors

NewppEnv

Fields

neCertState ∷ !(CertState era)

neUTxOState ∷ !(UTxOState era)

data ShelleyNewppPredFailure era

Constructors

UnexpectedDepositPot !Coin !Coin

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Orphan instances

(EraGov era, GovState era ~ ShelleyGovState era, ProtVerAtMost era 8) ⇒ STS (ShelleyNEWPP era)

Instance details

Associated Types

type State (ShelleyNEWPP era)

type Signal (ShelleyNEWPP era)

type Environment (ShelleyNEWPP era)

type BaseM (ShelleyNEWPP era) ∷ Type → Type

type Event (ShelleyNEWPP era)

type PredicateFailure (ShelleyNEWPP era)

Methods

initialRules ∷ [InitialRule (ShelleyNEWPP era)]

transitionRules ∷ [TransitionRule (ShelleyNEWPP era)]

assertions ∷ [Assertion (ShelleyNEWPP era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyNEWPP era) → String