Cardano.Ledger.Shelley.Rules.PoolReap

Documentation

data ShelleyPOOLREAP era

data ShelleyPoolreapEvent era

Constructors

RetiredPools

Fields

refundPools ∷ Map (Credential 'Staking (EraCrypto era)) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

unclaimedPools ∷ Map (Credential 'Staking (EraCrypto era)) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

epochNo ∷ EpochNo

data ShelleyPoolreapState era

Constructors

PoolreapState

Fields

prUTxOSt ∷ UTxOState era

prAccountState ∷ AccountState

prDState ∷ DState era

prPState ∷ PState era

newtype ShelleyPoolreapEnv era

Constructors

ShelleyPoolreapEnv

Fields

speVState ∷ VState era

This enviroment field is only needed for assertions.

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyPoolreapPredFailure era

Deprecations

prAcnt ∷ ShelleyPoolreapState era → AccountState

Deprecated: Use prAccountState instead

Orphan instances

(Default (ShelleyPoolreapState era), EraPParams era, EraGov era) ⇒ STS (ShelleyPOOLREAP era)

Instance details

Associated Types

type State (ShelleyPOOLREAP era)

type Signal (ShelleyPOOLREAP era)

type Environment (ShelleyPOOLREAP era)

type BaseM (ShelleyPOOLREAP era) ∷ Type → Type

type Event (ShelleyPOOLREAP era)

type PredicateFailure (ShelleyPOOLREAP era)

Methods

initialRules ∷ [InitialRule (ShelleyPOOLREAP era)]

transitionRules ∷ [TransitionRule (ShelleyPOOLREAP era)]

assertions ∷ [Assertion (ShelleyPOOLREAP era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyPOOLREAP era) → String