Cardano.Ledger.Shelley.Rules.Ppup

Documentation

data ShelleyPPUP era

data PpupEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs (EraCrypto era))

data ShelleyPpupPredFailure era

Constructors

NonGenesisUpdatePPUP !(Set (KeyHash 'Genesis (EraCrypto era))) !(Set (KeyHash 'Genesis (EraCrypto era)))
An update was proposed by a key hash that is not one of the genesis keys. The first set contains the key hashes which were a part of the update. The second set contains the key hashes of the genesis keys.

PPUpdateWrongEpoch !EpochNo !EpochNo !VotingPeriod
An update was proposed for the wrong epoch. The first EpochNo is the current epoch. The second EpochNo is the epoch listed in the update. The last parameter indicates if the update was intended for the current or the next epoch.

PVCannotFollowPPUP !ProtVer
An update was proposed which contains an invalid protocol version. New protocol versions must either increase the major number by exactly one and set the minor version to zero, or keep the major version the same and increase the minor version by exactly one.

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

sgsFuturePParams ∷ !(FuturePParams era)

Prediction of parameter changes that might happen on the epoch boundary.

newtype PpupEvent era

Constructors

PpupNewEpoch EpochNo

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data VotingPeriod

Constructors

VoteForThisEpoch

VoteForNextEpoch

type PPUPPredFailure era = EraRuleFailure "PPUP" era

Deprecated: In favor of EraRuleFailure PPUP era

votedFuturePParams

Arguments

∷ ∀ era. EraPParams era

⇒ ProposedPPUpdates era

→ PParams era
Protocol parameters to which the change will be applied.

→ Word64
Quorum needed to change the protocol parameters.

→ Maybe (PParams era)

If at least n nodes voted to change the same protocol parameters to the same values, return the given protocol parameters updated to these values. Here n is the quorum needed.

Orphan instances

(EraPParams era, ProtVerAtMost era 8) ⇒ STS (ShelleyPPUP era)

Instance details

Associated Types

type State (ShelleyPPUP era)

type Signal (ShelleyPPUP era)

type Environment (ShelleyPPUP era)

type BaseM (ShelleyPPUP era) ∷ Type → Type

type Event (ShelleyPPUP era)

type PredicateFailure (ShelleyPPUP era)

Methods

initialRules ∷ [InitialRule (ShelleyPPUP era)]

transitionRules ∷ [TransitionRule (ShelleyPPUP era)]

assertions ∷ [Assertion (ShelleyPPUP era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyPPUP era) → String