Cardano.Ledger.Shelley.Rules.Rupd

Documentation

data ShelleyRUPD era

data RupdEnv era

Constructors

RupdEnv !(BlocksMade (EraCrypto era)) !(EpochState era)

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleyRupdPredFailure era

epochInfoRange ∷ Monad m ⇒ EpochInfo m → EpochNo → m (SlotNo, SlotNo)

data PulsingRewUpdate c

The state used in the STS rules

Constructors

Pulsing !(RewardSnapShot c) !(Pulser c)

Complete !(RewardUpdate c)

startStep ∷ ∀ era. EraGov era ⇒ EpochSize → BlocksMade (EraCrypto era) → EpochState era → Coin → ActiveSlotCoeff → Word64 → PulsingRewUpdate (EraCrypto era)

pulseStep ∷ PulsingRewUpdate c → ShelleyBase (PulsingRewUpdate c, RewardEvent c)

Run the pulser for a bit. If is has nothing left to do, complete it.

completeStep ∷ PulsingRewUpdate c → ShelleyBase (PulsingRewUpdate c, RewardEvent c)

lift ∷ (MonadTrans t, Monad m) ⇒ m a → t m a

Lift a computation from the argument monad to the constructed monad.

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity ∷ a

data RupdEvent c

Constructors

RupdEvent !EpochNo !(Map (Credential 'Staking c) (Set (Reward c)))

Orphan instances

(Era era, EraGov era) ⇒ STS (ShelleyRUPD era)

Instance details

Associated Types

type State (ShelleyRUPD era)

type Signal (ShelleyRUPD era)

type Environment (ShelleyRUPD era)

type BaseM (ShelleyRUPD era) ∷ Type → Type

type Event (ShelleyRUPD era)

type PredicateFailure (ShelleyRUPD era)

Methods

initialRules ∷ [InitialRule (ShelleyRUPD era)]

transitionRules ∷ [TransitionRule (ShelleyRUPD era)]

assertions ∷ [Assertion (ShelleyRUPD era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyRUPD era) → String