Cardano.Ledger.Shelley.Rules.Snap

Documentation

data ShelleySNAP era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ShelleySnapPredFailure era

newtype SnapEvent era

Constructors

StakeDistEvent (Map (Credential 'Staking (EraCrypto era)) (Coin, KeyHash 'StakePool (EraCrypto era)))

data SnapEnv era

Constructors

SnapEnv !(LedgerState era) !(PParams era)

Orphan instances

EraTxOut era ⇒ STS (ShelleySNAP era)

Instance details

Associated Types

type State (ShelleySNAP era)

type Signal (ShelleySNAP era)

type Environment (ShelleySNAP era)

type BaseM (ShelleySNAP era) ∷ Type → Type

type Event (ShelleySNAP era)

type PredicateFailure (ShelleySNAP era)

Methods

initialRules ∷ [InitialRule (ShelleySNAP era)]

transitionRules ∷ [TransitionRule (ShelleySNAP era)]

assertions ∷ [Assertion (ShelleySNAP era)]

renderAssertionViolation ∷ AssertionViolation (ShelleySNAP era) → String