Cardano.Ledger.Shelley.Rules.Tick

Documentation

data ShelleyTICK era

type family State a

Type of the state which the system transitions between.

data ShelleyTickPredFailure era

Constructors

NewEpochFailure (PredicateFailure (EraRule "NEWEPOCH" era))

RupdFailure (PredicateFailure (EraRule "RUPD" era))

data ShelleyTickEvent era

Constructors

TickNewEpochEvent (Event (EraRule "NEWEPOCH" era))

TickRupdEvent (Event (EraRule "RUPD" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

adoptGenesisDelegs ∷ EpochState era → SlotNo → EpochState era

data ShelleyTICKF era

data ShelleyTickfPredFailure era

validatingTickTransition ∷ ∀ tick era. (EraGov era, Embed (EraRule "NEWEPOCH" era) (tick era), STS (tick era), State (tick era) ~ NewEpochState era, BaseM (tick era) ~ ShelleyBase, Environment (EraRule "NEWEPOCH" era) ~ (), State (EraRule "NEWEPOCH" era) ~ NewEpochState era, Signal (EraRule "NEWEPOCH" era) ~ EpochNo) ⇒ NewEpochState era → SlotNo → TransitionRule (tick era)

This is a limited version of bheadTransition which is suitable for the future ledger view.

validatingTickTransitionFORECAST ∷ ∀ tick era. (State (tick era) ~ NewEpochState era, BaseM (tick era) ~ ShelleyBase, State (EraRule "UPEC" era) ~ UpecState era, Signal (EraRule "UPEC" era) ~ (), Environment (EraRule "UPEC" era) ~ LedgerState era, Embed (EraRule "UPEC" era) (tick era), STS (tick era), GovState era ~ ShelleyGovState era, EraGov era) ⇒ NewEpochState era → SlotNo → TransitionRule (tick era)

This is a limited version of validatingTickTransition which is only suitable for the future ledger view.

solidifyNextEpochPParams ∷ EraGov era ⇒ NewEpochState era → SlotNo → ShelleyBase (EpochNo, NewEpochState era)

This action ensures that once the current slot number is at the point of no return we mark the future PParams to be updated at the next epoch boundary. Also returns the current epoch number for convenience.

Orphan instances

(EraGov era, Embed (EraRule "NEWEPOCH" era) (ShelleyTICK era), Embed (EraRule "RUPD" era) (ShelleyTICK era), State (ShelleyTICK era) ~ NewEpochState era, BaseM (ShelleyTICK era) ~ ShelleyBase, Environment (EraRule "RUPD" era) ~ RupdEnv era, State (EraRule "RUPD" era) ~ StrictMaybe (PulsingRewUpdate (EraCrypto era)), Signal (EraRule "RUPD" era) ~ SlotNo, Environment (EraRule "NEWEPOCH" era) ~ (), State (EraRule "NEWEPOCH" era) ~ NewEpochState era, Signal (EraRule "NEWEPOCH" era) ~ EpochNo) ⇒ STS (ShelleyTICK era)

Instance details

Associated Types

type State (ShelleyTICK era)

type Signal (ShelleyTICK era)

type Environment (ShelleyTICK era)

type BaseM (ShelleyTICK era) ∷ Type → Type

type Event (ShelleyTICK era)

type PredicateFailure (ShelleyTICK era)

Methods

initialRules ∷ [InitialRule (ShelleyTICK era)]

transitionRules ∷ [TransitionRule (ShelleyTICK era)]

assertions ∷ [Assertion (ShelleyTICK era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyTICK era) → String

(EraGov era, GovState era ~ ShelleyGovState era, State (EraRule "PPUP" era) ~ ShelleyGovState era, Signal (EraRule "UPEC" era) ~ (), State (EraRule "UPEC" era) ~ UpecState era, Environment (EraRule "UPEC" era) ~ LedgerState era, Embed (EraRule "UPEC" era) (ShelleyTICKF era)) ⇒ STS (ShelleyTICKF era)

Instance details

Associated Types

type State (ShelleyTICKF era)

type Signal (ShelleyTICKF era)

type Environment (ShelleyTICKF era)

type BaseM (ShelleyTICKF era) ∷ Type → Type

type Event (ShelleyTICKF era)

type PredicateFailure (ShelleyTICKF era)

Methods

initialRules ∷ [InitialRule (ShelleyTICKF era)]

transitionRules ∷ [TransitionRule (ShelleyTICKF era)]

assertions ∷ [Assertion (ShelleyTICKF era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyTICKF era) → String

(STS (ShelleyNEWEPOCH era), PredicateFailure (EraRule "NEWEPOCH" era) ~ ShelleyNewEpochPredFailure era, Event (EraRule "NEWEPOCH" era) ~ ShelleyNewEpochEvent era) ⇒ Embed (ShelleyNEWEPOCH era) (ShelleyTICK era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyNEWEPOCH era) → PredicateFailure (ShelleyTICK era)

wrapEvent ∷ Event (ShelleyNEWEPOCH era) → Event (ShelleyTICK era)

(Era era, STS (ShelleyRUPD era), PredicateFailure (EraRule "RUPD" era) ~ ShelleyRupdPredFailure era, Event (EraRule "RUPD" era) ~ RupdEvent (EraCrypto era)) ⇒ Embed (ShelleyRUPD era) (ShelleyTICK era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyRUPD era) → PredicateFailure (ShelleyTICK era)

wrapEvent ∷ Event (ShelleyRUPD era) → Event (ShelleyTICK era)

(Era era, STS (ShelleyUPEC era), PredicateFailure (EraRule "UPEC" era) ~ ShelleyUpecPredFailure era, Event (EraRule "UPEC" era) ~ Void) ⇒ Embed (ShelleyUPEC era) (ShelleyTICKF era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyUPEC era) → PredicateFailure (ShelleyTICKF era)

wrapEvent ∷ Event (ShelleyUPEC era) → Event (ShelleyTICKF era)