Cardano.Ledger.Shelley.Rules.Upec

Description

Epoch change registration.

The rules of this module determine how the update subsystem of the ledger handles the epoch transitions.

Documentation

data ShelleyUPEC era

data UpecState era

Constructors

UpecState

Fields

usCurPParams ∷ !(PParams era)

Current protocol parameters.

usGovState ∷ !(ShelleyGovState era)

State of the protocol update transition system.

newtype ShelleyUpecPredFailure era

Constructors

NewPpFailure (PredicateFailure (ShelleyNEWPP era))

votedValue ∷ ∀ era. EraPParams era ⇒ ProposedPPUpdates era → PParams era → Int → Maybe (PParams era)

Deprecated: In favor of votedFuturePParams

Orphan instances

(EraGov era, Default (PParams era), GovState era ~ ShelleyGovState era, ProtVerAtMost era 8) ⇒ STS (ShelleyUPEC era)

Instance details

Associated Types

type State (ShelleyUPEC era)

type Signal (ShelleyUPEC era)

type Environment (ShelleyUPEC era)

type BaseM (ShelleyUPEC era) ∷ Type → Type

type Event (ShelleyUPEC era)

type PredicateFailure (ShelleyUPEC era)

Methods

initialRules ∷ [InitialRule (ShelleyUPEC era)]

transitionRules ∷ [TransitionRule (ShelleyUPEC era)]

assertions ∷ [Assertion (ShelleyUPEC era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyUPEC era) → String

(Era era, STS (ShelleyNEWPP era)) ⇒ Embed (ShelleyNEWPP era) (ShelleyUPEC era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyNEWPP era) → PredicateFailure (ShelleyUPEC era)

wrapEvent ∷ Event (ShelleyNEWPP era) → Event (ShelleyUPEC era)