Cardano.Ledger.Shelley.Rules.Utxo

Documentation

data ShelleyUTXO era

data UtxoEnv era

Constructors

UtxoEnv

Fields

ueSlot ∷ SlotNo

uePParams ∷ PParams era

ueCertState ∷ CertState era

data ShelleyUtxoPredFailure era

Constructors

BadInputsUTxO !(Set (TxIn (EraCrypto era)))

ExpiredUTxO !SlotNo !SlotNo

MaxTxSizeUTxO !Integer !Integer

InputSetEmptyUTxO

FeeTooSmallUTxO !Coin !Coin

ValueNotConservedUTxO !(Value era) !(Value era)

WrongNetwork !Network !(Set (Addr (EraCrypto era)))

WrongNetworkWithdrawal !Network !(Set (RewardAccount (EraCrypto era)))

OutputTooSmallUTxO ![TxOut era]

UpdateFailure (EraRuleFailure "PPUP" era)

OutputBootAddrAttrsTooBig ![TxOut era]

data UtxoEvent era

Constructors

TotalDeposits (SafeHash (EraCrypto era) EraIndependentTxBody) Coin

UpdateEvent (Event (EraRule "PPUP" era))

TxUTxODiff
The UTxOs consumed and created by a signal tx

Fields

(UTxO era)

UTxO consumed

(UTxO era)

UTxO created

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

updateUTxOState ∷ (EraTxBody era, Monad m) ⇒ PParams era → UTxOState era → TxBody era → CertState era → GovState era → (Coin → m ()) → (UTxO era → UTxO era → m ()) → m (UTxOState era)

This monadic action captures the final stages of the UTXO(S) rule. In particular it applies all of the UTxO related aditions and removals, gathers all of the fees into the fee pot utxosFees and updates the utxosDeposited field. Continuation supplied will be called on the deposit - refund change, which is normally used to emit the TotalDeposits event.

Validations

validateInputSetEmptyUTxO ∷ EraTxBody era ⇒ TxBody era → Test (ShelleyUtxoPredFailure era)

Ensure that there is at least one input in the TxBody

txins txb ≠ ∅

validateFeeTooSmallUTxO ∷ EraUTxO era ⇒ PParams era → Tx era → UTxO era → Test (ShelleyUtxoPredFailure era)

Ensure that the fee is at least the amount specified by the minfee

minfee pp tx ≤ txfee txb

validateBadInputsUTxO ∷ UTxO era → Set (TxIn (EraCrypto era)) → Test (ShelleyUtxoPredFailure era)

Ensure all transaction inputs are present in UTxO

inputs ⊆ dom utxo

validateWrongNetwork ∷ (EraTxOut era, Foldable f) ⇒ Network → f (TxOut era) → Test (ShelleyUtxoPredFailure era)

Make sure all addresses match the supplied NetworkId

∀(_ → (a, _)) ∈ txouts txb, netId a = NetworkId

validateWrongNetworkWithdrawal ∷ EraTxBody era ⇒ Network → TxBody era → Test (ShelleyUtxoPredFailure era)

Make sure all addresses match the supplied NetworkId

∀(a → ) ∈ txwdrls txb, netId a = NetworkId

validateOutputBootAddrAttrsTooBig ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Test (ShelleyUtxoPredFailure era)

Bootstrap (i.e. Byron) addresses have variable sized attributes in them. It is important to limit their overall size.

∀ ( _ ↦ (a,_)) ∈ txoutstxb,  a ∈ Addrbootstrap → bootstrapAttrsSize a ≤ 64

validateMaxTxSizeUTxO ∷ EraTx era ⇒ PParams era → Tx era → Test (ShelleyUtxoPredFailure era)

Ensure that the size of the transaction does not exceed the maxTxSize protocol parameter

txsize tx ≤ maxTxSize pp

validateValueNotConservedUTxO ∷ EraUTxO era ⇒ PParams era → UTxO era → CertState era → TxBody era → Test (ShelleyUtxoPredFailure era)

Ensure that value consumed and produced matches up exactly

consumed pp utxo txb = produced pp poolParams txb

utxoEnvSlotL ∷ Lens' (UtxoEnv era) SlotNo

utxoEnvPParamsL ∷ Lens' (UtxoEnv era) (PParams era)

utxoEnvCertStateL ∷ Lens' (UtxoEnv era) (CertState era)

Orphan instances

InjectRuleFailure "UTXO" ShelleyPpupPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (ShelleyEra c) → EraRuleFailure "UTXO" (ShelleyEra c)

(EraTx era, EraUTxO era, ShelleyEraTxBody era, EraGov era, GovState era ~ ShelleyGovState era, ExactEra ShelleyEra era, Embed (EraRule "PPUP" era) (ShelleyUTXO era), Environment (EraRule "PPUP" era) ~ PpupEnv era, Signal (EraRule "PPUP" era) ~ StrictMaybe (Update era), State (EraRule "PPUP" era) ~ ShelleyGovState era, Eq (EraRuleFailure "PPUP" era), Show (EraRuleFailure "PPUP" era), EraRule "UTXO" era ~ ShelleyUTXO era, InjectRuleFailure "UTXO" ShelleyUtxoPredFailure era) ⇒ STS (ShelleyUTXO era)

Instance details

Associated Types

type State (ShelleyUTXO era)

type Signal (ShelleyUTXO era)

type Environment (ShelleyUTXO era)

type BaseM (ShelleyUTXO era) ∷ Type → Type

type Event (ShelleyUTXO era)

type PredicateFailure (ShelleyUTXO era)

Methods

initialRules ∷ [InitialRule (ShelleyUTXO era)]

transitionRules ∷ [TransitionRule (ShelleyUTXO era)]

assertions ∷ [Assertion (ShelleyUTXO era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyUTXO era) → String

(Era era, STS (ShelleyPPUP era), EraRuleFailure "PPUP" era ~ ShelleyPpupPredFailure era, Event (EraRule "PPUP" era) ~ PpupEvent era) ⇒ Embed (ShelleyPPUP era) (ShelleyUTXO era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyPPUP era) → PredicateFailure (ShelleyUTXO era)

wrapEvent ∷ Event (ShelleyPPUP era) → Event (ShelleyUTXO era)