Cardano.Ledger.Shelley.Rules.Utxow

Documentation

data ShelleyUTXOW era

data ShelleyUtxowPredFailure era

Constructors

InvalidWitnessesUTXOW ![VKey 'Witness (EraCrypto era)]

MissingVKeyWitnessesUTXOW !(Set (KeyHash 'Witness (EraCrypto era)))

MissingScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))

ScriptWitnessNotValidatingUTXOW !(Set (ScriptHash (EraCrypto era)))

UtxoFailure (PredicateFailure (EraRule "UTXO" era))

MIRInsufficientGenesisSigsUTXOW (Set (KeyHash 'Witness (EraCrypto era)))

MissingTxBodyMetadataHash !(AuxiliaryDataHash (EraCrypto era))

MissingTxMetadata !(AuxiliaryDataHash (EraCrypto era))

ConflictingMetadataHash !(AuxiliaryDataHash (EraCrypto era)) !(AuxiliaryDataHash (EraCrypto era))

InvalidMetadata

ExtraneousScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))

newtype ShelleyUtxowEvent era

Constructors

UtxoEvent (Event (EraRule "UTXO" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

transitionRulesUTXOW ∷ ∀ era. (EraUTxO era, ShelleyEraTxBody era, ScriptsNeeded era ~ ShelleyScriptsNeeded era, BaseM (EraRule "UTXOW" era) ~ ShelleyBase, Embed (EraRule "UTXO" era) (EraRule "UTXOW" era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Signal (EraRule "UTXOW" era) ~ Tx era, InjectRuleFailure "UTXOW" ShelleyUtxowPredFailure era, STS (EraRule "UTXOW" era), DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody)) ⇒ TransitionRule (EraRule "UTXOW" era)

A generic Utxow witnessing function designed to be used across many Eras. Note the embed argument lifts from the simple Shelley (ShelleyUtxowPredFailure) to the PredicateFailure (type family) of the context of where it is called.

shelleyWitsVKeyNeeded ∷ ∀ era. (EraTx era, ShelleyEraTxBody era) ⇒ UTxO era → TxBody era → GenDelegs (EraCrypto era) → Set (KeyHash 'Witness (EraCrypto era))

Deprecated: Use getShelleyWitsVKeyNeeded instead

witsVKeyNeededGov ∷ ∀ era. ShelleyEraTxBody era ⇒ TxBody era → GenDelegs (EraCrypto era) → Set (KeyHash 'Witness (EraCrypto era))

Deprecated: As unnecessary. Use getWitsVKeyNeeded instead

Collect the set of hashes of keys that needs to sign a given transaction. This set consists of the txin owners, certificate authors, and withdrawal reward accounts.

witsVKeyNeededNoGov ∷ ∀ era. EraTx era ⇒ UTxO era → TxBody era → Set (KeyHash 'Witness (EraCrypto era))

Deprecated: Use getShelleyWitsVKeyNeededNoGov instead

Individual validation steps

validateFailedNativeScripts ∷ EraTx era ⇒ ScriptsProvided era → Tx era → Test (ShelleyUtxowPredFailure era)

validateMissingScripts ∷ ShelleyScriptsNeeded era → ScriptsProvided era → Test (ShelleyUtxowPredFailure era)

validateVerifiedWits ∷ (EraTx era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody)) ⇒ Tx era → Test (ShelleyUtxowPredFailure era)

Determine if the UTxO witnesses in a given transaction are correct.

validateMetadata ∷ EraTx era ⇒ PParams era → Tx era → Test (ShelleyUtxowPredFailure era)

check metadata hash ((adh = ◇) ∧ (ad= ◇)) ∨ (adh = hashAD ad)

validateMIRInsufficientGenesisSigs ∷ (EraTx era, ShelleyEraTxBody era) ⇒ GenDelegs (EraCrypto era) → Word64 → Set (KeyHash 'Witness (EraCrypto era)) → Tx era → Test (ShelleyUtxowPredFailure era)

check genesis keys signatures for instantaneous rewards certificates

genSig := { hashKey gkey | gkey ∈ dom(genDelegs)} ∩ witsKeyHashes { c ∈ txcerts txb ∩ TxCert_mir} ≠ ∅ ⇒ |genSig| ≥ Quorum

validateNeededWitnesses

Arguments

∷ EraUTxO era

⇒ Set (KeyHash 'Witness (EraCrypto era))
Provided witness

→ CertState era

→ UTxO era

→ TxBody era

→ Test (ShelleyUtxowPredFailure era)

Verify that we provide at least all of the required witnesses

witsVKeyNeeded utxo tx ⊆ witsKeyHashes

propWits ∷ Maybe (Update era) → GenDelegs (EraCrypto era) → Set (KeyHash 'Witness (EraCrypto era))

Deprecated: This will become an internal function in the future. Submit an issue if you still need it.

Calculate the set of hash keys of the required witnesses for update proposals.

Orphan instances

InjectRuleFailure "UTXOW" ShelleyPpupPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (ShelleyEra c) → EraRuleFailure "UTXOW" (ShelleyEra c)

InjectRuleFailure "UTXOW" ShelleyUtxoPredFailure (ShelleyEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (ShelleyEra c) → EraRuleFailure "UTXOW" (ShelleyEra c)

(EraTx era, EraUTxO era, ShelleyEraTxBody era, ScriptsNeeded era ~ ShelleyScriptsNeeded era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), Embed (EraRule "UTXO" era) (ShelleyUTXOW era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, EraRule "UTXOW" era ~ ShelleyUTXOW era, InjectRuleFailure "UTXOW" ShelleyUtxowPredFailure era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), EraGov era) ⇒ STS (ShelleyUTXOW era)

Instance details

Associated Types

type State (ShelleyUTXOW era)

type Signal (ShelleyUTXOW era)

type Environment (ShelleyUTXOW era)

type BaseM (ShelleyUTXOW era) ∷ Type → Type

type Event (ShelleyUTXOW era)

type PredicateFailure (ShelleyUTXOW era)

Methods

initialRules ∷ [InitialRule (ShelleyUTXOW era)]

transitionRules ∷ [TransitionRule (ShelleyUTXOW era)]

assertions ∷ [Assertion (ShelleyUTXOW era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyUTXOW era) → String

(Era era, STS (ShelleyUTXO era), PredicateFailure (EraRule "UTXO" era) ~ ShelleyUtxoPredFailure era, Event (EraRule "UTXO" era) ~ UtxoEvent era) ⇒ Embed (ShelleyUTXO era) (ShelleyUTXOW era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyUTXO era) → PredicateFailure (ShelleyUTXOW era)

wrapEvent ∷ Event (ShelleyUTXO era) → Event (ShelleyUTXOW era)