Cardano.Ledger.Shelley.StabilityWindow

Documentation

computeStabilityWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the stability window (e.g. the number of slots needed for a block to become stable) from the security param and the active slot coefficient.

The value 3k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

computeRandomnessStabilisationWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the randomness stabilisation window from the security param and the active slot coefficient.

The value 4k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm