Cardano.Ledger.Shelley.State

Documentation

module Cardano.Ledger.State

data ShelleyAccounts era

Constructors

ShelleyAccounts

Fields

saStates ∷ !(Map (Credential 'Staking) (ShelleyAccountState era))

Map from a staking credential to the account state.

saPtrs ∷ !(Map Ptr (Credential 'Staking))

A Map from a pointer, to the staking credential. Pointer points to the certificate which registered the staking credential.

data ShelleyAccountState era

Constructors

ShelleyAccountState

Fields

sasPtr ∷ !Ptr

Pointer to the certificate in which the stake credential was registered in.

sasBalance ∷ !(CompactForm Coin)

Current balance of the account

sasDeposit ∷ !(CompactForm Coin)

Deposit amount that was left when staking credential was registered

sasStakePoolDelegation ∷ !(StrictMaybe (KeyHash 'StakePool))

Potential delegation to a stake pool

class EraAccounts era ⇒ ShelleyEraAccounts era where

Minimal complete definition

Nothing

Methods

mkShelleyAccountState ∷ Ptr → CompactForm Coin → AccountState era

default mkShelleyAccountState ∷ AccountState era ~ ShelleyAccountState era ⇒ Ptr → CompactForm Coin → AccountState era

accountsPtrsMapG ∷ SimpleGetter (Accounts era) (Map Ptr (Credential 'Staking))

Get the map with all of the pointers and their respective credentials.

ptrAccountStateG ∷ SimpleGetter (AccountState era) Ptr

This is a getter for a Ptr. It is not a full lens, because it is not only unsafe to modify a pointer for an existing AccountState due to violation of an invariant in the ShelleyAccounts, but also because once account is registered pointer cannot change. Pointer describes unique point on chain when registration has occured, which means it can't change.

default ptrAccountStateG ∷ AccountState era ~ ShelleyAccountState era ⇒ SimpleGetter (AccountState era) Ptr

shelleyAddAccountState ∷ ShelleyEraAccounts era ⇒ Credential Staking → AccountState era → Accounts era → Accounts era

registerShelleyAccount

Arguments

∷ ShelleyEraAccounts era

⇒ Credential 'Staking

→ Ptr
Pointer to the certificate that registered the credential

→ CompactForm Coin
Deposit

→ Maybe (KeyHash 'StakePool)

→ Accounts era

→ Accounts era

unregisterShelleyAccount

Arguments

∷ ShelleyEraAccounts era

⇒ Credential 'Staking
Credential to unregister

→ Accounts era
Accounts to remove the account state from

→ (Maybe (AccountState era), Accounts era)
Returns Just whenever account was registered and Nothing otherwise. Produced Accounts will have the account state removed, if it was present there to begin with.

data ShelleyCertState era

Constructors

ShelleyCertState

Fields

shelleyCertPState ∷ !(PState era)

shelleyCertDState ∷ !(DState era)

mkShelleyCertState ∷ EraCertState era ⇒ PState era → DState era → CertState era

shelleyCertDStateL ∷ Lens' (ShelleyCertState era) (DState era)

shelleyCertPStateL ∷ Lens' (ShelleyCertState era) (PState era)

shelleyObligationCertState ∷ EraCertState era ⇒ CertState era → Obligations

shelleyCertsTotalDepositsTxBody ∷ EraTxBody era ⇒ PParams era → ShelleyCertState era → TxBody era → Coin

shelleyCertsTotalRefundsTxBody ∷ (EraTxBody era, EraAccounts era) ⇒ PParams era → ShelleyCertState era → TxBody era → Coin

data ShelleyInstantStake era

Constructors

ShelleyInstantStake

Fields

sisCredentialStake ∷ !(Map (Credential 'Staking) (CompactForm Coin))

sisPtrStake ∷ !(Map Ptr (CompactForm Coin))

shelleyInstantStakeCredentialsL ∷ Lens' (ShelleyInstantStake era) (Map (Credential 'Staking) (CompactForm Coin))

addShelleyInstantStake ∷ EraTxOut era ⇒ UTxO era → ShelleyInstantStake era → ShelleyInstantStake era

deleteShelleyInstantStake ∷ EraTxOut era ⇒ UTxO era → ShelleyInstantStake era → ShelleyInstantStake era

resolveShelleyInstantStake ∷ ∀ era. (EraStake era, InstantStake era ~ ShelleyInstantStake era, AccountState era ~ ShelleyAccountState era, Accounts era ~ ShelleyAccounts era) ⇒ ShelleyInstantStake era → ShelleyAccounts era → Stake