Cardano.Ledger.Shelley.Tx

Documentation

data Tx era where

Bundled Patterns

pattern Tx :: (Era era, ToCBOR (AuxiliaryData era), ToCBOR (TxBody era), ToCBOR (Witnesses era)) => TxBody era -> Witnesses era -> StrictMaybe (AuxiliaryData era) -> Tx era

newtype TxBody era

Constructors

TxBodyConstr (MemoBytes (TxBodyRaw era))

Bundled Patterns

pattern TxBody :: (Era era, FromCBOR (PParamsDelta era), TransTxBody ToCBOR era) => Set (TxIn (Crypto era)) -> StrictSeq (TxOut era) -> StrictSeq (DCert (Crypto era)) -> Wdrl (Crypto era) -> Coin -> SlotNo -> StrictMaybe (Update era) -> StrictMaybe (AuxiliaryDataHash (Crypto era)) -> TxBody era
Pattern for use by external users

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !(CompactAddr (Crypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern TxOut :: (Era era, Show (Value era), Compactible (Value era)) => Addr (Crypto era) -> Value era -> TxOut era

data TxIn crypto

The input of a UTxO.

Constructors

TxIn !(TxId crypto) !TxIx

newtype TxId crypto

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId :: SafeHash crypto EraIndependentTxBody

decodeWits :: forall era s. (FromCBOR (Annotator (Script era)), ValidateScript era) => Decoder s (Annotator (WitnessSet era))

segwitTx :: (ToCBOR (TxBody era), ToCBOR (Witnesses era), ToCBOR (AuxiliaryData era)) => Annotator (TxBody era) -> Annotator (Witnesses era) -> Maybe (Annotator (AuxiliaryData era)) -> Annotator (Tx era)

type WitnessSet = WitnessSetHKD Identity

data WitnessSetHKD f era where

Bundled Patterns

pattern WitnessSet :: (Era era, AnnotatedData (Script era)) => Set (WitVKey 'Witness (Crypto era)) -> Map (ScriptHash (Crypto era)) (Script era) -> Set (BootstrapWitness (Crypto era)) -> WitnessSet era

data WitVKey kr crypto where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey :: (Typeable kr, Crypto crypto) => VKey kr crypto -> SignedDSIGN crypto (Hash crypto EraIndependentTxBody) -> WitVKey kr crypto

class (Era era, SafeToHash (Script era), HasField "body" (Tx era) (TxBody era)) => ValidateScript era where

Typeclass for script data types. Allows for script validation and hashing. You must understand the role of SafeToHash and scriptPrefixTag to make new instances. scriptPrefixTag is a magic number representing the tag of the script language. For each new script language defined, a new tag is chosen and the tag is included in the script hash for a script. The safeToHash constraint ensures that Scripts are never reserialised.

Minimal complete definition

scriptPrefixTag, validateScript

Methods

scriptPrefixTag :: Script era -> ByteString

validateScript :: Script era -> Tx era -> Bool

hashScript :: Script era -> ScriptHash (Crypto era)

isNativeScript :: Script era -> Bool

txwitsScript :: Witnesses era ~ WitnessSet era => Tx era -> Map (ScriptHash (Crypto era)) (Script era)

Multi-signature script witness accessor function for Transactions

extractKeyHashWitnessSet :: forall (r :: KeyRole) crypto. [Credential r crypto] -> Set (KeyHash 'Witness crypto)

addrWits' :: WitnessSetHKD f era -> HKD f (Set (WitVKey 'Witness (Crypto era)))

evalNativeMultiSigScript :: Crypto crypto => MultiSig crypto -> Set (KeyHash 'Witness crypto) -> Bool

Script evaluator for native multi-signature scheme. vhks is the set of key hashes that signed the transaction to be validated.

hashMultiSigScript :: forall era. (ValidateScript era, Script era ~ MultiSig (Crypto era)) => MultiSig (Crypto era) -> ScriptHash (Crypto era)

Hashes native multi-signature script.

validateNativeMultiSigScript :: (TransTx ToCBOR era, Witnesses era ~ WitnessSet era) => MultiSig (Crypto era) -> Tx era -> Bool

Script validator for native multi-signature scheme.

type TransTx (c :: Type -> Constraint) era = (Era era, c (Script era), c (TxBody era), c (AuxiliaryData era))

type TransWitnessSet (c :: Type -> Constraint) era = c (Script era)

prettyWitnessSetParts :: WitnessSetHKD Identity era -> (Set (WitVKey 'Witness (Crypto era)), Map (ScriptHash (Crypto era)) (Script era), Set (BootstrapWitness (Crypto era)))

Exports the relevant parts from a (WintessSetHKD Identity era) for use by the pretty printer without all the horrible constraints. Uses the non-exported WitnessSet' constructor.