Cardano.Ledger.Shelley.Tx.Internal

Description

Provides Shelley Tx internals

Warning

This module is considered internal.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Transaction

newtype ShelleyTx era

Constructors

MkShelleyTx (MemoBytes (ShelleyTxRaw era))

Bundled Patterns

pattern ShelleyTx ∷ EraTx era ⇒ TxBody era → TxWits era → StrictMaybe (TxAuxData era) → ShelleyTx era

data ShelleyTxRaw era

Constructors

ShelleyTxRaw

Fields

strBody ∷ !(TxBody era)

strWits ∷ !(TxWits era)

strAuxData ∷ !(StrictMaybe (TxAuxData era))

bodyShelleyTxL ∷ ∀ era. EraTx era ⇒ Lens' (ShelleyTx era) (TxBody era)

TxBody setter and getter for ShelleyTx. The setter does update memoized binary representation.

witsShelleyTxL ∷ ∀ era. EraTx era ⇒ Lens' (ShelleyTx era) (TxWits era)

TxWits setter and getter for ShelleyTx. The setter does update memoized binary representation.

auxDataShelleyTxL ∷ ∀ era. EraTx era ⇒ Lens' (ShelleyTx era) (StrictMaybe (TxAuxData era))

TxAuxData setter and getter for ShelleyTx. The setter does update memoized binary representation.

sizeShelleyTxF ∷ SimpleGetter (ShelleyTx era) Integer

Size getter for ShelleyTx.

wireSizeShelleyTxF ∷ SimpleGetter (ShelleyTx era) Word32

segWitTx ∷ ∀ era. EraTx era ⇒ TxBody era → TxWits era → Maybe (TxAuxData era) → ShelleyTx era

mkBasicShelleyTx ∷ EraTx era ⇒ TxBody era → ShelleyTx era

shelleyMinFeeTx ∷ EraTx era ⇒ PParams era → Tx era → Coin

Minimum fee calculation

witsFromTxWitnesses ∷ EraTx era ⇒ Tx era → Set (KeyHash 'Witness)

Deprecated: In favor ot keyHashWitnessesTxWits

Extract the witness hashes from the Transaction.

shelleyEqTxRaw ∷ EraTx era ⇒ Tx era → Tx era → Bool

unsafeConstructTxWithBytes ∷ TxBody era → TxWits era → StrictMaybe (TxAuxData era) → ByteString → ShelleyTx era

Construct a Tx containing the explicit serialised bytes.

This function is marked as unsafe since it makes no guarantee that the represented bytes are indeed the correct serialisation of the transaction. Thus, when calling this function, the caller is responsible for making this guarantee.

The only intended use case for this is for segregated witness.

Orphan instances

EraTx ShelleyEra

Instance details

Associated Types

type Tx ShelleyEra = (r ∷ Type)

type TxUpgradeError ShelleyEra

Methods

mkBasicTx ∷ TxBody ShelleyEra → Tx ShelleyEra

bodyTxL ∷ Lens' (Tx ShelleyEra) (TxBody ShelleyEra)

witsTxL ∷ Lens' (Tx ShelleyEra) (TxWits ShelleyEra)

auxDataTxL ∷ Lens' (Tx ShelleyEra) (StrictMaybe (TxAuxData ShelleyEra))

sizeTxF ∷ SimpleGetter (Tx ShelleyEra) Integer

wireSizeTxF ∷ SimpleGetter (Tx ShelleyEra) Word32

sizeTxForFeeCalculation ∷ Tx ShelleyEra → Integer

validateNativeScript ∷ Tx ShelleyEra → NativeScript ShelleyEra → Bool

getMinFeeTx ∷ PParams ShelleyEra → Tx ShelleyEra → Int → Coin

upgradeTx ∷ Tx (PreviousEra ShelleyEra) → Either (TxUpgradeError ShelleyEra) (Tx ShelleyEra)