Cardano.Ledger.Shelley.TxBody

Documentation

data DCert crypto

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert crypto)

DCertPool !(PoolCert crypto)

DCertGenesis !(GenesisDelegCert crypto)

DCertMir !(MIRCert crypto)

data DelegCert crypto

Constructors

RegKey !(StakeCredential crypto)
A stake key registration certificate.

DeRegKey !(StakeCredential crypto)
A stake key deregistration certificate.

Delegate !(Delegation crypto)
A stake delegation certificate.

data Delegation crypto

The delegation of one stake key to another.

Constructors

Delegation

Fields

_delegator :: !(StakeCredential crypto)

_delegatee :: !(KeyHash 'StakePool crypto)

data GenesisDelegCert crypto

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis crypto) !(KeyHash 'GenesisDelegate crypto) !(Hash crypto (VerKeyVRF crypto))

data MIRCert crypto

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot :: MIRPot

mirRewards :: MIRTarget crypto

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget crypto

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR (Map (Credential 'Staking crypto) DeltaCoin)

SendToOppositePotMIR Coin

data PoolCert crypto

Constructors

RegPool !(PoolParams crypto)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool crypto) !EpochNo
A stake pool retirement certificate.

data PoolMetadata

Constructors

PoolMetadata

Fields

_poolMDUrl :: !Url

_poolMDHash :: !ByteString

data PoolParams crypto

A stake pool.

Constructors

PoolParams

Fields

_poolId :: !(KeyHash 'StakePool crypto)

_poolVrf :: !(Hash crypto (VerKeyVRF crypto))

_poolPledge :: !Coin

_poolCost :: !Coin

_poolMargin :: !UnitInterval

_poolRAcnt :: !(RewardAcnt crypto)

_poolOwners :: !(Set (KeyHash 'Staking crypto))

_poolRelays :: !(StrictSeq StakePoolRelay)

_poolMD :: !(StrictMaybe PoolMetadata)

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo !TxIx !CertIx

data RewardAcnt crypto

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork :: !Network

getRwdCred :: !(Credential 'Staking crypto)

newtype StakeCreds crypto

Constructors

StakeCreds

Fields

unStakeCreds :: Map (Credential 'Staking crypto) SlotNo

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

newtype TxBody era

Constructors

TxBodyConstr (MemoBytes (TxBodyRaw era))

Bundled Patterns

pattern TxBody :: (Era era, FromCBOR (PParamsDelta era), TransTxBody ToCBOR era) => Set (TxIn (Crypto era)) -> StrictSeq (TxOut era) -> StrictSeq (DCert (Crypto era)) -> Wdrl (Crypto era) -> Coin -> SlotNo -> StrictMaybe (Update era) -> StrictMaybe (AuxiliaryDataHash (Crypto era)) -> TxBody era
Pattern for use by external users

data TxBodyRaw era

Constructors

TxBodyRaw

Fields

_inputsX :: !(Set (TxIn (Crypto era)))

_outputsX :: !(StrictSeq (TxOut era))

_certsX :: !(StrictSeq (DCert (Crypto era)))

_wdrlsX :: !(Wdrl (Crypto era))

_txfeeX :: !Coin

_ttlX :: !SlotNo

_txUpdateX :: !(StrictMaybe (Update era))

_mdHashX :: !(StrictMaybe (AuxiliaryDataHash (Crypto era)))

data EraIndependentTxBody

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !(CompactAddr (Crypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern TxOut :: (Era era, Show (Value era), Compactible (Value era)) => Addr (Crypto era) -> Value era -> TxOut era

data Url

newtype Wdrl crypto

Constructors

Wdrl

Fields

unWdrl :: Map (RewardAcnt crypto) Coin

data WitVKey kr crypto where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey :: (Typeable kr, Crypto crypto) => VKey kr crypto -> SignedDSIGN crypto (Hash crypto EraIndependentTxBody) -> WitVKey kr crypto

witKeyHash :: WitVKey kr crypto -> KeyHash 'Witness crypto

data SizeOfPoolOwners

The size of the _poolOwners Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolOwners

data SizeOfPoolRelays

The size of the _poolRelays Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolRelays

type TransTxId (c :: Type -> Constraint) era = (HashAnnotated (TxBody era) EraIndependentTxBody (Crypto era), ToCBOR (TxBody era), ToCBOR (TxOut era), TransValue ToCBOR era, TransValue c era)

type TransTxOut (c :: Type -> Constraint) era = (c (Value era), Compactible (Value era))

type TransTxBody (c :: Type -> Constraint) era = (c (TxOut era), c (PParamsDelta era), HashAnnotated (TxBody era) EraIndependentTxBody (Crypto era))