Cardano.Ledger.Shelley.TxCert

Documentation

class EraTxCert era ⇒ ShelleyEraTxCert era where

Methods

mkRegTxCert ∷ StakeCredential → TxCert era

getRegTxCert ∷ TxCert era → Maybe StakeCredential

mkUnRegTxCert ∷ StakeCredential → TxCert era

getUnRegTxCert ∷ TxCert era → Maybe StakeCredential

mkDelegStakeTxCert ∷ StakeCredential → KeyHash 'StakePool → TxCert era

getDelegStakeTxCert ∷ TxCert era → Maybe (StakeCredential, KeyHash 'StakePool)

mkGenesisDelegTxCert ∷ AtMostEra "Babbage" era ⇒ GenesisDelegCert → TxCert era

getGenesisDelegTxCert ∷ AtMostEra "Babbage" era ⇒ TxCert era → Maybe GenesisDelegCert

mkMirTxCert ∷ AtMostEra "Babbage" era ⇒ MIRCert → TxCert era

getMirTxCert ∷ AtMostEra "Babbage" era ⇒ TxCert era → Maybe MIRCert

pattern MirTxCert ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ MIRCert → TxCert era

pattern GenesisDelegTxCert ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ KeyHash 'Genesis → KeyHash 'GenesisDelegate → VRFVerKeyHash 'GenDelegVRF → TxCert era

pattern RegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential → TxCert era

pattern UnRegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential → TxCert era

pattern DelegStakeTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential → KeyHash 'StakePool → TxCert era

data ShelleyDelegCert

Constructors

ShelleyRegCert !StakeCredential
A stake credential registration certificate.

ShelleyUnRegCert !StakeCredential
A stake credential deregistration certificate.

ShelleyDelegCert !StakeCredential !(KeyHash 'StakePool)
A stake delegation certificate.

getVKeyWitnessShelleyTxCert ∷ ShelleyTxCert era → Maybe (KeyHash 'Witness)

getScriptWitnessShelleyTxCert ∷ ShelleyTxCert era → Maybe ScriptHash

data ShelleyTxCert era

A heavyweight certificate.

Constructors

ShelleyTxCertDelegCert !ShelleyDelegCert

ShelleyTxCertPool !PoolCert

ShelleyTxCertGenesisDeleg !GenesisDelegCert

ShelleyTxCertMir !MIRCert

upgradeShelleyTxCert ∷ ShelleyTxCert era1 → ShelleyTxCert era2

GenesisDelegCert

data GenesisDelegCert

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis) !(KeyHash 'GenesisDelegate) !(VRFVerKeyHash 'GenDelegVRF)

genesisCWitness ∷ GenesisDelegCert → KeyHash 'Genesis

genesisKeyHashWitness ∷ GenesisDelegCert → KeyHash 'Witness

MIRCert

data MIRCert

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ !MIRPot

mirRewards ∷ !MIRTarget

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking) DeltaCoin)

SendToOppositePotMIR !Coin

isDelegation ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Check for ShelleyDelegCert constructor

isRegPool ∷ EraTxCert era ⇒ TxCert era → Bool

Check for RegPool constructor

isRetirePool ∷ EraTxCert era ⇒ TxCert era → Bool

Check for RetirePool constructor

isGenesisDelegation ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ TxCert era → Bool

Check for GenesisDelegate constructor

isInstantaneousRewards ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ TxCert era → Bool

isReservesMIRCert ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ TxCert era → Bool

isTreasuryMIRCert ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ TxCert era → Bool

Serialization helpers

shelleyTxCertDelegDecoder ∷ ShelleyEraTxCert era ⇒ Word → Decoder s (Int, TxCert era)

poolTxCertDecoder ∷ EraTxCert era ⇒ Word → Decoder s (Int, TxCert era)

encodeShelleyDelegCert ∷ ShelleyDelegCert → Encoding

encodePoolCert ∷ PoolCert → Encoding

encodeGenesisDelegCert ∷ GenesisDelegCert → Encoding

Deposits and Refunds

shelleyTotalDepositsTxCerts

Arguments

∷ (EraPParams era, Foldable f, EraTxCert era)

⇒ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ f (TxCert era)

→ Coin

Determine the total deposit amount needed from a TxBody. The block may (legitimately) contain multiple registration certificates for the same pool, where the first will be treated as a registration and any subsequent ones as re-registration. As such, we must only take a deposit for the first such registration. It is even possible for a single transaction to have multiple pool registration for the same pool, so as we process pool registrations, we must keep track of those that are already registered, so we do not add a Deposit for the same pool twice.

Note that this is not an issue for key registrations since subsequent registration certificates would be invalid.

shelleyTotalRefundsTxCerts

Arguments

∷ (EraPParams era, Foldable f, EraTxCert era)

⇒ PParams era

→ (StakeCredential → Maybe Coin)
Function that can lookup current deposit, in case when the stake key is registered.

→ f (TxCert era)

→ Coin

Compute the key deregistration refunds in a transaction

Re-exports

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Ord (TxCert era), Eq (TxCert era)) ⇒ EraTxCert era where

Associated Types

type TxCert era = (r ∷ Type) | r → era

type TxCertUpgradeError era

type TxCertUpgradeError era = Void

Methods

upgradeTxCert ∷ TxCert (PreviousEra era) → Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert ∷ TxCert era → Maybe (KeyHash 'Witness)

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert ∷ TxCert era → Maybe ScriptHash

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert ∷ PoolParams → TxCert era

getRegPoolTxCert ∷ TxCert era → Maybe PoolParams

mkRetirePoolTxCert ∷ KeyHash 'StakePool → EpochNo → TxCert era

getRetirePoolTxCert ∷ TxCert era → Maybe (KeyHash 'StakePool, EpochNo)

lookupRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking)

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking)

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether stake pool is registered or not

→ f (TxCert era)

→ Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (Credential 'Staking → Maybe Coin)
Lookup current deposit for Staking credential if one is registered

→ (Credential 'DRepRole → Maybe Coin)
Lookup current deposit for DRep credential if one is registered

→ f (TxCert era)

→ Coin

Compute the total refunds from a list of certificates.

pattern RegPoolTxCert ∷ EraTxCert era ⇒ PoolParams → TxCert era

pattern RetirePoolTxCert ∷ EraTxCert era ⇒ KeyHash 'StakePool → EpochNo → TxCert era

data PoolCert

Constructors

RegPool !PoolParams
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool) !EpochNo
A stake pool retirement certificate.

isRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake registering certificate

isUnRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake un-registering certificate

Orphan instances

EraTxCert ShelleyEra

Instance details

Associated Types

type TxCert ShelleyEra = (r ∷ Type)

type TxCertUpgradeError ShelleyEra

Methods

upgradeTxCert ∷ TxCert (PreviousEra ShelleyEra) → Either (TxCertUpgradeError ShelleyEra) (TxCert ShelleyEra)

getVKeyWitnessTxCert ∷ TxCert ShelleyEra → Maybe (KeyHash 'Witness)

getScriptWitnessTxCert ∷ TxCert ShelleyEra → Maybe ScriptHash

mkRegPoolTxCert ∷ PoolParams → TxCert ShelleyEra

getRegPoolTxCert ∷ TxCert ShelleyEra → Maybe PoolParams

mkRetirePoolTxCert ∷ KeyHash 'StakePool → EpochNo → TxCert ShelleyEra

getRetirePoolTxCert ∷ TxCert ShelleyEra → Maybe (KeyHash 'StakePool, EpochNo)

lookupRegStakeTxCert ∷ TxCert ShelleyEra → Maybe (Credential 'Staking)

lookupUnRegStakeTxCert ∷ TxCert ShelleyEra → Maybe (Credential 'Staking)

getTotalDepositsTxCerts ∷ Foldable f ⇒ PParams ShelleyEra → (KeyHash 'StakePool → Bool) → f (TxCert ShelleyEra) → Coin

getTotalRefundsTxCerts ∷ Foldable f ⇒ PParams ShelleyEra → (Credential 'Staking → Maybe Coin) → (Credential 'DRepRole → Maybe Coin) → f (TxCert ShelleyEra) → Coin