Cardano.Ledger.Shelley.TxWits

Documentation

keyBy ∷ Ord k ⇒ (a → k) → [a] → Map k a

decodeWits ∷ ∀ era s. EraScript era ⇒ Decoder s (Annotator (ShelleyTxWits era))

data ShelleyTxWits era where

Bundled Patterns

pattern ShelleyTxWits ∷ ∀ era. EraScript era ⇒ Set (WitVKey 'Witness (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era) → Set (BootstrapWitness (EraCrypto era)) → ShelleyTxWits era

data WitnessSetHKD f era

scriptShelleyTxWitsL ∷ EraScript era ⇒ Lens' (ShelleyTxWits era) (Map (ScriptHash (EraCrypto era)) (Script era))

Script witness setter and getter for ShelleyTxWits. The setter does update memoized binary representation.

addrShelleyTxWitsL ∷ EraScript era ⇒ Lens' (ShelleyTxWits era) (Set (WitVKey 'Witness (EraCrypto era)))

Addresses witness setter and getter for ShelleyTxWits. The setter does update memoized binary representation.

bootAddrShelleyTxWitsL ∷ EraScript era ⇒ Lens' (ShelleyTxWits era) (Set (BootstrapWitness (EraCrypto era)))

Bootstrap Addresses witness setter and getter for ShelleyTxWits. The setter does update memoized binary representation.

addrWits' ∷ WitnessSetHKD f era → HKD f (Set (WitVKey 'Witness (EraCrypto era)))

prettyWitnessSetParts ∷ ShelleyTxWits era → (Set (WitVKey 'Witness (EraCrypto era)), Map (ScriptHash (EraCrypto era)) (Script era), Set (BootstrapWitness (EraCrypto era)))

Exports the relevant parts from a (WintessSetHKD Identity era) for use by the pretty printer without all the horrible constraints. Uses the non-exported WitnessSet' constructor.

Re-exports

data WitVKey (kr ∷ KeyRole) c where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Crypto c) ⇒ VKey kr c → SignedDSIGN c (Hash c EraIndependentTxBody) → WitVKey kr c

Orphan instances

Crypto c ⇒ EraTxWits (ShelleyEra c)

Instance details

Associated Types

type TxWits (ShelleyEra c) = (r ∷ Type)

Methods

mkBasicTxWits ∷ TxWits (ShelleyEra c)

addrTxWitsL ∷ Lens' (TxWits (ShelleyEra c)) (Set (WitVKey 'Witness (EraCrypto (ShelleyEra c))))

bootAddrTxWitsL ∷ Lens' (TxWits (ShelleyEra c)) (Set (BootstrapWitness (EraCrypto (ShelleyEra c))))

scriptTxWitsL ∷ Lens' (TxWits (ShelleyEra c)) (Map (ScriptHash (EraCrypto (ShelleyEra c))) (Script (ShelleyEra c)))