Cardano.Ledger.Shelley.UTxO

Documentation

class EraTx era ⇒ EraUTxO era where

Associated Types

type ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

Methods

consumed ∷ PParams era → CertState era → UTxO era → TxBody era → Value era

getConsumedValue

Arguments

∷ PParams era

→ (Credential 'Staking → Maybe Coin)
Function that can lookup current delegation deposits

→ (Credential 'DRepRole → Maybe Coin)
Function that can lookup current drep deposits

→ UTxO era

→ TxBody era

→ Value era

Calculate all the value that is being consumed by the transaction.

getProducedValue

Arguments

∷ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

getScriptsProvided

Arguments

∷ UTxO era
For some era it is necessary to look into the UTxO to find all of the available scripts for the transaction

→ Tx era

→ ScriptsProvided era

Initial eras will look into witness set to find all of the available scripts, but starting with Babbage we can look for available scripts in the UTxO using reference inputs.

getScriptsNeeded ∷ UTxO era → TxBody era → ScriptsNeeded era

Produce all the information required for figuring out which scripts are required for the transaction to be valid, once those scripts are evaluated

getScriptsHashesNeeded ∷ ScriptsNeeded era → Set ScriptHash

Extract the set of all script hashes that are needed for script validation.

getWitsVKeyNeeded ∷ CertState era → UTxO era → TxBody era → Set (KeyHash 'Witness)

Extract all of the KeyHash witnesses that are required for validating the transaction

getMinFeeTxUtxo ∷ PParams era → Tx era → UTxO era → Coin

Minimum fee computation, excluding witnesses and including ref scripts size

newtype ShelleyScriptsNeeded era

Constructors

ShelleyScriptsNeeded (Set ScriptHash)

getShelleyScriptsNeeded ∷ EraTxBody era ⇒ UTxO era → TxBody era → ShelleyScriptsNeeded era

getConsumedCoin ∷ EraTxBody era ⇒ PParams era → (Credential 'Staking → Maybe Coin) → UTxO era → TxBody era → Coin

Compute the lovelace which are destroyed by the transaction. This implementation is suitable for Shelley and Allegra only.

shelleyProducedValue

Arguments

∷ EraTxBody era

⇒ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

shelleyConsumed ∷ (EraUTxO era, EraCertState era) ⇒ PParams era → CertState era → UTxO era → TxBody era → Value era

For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

produced ∷ (EraUTxO era, EraCertState era) ⇒ PParams era → CertState era → TxBody era → Value era

Compute the lovelace which are created by the transaction For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

getShelleyMinFeeTxUtxo ∷ EraTx era ⇒ PParams era → Tx era → Coin

getShelleyWitsVKeyNeeded ∷ ∀ era. (EraTx era, ShelleyEraTxBody era, EraCertState era) ⇒ CertState era → UTxO era → TxBody era → Set (KeyHash 'Witness)

getShelleyWitsVKeyNeededNoGov ∷ ∀ era. EraTx era ⇒ UTxO era → TxBody era → Set (KeyHash 'Witness)

Extract witnesses from UTxO and TxBody. Does not enforce witnesses for governance related Keys, i.e. GenDelegs

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map TxIn (TxOut era)

class EraTx era ⇒ EraUTxO era where

Associated Types

type ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

Methods

consumed ∷ PParams era → CertState era → UTxO era → TxBody era → Value era

getConsumedValue

Arguments

∷ PParams era

→ (Credential 'Staking → Maybe Coin)
Function that can lookup current delegation deposits

→ (Credential 'DRepRole → Maybe Coin)
Function that can lookup current drep deposits

→ UTxO era

→ TxBody era

→ Value era

Calculate all the value that is being consumed by the transaction.

getProducedValue

Arguments

∷ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

getScriptsProvided

Arguments

∷ UTxO era
For some era it is necessary to look into the UTxO to find all of the available scripts for the transaction

→ Tx era

→ ScriptsProvided era

Initial eras will look into witness set to find all of the available scripts, but starting with Babbage we can look for available scripts in the UTxO using reference inputs.

getScriptsNeeded ∷ UTxO era → TxBody era → ScriptsNeeded era

Produce all the information required for figuring out which scripts are required for the transaction to be valid, once those scripts are evaluated

getScriptsHashesNeeded ∷ ScriptsNeeded era → Set ScriptHash

Extract the set of all script hashes that are needed for script validation.

getWitsVKeyNeeded ∷ CertState era → UTxO era → TxBody era → Set (KeyHash 'Witness)

Extract all of the KeyHash witnesses that are required for validating the transaction

getMinFeeTxUtxo ∷ PParams era → Tx era → UTxO era → Coin

Minimum fee computation, excluding witnesses and including ref scripts size

type family ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

newtype ScriptsProvided era

The only reason it is a newtype instead of just a Map is becuase for later eras is expensive to compute the actual map, so we want to use the type safety guidance to avoid redundant work.

Constructors

ScriptsProvided

Fields

unScriptsProvided ∷ Map ScriptHash (Script era)

class CanGetUTxO t ⇒ CanSetUTxO (t ∷ Type → Type) where

Methods

utxoL ∷ Lens' (t era) (UTxO era)

class CanGetUTxO (t ∷ Type → Type) where

Minimal complete definition

Nothing

Methods

utxoG ∷ SimpleGetter (t era) (UTxO era)

txins ∷ EraTxBody era ⇒ TxBody era → Set TxIn

Compute the UTxO inputs of a transaction. txins has the same problems as txouts, see notes below.

txouts ∷ EraTxBody era ⇒ TxBody era → UTxO era

Compute the transaction outputs of a transaction.

txinLookup ∷ TxIn → UTxO era → Maybe (TxOut era)

Lookup a txin for a given UTxO collection

txInsFilter

Arguments

∷ UTxO era
Source UTxO

→ Set TxIn
Which of the TxIns you would like to keep.

→ UTxO era

Filter out TxIn's from the UTxO map

verifyWitVKey ∷ ∀ (kr ∷ KeyRole). Hash HASH EraIndependentTxBody → WitVKey kr → Bool

Verify a transaction body witness

sumUTxO ∷ EraTxOut era ⇒ UTxO era → Value era

Determine the total balance contained in the UTxO.

sumCoinUTxO ∷ EraTxOut era ⇒ UTxO era → Coin

Determine the total Ada only balance contained in the UTxO. This is equivalent to coin . sumUTxO, but it will be more efficient.

Warning - This function cannot be applied to an untrusted UTxO, since it is susceptible to overflow

sumAllValue ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Value era

Sum all the value in any Foldable with TxOuts

sumAllCoin ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Coin

Sum all the Coins in any Foldable with with TxOuts.

Warning - Care should be taken since it is susceptible to integer overflow, therefore make sure this function is not applied to unvalidated TxOuts

areAllAdaOnly ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Bool

Check whether any of the supplied TxOuts contain any MultiAssets. Returns True if non of them do.

getScriptHash ∷ Addr → Maybe ScriptHash

Extract script hash from value address with script.

Orphan instances

EraUTxO ShelleyEra

Instance details

Associated Types

type ScriptsNeeded ShelleyEra = (r ∷ Type)

Methods

consumed ∷ PParams ShelleyEra → CertState ShelleyEra → UTxO ShelleyEra → TxBody ShelleyEra → Value ShelleyEra

getConsumedValue ∷ PParams ShelleyEra → (Credential 'Staking → Maybe Coin) → (Credential 'DRepRole → Maybe Coin) → UTxO ShelleyEra → TxBody ShelleyEra → Value ShelleyEra

getProducedValue ∷ PParams ShelleyEra → (KeyHash 'StakePool → Bool) → TxBody ShelleyEra → Value ShelleyEra

getScriptsProvided ∷ UTxO ShelleyEra → Tx ShelleyEra → ScriptsProvided ShelleyEra

getScriptsNeeded ∷ UTxO ShelleyEra → TxBody ShelleyEra → ScriptsNeeded ShelleyEra

getScriptsHashesNeeded ∷ ScriptsNeeded ShelleyEra → Set ScriptHash

getWitsVKeyNeeded ∷ CertState ShelleyEra → UTxO ShelleyEra → TxBody ShelleyEra → Set (KeyHash 'Witness)

getMinFeeTxUtxo ∷ PParams ShelleyEra → Tx ShelleyEra → UTxO ShelleyEra → Coin