Test.Cardano.Ledger.Shelley.ImpTest

Documentation

data ImpTestM era a

data SomeSTSEvent era

Constructors

∀ (rule ∷ Symbol).(Typeable (Event (EraRule rule era)), Eq (Event (EraRule rule era)), ToExpr (Event (EraRule rule era))) ⇒ SomeSTSEvent (Event (EraRule rule era))

runImpTestM ∷ Maybe (QCGen, Int) → ImpTestState era → ImpTestM era b → IO (b, ImpTestState era)

runImpTestM_ ∷ Maybe (QCGen, Int) → ImpTestState era → ImpTestM era b → IO ()

evalImpTestM ∷ Maybe (QCGen, Int) → ImpTestState era → ImpTestM era b → IO b

execImpTestM ∷ Maybe (QCGen, Int) → ImpTestState era → ImpTestM era b → IO (ImpTestState era)

runImpTestGenM ∷ ImpTestState era → ImpTestM era b → Gen (IO (b, ImpTestState era))

runImpTestGenM_ ∷ ImpTestState era → ImpTestM era b → Gen (IO ())

evalImpTestGenM ∷ ImpTestState era → ImpTestM era b → Gen (IO b)

execImpTestGenM ∷ ImpTestState era → ImpTestM era b → Gen (IO (ImpTestState era))

data ImpTestState era

data ImpTestEnv era

Constructors

ImpTestEnv

Fields

iteState ∷ !(IORef (ImpTestState era))

iteDoTxFixup ∷ !Bool

This flag is used to toggle the fixing up of transactions. If it is set to False then any transaction should be submitted as-is.

data ImpException

Stores extra information about the failure of the unit test

Constructors

ImpException

Fields

ieAnnotation ∷ [String]

Description of the IO action that caused the failure

ieThrownException ∷ SomeException

Exception that caused the test to fail

class (Show (NewEpochState era), ToExpr (NewEpochState era), ToExpr (Tx era), ToExpr (TxOut era), ToExpr (PParams era), ToExpr (StashedAVVMAddresses era), ToExpr (GovState era), ToExpr (PParamsHKD StrictMaybe era), ToExpr (PParamsHKD Identity era), Show (TxOut era), Show (PParams era), Show (StashedAVVMAddresses era), Show (GovState era), Eq (TxOut era), Eq (PParams era), Eq (StashedAVVMAddresses era), Eq (GovState era), Eq (Event (EraRule "LEDGER" era)), Eq (Event (EraRule "TICK" era)), Signal (EraRule "LEDGER" era) ~ Tx era, BaseM (EraRule "LEDGER" era) ~ ShelleyBase, STS (EraRule "LEDGER" era), Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody), DSIGN (EraCrypto era) ~ Ed25519DSIGN, ToExpr (PredicateFailure (EraRule "LEDGER" era)), ToExpr (PredicateFailure (EraRule "UTXOW" era)), ToExpr (Event (EraRule "LEDGER" era)), ToExpr (Event (EraRule "TICK" era)), EraUTxO era, ShelleyEraTxCert era, State (EraRule "LEDGER" era) ~ LedgerState era, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, EraGov era, BaseM (EraRule "TICK" era) ~ ShelleyBase, Signal (EraRule "TICK" era) ~ SlotNo, Environment (EraRule "TICK" era) ~ (), State (EraRule "TICK" era) ~ NewEpochState era, STS (EraRule "TICK" era), NFData (PredicateFailure (EraRule "TICK" era)), NFData (StashedAVVMAddresses era), NFData (Tx era), NFData (VerKeyDSIGN (DSIGN (EraCrypto era))), NFData (PredicateFailure (EraRule "LEDGER" era)), NFData (Event (EraRule "LEDGER" era)), NFData (Event (EraRule "TICK" era)), VRFAlgorithm (VRF (EraCrypto era)), HashAlgorithm (HASH (EraCrypto era)), DSIGNAlgorithm (DSIGN (EraCrypto era)), EraRuleEvent "LEDGER" era ~ Event (EraRule "LEDGER" era), EraRuleEvent "TICK" era ~ Event (EraRule "TICK" era), Typeable (EraRuleEvent "LEDGER" era), Typeable (EraRuleEvent "TICK" era), ToExpr (EraRuleEvent "LEDGER" era), ToExpr (EraRuleEvent "TICK" era)) ⇒ ShelleyEraImp era where

Minimal complete definition

initImpNES, impSatisfyNativeScript, fixupTx

Methods

initImpNES ∷ NewEpochState era

impSatisfyNativeScript

Arguments

∷ Set (KeyHash 'Witness (EraCrypto era))
Set of Witnesses that have already been satisfied

→ NativeScript era

→ ImpTestM era (Maybe (Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era))))

Try to find a sufficient number of KeyPairs that would satisfy a native script. Whenever script can't be satisfied, Nothing is returned

modifyPParams ∷ (PParams era → PParams era) → ImpTestM era ()

This modifer should change not only the current PParams, but also the future PParams. If the future PParams are not updated, then they will overwrite the mofication of the current PParams at the next epoch.

fixupTx ∷ HasCallStack ⇒ Tx era → ImpTestM era (Tx era)

data PlutusArgs

data ScriptTestContext

initShelleyImpNES ∷ ∀ era. (Default (StashedAVVMAddresses era), ShelleyEraImp era) ⇒ NewEpochState era

impWitsVKeyNeeded ∷ EraUTxO era ⇒ TxBody era → ImpTestM era (Set (BootstrapAddress (EraCrypto era)), Set (KeyHash 'Witness (EraCrypto era)))

Figure out all the Byron Addresses that need witnesses as well as all of the KeyHashes for Shelley Key witnesses that are required.

modifyPrevPParams ∷ EraGov era ⇒ (PParams era → PParams era) → ImpTestM era ()

Modify the previous PParams in the current state with the given function. For current and future PParams, use modifyPParams

passEpoch ∷ ∀ era. ShelleyEraImp era ⇒ ImpTestM era ()

Runs the TICK rule until the next epoch is reached

passNEpochs ∷ ∀ era. ShelleyEraImp era ⇒ Natural → ImpTestM era ()

Runs the TICK rule until the n epochs are passed

passTick ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ ImpTestM era ()

Runs the TICK rule once

freshKeyHash ∷ Era era ⇒ ImpTestM era (KeyHash r (EraCrypto era))

Generates a fresh KeyHash and stores the corresponding KeyPair in the ImpTestState. If you also need the KeyPair consider using freshKeyPair for generation or lookupKeyPair to look up the KeyPair corresponding to the KeyHash

freshKeyPair ∷ (Era era, MonadState (ImpTestState era) m, MonadGen m) ⇒ m (KeyHash r (EraCrypto era), KeyPair r (EraCrypto era))

Generate a random KeyPair and add it to the known keys in the Imp state

freshKeyAddr ∷ Era era ⇒ ImpTestM era (KeyHash r (EraCrypto era), Addr (EraCrypto era))

lookupKeyPair ∷ HasCallStack ⇒ KeyHash r (EraCrypto era) → ImpTestM era (KeyPair r (EraCrypto era))

Looks up the KeyPair corresponding to the KeyHash. The KeyHash must be created with freshKeyHash for this to work.

freshByronKeyHash ∷ Era era ⇒ ImpTestM era (KeyHash r (EraCrypto era))

Generates a fresh KeyHash and stores the corresponding ByronKeyPair in the ImpTestState. If you also need the ByronKeyPair consider using freshByronKeyPair for generation or lookupByronKeyPair to look up the ByronKeyPair corresponding to the KeyHash

freshBootstapAddress ∷ ImpTestM era (BootstrapAddress (EraCrypto era))

lookupByronKeyPair ∷ HasCallStack ⇒ BootstrapAddress (EraCrypto era) → ImpTestM era ByronKeyPair

Looks up the keypair corresponding to the BootstrapAddress. The BootstrapAddress must be created with freshBootstrapAddess for this to work.

freshSafeHash ∷ Era era ⇒ ImpTestM era (SafeHash (EraCrypto era) a)

Creates a fresh SafeHash

freshKeyHashVRF ∷ Era era ⇒ ImpTestM era (Hash (HASH (EraCrypto era)) (VerKeyVRF (EraCrypto era)))

submitTx ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era (Tx era)

submitTx_ ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era ()

submitTxAnn ∷ (HasCallStack, ShelleyEraImp era) ⇒ String → Tx era → ImpTestM era (Tx era)

submitTxAnn_ ∷ (HasCallStack, ShelleyEraImp era) ⇒ String → Tx era → ImpTestM era ()

submitFailingTx ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

Submit a transaction that is expected to be rejected. The inputs and outputs are automatically balanced.

trySubmitTx ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (Tx era))

modifyNES ∷ (NewEpochState era → NewEpochState era) → ImpTestM era ()

Modify the current new epoch state with a function

getsNES ∷ SimpleGetter (NewEpochState era) a → ImpTestM era a

Get a value from the current new epoch state using the lens

getUTxO ∷ ImpTestM era (UTxO era)

impAddNativeScript ∷ ∀ era. EraScript era ⇒ NativeScript era → ImpTestM era (ScriptHash (EraCrypto era))

impAnn ∷ NFData a ⇒ String → ImpTestM era a → ImpTestM era a

Annotation for when failure happens. All the logging done within annotation will be discarded if there no failures within the annotation.

runImpRule ∷ ∀ rule era. (HasCallStack, KnownSymbol rule, STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase, NFData (State (EraRule rule era)), NFData (Event (EraRule rule era)), ToExpr (Event (EraRule rule era)), Eq (Event (EraRule rule era)), Typeable (Event (EraRule rule era))) ⇒ Environment (EraRule rule era) → State (EraRule rule era) → Signal (EraRule rule era) → ImpTestM era (State (EraRule rule era))

tryRunImpRule ∷ ∀ rule era. (STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase) ⇒ Environment (EraRule rule era) → State (EraRule rule era) → Signal (EraRule rule era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule rule era))) (State (EraRule rule era), [Event (EraRule rule era)]))

registerRewardAccount ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ ImpTestM era (RewardAccount (EraCrypto era))

lookupReward ∷ HasCallStack ⇒ Credential 'Staking (EraCrypto era) → ImpTestM era Coin

registerPool ∷ ShelleyEraImp era ⇒ ImpTestM era (KeyHash 'StakePool (EraCrypto era))

getRewardAccountAmount ∷ RewardAccount (EraCrypto era) → ImpTestM era Coin

withImpState ∷ ∀ era. ShelleyEraImp era ⇒ SpecWith (ImpTestState era) → Spec

shelleyFixupTx ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Tx era)

lookupImpRootTxOut ∷ ImpTestM era (TxIn (EraCrypto era), TxOut era)

sendValueTo ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr (EraCrypto era) → Value era → ImpTestM era (TxIn (EraCrypto era))

sendCoinTo ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr (EraCrypto era) → Coin → ImpTestM era (TxIn (EraCrypto era))

expectRegisteredRewardAddress ∷ RewardAccount (EraCrypto era) → ImpTestM era ()

expectNotRegisteredRewardAddress ∷ RewardAccount (EraCrypto era) → ImpTestM era ()

expectTreasury ∷ HasCallStack ⇒ Coin → ImpTestM era ()

updateAddrTxWits ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era (Tx era)

Adds TxWits that will satisfy all of the required key witnesses

addNativeScriptTxWits ∷ ShelleyEraImp era ⇒ Tx era → ImpTestM era (Tx era)

Modifies transaction by adding necessary scripts

addRootTxIn ∷ EraTx era ⇒ Tx era → ImpTestM era (Tx era)

This fixup step ensures that there are enough funds in the transaction.

fixupFees ∷ (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Tx era)

impGetNativeScript ∷ ScriptHash (EraCrypto era) → ImpTestM era (Maybe (NativeScript era))

impLookupUTxO ∷ ShelleyEraImp era ⇒ TxIn (EraCrypto era) → ImpTestM era (TxOut era)

Logging

logEntry ∷ HasCallStack ⇒ String → ImpTestM era ()

Adds a string to the log, which is only shown if the test fails

logToExpr ∷ (HasCallStack, ToExpr a) ⇒ a → ImpTestM era ()

logStakeDistr ∷ ImpTestM era ()

Logs the current stake distribution

logFeeMismatch ∷ (EraGov era, EraUTxO era) ⇒ Tx era → ImpTestM era ()

Combinators

withNoFixup ∷ ImpTestM era a → ImpTestM era a

Performs the action without running the fix-up function on any transactions

impNESG ∷ SimpleGetter (ImpTestState era) (NewEpochState era)

impLastTickG ∷ SimpleGetter (ImpTestState era) SlotNo

impKeyPairsG ∷ SimpleGetter (ImpTestState era) (Map (KeyHash k (EraCrypto era)) (KeyPair k (EraCrypto era)))

impNativeScriptsG ∷ SimpleGetter (ImpTestState era) (Map (ScriptHash (EraCrypto era)) (NativeScript era))