Cardano.Ledger.Shelley.API.Mempool

Description

Interface to the Shelley ledger for the purposes of managing a Shelley mempool.

Documentation

class (ChainData (Tx era), AnnotatedData (Tx era), Eq (ApplyTxError era), Show (ApplyTxError era), Typeable (ApplyTxError era), SerialisableData (ApplyTxError era), STS (EraRule "LEDGER" era), BaseM (EraRule "LEDGER" era) ~ ShelleyBase, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, State (EraRule "LEDGER" era) ~ MempoolState era, Signal (EraRule "LEDGER" era) ~ Tx era, PredicateFailure (EraRule "LEDGER" era) ~ LedgerPredicateFailure era) => ApplyTx era where

Minimal complete definition

Nothing

Methods

applyTx :: MonadError (ApplyTxError era) m => Globals -> MempoolEnv era -> MempoolState era -> Tx era -> m (MempoolState era, Validated (Tx era))

Validate a transaction against a mempool state, and return both the new mempool state and a "validated" TxInBlock.

The meaning of being "validated" depends on the era. In general, a TxInBlock has had all checks run, and can now only fail due to checks which depend on the state; most notably, that UTxO inputs disappear.

reapplyTx :: MonadError (ApplyTxError era) m => Globals -> MempoolEnv era -> MempoolState era -> Validated (Tx era) -> m (MempoolState era)

Reapply a previously validated Tx.

This applies the (validated) transaction to a new mempool state. It may fail due to the mempool state changing (for example, a needed output having already been spent). It should not fail due to any static check (such as cryptographic checks).

Implementations of this function may optionally skip the performance of any static checks. This is not required, but strongly encouraged since this function will be called each time the mempool revalidates transactions against a new mempool state.

newtype ApplyTxError era

Constructors

ApplyTxError [PredicateFailure (EraRule "LEDGER" era)]

data Validated tx

A newtype which indicates that a transaction has been validated against some chain state.

extractTx :: Validated tx -> tx

Extract the underlying unvalidated Tx.

coerceValidated :: Coercible a b => Validated a -> Validated b

translateValidated :: forall era f. TranslateEra era f => TranslationContext era -> Validated (f (PreviousEra era)) -> Except (TranslationError era f) (Validated (f era))

Translate a validated transaction across eras.

This is not a TranslateEra instance since Validated is not itself era-parametrised.

Exports for testing

type MempoolEnv era = LedgerEnv era

type MempoolState era = LedgerState era

applyTxsTransition :: forall era m. (ApplyTx era, MonadError (ApplyTxError era) m) => Globals -> MempoolEnv era -> Seq (Tx era) -> MempoolState era -> m (MempoolState era)

unsafeMakeValidated :: tx -> Validated tx

Exports for compatibility

applyTxs :: (ApplyTx era, MonadError (ApplyTxError era) m) => Globals -> SlotNo -> Seq (Tx era) -> NewEpochState era -> m (NewEpochState era)

Old applyTxs

mkMempoolEnv :: NewEpochState era -> SlotNo -> MempoolEnv era

Construct the environment used to validate transactions from the full ledger state.

Note that this function also takes a slot. During slot validation, the slot given here is the slot of the block containing the transactions. This slot is used for quite a number of things, but in general these do not determine the validity of the transaction. There are two exceptions:

Each transaction has a ttl (time-to-live) value. If the slot is beyond this value, then the transaction is invalid.

If the transaction contains a protocol update proposal, then it may only be included until a certain number of slots before the end of the epoch. A protocol update proposal submitted after this is considered invalid.

mkMempoolState :: NewEpochState era -> MempoolState era

Construct a mempool state from the wider ledger state.

The given mempool state may then be evolved using applyTxs, but should be regenerated when the ledger state gets updated (e.g. through application of a new block).

overNewEpochState :: Functor f => (MempoolState era -> f (MempoolState era)) -> NewEpochState era -> f (NewEpochState era)

Transform a function over mempool states to one over the full NewEpochState.