Cardano.Ledger.Shelley.API.Types

Documentation

hashVerKeyVRF :: (VRFAlgorithm v, HashAlgorithm h) => VerKeyVRF v -> Hash h (VerKeyVRF v)

data StrictMaybe a

Constructors

SNothing

SJust !a

bbody :: Block h era -> TxSeq era

bheader :: Block h era -> h

data Block h era

Constructors

Block' !h !(TxSeq era) ByteString

Bundled Patterns

pattern UnserialisedBlock :: h -> TxSeq era -> Block h era
Access a block without its serialised bytes. This is often useful when we're using a BHeaderView in place of the concrete header.

pattern UnsafeUnserialisedBlock :: h -> TxSeq era -> Block h era
Unsafely construct a block without the ability to serialise its bytes.

Anyone calling this pattern must ensure that the resulting block is never serialised. Any uses of this pattern outside of testing code should be regarded with suspicion.

pattern Block :: (Era era, ToCBORGroup (TxSeq era), ToCBOR h) => h -> TxSeq era -> Block h era

newtype TxId crypto

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId :: SafeHash crypto EraIndependentTxBody

data TxIn crypto

The input of a UTxO.

Constructors

TxIn !(TxId crypto) !TxIx

data Addr crypto

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential crypto) (StakeReference crypto)

AddrBootstrap (BootstrapAddress crypto)

data RewardAcnt crypto

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork :: !Network

getRwdCred :: !(Credential 'Staking crypto)

isOverlaySlot

Arguments

:: SlotNo
The first slot of the given epoch.

-> UnitInterval
The decentralization parameter.

-> SlotNo
The slot to check.

-> Bool

Determine if the given slot is reserved for the overlay schedule.

data Credential (kr :: KeyRole) crypto

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !(ScriptHash crypto)

KeyHashObj !(KeyHash kr crypto)

data StakeReference crypto

Constructors

StakeRefBase !(StakeCredential crypto)

StakeRefPtr !Ptr

StakeRefNull

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo !TxIx !CertIx

certIxFromIntegral :: Integral a => a -> Maybe CertIx

certIxToInt :: CertIx -> Int

txIxFromIntegral :: Integral a => a -> Maybe TxIx

txIxToInt :: TxIx -> Int

data ProtVer

Constructors

ProtVer

Fields

pvMajor :: !Natural

pvMinor :: !Natural

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

newtype Port

Constructors

Port

Fields

portToWord16 :: Word16

data Globals

Constructors

Globals

Fields

epochInfo :: !(EpochInfo (Either Text))

slotsPerKESPeriod :: !Word64

stabilityWindow :: !Word64

The window size in which our chosen chain growth property guarantees at least k blocks. From the paper "Ouroboros praos: An adaptively-secure, semi-synchronous proof-of-stake protocol". The stabilityWindow constant is used in a number of places; for example, protocol updates must be submitted at least twice this many slots before an epoch boundary.

randomnessStabilisationWindow :: !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

securityParameter :: !Word64

Maximum number of blocks we are allowed to roll back

maxKESEvo :: !Word64

Maximum number of KES iterations

quorum :: !Word64

Quorum for update system votes and MIR certificates

maxMajorPV :: !Natural

All blocks invalid after this protocol version

maxLovelaceSupply :: !Word64

Maximum number of lovelace in the system

activeSlotCoeff :: !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

networkId :: !Network

The network ID

systemStart :: !SystemStart

System start time

data Network

Constructors

Testnet

Mainnet

data TxIx

Transaction index.

data CertIx

Certificate index. Use certIxFromIntegral in order to construct this index safely from anything other than Word16. There is also mkCertIxPartial that can be used for testing.

individualPoolStake :: IndividualPoolStake crypto -> Rational

newtype PoolDistr crypto

A map of stake pool IDs (the hash of the stake pool operator's verification key) to IndividualPoolStake.

Constructors

PoolDistr

Fields

unPoolDistr :: Map (KeyHash 'StakePool crypto) (IndividualPoolStake crypto)

hashKey :: forall crypto (kd :: KeyRole). Crypto crypto => VKey kd crypto -> KeyHash kd crypto

Hash a given public key

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

coerceKeyRole :: forall (r :: KeyRole) crypto (r' :: KeyRole). HasKeyRole a => a r crypto -> a r' crypto

General coercion of key roles.

The presence of this function is mostly to help the user realise where they are converting key roles.

newtype VKey (kd :: KeyRole) crypto

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey :: VerKeyDSIGN (DSIGN crypto)

data KeyPair (kd :: KeyRole) crypto

Pair of signing key and verification key, with a usage role.

Constructors

KeyPair

Fields

vKey :: !(VKey kd crypto)

sKey :: !(SignKeyDSIGN (DSIGN crypto))

newtype KeyHash (discriminator :: KeyRole) crypto

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH crypto) (VerKeyDSIGN (DSIGN crypto)))

type KESignable c = Signable (KES c)

data GenDelegPair crypto

Constructors

GenDelegPair

Fields

genDelegKeyHash :: !(KeyHash 'GenesisDelegate crypto)

genDelegVrfHash :: !(Hash crypto (VerKeyVRF crypto))

newtype GenDelegs crypto

Constructors

GenDelegs

Fields

unGenDelegs :: Map (KeyHash 'Genesis crypto) (GenDelegPair crypto)

type Hash c = Hash (HASH c)

type SignedDSIGN c = SignedDSIGN (DSIGN c)

type SignKeyDSIGN c = SignKeyDSIGN (DSIGN c)

type SignedKES c = SignedKES (KES c)

type SignKeyKES c = SignKeyKES (KES c)

type VerKeyKES c = VerKeyKES (KES c)

type CertifiedVRF c = CertifiedVRF (VRF c)

type SignKeyVRF c = SignKeyVRF (VRF c)

type VerKeyVRF c = VerKeyVRF (VRF c)

newtype ScriptHash crypto

Constructors

ScriptHash (Hash (ADDRHASH crypto) EraIndependentScript)

word64ToCoin :: Word64 -> Coin

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin :: Integer

data BootstrapWitness crypto where

Bundled Patterns

pattern BootstrapWitness :: Crypto crypto => VKey 'Witness crypto -> SignedDSIGN crypto (Hash crypto EraIndependentTxBody) -> ChainCode -> ByteString -> BootstrapWitness crypto

data Metadata era where

Bundled Patterns

pattern Metadata :: Map Word64 Metadatum -> Metadata era

data Metadatum

A generic metadatum type.

Constructors

Map ![(Metadatum, Metadatum)]

List ![Metadatum]

I !Integer

B !ByteString

S !Text

newtype ProposedPPUpdates era

Update operation for protocol parameters structure @PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis (Crypto era)) (PParamsDelta era))

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

type PParams era = PParams' Identity era

data PParams' f era

Protocol parameters.

We use the HKD type family so that the protocol parameters type and the type for the updates to the protocol parameters can share records fields. The protocol parameters will have type PParams' Identity, and the updates will have type PParams' StrictMaybe, though Identity will be hidden from use.

For example:

  myParameters =
    PParams
      { _minfeeA = 0,
        _minfeeB = 0,
        ...
      }

  myUpdate =
    PParamsUpdate
      { _minfeeA = SNothing,
        _minfeeB = SJust 42,
        ...
      }

Constructors

PParams

Fields

_minfeeA :: !(HKD f Natural)

The linear factor for the minimum fee calculation

_minfeeB :: !(HKD f Natural)

The constant factor for the minimum fee calculation

_maxBBSize :: !(HKD f Natural)

Maximal block body size

_maxTxSize :: !(HKD f Natural)

Maximal transaction size

_maxBHSize :: !(HKD f Natural)

Maximal block header size

_keyDeposit :: !(HKD f Coin)

The amount of a key registration deposit

_poolDeposit :: !(HKD f Coin)

The amount of a pool registration deposit

_eMax :: !(HKD f EpochNo)

epoch bound on pool retirement

_nOpt :: !(HKD f Natural)

Desired number of pools

_a0 :: !(HKD f NonNegativeInterval)

Pool influence

_rho :: !(HKD f UnitInterval)

Monetary expansion

_tau :: !(HKD f UnitInterval)

Treasury expansion

_d :: !(HKD f UnitInterval)

Decentralization parameter

_extraEntropy :: !(HKD f Nonce)

Extra entropy

_protocolVersion :: !(HKD f ProtVer)

Protocol version

_minUTxOValue :: !(HKD f Coin)

Minimum UTxO value

_minPoolCost :: !(HKD f Coin)

Minimum Stake Pool Cost

data MultiSig crypto where

Bundled Patterns

pattern RequireMOf :: Crypto crypto => Int -> [MultiSig crypto] -> MultiSig crypto

pattern RequireSignature :: Crypto crypto => KeyHash 'Witness crypto -> MultiSig crypto

pattern RequireAllOf :: Crypto crypto => [MultiSig crypto] -> MultiSig crypto

pattern RequireAnyOf :: Crypto crypto => [MultiSig crypto] -> MultiSig crypto

computeStabilityWindow :: Word64 -> ActiveSlotCoeff -> Word64

Calculate the stability window (e.g. the number of slots needed for a block to become stable) from the security param and the active slot coefficient.

The value 3k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

computeRandomnessStabilisationWindow :: Word64 -> ActiveSlotCoeff -> Word64

Calculate the randomness stabilisation window from the security param and the active slot coefficient.

The value 4k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

newtype StakeCreds crypto

Constructors

StakeCreds

Fields

unStakeCreds :: Map (Credential 'Staking crypto) SlotNo

data WitVKey kr crypto where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey :: (Typeable kr, Crypto crypto) => VKey kr crypto -> SignedDSIGN crypto (Hash crypto EraIndependentTxBody) -> WitVKey kr crypto

newtype TxBody era

Constructors

TxBodyConstr (MemoBytes (TxBodyRaw era))

Bundled Patterns

pattern TxBody :: (Era era, FromCBOR (PParamsDelta era), TransTxBody ToCBOR era) => Set (TxIn (Crypto era)) -> StrictSeq (TxOut era) -> StrictSeq (DCert (Crypto era)) -> Wdrl (Crypto era) -> Coin -> SlotNo -> StrictMaybe (Update era) -> StrictMaybe (AuxiliaryDataHash (Crypto era)) -> TxBody era
Pattern for use by external users

data DCert crypto

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert crypto)

DCertPool !(PoolCert crypto)

DCertGenesis !(GenesisDelegCert crypto)

DCertMir !(MIRCert crypto)

data MIRCert crypto

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot :: MIRPot

mirRewards :: MIRTarget crypto

data MIRTarget crypto

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR (Map (Credential 'Staking crypto) DeltaCoin)

SendToOppositePotMIR Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data GenesisDelegCert crypto

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis crypto) !(KeyHash 'GenesisDelegate crypto) !(Hash crypto (VerKeyVRF crypto))

data PoolCert crypto

Constructors

RegPool !(PoolParams crypto)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool crypto) !EpochNo
A stake pool retirement certificate.

data DelegCert crypto

Constructors

RegKey !(StakeCredential crypto)
A stake key registration certificate.

DeRegKey !(StakeCredential crypto)
A stake key deregistration certificate.

Delegate !(Delegation crypto)
A stake delegation certificate.

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !(CompactAddr (Crypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern TxOut :: (Era era, Show (Value era), Compactible (Value era)) => Addr (Crypto era) -> Value era -> TxOut era

newtype Wdrl crypto

Constructors

Wdrl

Fields

unWdrl :: Map (RewardAcnt crypto) Coin

data PoolParams crypto

A stake pool.

Constructors

PoolParams

Fields

_poolId :: !(KeyHash 'StakePool crypto)

_poolVrf :: !(Hash crypto (VerKeyVRF crypto))

_poolPledge :: !Coin

_poolCost :: !Coin

_poolMargin :: !UnitInterval

_poolRAcnt :: !(RewardAcnt crypto)

_poolOwners :: !(Set (KeyHash 'Staking crypto))

_poolRelays :: !(StrictSeq StakePoolRelay)

_poolMD :: !(StrictMaybe PoolMetadata)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data PoolMetadata

Constructors

PoolMetadata

Fields

_poolMDUrl :: !Url

_poolMDHash :: !ByteString

data Delegation crypto

The delegation of one stake key to another.

Constructors

Delegation

Fields

_delegator :: !(StakeCredential crypto)

_delegatee :: !(KeyHash 'StakePool crypto)

type WitnessSet = WitnessSetHKD Identity

data Tx era where

Bundled Patterns

pattern Tx :: (Era era, ToCBOR (AuxiliaryData era), ToCBOR (TxBody era), ToCBOR (Witnesses era)) => TxBody era -> Witnesses era -> StrictMaybe (AuxiliaryData era) -> Tx era

bbHash :: forall era. Era era => TxSeq era -> Hash (Crypto era) EraIndependentBlockBody

Hash a given block body

data SnapShots crypto

Snapshots of the stake distribution.

Constructors

SnapShots

Fields

_pstakeMark :: SnapShot crypto

_pstakeSet :: !(SnapShot crypto)

_pstakeGo :: !(SnapShot crypto)

_feeSS :: !Coin

data SnapShot crypto

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

_stake :: !(Stake crypto)

_delegations :: !(VMap VB VB (Credential 'Staking crypto) (KeyHash 'StakePool crypto))

_poolParams :: !(VMap VB VB (KeyHash 'StakePool crypto) (PoolParams crypto))

newtype Stake crypto

Type of stake as map from hash key to coins associated.

Constructors

Stake

Fields

unStake :: VMap VB VP (Credential 'Staking crypto) (CompactForm Coin)

data NonMyopic crypto

data RewardUpdate crypto

The ultiate goal of a reward update computation. Aggregating rewards for each staking credential.

Constructors

RewardUpdate

Fields

deltaT :: !DeltaCoin

deltaR :: !DeltaCoin

rs :: !(Map (Credential 'Staking crypto) (Set (Reward crypto)))

deltaF :: !DeltaCoin

nonMyopic :: !(NonMyopic crypto)

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO :: Map (TxIn (Crypto era)) (TxOut era)

balance :: forall era. Era era => UTxO era -> Value era

Determine the total balance contained in the UTxO.

newtype WitHashes crypto

Constructors

WitHashes

Fields

unWitHashes :: Set (KeyHash 'Witness crypto)

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState :: !(UTxOState era)

The current unspent transaction outputs.

lsDPState :: !(DPState (Crypto era))

The current delegation state

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL :: !EpochNo

Last epoch

nesBprev :: !(BlocksMade (Crypto era))

Blocks made before current epoch

nesBcur :: !(BlocksMade (Crypto era))

Blocks made in current epoch

nesEs :: !(EpochState era)

Epoch state before current

nesRu :: !(StrictMaybe (PulsingRewUpdate (Crypto era)))

Possible reward update

nesPd :: !(PoolDistr (Crypto era))

Stake distribution within the stake pool

stashedAVVMAddresses :: !(StashedAVVMAddresses era)

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that istake == (updateStakeDistribution (UTxO Map.empty) (UTxO Map.empty) utxo) Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very carefull, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

_utxo :: !(UTxO era)

_deposited :: !Coin

_fees :: !Coin

_ppups :: !(State (EraRule "PPUP" era))

_stakeDistro :: !(IncrementalStake (Crypto era))

data IncrementalStake crypto

Incremental Stake, Stake along with possible missed coins from danging Ptrs. Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr does not have to point to anything until the epoch boundary, when we compute rewards and aggregate staking information for ranking. This is unusual but legal. In a non incremental system, we use whatever legal Ptrs exist at the epoch boundary. Here we are computing things incrementally, so we need to remember Ptrs that might point to something by the time the epoch boundary is reached. When the epoch boundary is reached we resolve these pointers, to see if any have become non-dangling since the time they were first used in the incremental computation.

Constructors

IStake

Fields

credMap :: !(Map (Credential 'Staking crypto) Coin)

ptrMap :: !(Map Ptr Coin)

data PPUPState era

Constructors

PPUPState

Fields

proposals :: !(ProposedPPUpdates era)

futureProposals :: !(ProposedPPUpdates era)

data EpochState era

Constructors

EpochState

Fields

esAccountState :: !AccountState

esSnapshots :: !(SnapShots (Crypto era))

esLState :: !(LedgerState era)

esPrevPp :: !(PParams era)

esPp :: !(PParams era)

esNonMyopic :: !(NonMyopic (Crypto era))

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://hydra.iohk.io/job/Cardano/cardano-ledger/specs.pool-ranking/latest/download-by-type/doc-pdf/pool-ranking

data AccountState

Constructors

AccountState

Fields

_treasury :: !Coin

_reserves :: !Coin

data DPState crypto

The state associated with the current stake delegation.

Constructors

DPState

Fields

dpsDState :: !(DState crypto)

dpsPState :: !(PState crypto)

data PState crypto

Current state of staking pools and their certificate counters.

Constructors

PState

Fields

_pParams :: !(Map (KeyHash 'StakePool crypto) (PoolParams crypto))

The pool parameters.

_fPParams :: !(Map (KeyHash 'StakePool crypto) (PoolParams crypto))

The future pool parameters.

_retiring :: !(Map (KeyHash 'StakePool crypto) EpochNo)

A map of retiring stake pools to the epoch when they retire.

data DState crypto

State of staking pool delegations and rewards

Constructors

DState

Fields

_unified :: !(UnifiedMap crypto)

Unified Reward Maps

_fGenDelegs :: !(Map (FutureGenDeleg crypto) (GenDelegPair crypto))

Future genesis key delegations

_genDelegs :: !(GenDelegs crypto)

Genesis key delegations

_irwd :: !(InstantaneousRewards crypto)

Instantaneous Rewards

data InstantaneousRewards crypto

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves :: !(Map (Credential 'Staking crypto) Coin)

iRTreasury :: !(Map (Credential 'Staking crypto) Coin)

deltaReserves :: !DeltaCoin

deltaTreasury :: !DeltaCoin

type KeyPairs crypto = [(KeyPair 'Payment crypto, KeyPair 'Staking crypto)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

data PPUPEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs (Crypto era))

data PPUP era

data UtxoEnv era

Constructors

UtxoEnv SlotNo (PParams era) (Map (KeyHash 'StakePool (Crypto era)) (PoolParams (Crypto era))) (GenDelegs (Crypto era))

data UTXO era

data UTXOW era

data POOLREAP era

data PoolEnv era

Constructors

PoolEnv SlotNo (PParams era)

data POOL (era :: Type)

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo :: SlotNo

ptr_ :: Ptr

acnt_ :: AccountState

ppDE :: PParams era

data DELEG era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo :: SlotNo

delPlPtr :: Ptr

delPlPp :: PParams era

delPlAcnt :: AccountState

data DELPL era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo :: !SlotNo

delegsIx :: !TxIx

delegspp :: !(PParams era)

delegsTx :: !(Tx era)

delegsAccount :: !AccountState

data DELEGS era

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo :: !SlotNo

ledgerIx :: !TxIx

ledgerPp :: !(PParams era)

ledgerAccount :: !AccountState

data LEDGER era

data LedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo :: SlotNo

ledgersPp :: PParams era

ledgersAccount :: AccountState

data LEDGERS era

data ValidationErr

Constructors

EpochNotLongEnough EpochSize Word64 Rational EpochSize

MaxKESEvolutionsUnsupported Word64 Word

QuorumTooSmall Word64 Word64 Word64

data ShelleyGenesis era

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTime

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams era)

sgGenDelegs :: !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds :: !(Map (Addr (Crypto era)) Coin)

sgStaking :: !(ShelleyGenesisStaking (Crypto era))

data ShelleyGenesisStaking crypto

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools :: !(Map (KeyHash 'StakePool crypto) (PoolParams crypto))

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake :: !(Map (KeyHash 'Staking crypto) (KeyHash 'StakePool crypto))

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

emptyGenesisStaking :: ShelleyGenesisStaking crypto

Empty genesis staking

sgActiveSlotCoeff :: ShelleyGenesis era -> ActiveSlotCoeff

genesisUTxO :: forall era. (Era era, UsesTxOut era) => ShelleyGenesis era -> UTxO era

initialFundsPseudoTxIn :: forall crypto. Crypto crypto => Addr crypto -> TxIn crypto

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

describeValidationErr :: ValidationErr -> Text

validateGenesis :: forall era. Era era => ShelleyGenesis era -> Either [ValidationErr] ()

Do some basic sanity checking on the Shelley genesis file.

mkShelleyGlobals :: ShelleyGenesis era -> EpochInfo (Either Text) -> Natural -> Globals

data NEWEPOCH era

calculatePoolDistr :: SnapShot crypto -> PoolDistr crypto

data TICK era