Cardano.Ledger.Shelley.API.Types

Documentation

data Block h era

Constructors

Block' !h !(TxSeq era) ByteString

Bundled Patterns

pattern Block :: (Era era, EncCBORGroup (TxSeq era), EncCBOR h) => h -> TxSeq era -> Block h era

pattern UnsafeUnserialisedBlock :: h -> TxSeq era -> Block h era

pattern UnserialisedBlock :: h -> TxSeq era -> Block h era

data Ptr

Constructors

Ptr !SlotNo !TxIx !CertIx

data Addr c

Constructors

Addr Network (PaymentCredential c) (StakeReference c)

AddrBootstrap (BootstrapAddress c)

type Metadata era = ShelleyTxAuxData era

Deprecated: Use ShelleyTxAuxData instead

data ProtVer

Constructors

ProtVer

Fields

pvMajor :: !Version

pvMinor :: !Natural

data ShelleyTx era where

Bundled Patterns

pattern ShelleyTx :: EraTx era => TxBody era -> TxWits era -> StrictMaybe (TxAuxData era) -> ShelleyTx era

data ShelleyTxOut era

Constructors

TxOutCompact !(CompactAddr (EraCrypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern ShelleyTxOut :: (HasCallStack, Era era, Val (Value era)) => Addr (EraCrypto era) -> Value era -> ShelleyTxOut era

newtype ShelleyTxBody era

Constructors

TxBodyConstr (MemoBytes ShelleyTxBodyRaw era)

Bundled Patterns

pattern ShelleyTxBody :: (EraTxOut era, EncCBOR (TxCert era)) => Set (TxIn (EraCrypto era)) -> StrictSeq (TxOut era) -> StrictSeq (TxCert era) -> Withdrawals (EraCrypto era) -> Coin -> SlotNo -> StrictMaybe (Update era) -> StrictMaybe (AuxiliaryDataHash (EraCrypto era)) -> ShelleyTxBody era
Pattern for use by external users

data ShelleyTxAuxData era where

Bundled Patterns

pattern ShelleyTxAuxData :: forall era. Era era => Map Word64 Metadatum -> ShelleyTxAuxData era

newtype UTxO era

Constructors

UTxO

Fields

unUTxO :: Map (TxIn (EraCrypto era)) (TxOut era)

type Hash c = Hash (HASH c)

data TxIx

newtype Coin

Constructors

Coin

Fields

unCoin :: Integer

data AccountState

Constructors

AccountState

Fields

asTreasury :: !Coin

asReserves :: !Coin

data CertState era

Constructors

CertState

Fields

certVState :: !(VState era)

certPState :: !(PState era)

certDState :: !(DState era)

data DState era

Constructors

DState

Fields

dsUnified :: !(UMap (EraCrypto era))

dsFutureGenDelegs :: !(Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

dsGenDelegs :: !(GenDelegs (EraCrypto era))

dsIRewards :: !(InstantaneousRewards (EraCrypto era))

data EpochState era

Constructors

EpochState

Fields

esAccountState :: !AccountState

esLState :: !(LedgerState era)

esSnapshots :: !(SnapShots (EraCrypto era))

esNonMyopic :: !(NonMyopic (EraCrypto era))

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://github.com/input-output-hk/cardano-ledger/releases/latest/download/pool-ranking.pdf

newtype GenDelegs c

Constructors

GenDelegs

Fields

unGenDelegs :: Map (KeyHash 'Genesis c) (GenDelegPair c)

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState :: !(UTxOState era)

The current unspent transaction outputs.

lsCertState :: !(CertState era)

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL :: !EpochNo

Last epoch

nesBprev :: !(BlocksMade (EraCrypto era))

Blocks made before current epoch

nesBcur :: !(BlocksMade (EraCrypto era))

Blocks made in current epoch

nesEs :: !(EpochState era)

Epoch state before current

nesRu :: !(StrictMaybe (PulsingRewUpdate (EraCrypto era)))

Possible reward update

nesPd :: !(PoolDistr (EraCrypto era))

Stake distribution within the stake pool

stashedAVVMAddresses :: !(StashedAVVMAddresses era)

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

newtype PoolDistr c

Constructors

PoolDistr

Fields

unPoolDistr :: Map (KeyHash 'StakePool c) (IndividualPoolStake c)

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams (ShelleyEra c))

sgGenDelegs :: !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds :: ListMap (Addr c) Coin

sgInitialFunds is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

sgStaking :: ShelleyGenesisStaking c

sgStaking is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

data StrictMaybe a

Constructors

SNothing

SJust !a

data Globals

Constructors

Globals

Fields

epochInfo :: !(EpochInfo (Either Text))

slotsPerKESPeriod :: !Word64

stabilityWindow :: !Word64

randomnessStabilisationWindow :: !Word64

securityParameter :: !Word64

maxKESEvo :: !Word64

quorum :: !Word64

maxMajorPV :: !Version

maxLovelaceSupply :: !Word64

activeSlotCoeff :: !ActiveSlotCoeff

networkId :: !Network

systemStart :: !SystemStart

data Credential (kr :: KeyRole) c

Constructors

ScriptHashObj !(ScriptHash c)

KeyHashObj !(KeyHash kr c)

newtype KeyHash (discriminator :: KeyRole) c

Constructors

KeyHash (Hash (ADDRHASH c) (VerKeyDSIGN (DSIGN c)))

data KeyRole

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

data PState era

Constructors

PState

Fields

psStakePoolParams :: !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

psFutureStakePoolParams :: !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

psRetiring :: !(Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

psDeposits :: !(Map (KeyHash 'StakePool (EraCrypto era)) Coin)

data RewardUpdate c

The ultiate goal of a reward update computation. Aggregating rewards for each staking credential.

Constructors

RewardUpdate

Fields

deltaT :: !DeltaCoin

deltaR :: !DeltaCoin

rs :: !(Map (Credential 'Staking c) (Set (Reward c)))

deltaF :: !DeltaCoin

nonMyopic :: !(NonMyopic c)

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that istake == (updateStakeDistribution (UTxO Map.empty) (UTxO Map.empty) utxo) Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very careful, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

utxosUtxo :: !(UTxO era)

utxosDeposited :: Coin

This field is left lazy, because we only use it for assertions

utxosFees :: !Coin

utxosGovState :: !(GovState era)

utxosStakeDistr :: !(IncrementalStake (EraCrypto era))

utxosDonation :: !Coin

data NonMyopic c

data PoolParams c

Constructors

PoolParams

Fields

ppId :: !(KeyHash 'StakePool c)

ppVrf :: !(Hash c (VerKeyVRF c))

ppPledge :: !Coin

ppCost :: !Coin

ppMargin :: !UnitInterval

ppRewardAcnt :: !(RewardAcnt c)

ppOwners :: !(Set (KeyHash 'Staking c))

ppRelays :: !(StrictSeq StakePoolRelay)

ppMetadata :: !(StrictMaybe PoolMetadata)

data WitVKey (kr :: KeyRole) c where

Bundled Patterns

pattern WitVKey :: (Typeable kr, Crypto c) => VKey kr c -> SignedDSIGN c (Hash c EraIndependentTxBody) -> WitVKey kr c

data TxIn c

Constructors

TxIn !(TxId c) !TxIx

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo :: !SlotNo

ledgerIx :: !TxIx

ledgerPp :: !(PParams era)

ledgerAccount :: !AccountState

data RewardAcnt c

Constructors

RewardAcnt

Fields

getRwdNetwork :: !Network

getRwdCred :: !(Credential 'Staking c)

data CertIx

data Network

Constructors

Testnet

Mainnet

data Nonce

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce

newtype Port

Constructors

Port

Fields

portToWord16 :: Word16

data StakeReference c

Constructors

StakeRefBase !(StakeCredential c)

StakeRefPtr !Ptr

StakeRefNull

data SnapShot c

Constructors

SnapShot

Fields

ssStake :: !(Stake c)

ssDelegations :: !(VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c))

ssPoolParams :: !(VMap VB VB (KeyHash 'StakePool c) (PoolParams c))

data SnapShots c

Constructors

SnapShots

Fields

ssStakeMark :: SnapShot c

ssStakeMarkPoolDistr :: PoolDistr c

ssStakeSet :: !(SnapShot c)

ssStakeGo :: !(SnapShot c)

ssFee :: !Coin

newtype Stake c

Constructors

Stake

Fields

unStake :: VMap VB VP (Credential 'Staking c) (CompactForm Coin)

type CertifiedVRF c = CertifiedVRF (VRF c)

data GenDelegPair c

Constructors

GenDelegPair

Fields

genDelegKeyHash :: !(KeyHash 'GenesisDelegate c)

genDelegVrfHash :: !(Hash c (VerKeyVRF c))

type KESignable c = Signable (KES c)

data KeyPair (kd :: KeyRole) c

Constructors

KeyPair

Fields

vKey :: !(VKey kd c)

sKey :: !(SignKeyDSIGN (DSIGN c))

type SignKeyDSIGN c = SignKeyDSIGN (DSIGN c)

type SignKeyKES c = SignKeyKES (KES c)

type SignKeyVRF c = SignKeyVRF (VRF c)

type SignedDSIGN c = SignedDSIGN (DSIGN c)

type SignedKES c = SignedKES (KES c)

newtype VKey (kd :: KeyRole) c

Constructors

VKey

Fields

unVKey :: VerKeyDSIGN (DSIGN c)

type VerKeyKES c = VerKeyKES (KES c)

type VerKeyVRF c = VerKeyVRF (VRF c)

data BootstrapWitness c where

Bundled Patterns

pattern BootstrapWitness :: Crypto c => VKey 'Witness c -> SignedDSIGN c (Hash c EraIndependentTxBody) -> ChainCode -> ByteString -> BootstrapWitness c

data IncrementalStake c

Incremental Stake, Stake along with possible missed coins from danging Ptrs. Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr does not have to point to anything until the epoch boundary, when we compute rewards and aggregate staking information for ranking. This is unusual but legal. In a non incremental system, we use whatever legal Ptrs exist at the epoch boundary. Here we are computing things incrementally, so we need to remember Ptrs that might point to something by the time the epoch boundary is reached. When the epoch boundary is reached we resolve these pointers, to see if any have become non-dangling since the time they were first used in the incremental computation.

Constructors

IStake

Fields

credMap :: !(Map (Credential 'Staking c) (CompactForm Coin))

ptrMap :: !(Map Ptr (CompactForm Coin))

data InstantaneousRewards c

Constructors

InstantaneousRewards

Fields

iRReserves :: !(Map (Credential 'Staking c) Coin)

iRTreasury :: !(Map (Credential 'Staking c) Coin)

deltaReserves :: !DeltaCoin

deltaTreasury :: !DeltaCoin

type KeyPairs c = [(KeyPair 'Payment c, KeyPair 'Staking c)]

Deprecated: Use `Test.Cardano.Ledger.Core.KeyPair (KeyPairs)` instead

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

proposals :: !(ProposedPPUpdates era)

futureProposals :: !(ProposedPPUpdates era)

sgovPp :: !(PParams era)

sgovPrevPp :: !(PParams era)

data VState era

Constructors

VState

Fields

vsDReps :: !(Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

vsCommitteeState :: !(CommitteeState era)

vsNumDormantEpochs :: EpochNo

newtype ProposedPPUpdates era

Update operation for protocol parameters structure PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdate era))

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo :: SlotNo

ptr_ :: Ptr

acnt_ :: AccountState

ppDE :: PParams era

data ShelleyDELEG era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo :: !SlotNo

delegsIx :: !TxIx

delegspp :: !(PParams era)

delegsTx :: !(Tx era)

delegsAccount :: !AccountState

data ShelleyDELEGS era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo :: SlotNo

delPlPtr :: Ptr

delPlPp :: PParams era

delPlAcnt :: AccountState

data ShelleyDELPL era

data ShelleyLEDGER era

data ShelleyLEDGERS era

data ShelleyLedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo :: SlotNo

ledgersPp :: PParams era

ledgersAccount :: AccountState

data ShelleyNEWEPOCH era

data PoolEnv era

Constructors

PoolEnv !SlotNo !(PParams era)

data ShelleyPOOL era

data ShelleyPOOLREAP era

data PpupEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs (EraCrypto era))

data ShelleyPPUP era

data SnapEnv era

Constructors

SnapEnv !(LedgerState era) !(PParams era)

data ShelleyTICK era

data ShelleyTICKF era

data ShelleyUTXO era

data UtxoEnv era

Constructors

UtxoEnv

Fields

ueSlot :: SlotNo

uePParams :: PParams era

ueCertState :: CertState era

ueGenDelegs :: GenDelegs (EraCrypto era)

data ShelleyUTXOW era

data MultiSig era where

Bundled Patterns

pattern RequireAllOf :: Era era => [MultiSig era] -> MultiSig era

pattern RequireAnyOf :: Era era => [MultiSig era] -> MultiSig era

pattern RequireSignature :: Era era => KeyHash 'Witness (EraCrypto era) -> MultiSig era

pattern RequireMOf :: Era era => Int -> [MultiSig era] -> MultiSig era

newtype ScriptHash c

Constructors

ScriptHash (Hash (ADDRHASH c) EraIndependentScript)

data Metadatum

A generic metadatum type.

Constructors

Map ![(Metadatum, Metadatum)]

List ![Metadatum]

I !Integer

B !ByteString

S !Text

data Delegation c

Constructors

Delegation

Fields

dDelegator :: !(StakeCredential c)

dDelegatee :: !(KeyHash 'StakePool c)

data GenesisDelegCert c

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot :: !MIRPot

mirRewards :: !(MIRTarget c)

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl :: !Url

pmHash :: !ByteString

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)

SingleHostName !(StrictMaybe Port) !DnsName

MultiHostName !DnsName

newtype Withdrawals c

Constructors

Withdrawals

Fields

unWithdrawals :: Map (RewardAcnt c) Coin

data PoolCert c

Constructors

RegPool !(PoolParams c)

RetirePool !(KeyHash 'StakePool c) !EpochNo

data ShelleyDelegCert c

Constructors

ShelleyRegCert !(StakeCredential c)
A stake credential registration certificate.

ShelleyUnRegCert !(StakeCredential c)
A stake credential deregistration certificate.

ShelleyDelegCert !(StakeCredential c) !(KeyHash 'StakePool c)
A stake delegation certificate.

Bundled Patterns

pattern RegKey :: StakeCredential c -> ShelleyDelegCert c

Deprecated: In favor of ShelleyRegCert

pattern DeRegKey :: StakeCredential c -> ShelleyDelegCert c

Deprecated: In favor of ShelleyUnRegCert

pattern Delegate :: Delegation c -> ShelleyDelegCert c

Deprecated: In favor of ShelleyDelegCert

data ShelleyTxWits era

newtype TxId c

Constructors

TxId

Fields

unTxId :: SafeHash c EraIndependentTxBody

data ShelleyGenesisStaking c

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools :: ListMap (KeyHash 'StakePool c) (PoolParams c)

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake :: ListMap (KeyHash 'Staking c) (KeyHash 'StakePool c)

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

data ValidationErr

Constructors

EpochNotLongEnough EpochSize Word64 Rational EpochSize

MaxKESEvolutionsUnsupported Word64 Word

QuorumTooSmall Word64 Word64 Word64

newtype NominalDiffTimeMicro

Unlike NominalDiffTime that supports Pico precision, this type only supports Micro precision.

Constructors

NominalDiffTimeMicro Micro

balance :: EraTxOut era => UTxO era -> Value era

genesisUTxO :: forall era. EraTxOut era => ShelleyGenesis (EraCrypto era) -> UTxO era

word64ToCoin :: Word64 -> Coin

calculatePoolDistr :: SnapShot c -> PoolDistr c

isOverlaySlot :: SlotNo -> UnitInterval -> SlotNo -> Bool

certIxFromIntegral :: Integral a => a -> Maybe CertIx

certIxToInt :: CertIx -> Int

txIxFromIntegral :: Integral a => a -> Maybe TxIx

txIxToInt :: TxIx -> Int

bbody :: Block h era -> TxSeq era

bheader :: Block h era -> h

coerceKeyRole :: forall (r :: KeyRole) c (r' :: KeyRole). HasKeyRole a => a r c -> a r' c

hashKey :: forall c (kd :: KeyRole). Crypto c => VKey kd c -> KeyHash kd c

hashVerKeyVRF :: (VRFAlgorithm v, HashAlgorithm h) => VerKeyVRF v -> Hash h (VerKeyVRF v)

individualPoolStake :: IndividualPoolStake c -> Rational

bbHash :: forall era. Era era => ShelleyTxSeq era -> Hash (EraCrypto era) EraIndependentBlockBody

Hash a given block body

calculatePoolDistr' :: (KeyHash 'StakePool c -> Bool) -> SnapShot c -> PoolDistr c

computeRandomnessStabilisationWindow :: Word64 -> ActiveSlotCoeff -> Word64

Calculate the randomness stabilisation window from the security param and the active slot coefficient.

The value 4k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

computeStabilityWindow :: Word64 -> ActiveSlotCoeff -> Word64

Calculate the stability window (e.g. the number of slots needed for a block to become stable) from the security param and the active slot coefficient.

The value 3k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

toShelleyGenesisPairs :: (KeyValue a, Crypto c) => ShelleyGenesis c -> [a]

emptyGenesisStaking :: ShelleyGenesisStaking c

Empty genesis staking

sgActiveSlotCoeff :: ShelleyGenesis c -> ActiveSlotCoeff

initialFundsPseudoTxIn :: forall c. Crypto c => Addr c -> TxIn c

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

validateGenesis :: forall c. Crypto c => ShelleyGenesis c -> Either [ValidationErr] ()

Do some basic sanity checking on the Shelley genesis file.

describeValidationErr :: ValidationErr -> Text

mkShelleyGlobals :: ShelleyGenesis c -> EpochInfo (Either Text) -> Version -> Globals

nominalDiffTimeMicroToMicroseconds :: NominalDiffTimeMicro -> Micro

nominalDiffTimeMicroToSeconds :: NominalDiffTimeMicro -> Pico

toNominalDiffTimeMicro :: NominalDiffTime -> Maybe NominalDiffTimeMicro

toNominalDiffTimeMicroWithRounding :: NominalDiffTime -> NominalDiffTimeMicro

fromNominalDiffTimeMicro :: NominalDiffTimeMicro -> NominalDiffTime

secondsToNominalDiffTimeMicro :: Micro -> NominalDiffTimeMicro