Cardano.Ledger.Shelley.API.Types

Documentation

hashVerKeyVRF ∷ (VRFAlgorithm v, HashAlgorithm h) ⇒ VerKeyVRF v → Hash h (VerKeyVRF v)

data StrictMaybe a

Constructors

SNothing

SJust !a

balance ∷ EraTxOut era ⇒ UTxO era → Value era

Determine the total balance contained in the UTxO.

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map (TxIn (EraCrypto era)) (TxOut era)

bbody ∷ Block h era → TxSeq era

bheader ∷ Block h era → h

data Block h era

Constructors

Block' !h !(TxSeq era) ByteString

Bundled Patterns

pattern UnserialisedBlock ∷ h → TxSeq era → Block h era
Access a block without its serialised bytes. This is often useful when we're using a BHeaderView in place of the concrete header.

pattern UnsafeUnserialisedBlock ∷ h → TxSeq era → Block h era
Unsafely construct a block without the ability to serialise its bytes.

Anyone calling this pattern must ensure that the resulting block is never serialised. Any uses of this pattern outside of testing code should be regarded with suspicion.

pattern Block ∷ (Era era, EncCBORGroup (TxSeq era), EncCBOR h) ⇒ h → TxSeq era → Block h era

data InstantaneousRewards c

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking c) Coin)

iRTreasury ∷ !(Map (Credential 'Staking c) Coin)

deltaReserves ∷ !DeltaCoin

deltaTreasury ∷ !DeltaCoin

data DState era

The state used by the DELEG rule, which roughly tracks stake delegation and some governance features.

Constructors

DState

Fields

dsUnified ∷ !(UMap (EraCrypto era))

Unified Reward Maps. This contains the reward map (which is the source of truth regarding the registered stake credentials, the deposit map, the delegation map, and the stake credential pointer map.

dsFutureGenDelegs ∷ !(Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

Future genesis key delegations

dsGenDelegs ∷ !(GenDelegs (EraCrypto era))

Genesis key delegations

dsIRewards ∷ !(InstantaneousRewards (EraCrypto era))

Instantaneous Rewards

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The stake pool parameters.

psFutureStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The future stake pool parameters. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

A map of retiring stake pools to the epoch when they retire.

psDeposits ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) Coin)

A map of the deposits for each pool

data VState era

Constructors

VState

Fields

vsDReps ∷ !(Set (Credential 'Voting (EraCrypto era)))

vsCCHotKeys ∷ !(Map (KeyHash 'Voting (EraCrypto era)) (KeyHash 'Voting (EraCrypto era)))

data CertState era

The state associated with the DELPL rule, which combines the DELEG rule and the POOL rule.

Constructors

CertState

Fields

certVState ∷ !(VState era)

certPState ∷ !(PState era)

certDState ∷ !(DState era)

calculatePoolDistr' ∷ (KeyHash 'StakePool c → Bool) → SnapShot c → PoolDistr c

calculatePoolDistr ∷ SnapShot c → PoolDistr c

newtype Stake c

Type of stake as map from hash key to coins associated.

Constructors

Stake

Fields

unStake ∷ VMap VB VP (Credential 'Staking c) (CompactForm Coin)

data SnapShot c

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

ssStake ∷ !(Stake c)

ssDelegations ∷ !(VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c))

ssPoolParams ∷ !(VMap VB VB (KeyHash 'StakePool c) (PoolParams c))

data SnapShots c

Snapshots of the stake distribution.

Note that ssStakeMark and ssStakeMarkPoolDistr are lazy on purpose since we only want to force the thunk after one stability window when we know that they are stable (so that we do not compute them if we do not have to). See more info in the Optimize TICKF ADR

Constructors

SnapShots

Fields

ssStakeMark ∷ SnapShot c

ssStakeMarkPoolDistr ∷ PoolDistr c

ssStakeSet ∷ !(SnapShot c)

ssStakeGo ∷ !(SnapShot c)

ssFee ∷ !Coin

newtype TxId c

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

unTxId ∷ SafeHash c EraIndependentTxBody

data TxIn c

The input of a UTxO.

Constructors

TxIn !(TxId c) !TxIx

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl ∷ !Url

pmHash ∷ !ByteString

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data PoolParams c

A stake pool.

Constructors

PoolParams

Fields

ppId ∷ !(KeyHash 'StakePool c)

ppVrf ∷ !(Hash c (VerKeyVRF c))

ppPledge ∷ !Coin

ppCost ∷ !Coin

ppMargin ∷ !UnitInterval

ppRewardAcnt ∷ !(RewardAcnt c)

ppOwners ∷ !(Set (KeyHash 'Staking c))

ppRelays ∷ !(StrictSeq StakePoolRelay)

ppMetadata ∷ !(StrictMaybe PoolMetadata)

data Addr c

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential c) (StakeReference c)

AddrBootstrap (BootstrapAddress c)

data RewardAcnt c

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork ∷ !Network

getRwdCred ∷ !(Credential 'Staking c)

newtype Withdrawals c

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map (RewardAcnt c) Coin

isOverlaySlot

Arguments

∷ SlotNo
The first slot of the given epoch.

→ UnitInterval
The decentralization parameter.

→ SlotNo
The slot to check.

→ Bool

Determine if the given slot is reserved for the overlay schedule.

data Credential (kr ∷ KeyRole) c

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !(ScriptHash c)

KeyHashObj !(KeyHash kr c)

data StakeReference c

Constructors

StakeRefBase !(StakeCredential c)

StakeRefPtr !Ptr

StakeRefNull

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo !TxIx !CertIx

data BootstrapWitness c where

Bundled Patterns

pattern BootstrapWitness ∷ Crypto c ⇒ VKey 'Witness c → SignedDSIGN c (Hash c EraIndependentTxBody) → ChainCode → ByteString → BootstrapWitness c

data WitVKey (kr ∷ KeyRole) c where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Crypto c) ⇒ VKey kr c → SignedDSIGN c (Hash c EraIndependentTxBody) → WitVKey kr c

newtype ScriptHash c

Constructors

ScriptHash (Hash (ADDRHASH c) EraIndependentScript)

individualPoolStake ∷ IndividualPoolStake c → Rational

newtype PoolDistr c

A map of stake pool IDs (the hash of the stake pool operator's verification key) to IndividualPoolStake.

Constructors

PoolDistr

Fields

unPoolDistr ∷ Map (KeyHash 'StakePool c) (IndividualPoolStake c)

certIxFromIntegral ∷ Integral a ⇒ a → Maybe CertIx

certIxToInt ∷ CertIx → Int

txIxFromIntegral ∷ Integral a ⇒ a → Maybe TxIx

txIxToInt ∷ TxIx → Int

data ProtVer

Constructors

ProtVer

Fields

pvMajor ∷ !Version

pvMinor ∷ !Natural

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

newtype Port

Constructors

Port

Fields

portToWord16 ∷ Word16

data Globals

Constructors

Globals

Fields

epochInfo ∷ !(EpochInfo (Either Text))

slotsPerKESPeriod ∷ !Word64

stabilityWindow ∷ !Word64

The window size in which our chosen chain growth property guarantees at least k blocks. From the paper "Ouroboros praos: An adaptively-secure, semi-synchronous proof-of-stake protocol". The stabilityWindow constant is used in a number of places; for example, protocol updates must be submitted at least twice this many slots before an epoch boundary.

randomnessStabilisationWindow ∷ !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

securityParameter ∷ !Word64

Maximum number of blocks we are allowed to roll back

maxKESEvo ∷ !Word64

Maximum number of KES iterations

quorum ∷ !Word64

Quorum for update system votes and MIR certificates

maxMajorPV ∷ !Version

All blocks invalid after this protocol version

maxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system

activeSlotCoeff ∷ !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

networkId ∷ !Network

The network ID

systemStart ∷ !SystemStart

System start time

data Network

Constructors

Testnet

Mainnet

data TxIx

Transaction index.

data CertIx

Certificate index. Use certIxFromIntegral in order to construct this index safely from anything other than Word16. There is also mkCertIxPartial that can be used for testing.

hashKey ∷ ∀ c (kd ∷ KeyRole). Crypto c ⇒ VKey kd c → KeyHash kd c

Hash a given public key

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

Voting

coerceKeyRole ∷ ∀ (r ∷ KeyRole) c (r' ∷ KeyRole). HasKeyRole a ⇒ a r c → a r' c

General coercion of key roles.

The presence of this function is mostly to help the user realise where they are converting key roles.

newtype VKey (kd ∷ KeyRole) c

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey ∷ VerKeyDSIGN (DSIGN c)

data KeyPair (kd ∷ KeyRole) c

Constructors

KeyPair

Fields

vKey ∷ !(VKey kd c)

sKey ∷ !(SignKeyDSIGN (DSIGN c))

newtype KeyHash (discriminator ∷ KeyRole) c

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH c) (VerKeyDSIGN (DSIGN c)))

type KESignable c = Signable (KES c)

data GenDelegPair c

Constructors

GenDelegPair

Fields

genDelegKeyHash ∷ !(KeyHash 'GenesisDelegate c)

genDelegVrfHash ∷ !(Hash c (VerKeyVRF c))

newtype GenDelegs c

Constructors

GenDelegs

Fields

unGenDelegs ∷ Map (KeyHash 'Genesis c) (GenDelegPair c)

type Hash c = Hash (HASH c)

type SignedDSIGN c = SignedDSIGN (DSIGN c)

type SignKeyDSIGN c = SignKeyDSIGN (DSIGN c)

type SignedKES c = SignedKES (KES c)

type SignKeyKES c = SignKeyKES (KES c)

type VerKeyKES c = VerKeyKES (KES c)

type CertifiedVRF c = CertifiedVRF (VRF c)

type SignKeyVRF c = SignKeyVRF (VRF c)

type VerKeyVRF c = VerKeyVRF (VRF c)

word64ToCoin ∷ Word64 → Coin

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin ∷ Integer

data DCert c

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert c)

DCertPool !(PoolCert c)

DCertGenesis !(ConstitutionalDelegCert c)

DCertMir !(MIRCert c)

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ MIRPot

mirRewards ∷ MIRTarget c

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR (Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data ConstitutionalDelegCert c

Genesis key delegation certificate

Constructors

ConstitutionalDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

data DelegCert c

Constructors

RegKey !(StakeCredential c)
A stake key registration certificate.

DeRegKey !(StakeCredential c)
A stake key deregistration certificate.

Delegate !(Delegation c)
A stake delegation certificate.

data Delegation c

The delegation of one stake key to another.

Constructors

Delegation

Fields

dDelegator ∷ !(StakeCredential c)

dDelegatee ∷ !(KeyHash 'StakePool c)

data ShelleyUTXOW era

data ShelleyUTXO era

data ShelleyTICKF era

data ShelleyTICK era

data ShelleyPPUP era

data ShelleyPOOLREAP era

data ShelleyPOOL era

data ShelleyNEWEPOCH era

data ShelleyLEDGERS era

data ShelleyLEDGER era

data ShelleyDELPL era

data ShelleyDELEGS era

data ShelleyDELEG era

newtype ProposedPPUpdates era

Update operation for protocol parameters structure PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdate era))

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

data ShelleyPPUPState era

Constructors

ShelleyPPUPState

Fields

proposals ∷ !(ProposedPPUpdates era)

futureProposals ∷ !(ProposedPPUpdates era)

data PpupEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs (EraCrypto era))

data MultiSig era where

Bundled Patterns

pattern RequireMOf ∷ Era era ⇒ Int → [MultiSig era] → MultiSig era

pattern RequireSignature ∷ Era era ⇒ KeyHash 'Witness (EraCrypto era) → MultiSig era

pattern RequireAllOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

pattern RequireAnyOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

computeStabilityWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the stability window (e.g. the number of slots needed for a block to become stable) from the security param and the active slot coefficient.

The value 3k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

computeRandomnessStabilisationWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the randomness stabilisation window from the security param and the active slot coefficient.

The value 4k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

type Metadata era = ShelleyTxAuxData era

Deprecated: Use ShelleyTxAuxData instead

data ShelleyTxAuxData era where

Bundled Patterns

pattern ShelleyTxAuxData ∷ ∀ era. Era era ⇒ Map Word64 Metadatum → ShelleyTxAuxData era

data Metadatum

A generic metadatum type.

Constructors

Map ![(Metadatum, Metadatum)]

List ![Metadatum]

I !Integer

B !ByteString

S !Text

data ShelleyTxOut era

Constructors

TxOutCompact !(CompactAddr (EraCrypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern ShelleyTxOut ∷ (HasCallStack, Era era, Val (Value era)) ⇒ Addr (EraCrypto era) → Value era → ShelleyTxOut era

newtype ShelleyTxBody era

Constructors

TxBodyConstr (MemoBytes ShelleyTxBodyRaw era)

Bundled Patterns

pattern ShelleyTxBody ∷ EraTxOut era ⇒ Set (TxIn (EraCrypto era)) → StrictSeq (TxOut era) → StrictSeq (DCert (EraCrypto era)) → Withdrawals (EraCrypto era) → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → ShelleyTxBody era
Pattern for use by external users

data ValidationErr

Constructors

EpochNotLongEnough EpochSize Word64 Rational EpochSize

MaxKESEvolutionsUnsupported Word64 Word

QuorumTooSmall Word64 Word64 Word64

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !PositiveUnitInterval

sgSecurityParam ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTimeMicro

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(PParams (ShelleyEra c))

sgGenDelegs ∷ !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds ∷ ListMap (Addr c) Coin

sgStaking ∷ ShelleyGenesisStaking c

newtype NominalDiffTimeMicro

Unlike NominalDiffTime that supports Pico precision, this type only supports Micro precision.

Constructors

NominalDiffTimeMicro Micro

data ShelleyGenesisStaking c

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools ∷ ListMap (KeyHash 'StakePool c) (PoolParams c)

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake ∷ ListMap (KeyHash 'Staking c) (KeyHash 'StakePool c)

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

emptyGenesisStaking ∷ ShelleyGenesisStaking c

Empty genesis staking

fromNominalDiffTimeMicro ∷ NominalDiffTimeMicro → NominalDiffTime

toNominalDiffTimeMicroWithRounding ∷ NominalDiffTime → NominalDiffTimeMicro

toNominalDiffTimeMicro ∷ NominalDiffTime → Maybe NominalDiffTimeMicro

secondsToNominalDiffTimeMicro ∷ Micro → NominalDiffTimeMicro

nominalDiffTimeMicroToMicroseconds ∷ NominalDiffTimeMicro → Micro

nominalDiffTimeMicroToSeconds ∷ NominalDiffTimeMicro → Pico

sgActiveSlotCoeff ∷ ShelleyGenesis c → ActiveSlotCoeff

genesisUTxO ∷ ∀ era. EraTxOut era ⇒ ShelleyGenesis (EraCrypto era) → UTxO era

initialFundsPseudoTxIn ∷ ∀ c. Crypto c ⇒ Addr c → TxIn c

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

describeValidationErr ∷ ValidationErr → Text

validateGenesis ∷ ∀ c. Crypto c ⇒ ShelleyGenesis c → Either [ValidationErr] ()

Do some basic sanity checking on the Shelley genesis file.

mkShelleyGlobals ∷ ShelleyGenesis c → EpochInfo (Either Text) → Version → Globals

data NonMyopic c

data RewardUpdate c

The ultiate goal of a reward update computation. Aggregating rewards for each staking credential.

Constructors

RewardUpdate

Fields

deltaT ∷ !DeltaCoin

deltaR ∷ !DeltaCoin

rs ∷ !(Map (Credential 'Staking c) (Set (Reward c)))

deltaF ∷ !DeltaCoin

nonMyopic ∷ !(NonMyopic c)

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState ∷ !(UTxOState era)

The current unspent transaction outputs.

lsCertState ∷ !(CertState era)

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Last epoch

nesBprev ∷ !(BlocksMade (EraCrypto era))

Blocks made before current epoch

nesBcur ∷ !(BlocksMade (EraCrypto era))

Blocks made in current epoch

nesEs ∷ !(EpochState era)

Epoch state before current

nesRu ∷ !(StrictMaybe (PulsingRewUpdate (EraCrypto era)))

Possible reward update

nesPd ∷ !(PoolDistr (EraCrypto era))

Stake distribution within the stake pool

stashedAVVMAddresses ∷ !(StashedAVVMAddresses era)

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that istake == (updateStakeDistribution (UTxO Map.empty) (UTxO Map.empty) utxo) Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very careful, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

utxosUtxo ∷ !(UTxO era)

utxosDeposited ∷ !Coin

utxosFees ∷ !Coin

utxosGovernance ∷ !(GovernanceState era)

utxosStakeDistr ∷ !(IncrementalStake (EraCrypto era))

data IncrementalStake c

Incremental Stake, Stake along with possible missed coins from danging Ptrs. Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr does not have to point to anything until the epoch boundary, when we compute rewards and aggregate staking information for ranking. This is unusual but legal. In a non incremental system, we use whatever legal Ptrs exist at the epoch boundary. Here we are computing things incrementally, so we need to remember Ptrs that might point to something by the time the epoch boundary is reached. When the epoch boundary is reached we resolve these pointers, to see if any have become non-dangling since the time they were first used in the incremental computation.

Constructors

IStake

Fields

credMap ∷ !(Map (Credential 'Staking c) Coin)

ptrMap ∷ !(Map Ptr Coin)

data EpochState era

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esSnapshots ∷ !(SnapShots (EraCrypto era))

esLState ∷ !(LedgerState era)

esPrevPp ∷ !(PParams era)

esPp ∷ !(PParams era)

esNonMyopic ∷ !(NonMyopic (EraCrypto era))

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://github.com/input-output-hk/cardano-ledger/releases/latest/download/pool-ranking.pdf

data AccountState

Constructors

AccountState

Fields

asTreasury ∷ !Coin

asReserves ∷ !Coin

type KeyPairs c = [(KeyPair 'Payment c, KeyPair 'Staking c)]

Deprecated: Use `Test.Cardano.Ledger.Core.KeyPair (KeyPairs)` instead

data ShelleyTxWits era

data ShelleyTx era where

Bundled Patterns

pattern ShelleyTx ∷ EraTx era ⇒ TxBody era → TxWits era → StrictMaybe (TxAuxData era) → ShelleyTx era

bbHash ∷ ∀ era. Era era ⇒ ShelleyTxSeq era → Hash (EraCrypto era) EraIndependentBlockBody

Hash a given block body

data UtxoEnv era

Constructors

UtxoEnv SlotNo (PParams era) (CertState era) (GenDelegs (EraCrypto era))

data SnapEnv era

Constructors

SnapEnv !(LedgerState era) !(PParams era)

data PoolEnv era

Constructors

PoolEnv !SlotNo !(PParams era)

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo ∷ SlotNo

ptr_ ∷ Ptr

acnt_ ∷ AccountState

ppDE ∷ PParams era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo ∷ SlotNo

delPlPtr ∷ Ptr

delPlPp ∷ PParams era

delPlAcnt ∷ AccountState

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo ∷ !SlotNo

delegsIx ∷ !TxIx

delegspp ∷ !(PParams era)

delegsTx ∷ !(Tx era)

delegsAccount ∷ !AccountState

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo ∷ !SlotNo

ledgerIx ∷ !TxIx

ledgerPp ∷ !(PParams era)

ledgerAccount ∷ !AccountState

data ShelleyLedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo ∷ SlotNo

ledgersPp ∷ PParams era

ledgersAccount ∷ AccountState