Cardano.Ledger.Shelley.API.Validation

Description

Interface to the block validation and chain extension logic in the Shelley API.

Documentation

class (STS (EraRule "TICK" era), BaseM (EraRule "TICK" era) ~ ShelleyBase, Environment (EraRule "TICK" era) ~ (), State (EraRule "TICK" era) ~ NewEpochState era, Signal (EraRule "TICK" era) ~ SlotNo, STS (EraRule "BBODY" era), BaseM (EraRule "BBODY" era) ~ ShelleyBase, Environment (EraRule "BBODY" era) ~ BbodyEnv era, State (EraRule "BBODY" era) ~ ShelleyBbodyState era, Signal (EraRule "BBODY" era) ~ Block (BHeaderView (EraCrypto era)) era, EncCBORGroup (TxSeq era), State (EraRule "LEDGERS" era) ~ LedgerState era) => ApplyBlock era where

Minimal complete definition

Nothing

Methods

applyTickOpts :: ApplySTSOpts ep -> Globals -> NewEpochState era -> SlotNo -> EventReturnType ep (EraRule "TICK" era) (NewEpochState era)

Apply the header level ledger transition.

This handles checks and updates that happen on a slot tick, as well as a few header level checks, such as size constraints.

applyBlockOpts :: forall ep m. (EventReturnTypeRep ep, MonadError (BlockTransitionError era) m) => ApplySTSOpts ep -> Globals -> NewEpochState era -> Block (BHeaderView (EraCrypto era)) era -> m (EventReturnType ep (EraRule "BBODY" era) (NewEpochState era))

Apply the block level ledger transition.

default applyBlockOpts :: forall ep m. (EventReturnTypeRep ep, MonadError (BlockTransitionError era) m, EraGov era) => ApplySTSOpts ep -> Globals -> NewEpochState era -> Block (BHeaderView (EraCrypto era)) era -> m (EventReturnType ep (EraRule "BBODY" era) (NewEpochState era))

reapplyBlock :: Globals -> NewEpochState era -> Block (BHeaderView (EraCrypto era)) era -> NewEpochState era

Re-apply a ledger block to the same state it has been applied to before.

This function does no validation of whether the block applies successfully; the caller implicitly guarantees that they have previously called applyBlockTransition on the same block and that this was successful.

default reapplyBlock :: EraGov era => Globals -> NewEpochState era -> Block (BHeaderView (EraCrypto era)) era -> NewEpochState era

applyBlock :: (ApplyBlock era, MonadError (BlockTransitionError era) m) => Globals -> NewEpochState era -> Block (BHeaderView (EraCrypto era)) era -> m (NewEpochState era)

applyTick :: ApplyBlock era => Globals -> NewEpochState era -> SlotNo -> NewEpochState era

newtype TickTransitionError era

Constructors

TickTransitionError [PredicateFailure (EraRule "TICK" era)]

newtype BlockTransitionError era

Constructors

BlockTransitionError [PredicateFailure (EraRule "BBODY" era)]

chainChecks

Arguments

:: forall c m. MonadError ChainPredicateFailure m

=> Version
Max major protocol version

-> ChainChecksPParams

-> BHeaderView c

-> m ()

type ShelleyEraCrypto c = (Crypto c, DSignable c (Hash c EraIndependentTxBody))

Deprecated: Constraint synonyms are being removed