Cardano.Ledger.Shelley.API.Wallet

UTxOs

getUTxO ∷ NewEpochState era → UTxO era

Get the full UTxO.

getUTxOSubset ∷ NewEpochState era → Set (TxIn (EraCrypto era)) → UTxO era

getFilteredUTxO ∷ EraTxOut era ⇒ NewEpochState era → Set (Addr (EraCrypto era)) → UTxO era

Get the UTxO filtered by address.

Stake Pools

getPools ∷ NewEpochState era → Set (KeyHash 'StakePool (EraCrypto era))

Get the current registered stake pools.

getPoolParameters ∷ NewEpochState era → Set (KeyHash 'StakePool (EraCrypto era)) → Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era))

Get the current registered stake pool parameters for a given set of stake pools. The result map will contain entries for all the given stake pools that are currently registered.

getTotalStake ∷ Globals → NewEpochState era → Coin

Calculate the current total stake.

poolsByTotalStakeFraction ∷ ∀ era. EraGov era ⇒ Globals → NewEpochState era → PoolDistr (EraCrypto era)

Get pool sizes, but in terms of total stake

The stake distribution uses active stake (so that the leader schedule is not affected by undelegated stake), but the wallet wants to display pool saturation for rewards purposes. For that, it needs the fraction of total stake.

The fields individualTotalPoolStake and pdTotalActiveStake continue to remain based on active stake and not total stake.

This is not based on any snapshot, but uses the current ledger state.

data RewardInfoPool

Information about a stake pool

Constructors

RewardInfoPool

Fields

stake ∷ Coin

Absolute stake delegated to this pool

ownerPledge ∷ Coin

Pledge of pool owner(s)

ownerStake ∷ Coin

Absolute stake delegated by pool owner(s)

cost ∷ Coin

Pool cost

margin ∷ UnitInterval

Pool margin

performanceEstimate ∷ Double

Number of blocks produced divided by expected number of blocks. Can be larger than 1.0 for pool that gets lucky. (If some pools get unlucky, some pools must get lucky.)

data RewardParams

Global information that influences stake pool rewards

Constructors

RewardParams

Fields

nOpt ∷ Natural

Desired number of stake pools

a0 ∷ NonNegativeInterval

Influence of the pool owner's pledge on rewards

rPot ∷ Coin

Total rewards available for the given epoch

totalStake ∷ Coin

Maximum lovelace supply minus treasury

getRewardInfoPools ∷ EraGov era ⇒ Globals → NewEpochState era → (RewardParams, Map (KeyHash 'StakePool (EraCrypto era)) RewardInfoPool)

Retrieve the information necessary to calculate stake pool member rewards from the current stake distribution.

This information includes the current stake distribution aggregated by stake pools and pool owners, the current pool costs and margins, and performance estimates. Also included are global information such as the total stake or protocol parameters.

getRewardProvenance ∷ ∀ era. EraGov era ⇒ Globals → NewEpochState era → (RewardUpdate (EraCrypto era), RewardProvenance (EraCrypto era))

Calculate stake pool rewards from the snapshot labeled go. Also includes information on how the rewards were calculated (RewardProvenance).

For a calculation of rewards based on the current stake distribution, see getRewardInfoPools.

TODO: Deprecate getRewardProvenance, because wallets are more likely to use getRewardInfoPools for up-to-date information on stake pool rewards.

getNonMyopicMemberRewards ∷ EraGov era ⇒ Globals → NewEpochState era → Set (Either Coin (Credential 'Staking (EraCrypto era))) → Map (Either Coin (Credential 'Staking (EraCrypto era))) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

Calculate the Non-Myopic Pool Member Rewards for a set of credentials. For each given credential, this function returns a map from each stake pool (identified by the key hash of the pool operator) to the non-myopic pool member reward for that stake pool.

This is not based on any snapshot, but uses the current ledger state.

Transaction helpers

addKeyWitnesses ∷ EraTx era ⇒ Tx era → Set (WitVKey 'Witness (EraCrypto era)) → Tx era

evaluateTransactionBalance

Arguments

∷ EraUTxO era

⇒ PParams era
Current protocol parameters

→ CertState era
Where the deposit info is stored

→ UTxO era
The UTxO relevant to the transaction.

→ TxBody era
The transaction being evaluated for balance.

→ Value era
The difference between what the transaction consumes and what it produces.

Deprecated: In favor of new API function evalBalanceTxBody

Evaluate the difference between the value currently being consumed by a transaction and the number of lovelace being produced. This value will be zero for a valid transaction.

evaluateMinLovelaceOutput ∷ EraTxOut era ⇒ PParams era → TxOut era → Coin

Deprecated: In favor of getMinCoinTxOut

Evaluate the minimum lovelace that a given transaction output must contain.

addShelleyKeyWitnesses ∷ (EraTx era, Tx era ~ ShelleyTx era) ⇒ ShelleyTx era → Set (WitVKey 'Witness (EraCrypto era)) → ShelleyTx era

Deprecated: In favor of addKeyWitnesses

Ada pots

data AdaPots

Constructors

AdaPots

Fields

treasuryAdaPot ∷ Coin

reservesAdaPot ∷ Coin

rewardsAdaPot ∷ Coin

utxoAdaPot ∷ Coin

feesAdaPot ∷ Coin

obligationsPot ∷ Obligations

totalAdaES ∷ (EraTxOut era, EraGov era) ⇒ EpochState era → Coin

Calculate the total ada in the epoch state

totalAdaPotsES ∷ (EraTxOut era, EraGov era) ⇒ EpochState era → AdaPots

Calculate the total ada pots in the epoch state