Cardano.Ledger.Shelley.BlockChain

Documentation

data TxSeq era

Constructors

TxSeq' !(StrictSeq (Tx era)) ByteString ByteString ByteString

Bundled Patterns

pattern TxSeq :: forall era. (Era era, SafeToHash (Witnesses era)) => StrictSeq (Tx era) -> TxSeq era
Constuct a TxSeq (with all it bytes) from just Core.Tx's

constructMetadata :: forall era. Int -> Map Int (Annotator (AuxiliaryData era)) -> Seq (Maybe (Annotator (AuxiliaryData era)))

Given a size and a mapping from indices to maybe metadata, return a sequence whose size is the size paramater and whose non-Nothing values correspond to the values in the mapping.

txSeqTxns :: TxSeq era -> StrictSeq (Tx era)

bbHash :: forall era. Era era => TxSeq era -> Hash (Crypto era) EraIndependentBlockBody

Hash a given block body

bBodySize :: ToCBORGroup txSeq => txSeq -> Int

slotToNonce :: SlotNo -> Nonce

incrBlocks :: Bool -> KeyHash 'StakePool crypto -> BlocksMade crypto -> BlocksMade crypto

coreAuxDataBytes :: forall era. SafeToHash (AuxiliaryData era) => Tx era -> StrictMaybe ByteString

txSeqDecoder :: forall era. BlockAnn era => Bool -> forall s. Decoder s (Annotator (TxSeq era))

The parts of the Tx in Blocks that have to have FromCBOR(Annotator x) instances. These are exactly the parts that are SafeToHash. | Decode a TxSeq, used in decoding a Block.