Cardano.Ledger.Shelley.BlockChain

Documentation

data ShelleyTxSeq era

Constructors

TxSeq' !(StrictSeq (ShelleyTx era)) ByteString ByteString ByteString

Bundled Patterns

pattern ShelleyTxSeq :: forall era. (EraTx era, Tx era ~ ShelleyTx era, SafeToHash (TxWits era)) => StrictSeq (Tx era) -> ShelleyTxSeq era
Constuct a TxSeq (with all it bytes) from just Tx's

constructMetadata :: forall era. Int -> Map Int (Annotator (TxAuxData era)) -> Seq (Maybe (Annotator (TxAuxData era)))

Given a size and a mapping from indices to maybe metadata, return a sequence whose size is the size paramater and whose non-Nothing values correspond to the values in the mapping.

txSeqTxns :: ShelleyTxSeq era -> StrictSeq (ShelleyTx era)

bbHash :: forall era. Era era => ShelleyTxSeq era -> Hash (EraCrypto era) EraIndependentBlockBody

Hash a given block body

bBodySize :: EraSegWits era => ProtVer -> TxSeq era -> Int

slotToNonce :: SlotNo -> Nonce

incrBlocks :: Bool -> KeyHash 'StakePool c -> BlocksMade c -> BlocksMade c

coreAuxDataBytes :: EraTx era => Tx era -> StrictMaybe ByteString

txSeqDecoder :: forall era. EraTx era => Bool -> forall s. Decoder s (Annotator (ShelleyTxSeq era))

The parts of the Tx in Blocks that have to have DecCBOR(Annotator x) instances. These are exactly the parts that are SafeToHash. | Decode a TxSeq, used in decoding a Block.

Orphan instances

Crypto c => EraSegWits (ShelleyEra c)

Instance details

Associated Types

type TxSeq (ShelleyEra c) = (r :: Type)

Methods

fromTxSeq :: TxSeq (ShelleyEra c) -> StrictSeq (Tx (ShelleyEra c))

toTxSeq :: StrictSeq (Tx (ShelleyEra c)) -> TxSeq (ShelleyEra c)

hashTxSeq :: TxSeq (ShelleyEra c) -> Hash (HASH (EraCrypto (ShelleyEra c))) EraIndependentBlockBody

numSegComponents :: Word64