Cardano.Ledger.Shelley.Delegation.Certificates

Documentation

data Delegation c

The delegation of one stake key to another.

Constructors

Delegation

Fields

dDelegator ∷ !(StakeCredential c)

dDelegatee ∷ !(KeyHash 'StakePool c)

data DCert c

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert c)

DCertPool !(PoolCert c)

DCertGenesis !(ConstitutionalDelegCert c)

DCertMir !(MIRCert c)

data DelegCert c

Constructors

RegKey !(StakeCredential c)
A stake key registration certificate.

DeRegKey !(StakeCredential c)
A stake key deregistration certificate.

Delegate !(Delegation c)
A stake delegation certificate.

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

data ConstitutionalDelegCert c

Genesis key delegation certificate

Constructors

ConstitutionalDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ MIRPot

mirRewards ∷ MIRTarget c

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR (Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR Coin

delegCWitness ∷ DelegCert c → Credential 'Staking c

Determine the certificate author

poolCWitness ∷ PoolCert c → Credential 'StakePool c

genesisCWitness ∷ ConstitutionalDelegCert c → KeyHash 'Genesis c

isRegKey ∷ DCert c → Bool

Check for RegKey constructor

isDeRegKey ∷ DCert c → Bool

Check for DeRegKey constructor

isDelegation ∷ DCert c → Bool

Check for Delegation constructor

isGenesisDelegation ∷ DCert c → Bool

Check for GenesisDelegate constructor

isRegPool ∷ DCert c → Bool

Check for RegPool constructor

isRetirePool ∷ DCert c → Bool

Check for RetirePool constructor

isInstantaneousRewards ∷ DCert c → Bool

isReservesMIRCert ∷ DCert c → Bool

isTreasuryMIRCert ∷ DCert c → Bool

requiresVKeyWitness ∷ DCert c → Bool

Returns True for delegation certificates that require at least one witness, and False otherwise. It is mainly used to ensure that calling a variant of cwitness is safe.