Cardano.Ledger.Shelley.Genesis

Documentation

data ShelleyGenesisStaking crypto

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools :: !(Map (KeyHash 'StakePool crypto) (PoolParams crypto))

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake :: !(Map (KeyHash 'Staking crypto) (KeyHash 'StakePool crypto))

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

data ShelleyGenesis era

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTime

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams era)

sgGenDelegs :: !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds :: !(Map (Addr (Crypto era)) Coin)

sgStaking :: !(ShelleyGenesisStaking (Crypto era))

data ValidationErr

Constructors

EpochNotLongEnough EpochSize Word64 Rational EpochSize

MaxKESEvolutionsUnsupported Word64 Word

QuorumTooSmall Word64 Word64 Word64

emptyGenesisStaking :: ShelleyGenesisStaking crypto

Empty genesis staking

sgActiveSlotCoeff :: ShelleyGenesis era -> ActiveSlotCoeff

genesisUTxO :: forall era. (Era era, UsesTxOut era) => ShelleyGenesis era -> UTxO era

initialFundsPseudoTxIn :: forall crypto. Crypto crypto => Addr crypto -> TxIn crypto

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

validateGenesis :: forall era. Era era => ShelleyGenesis era -> Either [ValidationErr] ()

Do some basic sanity checking on the Shelley genesis file.

describeValidationErr :: ValidationErr -> Text

mkShelleyGlobals :: ShelleyGenesis era -> EpochInfo (Either Text) -> Natural -> Globals