Cardano.Ledger.Shelley.Governance

Documentation

class (EraPParams era, Eq (GovState era), Show (GovState era), NoThunks (GovState era), NFData (GovState era), EncCBOR (GovState era), DecCBOR (GovState era), DecShareCBOR (GovState era), ToCBOR (GovState era), FromCBOR (GovState era), Default (GovState era), ToJSON (GovState era)) => EraGov era where

Minimal complete definition

curPParamsGovStateL, prevPParamsGovStateL, obligationGovState, getDRepDistr

Associated Types

type GovState era = (r :: Type) | r -> era

Methods

emptyGovState :: GovState era

Construct empty governance state

getProposedPPUpdates :: GovState era -> Maybe (ProposedPPUpdates era)

Returns Nothing for all eras starting with Conway, otherwise returns proposed pparams updates

getConstitution :: GovState era -> Maybe (Constitution era)

Returns Nothing for all era preceding Conway, otherwise returns the hash of the constitution

getCommitteeMembers :: GovState era -> Maybe (Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo, UnitInterval)

curPParamsGovStateL :: Lens' (GovState era) (PParams era)

Lens for accessing current protocol parameters

prevPParamsGovStateL :: Lens' (GovState era) (PParams era)

Lens for accessing the previous protocol parameters

obligationGovState :: GovState era -> Coin

getDRepDistr :: GovState era -> Map (DRep (EraCrypto era)) (CompactForm Coin)

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

proposals :: !(ProposedPPUpdates era)

futureProposals :: !(ProposedPPUpdates era)

sgovPp :: !(PParams era)

sgovPrevPp :: !(PParams era)

data Constitution era

Constructors

Constitution

Fields

constitutionAnchor :: !(Anchor (EraCrypto era))

constitutionScript :: !(StrictMaybe (ScriptHash (EraCrypto era)))

proposalsL :: Lens' (ShelleyGovState era) (ProposedPPUpdates era)

futureProposalsL :: Lens' (ShelleyGovState era) (ProposedPPUpdates era)

curPParamsShelleyGovStateL :: Lens' (ShelleyGovState era) (PParams era)

prevPParamsShelleyGovStateL :: Lens' (ShelleyGovState era) (PParams era)

constitutionAnchorL :: Lens' (Constitution era) (Anchor (EraCrypto era))

constitutionScriptL :: Lens' (Constitution era) (StrictMaybe (ScriptHash (EraCrypto era)))