Cardano.Ledger.Shelley.LedgerState

Description

This module implements the operation rules for treating UTxO transactions (Tx) as state transformations on a ledger state (LedgerState), as specified in A Simplified Formal Specification of a UTxO Ledger.

Documentation

data AccountState

Constructors

AccountState

Fields

asTreasury ∷ !Coin

asReserves ∷ !Coin

data CertState era

The state associated with the DELPL rule, which combines the DELEG rule and the POOL rule.

Constructors

CertState

Fields

certVState ∷ !(VState era)

certPState ∷ !(PState era)

certDState ∷ !(DState era)

data DState era

The state used by the DELEG rule, which roughly tracks stake delegation and some governance features.

Constructors

DState

Fields

dsUnified ∷ !(UMap (EraCrypto era))

Unified Reward Maps. This contains the reward map (which is the source of truth regarding the registered stake credentials, the deposit map, the delegation map, and the stake credential pointer map.

dsFutureGenDelegs ∷ !(Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

Future genesis key delegations

dsGenDelegs ∷ !(GenDelegs (EraCrypto era))

Genesis key delegations

dsIRewards ∷ !(InstantaneousRewards (EraCrypto era))

Instantaneous Rewards

data VState era

The state that tracks the voting entities (DReps and Constitutional Committee members)

Constructors

VState

Fields

vsDReps ∷ !(Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

vsCommitteeState ∷ !(CommitteeState era)

vsNumDormantEpochs ∷ EpochNo

Number of contiguous epochs in which there are exactly zero active governance proposals to vote on. It is incremented in every EPOCH rule if the number of active governance proposals to vote on continues to be zero. It is reset to zero when a new governance action is successfully proposed. We need this counter in order to bump DRep expiries through dormant periods when DReps do not have an opportunity to vote on anything.

rewards ∷ DState era → UView (EraCrypto era) (Credential 'Staking (EraCrypto era)) RDPair

delegations ∷ DState era → UView (EraCrypto era) (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era))

ptrsMap ∷ DState era → Map Ptr (Credential 'Staking (EraCrypto era))

get the actual ptrs map, we don't need a view

data EpochState era

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esLState ∷ !(LedgerState era)

esSnapshots ∷ !(SnapShots (EraCrypto era))

esNonMyopic ∷ !(NonMyopic (EraCrypto era))

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://github.com/intersectmbo/cardano-ledger/releases/latest/download/pool-ranking.pdf

data PulsingRewUpdate c

The state used in the STS rules

Constructors

Pulsing !(RewardSnapShot c) !(Pulser c)

Complete !(RewardUpdate c)

data FutureGenDeleg c

Constructors

FutureGenDeleg

Fields

fGenDelegSlot ∷ !SlotNo

fGenDelegGenKeyHash ∷ !(KeyHash 'Genesis c)

data InstantaneousRewards c

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking c) Coin)

iRTreasury ∷ !(Map (Credential 'Staking c) Coin)

deltaReserves ∷ !DeltaCoin

deltaTreasury ∷ !DeltaCoin

type KeyPairs c = [(KeyPair 'Payment c, KeyPair 'Staking c)]

Deprecated: Use `Test.Cardano.Ledger.Core.KeyPair (KeyPairs)` instead

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState ∷ !(UTxOState era)

The current unspent transaction outputs.

lsCertState ∷ !(CertState era)

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The stake pool parameters.

psFutureStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The future stake pool parameters. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

A map of retiring stake pools to the epoch when they retire.

psDeposits ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) Coin)

A map of the deposits for each pool

type PPUPPredFailure era = EraRuleFailure "PPUP" era

Deprecated: In favor of EraRuleFailure PPUP era

type RewardAccounts c = Map (Credential 'Staking c) Coin

data RewardUpdate c

The ultimate goal of a reward update computation. Aggregating rewards for each staking credential.

Constructors

RewardUpdate

Fields

deltaT ∷ !DeltaCoin

deltaR ∷ !DeltaCoin

rs ∷ !(Map (Credential 'Staking c) (Set (Reward c)))

deltaF ∷ !DeltaCoin

nonMyopic ∷ !(NonMyopic c)

data RewardSnapShot c

To complete the reward update, we need a snap shot of the EpochState particular to this computation

Constructors

RewardSnapShot

Fields

rewFees ∷ !Coin

rewProtocolVersion ∷ !ProtVer

rewNonMyopic ∷ !(NonMyopic c)

rewDeltaR1 ∷ !Coin

rewR ∷ !Coin

rewDeltaT1 ∷ !Coin

rewLikelihoods ∷ !(Map (KeyHash 'StakePool c) Likelihood)

rewLeaders ∷ !(Map (Credential 'Staking c) (Set (Reward c)))

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that istake == (updateStakeDistribution (UTxO Map.empty) (UTxO Map.empty) utxo) Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very careful, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

utxosUtxo ∷ !(UTxO era)

utxosDeposited ∷ Coin

This field is left lazy, because we only use it for assertions

utxosFees ∷ !Coin

utxosGovState ∷ !(GovState era)

utxosStakeDistr ∷ !(IncrementalStake (EraCrypto era))

utxosDonation ∷ !Coin

smartUTxOState ∷ EraTxOut era ⇒ PParams era → UTxO era → Coin → Coin → GovState era → Coin → UTxOState era

A valid (or self-consistent) UTxOState{utxosUtxo, utxosDeposited , utxosFees , utxosPpups , utxosStakeDistr} maintains an invariant between the utxosUtxo and utxosStakeDistr fields. the utxosStakeDistr field is the aggregation of Coin over the StakeReferences in the UTxO. It can be computed by a pure function from the _utxo field. In some situations, mostly unit or example tests, or when initializing a small UTxO, we want to create a UTxOState that computes the utxosStakeDistr from the utxosUtxo. This is aways safe to do, but if the utxosUtxo field is big, this can be very expensive, which defeats the purpose of memoizing the utxosStakeDistr field. So use of this function should be restricted to tests and initializations, where the invariant should be maintained.

TO IncrementalStake

data IncrementalStake c

Incremental Stake, Stake along with possible missed coins from danging Ptrs. Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr does not have to point to anything until the epoch boundary, when we compute rewards and aggregate staking information for ranking. This is unusual but legal. In a non incremental system, we use whatever legal Ptrs exist at the epoch boundary. Here we are computing things incrementally, so we need to remember Ptrs that might point to something by the time the epoch boundary is reached. When the epoch boundary is reached we resolve these pointers, to see if any have become non-dangling since the time they were first used in the incremental computation.

Constructors

IStake

Fields

credMap ∷ !(Map (Credential 'Staking c) (CompactForm Coin))

ptrMap ∷ !(Map Ptr (CompactForm Coin))

Genesis State

genesisState ∷ ∀ era. EraGov era ⇒ Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)) → UTxO era → LedgerState era

Creates the ledger state for an empty ledger which contains the specified transaction outputs.

Validation

consumed ∷ EraUTxO era ⇒ PParams era → CertState era → UTxO era → TxBody era → Value era

For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

produced ∷ EraUTxO era ⇒ PParams era → CertState era → TxBody era → Value era

Compute the lovelace which are created by the transaction For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

DelegationState

payPoolDeposit ∷ EraPParams era ⇒ KeyHash 'StakePool (EraCrypto era) → PParams era → PState era → PState era

One only pays a deposit on the initial pool registration. So return the the Deposits unchanged if the keyhash already exists. There are legal situations where a pool may be registered multiple times.

refundPoolDeposit ∷ KeyHash 'StakePool (EraCrypto era) → PState era → (Coin, PState era)

totalObligation ∷ EraGov era ⇒ CertState era → GovState era → Coin

allObligations ∷ EraGov era ⇒ CertState era → GovState era → Obligations

Epoch boundary

incrementalStakeDistr ∷ ∀ era. EraPParams era ⇒ PParams era → IncrementalStake (EraCrypto era) → DState era → PState era → SnapShot (EraCrypto era)

This computes a Snapshot using IncrementalStake (which is an aggregate of the current UTxO) and UMap (which tracks Coin, SPoolUView, and Ptrs simultaneously). Note that logically: 1) IncrementalStake = (credStake, ptrStake) 2) UMap = (rewards, activeDelegs, ptrmap :: Map ptr cred)

Using this scheme the logic can do 3 things in one go, without touching the UTxO. 1) Resolve Pointers 2) Throw away things not actively delegated 3) Add up the coin

The Stake distribution function (Map cred coin) (the first component of a SnapShot) is defined by this SetAlgebra expression: (dom activeDelegs) ◁ (aggregate+ (credStake ∪ ptrStake ∪ rewards))

We can apply meaning preserving operations to get equivalent expressions

(dom activeDelegs) ◁ (aggregate+ (credStake ∪ ptrStake ∪ rewards)) aggregate+ (dom activeDelegs ◁ (credStake ∪ ptrStake ∪ rewards)) aggregate+ ((dom activeDelegs ◁ credStake) ∪ (dom activeDelegs ◁ ptrStake) ∪ (dom activeDelegs ◁ rewards))

We will compute this in several steps step1 = (dom activeDelegs ◁ credStake) ∪ (dom activeDelegs ◁ ptrStake) step2 = aggregate (dom activeDelegs ◁ rewards) step1 This function has a non-incremental analog, stakeDistr, mosty used in tests, which does use the UTxO.

updateStakeDistribution ∷ EraTxOut era ⇒ PParams era → IncrementalStake (EraCrypto era) → UTxO era → UTxO era → IncrementalStake (EraCrypto era)

Incrementally add the inserts utxoAdd and the deletes utxoDel to the IncrementalStake.

applyRUpd ∷ EraGov era ⇒ RewardUpdate (EraCrypto era) → EpochState era → EpochState era

Apply a RewardUpdate to the EpochState. Does several things 1) Adds reward coins to Rewards component of the UMap field of the DState, for actively delegated Stake 2) Adds to the Treasury of the AccountState for non-actively delegated stake 3) Adds fees to the UTxOState

applyRUpdFiltered ∷ EraGov era ⇒ RewardUpdate (EraCrypto era) → EpochState era → (EpochState era, FilteredRewards era)

filterAllRewards ∷ EraGov era ⇒ Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))) → EpochState era → FilteredRewards era

data FilteredRewards era

Constructors

FilteredRewards

Fields

frRegistered ∷ !(Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))))

These are registered, in the current Unified map of the CertState

frShelleyIgnored ∷ Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era)))

These are registered, but ignored in the ShelleyEra because of backward compatibility in non-Shelley Eras, this field will be Map.empty

frUnregistered ∷ Set (Credential 'Staking (EraCrypto era))

These are NOT registered in the current Unified map of the CertState

frTotalUnregistered ∷ Coin

Total Coin of the unregistered rewards. These will end up in the Treasury or Reserves.

createRUpd ∷ ∀ era. EraGov era ⇒ EpochSize → BlocksMade (EraCrypto era) → EpochState era → Coin → ActiveSlotCoeff → Word64 → ShelleyBase (RewardUpdate (EraCrypto era))

To create a reward update, run all 3 phases This function is not used in the rules, so it ignores RewardEvents

completeRupd ∷ PulsingRewUpdate c → ShelleyBase (RewardUpdate c, RewardEvent c)

Phase 3 of reward update has several parts a) completeM the pulser (in case there are still computions to run) b) Combine the pulser provenance with the RewardProvenance c) Construct the final RewardUpdate d) Add the leader rewards to both the events and the computed Rewards

startStep ∷ ∀ era. EraGov era ⇒ EpochSize → BlocksMade (EraCrypto era) → EpochState era → Coin → ActiveSlotCoeff → Word64 → PulsingRewUpdate (EraCrypto era)

pulseStep ∷ PulsingRewUpdate c → ShelleyBase (PulsingRewUpdate c, RewardEvent c)

Run the pulser for a bit. If is has nothing left to do, complete it.

completeStep ∷ PulsingRewUpdate c → ShelleyBase (PulsingRewUpdate c, RewardEvent c)

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState !EpochNo !(BlocksMade (EraCrypto era)) !(BlocksMade (EraCrypto era)) !(EpochState era) !(StrictMaybe (PulsingRewUpdate (EraCrypto era))) !(PoolDistr (EraCrypto era)) !(StashedAVVMAddresses era)

type family StashedAVVMAddresses era where ...

Equations

StashedAVVMAddresses (ShelleyEra c) = UTxO (ShelleyEra c)

StashedAVVMAddresses _ = ()

stashedAVVMAddresses ∷ NewEpochState era → StashedAVVMAddresses era

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

getGKeys ∷ NewEpochState era → Set (KeyHash 'Genesis (EraCrypto era))

updateNES ∷ EraGov era ⇒ NewEpochState era → BlocksMade (EraCrypto era) → LedgerState era → NewEpochState era

Update new epoch state

circulation ∷ EpochState era → Coin → Coin

Calculate the current circulation

This is used in the rewards calculation, and for API endpoints for pool ranking.

Decay

decayFactor ∷ Float

Remove Bootstrap Redeem Addresses

returnRedeemAddrsToReserves ∷ ∀ era. EraTxOut era ⇒ EpochState era → EpochState era

updateNonMyopic ∷ NonMyopic c → Coin → Map (KeyHash 'StakePool c) Likelihood → NonMyopic c

emptyRewardUpdate ∷ RewardUpdate c

pvCanFollow

Arguments

∷ ProtVer
Previous protocol version

→ ProtVer
New protocol version

→ Bool

Check whether the new protocol version is a legitimate version bump with respect to the previous one.

availableAfterMIR ∷ MIRPot → AccountState → InstantaneousRewards c → Coin

This function returns the coin balance of a given pot, either the reserves or the treasury, after the instantaneous rewards and pot transfers are accounted for.

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

sgsCurProposals ∷ !(ProposedPPUpdates era)

sgsFutureProposals ∷ !(ProposedPPUpdates era)

sgsCurPParams ∷ !(PParams era)

sgsPrevPParams ∷ !(PParams era)

sgsFuturePParams ∷ !(FuturePParams era)

Prediction of parameter changes that might happen on the epoch boundary.

Lenses from Types

nesPdL ∷ Lens' (NewEpochState era) (PoolDistr (EraCrypto era))

nesEsL ∷ Lens' (NewEpochState era) (EpochState era)

unifiedL ∷ Lens' (NewEpochState era) (UMap (EraCrypto era))

nesELL ∷ Lens' (NewEpochState era) EpochNo

nesBprevL ∷ Lens' (NewEpochState era) (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

nesBcurL ∷ Lens' (NewEpochState era) (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

nesRuL ∷ Lens' (NewEpochState era) (StrictMaybe (PulsingRewUpdate (EraCrypto era)))

nesStashedAVVMAddressesL ∷ Lens' (NewEpochState era) (StashedAVVMAddresses era)

nesEpochStateL ∷ Lens' (NewEpochState era) (EpochState era)

esAccountStateL ∷ Lens' (EpochState era) AccountState

esSnapshotsL ∷ Lens' (EpochState era) (SnapShots (EraCrypto era))

esLStateL ∷ Lens' (EpochState era) (LedgerState era)

esNonMyopicL ∷ Lens' (EpochState era) (NonMyopic (EraCrypto era))

curPParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (PParams era)

prevPParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (PParams era)

futurePParamsEpochStateL ∷ EraGov era ⇒ Lens' (EpochState era) (FuturePParams era)

asTreasuryL ∷ Lens' AccountState Coin

asReservesL ∷ Lens' AccountState Coin

lsUTxOStateL ∷ Lens' (LedgerState era) (UTxOState era)

lsCertStateL ∷ Lens' (LedgerState era) (CertState era)

utxosUtxoL ∷ Lens' (UTxOState era) (UTxO era)

utxosDepositedL ∷ Lens' (UTxOState era) Coin

utxosFeesL ∷ Lens' (UTxOState era) Coin

utxosGovStateL ∷ Lens' (UTxOState era) (GovState era)

utxosStakeDistrL ∷ Lens' (UTxOState era) (IncrementalStake (EraCrypto era))

utxosDonationL ∷ Lens' (UTxOState era) Coin

epochStateGovStateL ∷ Lens' (EpochState era) (GovState era)

epochStateStakeDistrL ∷ Lens' (EpochState era) (VMap VB VP (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

epochStateUMapL ∷ Lens' (EpochState era) (UMap (EraCrypto era))

epochStatePoolParamsL ∷ Lens' (EpochState era) (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

epochStateRegDrepL ∷ Lens' (EpochState era) (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

epochStateIncrStakeDistrL ∷ Lens' (EpochState era) (Map (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

epochStateDonationL ∷ Lens' (EpochState era) Coin

newEpochStateGovStateL ∷ Lens' (NewEpochState era) (GovState era)

epochStateTreasuryL ∷ Lens' (EpochState era) Coin

Lenses from CertState

certDStateL ∷ Lens' (CertState era) (DState era)

certPStateL ∷ Lens' (CertState era) (PState era)

certVStateL ∷ Lens' (CertState era) (VState era)

dsUnifiedL ∷ Lens' (DState era) (UMap (EraCrypto era))

dsGenDelegsL ∷ Lens' (DState era) (GenDelegs (EraCrypto era))

dsIRewardsL ∷ Lens' (DState era) (InstantaneousRewards (EraCrypto era))

dsFutureGenDelegsL ∷ Lens' (DState era) (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

psStakePoolParamsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

psFutureStakePoolParamsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

psRetiringL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

psDepositsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

vsDRepsL ∷ Lens' (VState era) (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

vsCommitteeStateL ∷ Lens' (VState era) (CommitteeState era)

credMapL ∷ Lens' (IncrementalStake c) (Map (Credential 'Staking c) (CompactForm Coin))

ptrMapL ∷ Lens' (IncrementalStake c) (Map Ptr (CompactForm Coin))

Lenses from SnapShot(s)

ssStakeMarkL ∷ Lens' (SnapShots c) (SnapShot c)

ssStakeMarkPoolDistrL ∷ Lens' (SnapShots c) (PoolDistr c)

ssStakeSetL ∷ Lens' (SnapShots c) (SnapShot c)

ssStakeGoL ∷ Lens' (SnapShots c) (SnapShot c)

ssFeeL ∷ Lens' (SnapShots c) Coin

ssStakeL ∷ Lens' (SnapShot c) (Stake c)

ssStakeDistrL ∷ Lens' (SnapShot c) (VMap VB VP (Credential 'Staking c) (CompactForm Coin))

ssDelegationsL ∷ Lens' (SnapShot c) (VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c))

ssPoolParamsL ∷ Lens' (SnapShot c) (VMap VB VB (KeyHash 'StakePool c) (PoolParams c))