Cardano.Ledger.Shelley.LedgerState.Types

Documentation

type KeyPairs c = [(KeyPair 'Payment c, KeyPair 'Staking c)]

Deprecated: Use `Test.Cardano.Ledger.Core.KeyPair (KeyPairs)` instead

type RewardAccounts c = Map (Credential 'Staking c) Coin

data AccountState

Constructors

AccountState

Fields

asTreasury ∷ !Coin

asReserves ∷ !Coin

toAccountStatePairs ∷ KeyValue a ⇒ AccountState → [a]

data EpochState era

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esSnapshots ∷ !(SnapShots (EraCrypto era))

esLState ∷ !(LedgerState era)

esPrevPp ∷ !(PParams era)

esPp ∷ !(PParams era)

esNonMyopic ∷ !(NonMyopic (EraCrypto era))

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://github.com/input-output-hk/cardano-ledger/releases/latest/download/pool-ranking.pdf

esAccountStateL ∷ Lens' (EpochState era) AccountState

esSnapshotsL ∷ Lens' (EpochState era) (SnapShots (EraCrypto era))

esLStateL ∷ Lens' (EpochState era) (LedgerState era)

esPrevPpL ∷ Lens' (EpochState era) (PParams era)

esPpL ∷ Lens' (EpochState era) (PParams era)

esNonMyopicL ∷ Lens' (EpochState era) (NonMyopic (EraCrypto era))

toEpochStatePairs ∷ (EraTxOut era, EraGovernance era, KeyValue a) ⇒ EpochState era → [a]

data IncrementalStake c

Incremental Stake, Stake along with possible missed coins from danging Ptrs. Transactions can use Ptrs to refer to a stake credential in a TxOut. The Ptr does not have to point to anything until the epoch boundary, when we compute rewards and aggregate staking information for ranking. This is unusual but legal. In a non incremental system, we use whatever legal Ptrs exist at the epoch boundary. Here we are computing things incrementally, so we need to remember Ptrs that might point to something by the time the epoch boundary is reached. When the epoch boundary is reached we resolve these pointers, to see if any have become non-dangling since the time they were first used in the incremental computation.

Constructors

IStake

Fields

credMap ∷ !(Map (Credential 'Staking c) Coin)

ptrMap ∷ !(Map Ptr Coin)

toIncrementalStakePairs ∷ (KeyValue a, Crypto crypto) ⇒ IncrementalStake crypto → [a]

data UTxOState era

There is a serious invariant that we must maintain in the UTxOState. Given (UTxOState utxo _ _ _ istake) it must be the case that istake == (updateStakeDistribution (UTxO Map.empty) (UTxO Map.empty) utxo) Of course computing the RHS of the above equality can be very expensive, so we only use this route in the testing function smartUTxO. But we are very careful, wherever we update the UTxO, we carefully make INCREMENTAL changes to istake to maintain this invariant. This happens in the UTxO rule.

Constructors

UTxOState

Fields

utxosUtxo ∷ !(UTxO era)

utxosDeposited ∷ !Coin

utxosFees ∷ !Coin

utxosGovernance ∷ !(GovernanceState era)

utxosStakeDistr ∷ !(IncrementalStake (EraCrypto era))

utxosUtxoL ∷ Lens' (UTxOState era) (UTxO era)

utxosDepositedL ∷ Lens' (UTxOState era) Coin

utxosFeesL ∷ Lens' (UTxOState era) Coin

utxosGovernanceL ∷ Lens' (UTxOState era) (GovernanceState era)

utxosStakeDistrL ∷ Lens' (UTxOState era) (IncrementalStake (EraCrypto era))

toUTxOStatePairs ∷ (EraTxOut era, EraGovernance era, KeyValue a) ⇒ UTxOState era → [a]

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Last epoch

nesBprev ∷ !(BlocksMade (EraCrypto era))

Blocks made before current epoch

nesBcur ∷ !(BlocksMade (EraCrypto era))

Blocks made in current epoch

nesEs ∷ !(EpochState era)

Epoch state before current

nesRu ∷ !(StrictMaybe (PulsingRewUpdate (EraCrypto era)))

Possible reward update

nesPd ∷ !(PoolDistr (EraCrypto era))

Stake distribution within the stake pool

stashedAVVMAddresses ∷ !(StashedAVVMAddresses era)

AVVM addresses to be removed at the end of the Shelley era. Note that the existence of this field is a hack, related to the transition of UTxO to disk. We remove AVVM addresses from the UTxO on the Shelley/Allegra boundary. However, by this point the UTxO will be moved to disk, and hence doing a scan of the UTxO for AVVM addresses will be expensive. Our solution to this is to do a scan of the UTxO on the Byron/Shelley boundary (since Byron UTxO are still on disk), stash the results here, and then remove them at the Shelley/Allegra boundary.

This is very much an awkward implementation hack, and hence we hide it from as many places as possible.

type family StashedAVVMAddresses era where ...

Equations

StashedAVVMAddresses (ShelleyEra c) = UTxO (ShelleyEra c)

StashedAVVMAddresses _ = ()

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

lsUTxOState ∷ !(UTxOState era)

The current unspent transaction outputs.

lsCertState ∷ !(CertState era)

lsUTxOStateL ∷ Lens' (LedgerState era) (UTxOState era)

lsCertStateL ∷ Lens' (LedgerState era) (CertState era)

toLedgerStatePairs ∷ (EraTxOut era, EraGovernance era, KeyValue a) ⇒ LedgerState era → [a]