Cardano.Ledger.Shelley.PParams

Description

This module contains just the type of protocol parameters.

Documentation

data PParams' f era

Protocol parameters.

We use the HKD type family so that the protocol parameters type and the type for the updates to the protocol parameters can share records fields. The protocol parameters will have type PParams' Identity, and the updates will have type PParams' StrictMaybe, though Identity will be hidden from use.

For example:

  myParameters =
    PParams
      { _minfeeA = 0,
        _minfeeB = 0,
        ...
      }

  myUpdate =
    PParamsUpdate
      { _minfeeA = SNothing,
        _minfeeB = SJust 42,
        ...
      }

Constructors

PParams

Fields

_minfeeA :: !(HKD f Natural)

The linear factor for the minimum fee calculation

_minfeeB :: !(HKD f Natural)

The constant factor for the minimum fee calculation

_maxBBSize :: !(HKD f Natural)

Maximal block body size

_maxTxSize :: !(HKD f Natural)

Maximal transaction size

_maxBHSize :: !(HKD f Natural)

Maximal block header size

_keyDeposit :: !(HKD f Coin)

The amount of a key registration deposit

_poolDeposit :: !(HKD f Coin)

The amount of a pool registration deposit

_eMax :: !(HKD f EpochNo)

epoch bound on pool retirement

_nOpt :: !(HKD f Natural)

Desired number of pools

_a0 :: !(HKD f NonNegativeInterval)

Pool influence

_rho :: !(HKD f UnitInterval)

Monetary expansion

_tau :: !(HKD f UnitInterval)

Treasury expansion

_d :: !(HKD f UnitInterval)

Decentralization parameter

_extraEntropy :: !(HKD f Nonce)

Extra entropy

_protocolVersion :: !(HKD f ProtVer)

Protocol version

_minUTxOValue :: !(HKD f Coin)

Minimum UTxO value

_minPoolCost :: !(HKD f Coin)

Minimum Stake Pool Cost

type PParams era = PParams' Identity era

emptyPParams :: PParams era

Returns a basic "empty" PParams structure with all zero values.

type family HKD f a where ...

Higher Kinded Data

Equations

HKD Identity a = a

HKD f a = f a

class HKDFunctor f where

Methods

hkdMap :: Proxy f -> (a -> b) -> HKD f a -> HKD f b

data PPUpdateEnv era

Constructors

PPUpdateEnv SlotNo (GenDelegs era)

newtype ProposedPPUpdates era

Update operation for protocol parameters structure @PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis (Crypto era)) (PParamsDelta era))

emptyPPPUpdates :: ProposedPPUpdates era

type PParamsUpdate era = PParams' StrictMaybe era

emptyPParamsUpdate :: PParamsUpdate era

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

updatePParams :: PParams era -> PParamsUpdate era -> PParams era