Cardano.Ledger.Shelley.PParams

Description

This module contains just the type of protocol parameters.

Documentation

emptyShelleyPParams :: forall era. Era era => ShelleyPParams Identity era

emptyShelleyPParamsUpdate :: ShelleyPParams StrictMaybe era

data ShelleyPParams f era

Protocol parameters.

Constructors

ShelleyPParams

Fields

sppMinFeeA :: !(HKD f Coin)

The linear factor for the minimum fee calculation

sppMinFeeB :: !(HKD f Coin)

The constant factor for the minimum fee calculation

sppMaxBBSize :: !(HKD f Natural)

Maximal block body size

sppMaxTxSize :: !(HKD f Natural)

Maximal transaction size

sppMaxBHSize :: !(HKD f Natural)

Maximal block header size

sppKeyDeposit :: !(HKD f Coin)

The amount of a key registration deposit

sppPoolDeposit :: !(HKD f Coin)

The amount of a pool registration deposit

sppEMax :: !(HKD f EpochNo)

epoch bound on pool retirement

sppNOpt :: !(HKD f Natural)

Desired number of pools

sppA0 :: !(HKD f NonNegativeInterval)

Pool influence

sppRho :: !(HKD f UnitInterval)

Monetary expansion

sppTau :: !(HKD f UnitInterval)

Treasury expansion

sppD :: !(HKD f UnitInterval)

Decentralization parameter

sppExtraEntropy :: !(HKD f Nonce)

Extra entropy

sppProtocolVersion :: !(HKD f ProtVer)

Protocol version

sppMinUTxOValue :: !(HKD f Coin)

Minimum UTxO value

sppMinPoolCost :: !(HKD f Coin)

Minimum Stake Pool Cost

emptyPParams :: EraPParams era => PParams era

type family HKD (f :: Type -> Type) a where ...

Equations

HKD Identity a = a

HKD f a = f a

data PPUpdateEnv era

Constructors

PPUpdateEnv SlotNo (GenDelegs era)

newtype ProposedPPUpdates era

Update operation for protocol parameters structure PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdate era))

emptyPPPUpdates :: ProposedPPUpdates era

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

pvCanFollow :: ProtVer -> StrictMaybe ProtVer -> Bool

JSON helpers

shelleyCommonPParamsHKDPairs :: forall f era. (HKDFunctor f, EraPParams era) => Proxy f -> PParamsHKD f era -> [(Key, HKD f Value)]

These are the fields that are common across all eras

shelleyCommonPParamsHKDPairsV6 :: forall f era. (HKDFunctor f, EraPParams era, ProtVerAtMost era 6) => Proxy f -> PParamsHKD f era -> [(Key, HKD f Value)]

These are the fields that are common only up to major protocol version 6

shelleyCommonPParamsHKDPairsV8 :: forall f era. (HKDFunctor f, EraPParams era, ProtVerAtMost era 8) => Proxy f -> PParamsHKD f era -> [(Key, HKD f Value)]

Deprecated

updatePParams :: EraPParams era => PParams era -> PParamsUpdate era -> PParams era

Deprecated: Use applyPPUpdates instead

upgradeUpdate :: forall era. (EraPParams era, EraPParams (PreviousEra era), EraCrypto (PreviousEra era) ~ EraCrypto era) => UpgradePParams StrictMaybe era -> Update (PreviousEra era) -> Update era

Orphan instances

Crypto c => EraPParams (ShelleyEra c)

Instance details

Associated Types

type PParamsHKD f (ShelleyEra c) = (r :: Type)

type UpgradePParams f (ShelleyEra c)

type DowngradePParams f (ShelleyEra c)

Methods

applyPPUpdates :: PParams (ShelleyEra c) -> PParamsUpdate (ShelleyEra c) -> PParams (ShelleyEra c)

emptyPParamsIdentity :: PParamsHKD Identity (ShelleyEra c)

emptyPParamsStrictMaybe :: PParamsHKD StrictMaybe (ShelleyEra c)

upgradePParamsHKD :: forall (f :: Type -> Type). (HKDFunctor f, EraPParams (PreviousEra (ShelleyEra c))) => UpgradePParams f (ShelleyEra c) -> PParamsHKD f (PreviousEra (ShelleyEra c)) -> PParamsHKD f (ShelleyEra c)

downgradePParamsHKD :: forall (f :: Type -> Type). (HKDFunctor f, EraPParams (PreviousEra (ShelleyEra c))) => DowngradePParams f (ShelleyEra c) -> PParamsHKD f (ShelleyEra c) -> PParamsHKD f (PreviousEra (ShelleyEra c))

hkdMinFeeAL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Coin)

hkdMinFeeBL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Coin)

hkdMaxBBSizeL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Natural)

hkdMaxTxSizeL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Natural)

hkdMaxBHSizeL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Natural)

hkdKeyDepositL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Coin)

hkdPoolDepositL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Coin)

hkdEMaxL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f EpochNo)

hkdNOptL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Natural)

hkdA0L :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f NonNegativeInterval)

hkdRhoL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f UnitInterval)

hkdTauL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f UnitInterval)

hkdDL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost (ShelleyEra c) 6) => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f UnitInterval)

ppDG :: SimpleGetter (PParams (ShelleyEra c)) UnitInterval

hkdExtraEntropyL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost (ShelleyEra c) 6) => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Nonce)

hkdProtocolVersionL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost (ShelleyEra c) 8) => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f ProtVer)

ppProtocolVersionL :: Lens' (PParams (ShelleyEra c)) ProtVer

ppuProtocolVersionL :: Lens' (PParamsUpdate (ShelleyEra c)) (StrictMaybe ProtVer)

hkdMinUTxOValueL :: forall (f :: Type -> Type). (HKDFunctor f, ProtVerAtMost (ShelleyEra c) 4) => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Coin)

hkdMinPoolCostL :: forall (f :: Type -> Type). HKDFunctor f => Lens' (PParamsHKD f (ShelleyEra c)) (HKD f Coin)