Cardano.Ledger.Shelley.PoolRank

Documentation

desirability ∷ (NonNegativeInterval, Natural) → Coin → PoolParams c → PerformanceEstimate → Coin → Double

Desirability calculation for non-myopic utility, corresponding to f^~ in section 5.6.1 of "Design Specification for Delegation and Incentives in Cardano"

newtype PerformanceEstimate

This is a estimate of the proportion of allowed blocks a pool will make in the future. It is used for ranking pools in delegation.

Constructors

PerformanceEstimate

Fields

unPerformanceEstimate ∷ Double

data NonMyopic c

Constructors

NonMyopic

Fields

likelihoodsNM ∷ !(Map (KeyHash 'StakePool c) Likelihood)

rewardPotNM ∷ !Coin

getTopRankedPools ∷ EraPParams era ⇒ Coin → Coin → PParams era → Map (KeyHash 'StakePool c) (PoolParams c) → Map (KeyHash 'StakePool c) PerformanceEstimate → Set (KeyHash 'StakePool c)

Computes the top ranked stake pools corresponding to section 5.6.1 of "Design Specification for Delegation and Incentives in Cardano"

getTopRankedPoolsVMap ∷ EraPParams era ⇒ Coin → Coin → PParams era → VMap VB VB (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)) → Map (KeyHash 'StakePool (EraCrypto era)) PerformanceEstimate → Set (KeyHash 'StakePool (EraCrypto era))

nonMyopicStake ∷ EraPParams era ⇒ PParams era → StakeShare → StakeShare → StakeShare → KeyHash 'StakePool c → Set (KeyHash 'StakePool c) → StakeShare

Compute the Non-Myopic Pool Stake

This function implements non-myopic stake calculation in section 5.6.2 of "Design Specification for Delegation and Incentives in Cardano". Note that the protocol parameters are implicit in the design document. Additionally, instead of passing a rank r to compare with k, we pass the top k desirable pools and check for membership.

nonMyopicMemberRew ∷ EraPParams era ⇒ PParams era → Coin → PoolParams c → StakeShare → StakeShare → StakeShare → Set (KeyHash 'StakePool c) → PerformanceEstimate → Coin

Compute the Non-Myopic Pool Member Reward

This function implements equation (3) in section 5.6.4 of "Design Specification for Delegation and Incentives in Cardano". Note that the protocol parameters and the reward pot are implicit in the design document. Additionally, instead of passing a rank r to compare with k, we pass the top k desirable pools and check for membership.

percentile' ∷ Likelihood → PerformanceEstimate

newtype Histogram

Constructors

Histogram

Fields

unHistogram ∷ StrictSeq LogWeight

newtype LogWeight

Constructors

LogWeight

Fields

unLogWeight ∷ Float

likelihood ∷ Natural → Double → EpochSize → Likelihood

applyDecay ∷ Float → Likelihood → Likelihood

Decay previous likelihood

newtype Likelihood

Constructors

Likelihood

Fields

unLikelihood ∷ StrictSeq LogWeight

leaderProbability ∷ ActiveSlotCoeff → Rational → UnitInterval → Double