Cardano.Ledger.Shelley.PoolRank

Documentation

desirability :: (NonNegativeInterval, Natural) -> Coin -> PoolParams c -> PerformanceEstimate -> Coin -> Double

Desirability calculation for non-myopic utility, corresponding to f^~ in section 5.6.1 of "Design Specification for Delegation and Incentives in Cardano"

newtype PerformanceEstimate

This is a estimate of the proportion of allowed blocks a pool will make in the future. It is used for ranking pools in delegation.

Constructors

PerformanceEstimate

Fields

unPerformanceEstimate :: Double

data NonMyopic crypto

Constructors

NonMyopic

Fields

likelihoodsNM :: !(Map (KeyHash 'StakePool crypto) Likelihood)

rewardPotNM :: !Coin

getTopRankedPools :: (HasField "_a0" pp NonNegativeInterval, HasField "_nOpt" pp Natural) => Coin -> Coin -> pp -> Map (KeyHash 'StakePool crypto) (PoolParams crypto) -> Map (KeyHash 'StakePool crypto) PerformanceEstimate -> Set (KeyHash 'StakePool crypto)

Computes the top ranked stake pools corresponding to section 5.6.1 of "Design Specification for Delegation and Incentives in Cardano"

getTopRankedPoolsVMap :: (HasField "_a0" pp NonNegativeInterval, HasField "_nOpt" pp Natural) => Coin -> Coin -> pp -> VMap VB VB (KeyHash 'StakePool crypto) (PoolParams crypto) -> Map (KeyHash 'StakePool crypto) PerformanceEstimate -> Set (KeyHash 'StakePool crypto)

nonMyopicStake :: HasField "_nOpt" pp Natural => pp -> StakeShare -> StakeShare -> StakeShare -> KeyHash 'StakePool c -> Set (KeyHash 'StakePool c) -> StakeShare

Compute the Non-Myopic Pool Stake

This function implements non-myopic stake calculation in section 5.6.2 of "Design Specification for Delegation and Incentives in Cardano". Note that the protocol parameters are implicit in the design document. Additionally, instead of passing a rank r to compare with k, we pass the top k desirable pools and check for membership.

nonMyopicMemberRew :: (HasField "_a0" pp NonNegativeInterval, HasField "_nOpt" pp Natural) => pp -> Coin -> PoolParams c -> StakeShare -> StakeShare -> StakeShare -> Set (KeyHash 'StakePool c) -> PerformanceEstimate -> Coin

Compute the Non-Myopic Pool Member Reward

This function implements equation (3) in section 5.6.4 of "Design Specification for Delegation and Incentives in Cardano". Note that the protocol parameters and the reward pot are implicit in the design document. Additionally, instead of passing a rank r to compare with k, we pass the top k desirable pools and check for membership.

percentile' :: Likelihood -> PerformanceEstimate

newtype Histogram

Constructors

Histogram

Fields

unHistogram :: StrictSeq LogWeight

newtype LogWeight

Constructors

LogWeight

Fields

unLogWeight :: Float

likelihood :: Natural -> Double -> EpochSize -> Likelihood

applyDecay :: Float -> Likelihood -> Likelihood

Decay previous likelihood

newtype Likelihood

Constructors

Likelihood

Fields

unLikelihood :: StrictSeq LogWeight

leaderProbability :: ActiveSlotCoeff -> Rational -> UnitInterval -> Double