Cardano.Ledger.Shelley.RewardProvenance

Documentation

data RewardProvenance c

RewardProvenenace captures some of the intermediate calculations when computing the staking reward distribution. Most of these fields are simple scalar values, computed from the current State, and are fixed before we start to compute the distribution. Two of them are aggregates computed when we compute the distribution (pools and desirabilities).

For more background, see "Figure 48: The Reward Calculation" and "Figure 51: Reward Update Creation" of the the formal specification. The variable names here align with those in the specification. See also Section 5 of the https://github.com/input-output-hk/cardano-ledger/releases/latest/download/shelley-delegation.pdf

Constructors

RewardProvenance

Fields

spe :: !Word64

The number of slots per epoch.

blocks :: !(BlocksMade c)

A map from pool ID (the key hash of the stake pool operator's verification key) to the number of blocks made in the given epoch.

maxLL :: !Coin

The maximum Lovelace supply. On mainnet, this value is equal to 45 * 10^15 (45 billion ADA).

deltaR1 :: !Coin

The maximum amount of Lovelace which can be removed from the reserves to be given out as rewards for the given epoch.

deltaR2 :: !Coin

The difference between the total Lovelace that could have been distributed as rewards this epoch (which is r) and what was actually distributed.

r :: !Coin

The total Lovelace available for rewards for the given epoch, equal to rPot less deltaT1.

totalStake :: !Coin

The maximum Lovelace supply (maxLL) less the current value of the reserves.

blocksCount :: !Integer

The total number of blocks produced during the given epoch.

d :: !Rational

The decentralization parameter.

expBlocks :: !Integer

The number of blocks expected to be produced during the given epoch.

eta :: !Rational

The ratio of the number of blocks actually made versus the number of blocks that were expected.

rPot :: !Coin

The reward pot for the given epoch, equal to deltaR1 plus the fee pot.

deltaT1 :: !Coin

The amount of Lovelace taken from the treasury for the given epoch.

activeStake :: !Coin

The amount of Lovelace that is delegated during the given epoch.

pools :: !(Map (KeyHash 'StakePool c) (RewardProvenancePool c))

Individual stake pool provenance.

desirabilities :: !(Map (KeyHash 'StakePool c) Desirability)

A map from pool ID to the desirability score. See the stake pool ranking document.

data RewardProvenancePool c

Provenance for an individual stake pool's reward calculation.

Constructors

RewardProvenancePool

Fields

poolBlocksP :: !Natural

The number of blocks the pool produced.

sigmaP :: !Rational

The stake pool's stake share (portion of the total stake).

sigmaAP :: !Rational

The stake pool's active stake share (portion of the active stake).

ownerStakeP :: !Coin

The number of Lovelace owned by the stake pool owners. If this value is not at least as large as the pledgeRatioP, the stake pool will not earn any rewards for the given epoch.

poolParamsP :: !(PoolParams c)

The stake pool's registered parameters.

pledgeRatioP :: !Rational

The stake pool's pledge.

maxPP :: !Coin

The maximum number of Lovelace this stake pool can earn.

appPerfP :: !Rational

The stake pool's apparent performance. See Section 5.5.2 of the https://github.com/input-output-hk/cardano-ledger/releases/latest/download/shelley-delegation.pdf

poolRP :: !Coin

The total Lovelace earned by the stake pool.

lRewardP :: !Coin

The total Lovelace earned by the stake pool leader.

data Desirability

The desirability score of a stake pool, as described in "Reward Sharing Schemes for Stake Pools". Additionally, the hit rate estimation described in the stake pool ranking document is included.

Constructors

Desirability

Fields

desirabilityScore :: !Double

hitRateEstimate :: !Double