Cardano.Ledger.Shelley.Rewards

Documentation

newtype StakeShare

StakeShare type

Constructors

StakeShare

Fields

unStakeShare ∷ Rational

data PoolRewardInfo c

Stake Pool specific information needed to compute the rewards for its members.

Constructors

PoolRewardInfo

Fields

poolRelativeStake ∷ !StakeShare

The stake pool's stake divided by the total stake

poolPot ∷ !Coin

The maximum rewards available for the entire pool

poolPs ∷ !(PoolParams c)

The stake pool parameters

poolBlocks ∷ !Natural

The number of blocks the stake pool produced

poolLeaderReward ∷ !(LeaderOnlyReward c)

The leader reward

mkApparentPerformance ∷ UnitInterval → Rational → Natural → Natural → Rational

Calculate pool reward

data RewardType

The staking rewards in Cardano are all either:

member rewards - rewards given to a registered stake credential which has delegated to a stake pool, or

leader rewards - rewards given to a registered stake pool (in particular, given to the stake credential in the stake pool registration certificate).

See Figure 47, "Functions used in the Reward Splitting", of the formal specification for more details.

Constructors

MemberReward

LeaderReward

data Reward c

The Reward type captures:

if the reward is a member or leader reward

the stake pool ID associated with the reward

the number of Lovelace in the reward

Constructors

Reward

Fields

rewardType ∷ !RewardType

rewardPool ∷ !(KeyHash 'StakePool c)

rewardAmount ∷ !Coin

data LeaderOnlyReward c

Constructors

LeaderOnlyReward

Fields

lRewardPool ∷ !(KeyHash 'StakePool c)

lRewardAmount ∷ !Coin

leaderRewardToGeneral ∷ LeaderOnlyReward c → Reward c

leaderRew ∷ Coin → PoolParams c → StakeShare → StakeShare → Coin

Calculate pool leader reward

memberRew ∷ Coin → PoolParams c → StakeShare → StakeShare → Coin

Calculate pool member reward

aggregateRewards ∷ ∀ c. ProtVer → Map (Credential 'Staking c) (Set (Reward c)) → Map (Credential 'Staking c) Coin

for each (Set (Reward c)) entry in the map, sum up the coin. In the ShelleyEra some of the coins are ignored (because of backward compatibility) see filterRewards Note that domain of the returned map is a subset of the input map rewards

filterRewards ∷ ∀ c. ProtVer → Map (Credential 'Staking c) (Set (Reward c)) → (Map (Credential 'Staking c) (Set (Reward c)), Map (Credential 'Staking c) (Set (Reward c)))

Filter the reward payments to those that will actually be delivered. This function exists since in Shelley, a stake credential earning rewards from multiple sources would only receive one reward. So some of the coins are ignored, because of this backward compatibility issue in early protocolVersions. Note that both of the domains of the returned maps are a subset of the the domain of the input map rewards

sumRewards ∷ ∀ c. ProtVer → Map (Credential 'Staking c) (Set (Reward c)) → Coin

aggregateCompactRewards ∷ ProtVer → Map (Credential 'Staking c) (Set (Reward c)) → Map (Credential 'Staking c) (CompactForm Coin)

for each (Set (Reward c)) entry in the map, sum up the coin. In the ShelleyEra some of the coins are ignored (because of backward compatibility) see filterRewards Note that the domain of the output map is a subset of the domain of the input rewards.

sumCompactRewards ∷ ProtVer → Map (Credential 'Staking c) (Set (Reward c)) → CompactForm Coin

rewardOnePoolMember

Arguments

∷ ProtVer
The protocol version

→ Coin
The total amount of stake in the system

→ Set (Credential 'Staking c)
The set of registered stake credentials

→ PoolRewardInfo c
Stake pool specific intermediate values needed to compute member rewards.

→ Credential 'Staking c
The stake credential whose reward is being calculated.

→ Coin
The stake controlled by the stake credential in the previous parameter above.

→ Maybe Coin
The reward for the given stake credential. This could be Nothing if the credential is no longer registered, if it is an owner, or if the reward is zero.

The stake pool member reward calculation

mkPoolRewardInfo ∷ EraPParams era ⇒ PParams era → Coin → BlocksMade (EraCrypto era) → Natural → Stake (EraCrypto era) → VMap VB VB (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)) → Map (KeyHash 'StakePool (EraCrypto era)) Coin → Coin → Coin → PoolParams (EraCrypto era) → Either StakeShare (PoolRewardInfo (EraCrypto era))

Calculate single stake pool specific values for the reward computation.

Note that if a stake pool has made no blocks in the given epoch, it will get no rewards, and so we do not need to return PoolRewardInfo. We do, however, need to return the relative stake of the pool in order to compute data for the stake pool ranking. Eventually we will remove the ranking information out of the ledger code and into a separate service, and at that point we can simplify this function to not care about ranking.