Cardano.Ledger.Shelley.Rules

Documentation

data ShelleyBBODY era

data ShelleyBbodyState era

Constructors

BbodyState (State (EraRule "LEDGERS" era)) (BlocksMade (EraCrypto era))

data BbodyEnv era

Constructors

BbodyEnv

Fields

bbodyPp :: PParams era

bbodyAccount :: AccountState

data ShelleyBbodyPredFailure era

Constructors

WrongBlockBodySizeBBODY !Int !Int

InvalidBodyHashBBODY !(Hash (EraCrypto era) EraIndependentBlockBody) !(Hash (EraCrypto era) EraIndependentBlockBody)

LedgersFailure (PredicateFailure (EraRule "LEDGERS" era))

newtype ShelleyBbodyEvent era

Constructors

LedgersEvent (Event (EraRule "LEDGERS" era))

type family PredicateFailure a

type family State a

data ShelleyDELEG era

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo :: SlotNo

ptr_ :: Ptr

acnt_ :: AccountState

ppDE :: PParams era

type family PredicateFailure a

data ShelleyDelegPredFailure era

Constructors

StakeKeyAlreadyRegisteredDELEG !(Credential 'Staking (EraCrypto era))

StakeKeyInRewardsDELEG !(Credential 'Staking (EraCrypto era))
Indicates that the stake key is somehow already in the rewards map. This error is now redundant with StakeKeyAlreadyRegisteredDELEG. We should remove it and replace its one use with StakeKeyAlreadyRegisteredDELEG.

StakeKeyNotRegisteredDELEG !(Credential 'Staking (EraCrypto era))

StakeKeyNonZeroAccountBalanceDELEG !(Maybe Coin)

StakeDelegationImpossibleDELEG !(Credential 'Staking (EraCrypto era))

WrongCertificateTypeDELEG

GenesisKeyNotInMappingDELEG !(KeyHash 'Genesis (EraCrypto era))

DuplicateGenesisDelegateDELEG !(KeyHash 'GenesisDelegate (EraCrypto era))

InsufficientForInstantaneousRewardsDELEG !MIRPot !Coin !Coin

MIRCertificateTooLateinEpochDELEG !SlotNo !SlotNo

DuplicateGenesisVRFDELEG !(Hash (EraCrypto era) (VerKeyVRF (EraCrypto era)))

MIRTransferNotCurrentlyAllowed

MIRNegativesNotCurrentlyAllowed

InsufficientForTransferDELEG !MIRPot !Coin !Coin

MIRProducesNegativeUpdate

MIRNegativeTransfer !MIRPot !Coin

data ShelleyDELEGS era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo :: !SlotNo

delegsIx :: !TxIx

delegspp :: !(PParams era)

delegsTx :: !(Tx era)

delegsAccount :: !AccountState

data ShelleyDelegsPredFailure era

Constructors

DelegateeNotRegisteredDELEG !(KeyHash 'StakePool (EraCrypto era))
Target pool which is not registered

WithdrawalsNotInRewardsDELEGS !(Map (RewardAcnt (EraCrypto era)) Coin)
Withdrawals that are missing or do not withdrawal the entire amount

DelplFailure !(PredicateFailure (EraRule "DELPL" era))
Subtransition Failures

newtype ShelleyDelegsEvent era

Constructors

DelplEvent (Event (EraRule "DELPL" era))

type family PredicateFailure a

validateZeroRewards :: forall era. DState era -> Withdrawals (EraCrypto era) -> Network -> Test (Map (RewardAcnt (EraCrypto era)) Coin)

validateStakePoolDelegateeRegistered :: PState era -> KeyHash 'StakePool (EraCrypto era) -> Test (KeyHash 'StakePool (EraCrypto era))

drainWithdrawals :: DState era -> Withdrawals (EraCrypto era) -> DState era

data ShelleyDELPL era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo :: SlotNo

delPlPtr :: Ptr

delPlPp :: PParams era

delPlAcnt :: AccountState

data ShelleyDelplPredFailure era

Constructors

PoolFailure (PredicateFailure (EraRule "POOL" era))

DelegFailure (PredicateFailure (EraRule "DELEG" era))

data ShelleyDelplEvent era

type family PredicateFailure a

data ShelleyEPOCH era

data ShelleyEpochPredFailure era

Constructors

PoolReapFailure (PredicateFailure (EraRule "POOLREAP" era))

SnapFailure (PredicateFailure (EraRule "SNAP" era))

UpecFailure (UpecPredFailure era)

data ShelleyEpochEvent era

Constructors

PoolReapEvent (Event (EraRule "POOLREAP" era))

SnapEvent (Event (EraRule "SNAP" era))

UpecEvent (Event (EraRule "UPEC" era))

type family PredicateFailure a

type UpecPredFailure era = UpecPredFailurePV (ProtVerLow era) era

data ShelleyLEDGER era

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo :: !SlotNo

ledgerIx :: !TxIx

ledgerPp :: !(PParams era)

ledgerAccount :: !AccountState

data ShelleyLedgerPredFailure era

Constructors

UtxowFailure (PredicateFailure (EraRule "UTXOW" era))

DelegsFailure (PredicateFailure (EraRule "DELEGS" era))

data ShelleyLedgerEvent era

Constructors

UtxowEvent (Event (EraRule "UTXOW" era))

DelegsEvent (Event (EraRule "DELEGS" era))

type family Event a

type family PredicateFailure a

epochFromSlot :: SlotNo -> Reader Globals EpochNo

depositEqualsObligation :: (EraTx era, ShelleyEraTxBody era, Environment t ~ LedgerEnv era, Signal t ~ Tx era, State t ~ LedgerState era) => AssertionViolation t -> String

shelleyLedgerAssertions :: (EraGov era, State (rule era) ~ LedgerState era) => [Assertion (rule era)]

data ShelleyLEDGERS era

data ShelleyLedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo :: SlotNo

ledgersPp :: PParams era

ledgersAccount :: AccountState

newtype ShelleyLedgersPredFailure era

Constructors

LedgerFailure (PredicateFailure (EraRule "LEDGER" era))

newtype ShelleyLedgersEvent era

Constructors

LedgerEvent (Event (EraRule "LEDGER" era))

type family PredicateFailure a

data ShelleyMIR era

type family PredicateFailure a

data ShelleyMirPredFailure era

data ShelleyMirEvent era

Constructors

MirTransfer (InstantaneousRewards (EraCrypto era))

NoMirTransfer (InstantaneousRewards (EraCrypto era)) Coin Coin
We were not able to perform an MIR transfer due to insufficient funds. This event gives the rewards we wanted to pay, plus the available reserves and treasury.

emptyInstantaneousRewards :: InstantaneousRewards c

data ShelleyNEWEPOCH era

data ShelleyNewEpochPredFailure era

Constructors

EpochFailure (PredicateFailure (EraRule "EPOCH" era))

CorruptRewardUpdate !(RewardUpdate (EraCrypto era))

MirFailure (PredicateFailure (EraRule "MIR" era))

data ShelleyNewEpochEvent era

Constructors

DeltaRewardEvent (Event (EraRule "RUPD" era))

RestrainedRewards EpochNo (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era)))) (Set (Credential 'Staking (EraCrypto era)))

TotalRewardEvent EpochNo (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))))

EpochEvent (Event (EraRule "EPOCH" era))

MirEvent (Event (EraRule "MIR" era))

TotalAdaPotsEvent AdaPots

type family PredicateFailure a

updateRewards :: EraGov era => EpochState era -> EpochNo -> RewardUpdate (EraCrypto era) -> Rule (ShelleyNEWEPOCH era) 'Transition (EpochState era)

calculatePoolDistr :: SnapShot c -> PoolDistr c

calculatePoolDistr' :: (KeyHash 'StakePool c -> Bool) -> SnapShot c -> PoolDistr c

data ShelleyNEWPP era

data ShelleyNewppState era

Constructors

NewppState (PParams era) (ShelleyGovState era)

data NewppEnv era

Constructors

NewppEnv (DState era) (PState era) (UTxOState era)

data ShelleyNewppPredFailure era

Constructors

UnexpectedDepositPot !Coin !Coin

type family PredicateFailure a

data ShelleyPOOL era

data PoolEvent era

Constructors

RegisterPool (KeyHash 'StakePool (EraCrypto era))

ReregisterPool (KeyHash 'StakePool (EraCrypto era))

data PoolEnv era

Constructors

PoolEnv !SlotNo !(PParams era)

type family PredicateFailure a

data ShelleyPoolPredFailure era

Constructors

StakePoolNotRegisteredOnKeyPOOL !(KeyHash 'StakePool (EraCrypto era))

StakePoolRetirementWrongEpochPOOL !EpochNo !EpochNo !EpochNo

StakePoolCostTooLowPOOL !Coin !Coin

WrongNetworkPOOL !Network !Network !(KeyHash 'StakePool (EraCrypto era))

PoolMedataHashTooBig !(KeyHash 'StakePool (EraCrypto era)) !Int

poolCertTransition :: forall (ledger :: Type -> Type) era c. (EraCrypto era ~ c, EraPParams era, State (ledger era) ~ PState era, STS (ledger era), Event (ledger era) ~ PoolEvent era, BaseM (ledger era) ~ ShelleyBase, PredicateFailure (ledger era) ~ ShelleyPoolPredFailure era) => PoolEnv era -> PState era -> PoolCert c -> TransitionRule (ledger era)

data ShelleyPOOLREAP era

data ShelleyPoolreapEvent era

Constructors

RetiredPools

Fields

refundPools :: Map (Credential 'Staking (EraCrypto era)) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

unclaimedPools :: Map (Credential 'Staking (EraCrypto era)) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

epochNo :: EpochNo

data ShelleyPoolreapState era

Constructors

PoolreapState

Fields

prUTxOSt :: UTxOState era

prAcnt :: AccountState

prDState :: DState era

prPState :: PState era

newtype ShelleyPoolreapEnv era

Constructors

ShelleyPoolreapEnv

Fields

speVState :: VState era

This enviroment field is only needed for assertions.

type family PredicateFailure a

data ShelleyPoolreapPredFailure era

data ShelleyPPUP era

data PpupEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs (EraCrypto era))

data ShelleyPpupPredFailure era

Constructors

NonGenesisUpdatePPUP !(Set (KeyHash 'Genesis (EraCrypto era))) !(Set (KeyHash 'Genesis (EraCrypto era)))
An update was proposed by a key hash that is not one of the genesis keys. The first set contains the key hashes which were a part of the update. The second set contains the key hashes of the genesis keys.

PPUpdateWrongEpoch !EpochNo !EpochNo !VotingPeriod
An update was proposed for the wrong epoch. The first EpochNo is the current epoch. The second EpochNo is the epoch listed in the update. The last parameter indicates if the update was intended for the current or the next epoch.

PVCannotFollowPPUP !ProtVer
An update was proposed which contains an invalid protocol version. New protocol versions must either increase the major number by exactly one and set the minor version to zero, or keep the major version the same and increase the minor version by exactly one.

data ShelleyGovState era

Constructors

ShelleyGovState

Fields

proposals :: !(ProposedPPUpdates era)

futureProposals :: !(ProposedPPUpdates era)

sgovPp :: !(PParams era)

sgovPrevPp :: !(PParams era)

newtype PpupEvent era

Constructors

NewEpoch EpochNo

type family PredicateFailure a

data VotingPeriod

Constructors

VoteForThisEpoch

VoteForNextEpoch

type PPUPPredFailure era = PPUPPredFailurePV (ProtVerLow era) era

data ShelleyRUPD era

data RupdEnv era

Constructors

RupdEnv (BlocksMade (EraCrypto era)) (EpochState era)

type family PredicateFailure a

data ShelleyRupdPredFailure era

epochInfoRange :: Monad m => EpochInfo m -> EpochNo -> m (SlotNo, SlotNo)

data PulsingRewUpdate c

The state used in the STS rules

Constructors

Pulsing !(RewardSnapShot c) !(Pulser c)

Complete !(RewardUpdate c)

startStep :: forall era. EraGov era => EpochSize -> BlocksMade (EraCrypto era) -> EpochState era -> Coin -> ActiveSlotCoeff -> Word64 -> PulsingRewUpdate (EraCrypto era)

pulseStep :: PulsingRewUpdate c -> ShelleyBase (PulsingRewUpdate c, RewardEvent c)

Run the pulser for a bit. If is has nothing left to do, complete it.

completeStep :: PulsingRewUpdate c -> ShelleyBase (PulsingRewUpdate c, RewardEvent c)

lift :: (MonadTrans t, Monad m) => m a -> t m a

Lift a computation from the argument monad to the constructed monad.

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity :: a

data RupdEvent c

Constructors

RupdEvent !EpochNo !(Map (Credential 'Staking c) (Set (Reward c)))

data ShelleySNAP era

type family PredicateFailure a

data ShelleySnapPredFailure era

newtype SnapEvent era

Constructors

StakeDistEvent (Map (Credential 'Staking (EraCrypto era)) (Coin, KeyHash 'StakePool (EraCrypto era)))

data SnapEnv era

Constructors

SnapEnv !(LedgerState era) !(PParams era)

data ShelleyTICK era

type family State a

data ShelleyTickPredFailure era

Constructors

NewEpochFailure (PredicateFailure (EraRule "NEWEPOCH" era))

RupdFailure (PredicateFailure (EraRule "RUPD" era))

data ShelleyTickEvent era

Constructors

TickNewEpochEvent (Event (EraRule "NEWEPOCH" era))

TickRupdEvent (Event (EraRule "RUPD" era))

type family PredicateFailure a

adoptGenesisDelegs :: EpochState era -> SlotNo -> EpochState era

data ShelleyTICKF era

data ShelleyTickfPredFailure era

validatingTickTransition :: forall tick era. (Embed (EraRule "NEWEPOCH" era) (tick era), STS (tick era), State (tick era) ~ NewEpochState era, BaseM (tick era) ~ ShelleyBase, Environment (EraRule "NEWEPOCH" era) ~ (), State (EraRule "NEWEPOCH" era) ~ NewEpochState era, Signal (EraRule "NEWEPOCH" era) ~ EpochNo) => NewEpochState era -> SlotNo -> TransitionRule (tick era)

This is a limited version of bheadTransition which is suitable for the future ledger view.

validatingTickTransitionFORECAST :: forall tick era. (State (tick era) ~ NewEpochState era, BaseM (tick era) ~ ShelleyBase, State (EraRule "UPEC" era) ~ UpecState era, Signal (EraRule "UPEC" era) ~ (), Environment (EraRule "UPEC" era) ~ EpochState era, Embed (EraRule "UPEC" era) (tick era), STS (tick era), GovState era ~ ShelleyGovState era, EraGov era) => NewEpochState era -> SlotNo -> TransitionRule (tick era)

This is a limited version of validatingTickTransition which is only suitable for the future ledger view.

data ShelleyUPEC era

data UpecState era

Constructors

UpecState

Fields

currentPp :: !(PParams era)

Current protocol parameters.

ppupState :: !(ShelleyGovState era)

State of the protocol update transition system.

newtype ShelleyUpecPredFailure era

Constructors

NewPpFailure (PredicateFailure (ShelleyNEWPP era))

votedValue

Arguments

:: forall era. EraPParams era

=> ProposedPPUpdates era

-> PParams era
Protocol parameters to which the change will be applied.

-> Int
Quorum needed to change the protocol parameters.

-> Maybe (PParams era)

If at least n nodes voted to change the same protocol parameters to the same values, return the given protocol parameters updated to these values. Here n is the quorum needed.

data ShelleyUTXO era

data UtxoEnv era

Constructors

UtxoEnv

Fields

ueSlot :: SlotNo

uePParams :: PParams era

ueCertState :: CertState era

ueGenDelegs :: GenDelegs (EraCrypto era)

data ShelleyUtxoPredFailure era

Constructors

BadInputsUTxO !(Set (TxIn (EraCrypto era)))

ExpiredUTxO !SlotNo !SlotNo

MaxTxSizeUTxO !Integer !Integer

InputSetEmptyUTxO

FeeTooSmallUTxO !Coin !Coin

ValueNotConservedUTxO !(Value era) !(Value era)

WrongNetwork !Network !(Set (Addr (EraCrypto era)))

WrongNetworkWithdrawal !Network !(Set (RewardAcnt (EraCrypto era)))

OutputTooSmallUTxO ![TxOut era]

UpdateFailure (PPUPPredFailure era)

OutputBootAddrAttrsTooBig ![TxOut era]

data UtxoEvent era

Constructors

TotalDeposits (SafeHash (EraCrypto era) EraIndependentTxBody) Coin

UpdateEvent (Event (EraRule "PPUP" era))

type family PredicateFailure a

updateUTxOState :: EraTxBody era => PParams era -> UTxOState era -> TxBody era -> Coin -> GovState era -> UTxOState era

Validations

validateInputSetEmptyUTxO :: EraTxBody era => TxBody era -> Test (ShelleyUtxoPredFailure era)

Ensure that there is at least one input in the TxBody

txins txb ≠ ∅

validateFeeTooSmallUTxO :: EraTx era => PParams era -> Tx era -> Test (ShelleyUtxoPredFailure era)

Ensure that the fee is at least the amount specified by the minfee

minfee pp tx ≤ txfee txb

validateBadInputsUTxO :: UTxO era -> Set (TxIn (EraCrypto era)) -> Test (ShelleyUtxoPredFailure era)

Ensure all transaction inputs are present in UTxO

inputs ⊆ dom utxo

validateWrongNetwork :: (EraTxOut era, Foldable f) => Network -> f (TxOut era) -> Test (ShelleyUtxoPredFailure era)

Make sure all addresses match the supplied NetworkId

∀(_ → (a, _)) ∈ txouts txb, netId a = NetworkId

validateWrongNetworkWithdrawal :: ShelleyEraTxBody era => Network -> TxBody era -> Test (ShelleyUtxoPredFailure era)

Make sure all addresses match the supplied NetworkId

∀(a → ) ∈ txwdrls txb, netId a = NetworkId

validateOutputBootAddrAttrsTooBig :: (EraTxOut era, Foldable f) => f (TxOut era) -> Test (ShelleyUtxoPredFailure era)

Bootstrap (i.e. Byron) addresses have variable sized attributes in them. It is important to limit their overall size.

∀ ( _ ↦ (a,_)) ∈ txoutstxb,  a ∈ Addrbootstrap → bootstrapAttrsSize a ≤ 64

validateMaxTxSizeUTxO :: EraTx era => PParams era -> Tx era -> Test (ShelleyUtxoPredFailure era)

Ensure that the size of the transaction does not exceed the maxTxSize protocol parameter

txsize tx ≤ maxTxSize pp

validateValueNotConservedUTxO :: (ShelleyEraTxBody era, EraUTxO era) => PParams era -> UTxO era -> CertState era -> TxBody era -> Test (ShelleyUtxoPredFailure era)

Ensure that value consumed and produced matches up exactly

consumed pp utxo txb = produced pp poolParams txb

utxoEnvSlotL :: Lens' (UtxoEnv era) SlotNo

utxoEnvPParamsL :: Lens' (UtxoEnv era) (PParams era)

utxoEnvCertStateL :: Lens' (UtxoEnv era) (CertState era)

utxoEnvGenDelegsL :: Lens' (UtxoEnv era) (GenDelegs (EraCrypto era))

data ShelleyUTXOW era

data ShelleyUtxowPredFailure era

Constructors

InvalidWitnessesUTXOW ![VKey 'Witness (EraCrypto era)]

MissingVKeyWitnessesUTXOW !(Set (KeyHash 'Witness (EraCrypto era)))

MissingScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))

ScriptWitnessNotValidatingUTXOW !(Set (ScriptHash (EraCrypto era)))

UtxoFailure (PredicateFailure (EraRule "UTXO" era))

MIRInsufficientGenesisSigsUTXOW (Set (KeyHash 'Witness (EraCrypto era)))

MissingTxBodyMetadataHash !(AuxiliaryDataHash (EraCrypto era))

MissingTxMetadata !(AuxiliaryDataHash (EraCrypto era))

ConflictingMetadataHash !(AuxiliaryDataHash (EraCrypto era)) !(AuxiliaryDataHash (EraCrypto era))

InvalidMetadata

ExtraneousScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))

newtype ShelleyUtxowEvent era

Constructors

UtxoEvent (Event (EraRule "UTXO" era))

type family PredicateFailure a

transitionRulesUTXOW :: forall era utxow. (EraUTxO era, ShelleyEraTxBody era, ProtVerAtMost era 8, ScriptsNeeded era ~ ShelleyScriptsNeeded era, BaseM (utxow era) ~ ShelleyBase, Embed (EraRule "UTXO" era) (utxow era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, Environment (utxow era) ~ UtxoEnv era, State (utxow era) ~ UTxOState era, Signal (utxow era) ~ Tx era, PredicateFailure (utxow era) ~ ShelleyUtxowPredFailure era, STS (utxow era), DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody)) => TransitionRule (utxow era)

A generic Utxow witnessing function designed to be used across many Eras. Note the embed argument lifts from the simple Shelley (ShelleyUtxowPredFailure) to the PredicateFailure (type family) of the context of where it is called.

shelleyWitsVKeyNeeded :: forall era. (EraTx era, ShelleyEraTxBody era, ProtVerAtMost era 8) => UTxO era -> TxBody era -> GenDelegs (EraCrypto era) -> Set (KeyHash 'Witness (EraCrypto era))

witsVKeyNeededGov :: forall era. (ShelleyEraTxBody era, ProtVerAtMost era 8) => TxBody era -> GenDelegs (EraCrypto era) -> Set (KeyHash 'Witness (EraCrypto era))

Collect the set of hashes of keys that needs to sign a given transaction. This set consists of the txin owners, certificate authors, and withdrawal reward accounts.

witsVKeyNeededNoGov :: forall era. EraTx era => UTxO era -> TxBody era -> Set (KeyHash 'Witness (EraCrypto era))

Individual validation steps

validateFailedNativeScripts :: EraTx era => ScriptsProvided era -> Tx era -> Test (ShelleyUtxowPredFailure era)

validateMissingScripts :: EraPParams era => PParams era -> ShelleyScriptsNeeded era -> ScriptsProvided era -> Test (ShelleyUtxowPredFailure era)

validateVerifiedWits :: (EraTx era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody)) => Tx era -> Test (ShelleyUtxowPredFailure era)

Given a ledger state, determine if the UTxO witnesses in a given transaction are correct.

validateMetadata :: EraTx era => PParams era -> Tx era -> Test (ShelleyUtxowPredFailure era)

check metadata hash ((adh = ◇) ∧ (ad= ◇)) ∨ (adh = hashAD ad)

validateMIRInsufficientGenesisSigs :: (EraTx era, ShelleyEraTxBody era, ProtVerAtMost era 8) => GenDelegs (EraCrypto era) -> Word64 -> Set (KeyHash 'Witness (EraCrypto era)) -> Tx era -> Test (ShelleyUtxowPredFailure era)

check genesis keys signatures for instantaneous rewards certificates

genSig := { hashKey gkey | gkey ∈ dom(genDelegs)} ∩ witsKeyHashes { c ∈ txcerts txb ∩ TxCert_mir} ≠ ∅ ⇒ |genSig| ≥ Quorum

validateNeededWitnesses :: Set (KeyHash 'Witness (EraCrypto era)) -> Set (KeyHash 'Witness (EraCrypto era)) -> Test (ShelleyUtxowPredFailure era)

propWits :: Maybe (Update era) -> GenDelegs (EraCrypto era) -> Set (KeyHash 'Witness (EraCrypto era))

Deprecated: This is will become an internal function in the future. Submit an issue if you still need it.

Calculate the set of hash keys of the required witnesses for update proposals.