Cardano.Ledger.Shelley.Rules.Bbody

Documentation

data BBODY era

data BbodyState era

Constructors

BbodyState (LedgerState era) (BlocksMade (Crypto era))

data BbodyEnv era

Constructors

BbodyEnv

Fields

bbodyPp :: PParams era

bbodyAccount :: AccountState

data BbodyPredicateFailure era

Constructors

WrongBlockBodySizeBBODY !Int !Int

InvalidBodyHashBBODY !(Hash (Crypto era) EraIndependentBlockBody) !(Hash (Crypto era) EraIndependentBlockBody)

LedgersFailure (PredicateFailure (EraRule "LEDGERS" era))

newtype BbodyEvent era

Constructors

LedgersEvent (Event (EraRule "LEDGERS" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

type family State a

Type of the state which the system transitions between.