Cardano.Ledger.Shelley.Rules.Deleg

Documentation

data DELEG era

data DelegEnv era

Constructors

DelegEnv

Fields

slotNo :: SlotNo

ptr_ :: Ptr

acnt_ :: AccountState

ppDE :: PParams era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data DelegPredicateFailure era

Constructors

StakeKeyAlreadyRegisteredDELEG !(Credential 'Staking (Crypto era))

StakeKeyInRewardsDELEG !(Credential 'Staking (Crypto era))
Indicates that the stake key is somehow already in the rewards map. This error is now redundant with StakeKeyAlreadyRegisteredDELEG. We should remove it and replace its one use with StakeKeyAlreadyRegisteredDELEG.

StakeKeyNotRegisteredDELEG !(Credential 'Staking (Crypto era))

StakeKeyNonZeroAccountBalanceDELEG !(Maybe Coin)

StakeDelegationImpossibleDELEG !(Credential 'Staking (Crypto era))

WrongCertificateTypeDELEG

GenesisKeyNotInMappingDELEG !(KeyHash 'Genesis (Crypto era))

DuplicateGenesisDelegateDELEG !(KeyHash 'GenesisDelegate (Crypto era))

InsufficientForInstantaneousRewardsDELEG !MIRPot !Coin !Coin

MIRCertificateTooLateinEpochDELEG !SlotNo !SlotNo

DuplicateGenesisVRFDELEG !(Hash (Crypto era) (VerKeyVRF (Crypto era)))

MIRTransferNotCurrentlyAllowed

MIRNegativesNotCurrentlyAllowed

InsufficientForTransferDELEG !MIRPot !Coin !Coin

MIRProducesNegativeUpdate

MIRNegativeTransfer !MIRPot !Coin