Cardano.Ledger.Shelley.Rules.Delegs

Documentation

data DELEGS era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo :: !SlotNo

delegsIx :: !TxIx

delegspp :: !(PParams era)

delegsTx :: !(Tx era)

delegsAccount :: !AccountState

data DelegsPredicateFailure era

Constructors

DelegateeNotRegisteredDELEG !(KeyHash 'StakePool (Crypto era))

WithdrawalsNotInRewardsDELEGS !(Map (RewardAcnt (Crypto era)) Coin)

DelplFailure (PredicateFailure (EraRule "DELPL" era))

newtype DelegsEvent era

Constructors

DelplEvent (Event (EraRule "DELPL" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.