Cardano.Ledger.Shelley.Rules.Ledger

Documentation

data LEDGER era

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo :: !SlotNo

ledgerIx :: !TxIx

ledgerPp :: !(PParams era)

ledgerAccount :: !AccountState

data LedgerPredicateFailure era

Constructors

UtxowFailure (PredicateFailure (EraRule "UTXOW" era))

DelegsFailure (PredicateFailure (EraRule "DELEGS" era))

data LedgerEvent era

Constructors

UtxowEvent (Event (EraRule "UTXOW" era))

DelegsEvent (Event (EraRule "DELEGS" era))

type family Event a

Event type.

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.