Cardano.Ledger.Shelley.Rules.Ledger

Documentation

data ShelleyLEDGER era

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo ∷ !SlotNo

ledgerIx ∷ !TxIx

ledgerPp ∷ !(PParams era)

ledgerAccount ∷ !AccountState

data ShelleyLedgerPredFailure era

Constructors

UtxowFailure (PredicateFailure (EraRule "UTXOW" era))

DelegsFailure (PredicateFailure (EraRule "DELEGS" era))

data ShelleyLedgerEvent era

Constructors

UtxowEvent (Event (EraRule "UTXOW" era))

DelegsEvent (Event (EraRule "DELEGS" era))

type family Event a

Event type.

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

epochFromSlot ∷ SlotNo → Reader Globals EpochNo

depositEqualsObligation ∷ (EraTx era, ShelleyEraTxBody era, Environment t ~ LedgerEnv era, Signal t ~ Tx era, State t ~ LedgerState era) ⇒ AssertionViolation t → String

Orphan instances

(DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), EraTx era, ShelleyEraTxBody era, Embed (EraRule "DELEGS" era) (ShelleyLEDGER era), Embed (EraRule "UTXOW" era) (ShelleyLEDGER era), Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Signal (EraRule "UTXOW" era) ~ Tx era, Environment (EraRule "DELEGS" era) ~ DelegsEnv era, State (EraRule "DELEGS" era) ~ CertState era, Signal (EraRule "DELEGS" era) ~ Seq (DCert (EraCrypto era)), ProtVerAtMost era 8) ⇒ STS (ShelleyLEDGER era)

Instance details

Associated Types

type State (ShelleyLEDGER era)

type Signal (ShelleyLEDGER era)

type Environment (ShelleyLEDGER era)

type BaseM (ShelleyLEDGER era) ∷ Type → Type

type Event (ShelleyLEDGER era)

type PredicateFailure (ShelleyLEDGER era)

Methods

initialRules ∷ [InitialRule (ShelleyLEDGER era)]

transitionRules ∷ [TransitionRule (ShelleyLEDGER era)]

assertions ∷ [Assertion (ShelleyLEDGER era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyLEDGER era) → String

(Era era, STS (ShelleyDELEGS era), PredicateFailure (EraRule "DELEGS" era) ~ ShelleyDelegsPredFailure era, Event (EraRule "DELEGS" era) ~ ShelleyDelegsEvent era) ⇒ Embed (ShelleyDELEGS era) (ShelleyLEDGER era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyDELEGS era) → PredicateFailure (ShelleyLEDGER era)

wrapEvent ∷ Event (ShelleyDELEGS era) → Event (ShelleyLEDGER era)

(STS (ShelleyUTXOW era), PredicateFailure (EraRule "UTXOW" era) ~ ShelleyUtxowPredFailure era, Event (EraRule "UTXOW" era) ~ Event (ShelleyUTXOW era)) ⇒ Embed (ShelleyUTXOW era) (ShelleyLEDGER era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyUTXOW era) → PredicateFailure (ShelleyLEDGER era)

wrapEvent ∷ Event (ShelleyUTXOW era) → Event (ShelleyLEDGER era)