Cardano.Ledger.Shelley.Rules.NewEpoch

Documentation

data ShelleyNEWEPOCH era

data ShelleyNewEpochPredFailure era

Constructors

EpochFailure (PredicateFailure (EraRule "EPOCH" era))

CorruptRewardUpdate !(RewardUpdate (EraCrypto era))

MirFailure (PredicateFailure (EraRule "MIR" era))

data ShelleyNewEpochEvent era

Constructors

DeltaRewardEvent (Event (EraRule "RUPD" era))

RestrainedRewards EpochNo (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era)))) (Set (Credential 'Staking (EraCrypto era)))

TotalRewardEvent EpochNo (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))))

EpochEvent (Event (EraRule "EPOCH" era))

MirEvent (Event (EraRule "MIR" era))

TotalAdaPotsEvent AdaPots

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

updateRewards ∷ EraPParams era ⇒ EpochState era → EpochNo → RewardUpdate (EraCrypto era) → Rule (ShelleyNEWEPOCH era) 'Transition (EpochState era)

calculatePoolDistr ∷ SnapShot c → PoolDistr c

calculatePoolDistr' ∷ (KeyHash 'StakePool c → Bool) → SnapShot c → PoolDistr c

Orphan instances

(EraTxOut era, EraGovernance era, Embed (EraRule "MIR" era) (ShelleyNEWEPOCH era), Embed (EraRule "EPOCH" era) (ShelleyNEWEPOCH era), Environment (EraRule "MIR" era) ~ (), State (EraRule "MIR" era) ~ EpochState era, Signal (EraRule "MIR" era) ~ (), Event (EraRule "RUPD" era) ~ RupdEvent (EraCrypto era), Environment (EraRule "EPOCH" era) ~ (), State (EraRule "EPOCH" era) ~ EpochState era, Signal (EraRule "EPOCH" era) ~ EpochNo, Default (EpochState era), Default (State (EraRule "PPUP" era)), Default (PParams era), Default (StashedAVVMAddresses era)) ⇒ STS (ShelleyNEWEPOCH era)

Instance details

Associated Types

type State (ShelleyNEWEPOCH era)

type Signal (ShelleyNEWEPOCH era)

type Environment (ShelleyNEWEPOCH era)

type BaseM (ShelleyNEWEPOCH era) ∷ Type → Type

type Event (ShelleyNEWEPOCH era)

type PredicateFailure (ShelleyNEWEPOCH era)

Methods

initialRules ∷ [InitialRule (ShelleyNEWEPOCH era)]

transitionRules ∷ [TransitionRule (ShelleyNEWEPOCH era)]

assertions ∷ [Assertion (ShelleyNEWEPOCH era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyNEWEPOCH era) → String

(STS (ShelleyEPOCH era), PredicateFailure (EraRule "EPOCH" era) ~ ShelleyEpochPredFailure era, Event (EraRule "EPOCH" era) ~ ShelleyEpochEvent era) ⇒ Embed (ShelleyEPOCH era) (ShelleyNEWEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyEPOCH era) → PredicateFailure (ShelleyNEWEPOCH era)

wrapEvent ∷ Event (ShelleyEPOCH era) → Event (ShelleyNEWEPOCH era)

(Era era, Default (EpochState era), PredicateFailure (EraRule "MIR" era) ~ ShelleyMirPredFailure era, Event (EraRule "MIR" era) ~ ShelleyMirEvent era) ⇒ Embed (ShelleyMIR era) (ShelleyNEWEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyMIR era) → PredicateFailure (ShelleyNEWEPOCH era)

wrapEvent ∷ Event (ShelleyMIR era) → Event (ShelleyNEWEPOCH era)