Cardano.Ledger.Shelley.Rules.Newpp

Documentation

data NEWPP era

data NewppState era

Constructors

NewppState (PParams era) (PPUPState era)

data NewppEnv era

Constructors

NewppEnv (DState (Crypto era)) (PState (Crypto era)) (UTxOState era) AccountState

data NewppPredicateFailure era

Constructors

UnexpectedDepositPot !Coin !Coin

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.