Cardano.Ledger.Shelley.Rules.Ppup

Documentation

data PPUP era

data PPUPEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs (Crypto era))

data PpupPredicateFailure era

Constructors

NonGenesisUpdatePPUP !(Set (KeyHash 'Genesis (Crypto era))) !(Set (KeyHash 'Genesis (Crypto era)))

PPUpdateWrongEpoch !EpochNo !EpochNo !VotingPeriod

PVCannotFollowPPUP !ProtVer

newtype PpupEvent era

Constructors

NewEpoch EpochNo

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data VotingPeriod

Constructors

VoteForThisEpoch

VoteForNextEpoch