Cardano.Ledger.Shelley.Rules.Rupd

Documentation

data RUPD era

data RupdEnv era

Constructors

RupdEnv (BlocksMade (Crypto era)) (EpochState era)

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data RupdPredicateFailure era

epochInfoRange :: Monad m => EpochInfo m -> EpochNo -> m (SlotNo, SlotNo)

data PulsingRewUpdate crypto

The state used in the STS rules

Constructors

Pulsing !(RewardSnapShot crypto) !(Pulser crypto)

Complete !(RewardUpdate crypto)

startStep :: forall era. UsesPP era => EpochSize -> BlocksMade (Crypto era) -> EpochState era -> Coin -> ActiveSlotCoeff -> Word64 -> (PulsingRewUpdate (Crypto era), RewardProvenance (Crypto era))

Assemble the components for, and then create, a Pulser.

pulseStep :: PulsingRewUpdate crypto -> ShelleyBase (PulsingRewUpdate crypto, RewardEvent crypto)

Run the pulser for a bit. If is has nothing left to do, complete it.

completeStep :: PulsingRewUpdate crypto -> ShelleyBase (PulsingRewUpdate crypto, RewardEvent crypto)

lift :: (MonadTrans t, Monad m) => m a -> t m a

Lift a computation from the argument monad to the constructed monad.

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity :: a

data RupdEvent crypto

Constructors

RupdEvent !EpochNo !(Map (Credential 'Staking crypto) (Set (Reward crypto)))