Cardano.Ledger.Shelley.Rules.Tick

Documentation

data TICK era

type family State a

Type of the state which the system transitions between.

data TickPredicateFailure era

Constructors

NewEpochFailure (PredicateFailure (EraRule "NEWEPOCH" era))

RupdFailure (PredicateFailure (EraRule "RUPD" era))

data TickEvent era

Constructors

NewEpochEvent (Event (EraRule "NEWEPOCH" era))

RupdEvent (Event (EraRule "RUPD" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

adoptGenesisDelegs :: EpochState era -> SlotNo -> EpochState era

data TICKF era

newtype TickfPredicateFailure era

Constructors

TickfNewEpochFailure (PredicateFailure (EraRule "NEWEPOCH" era))