Cardano.Ledger.Shelley.Rules.Upec

Description

Epoch change registration.

The rules of this module determine how the update subsystem of the ledger handles the epoch transitions.

Documentation

data UPEC era

Update epoch change

newtype UpecPredicateFailure era

Constructors

NewPpFailure (PredicateFailure (NEWPP era))

votedValue

Arguments

:: forall era. UsesPParams era

=> ProposedPPUpdates era

-> PParams era
Protocol parameters to which the change will be applied.

-> Int
Quorum needed to change the protocol parameters.

-> Maybe (PParams era)

If at least n nodes voted to change the same protocol parameters to the same values, return the given protocol parameters updated to these values. Here n is the quorum needed.