Cardano.Ledger.Shelley.Rules.Upec

Description

Epoch change registration.

The rules of this module determine how the update subsystem of the ledger handles the epoch transitions.

Documentation

data ShelleyUPEC era

data UpecState era

Constructors

UpecState

Fields

currentPp ∷ !(PParams era)

Current protocol parameters.

ppupState ∷ !(ShelleyPPUPState era)

State of the protocol update transition system.

newtype ShelleyUpecPredFailure era

Constructors

NewPpFailure (PredicateFailure (ShelleyNEWPP era))

votedValue

Arguments

∷ ∀ era. EraPParams era

⇒ ProposedPPUpdates era

→ PParams era
Protocol parameters to which the change will be applied.

→ Int
Quorum needed to change the protocol parameters.

→ Maybe (PParams era)

If at least n nodes voted to change the same protocol parameters to the same values, return the given protocol parameters updated to these values. Here n is the quorum needed.

Orphan instances

(EraPParams era, Default (PParams era), GovernanceState era ~ ShelleyPPUPState era) ⇒ STS (ShelleyUPEC era)

Instance details

Associated Types

type State (ShelleyUPEC era)

type Signal (ShelleyUPEC era)

type Environment (ShelleyUPEC era)

type BaseM (ShelleyUPEC era) ∷ Type → Type

type Event (ShelleyUPEC era)

type PredicateFailure (ShelleyUPEC era)

Methods

initialRules ∷ [InitialRule (ShelleyUPEC era)]

transitionRules ∷ [TransitionRule (ShelleyUPEC era)]

assertions ∷ [Assertion (ShelleyUPEC era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyUPEC era) → String

(Era era, STS (ShelleyNEWPP era)) ⇒ Embed (ShelleyNEWPP era) (ShelleyUPEC era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyNEWPP era) → PredicateFailure (ShelleyUPEC era)

wrapEvent ∷ Event (ShelleyNEWPP era) → Event (ShelleyUPEC era)