Cardano.Ledger.Shelley.Rules.Utxow

Documentation

data ShelleyUTXOW era

data ShelleyUtxowPredFailure era

Constructors

InvalidWitnessesUTXOW ![VKey 'Witness (EraCrypto era)]

MissingVKeyWitnessesUTXOW !(Set (KeyHash 'Witness (EraCrypto era)))

MissingScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))

ScriptWitnessNotValidatingUTXOW !(Set (ScriptHash (EraCrypto era)))

UtxoFailure (PredicateFailure (EraRule "UTXO" era))

MIRInsufficientGenesisSigsUTXOW (Set (KeyHash 'Witness (EraCrypto era)))

MissingTxBodyMetadataHash !(AuxiliaryDataHash (EraCrypto era))

MissingTxMetadata !(AuxiliaryDataHash (EraCrypto era))

ConflictingMetadataHash !(AuxiliaryDataHash (EraCrypto era)) !(AuxiliaryDataHash (EraCrypto era))

InvalidMetadata

ExtraneousScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))

newtype ShelleyUtxowEvent era

Constructors

UtxoEvent (Event (EraRule "UTXO" era))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

transitionRulesUTXOW ∷ ∀ era utxow. (EraTx era, EraUTxO era, ShelleyEraTxBody era, ScriptsNeeded era ~ ShelleyScriptsNeeded era, BaseM (utxow era) ~ ShelleyBase, Embed (EraRule "UTXO" era) (utxow era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, Environment (utxow era) ~ UtxoEnv era, State (utxow era) ~ UTxOState era, Signal (utxow era) ~ Tx era, PredicateFailure (utxow era) ~ ShelleyUtxowPredFailure era, STS (utxow era), DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), ProtVerAtMost era 8) ⇒ TransitionRule (utxow era)

A generic Utxow witnessing function designed to be used across many Eras. Note the embed argument lifts from the simple Shelley (ShelleyUtxowPredFailure) to the PredicateFailure (type family) of the context of where it is called.

Individual validation steps

validateFailedScripts ∷ ∀ era. EraTx era ⇒ Tx era → Test (ShelleyUtxowPredFailure era)

validateMissingScripts ∷ EraPParams era ⇒ PParams era → ShelleyScriptsNeeded era → Set (ScriptHash (EraCrypto era)) → Test (ShelleyUtxowPredFailure era)

validateVerifiedWits ∷ (EraTx era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody)) ⇒ Tx era → Test (ShelleyUtxowPredFailure era)

Given a ledger state, determine if the UTxO witnesses in a given transaction are correct.

validateMetadata ∷ EraTx era ⇒ PParams era → Tx era → Test (ShelleyUtxowPredFailure era)

check metadata hash ((adh = ◇) ∧ (ad= ◇)) ∨ (adh = hashAD ad)

validateMIRInsufficientGenesisSigs ∷ (EraTx era, ShelleyEraTxBody era) ⇒ GenDelegs (EraCrypto era) → Word64 → Set (KeyHash 'Witness (EraCrypto era)) → Tx era → Test (ShelleyUtxowPredFailure era)

check genesis keys signatures for instantaneous rewards certificates

genSig := { hashKey gkey | gkey ∈ dom(genDelegs)} ∩ witsKeyHashes { c ∈ txcerts txb ∩ DCert_mir} ≠ ∅ ⇒ |genSig| ≥ Quorum

validateNeededWitnesses ∷ (UTxO era → Tx era → GenDelegs (EraCrypto era) → Set (KeyHash 'Witness (EraCrypto era))) → GenDelegs (EraCrypto era) → UTxO era → Tx era → Set (KeyHash 'Witness (EraCrypto era)) → Test (ShelleyUtxowPredFailure era)

propWits ∷ Maybe (Update era) → GenDelegs (EraCrypto era) → Set (KeyHash 'Witness (EraCrypto era))

Calculate the set of hash keys of the required witnesses for update proposals.

Orphan instances

(EraTx era, EraUTxO era, ShelleyEraTxBody era, Tx era ~ ShelleyTx era, ScriptsNeeded era ~ ShelleyScriptsNeeded era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), Embed (EraRule "UTXO" era) (ShelleyUTXOW era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), ProtVerAtMost era 8) ⇒ STS (ShelleyUTXOW era)

Instance details

Associated Types

type State (ShelleyUTXOW era)

type Signal (ShelleyUTXOW era)

type Environment (ShelleyUTXOW era)

type BaseM (ShelleyUTXOW era) ∷ Type → Type

type Event (ShelleyUTXOW era)

type PredicateFailure (ShelleyUTXOW era)

Methods

initialRules ∷ [InitialRule (ShelleyUTXOW era)]

transitionRules ∷ [TransitionRule (ShelleyUTXOW era)]

assertions ∷ [Assertion (ShelleyUTXOW era)]

renderAssertionViolation ∷ AssertionViolation (ShelleyUTXOW era) → String

(Era era, STS (ShelleyUTXO era), PredicateFailure (EraRule "UTXO" era) ~ ShelleyUtxoPredFailure era, Event (EraRule "UTXO" era) ~ UtxoEvent era) ⇒ Embed (ShelleyUTXO era) (ShelleyUTXOW era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyUTXO era) → PredicateFailure (ShelleyUTXOW era)

wrapEvent ∷ Event (ShelleyUTXO era) → Event (ShelleyUTXOW era)