Cardano.Ledger.Shelley.Scripts

Documentation

data MultiSig era

class EraScript era ⇒ ShelleyEraScript era where

Methods

mkRequireSignature ∷ KeyHash 'Witness (EraCrypto era) → NativeScript era

getRequireSignature ∷ NativeScript era → Maybe (KeyHash 'Witness (EraCrypto era))

mkRequireAllOf ∷ StrictSeq (NativeScript era) → NativeScript era

getRequireAllOf ∷ NativeScript era → Maybe (StrictSeq (NativeScript era))

mkRequireAnyOf ∷ StrictSeq (NativeScript era) → NativeScript era

getRequireAnyOf ∷ NativeScript era → Maybe (StrictSeq (NativeScript era))

mkRequireMOf ∷ Int → StrictSeq (NativeScript era) → NativeScript era

getRequireMOf ∷ NativeScript era → Maybe (Int, StrictSeq (NativeScript era))

pattern RequireSignature ∷ ShelleyEraScript era ⇒ KeyHash 'Witness (EraCrypto era) → NativeScript era

pattern RequireAllOf ∷ ShelleyEraScript era ⇒ StrictSeq (NativeScript era) → NativeScript era

pattern RequireAnyOf ∷ ShelleyEraScript era ⇒ StrictSeq (NativeScript era) → NativeScript era

pattern RequireMOf ∷ ShelleyEraScript era ⇒ Int → StrictSeq (NativeScript era) → NativeScript era

evalMultiSig ∷ (ShelleyEraScript era, NativeScript era ~ MultiSig era) ⇒ Set (KeyHash 'Witness (EraCrypto era)) → NativeScript era → Bool

Script evaluator for native multi-signature scheme. vhks is the set of key hashes that signed the transaction to be validated.

validateMultiSig ∷ (ShelleyEraScript era, EraTx era, NativeScript era ~ MultiSig era) ⇒ Tx era → NativeScript era → Bool

Script validator for native multi-signature scheme.

newtype ScriptHash c

Constructors

ScriptHash (Hash (ADDRHASH c) EraIndependentScript)

nativeMultiSigTag ∷ ByteString

Magic number "memorialized" in the ValidateScript class under the method: scriptPrefixTag:: Core.Script era -> Bs.ByteString, for the Shelley Era.

eqMultiSigRaw ∷ MultiSig era → MultiSig era → Bool

Check the equality of two underlying types, while ignoring their binary representation, which Eq instance normally does. This is used for testing.

data MultiSigRaw era

A simple language for expressing conditions under which it is valid to withdraw from a normal UTxO payment address or to use a stake address.

The use case is for expressing multi-signature payment addresses and multi-signature stake addresses. These can be combined arbitrarily using logical operations:

multi-way "and";

multi-way "or";

multi-way "N of M".

This makes it easy to express multi-signature addresses, and provides an extension point to express other validity conditions, e.g., as needed for locking funds used with lightning.

Orphan instances

Crypto c ⇒ EraScript (ShelleyEra c)

Instance details

Associated Types

type Script (ShelleyEra c) = (r ∷ Type)

type NativeScript (ShelleyEra c) = (r ∷ Type)

Methods

upgradeScript ∷ Script (PreviousEra (ShelleyEra c)) → Script (ShelleyEra c)

scriptPrefixTag ∷ Script (ShelleyEra c) → ByteString

getNativeScript ∷ Script (ShelleyEra c) → Maybe (NativeScript (ShelleyEra c))

fromNativeScript ∷ NativeScript (ShelleyEra c) → Script (ShelleyEra c)