Cardano.Ledger.Shelley.Scripts

Documentation

data MultiSig era where

Bundled Patterns

pattern RequireAllOf :: Era era => [MultiSig era] -> MultiSig era

pattern RequireAnyOf :: Era era => [MultiSig era] -> MultiSig era

pattern RequireSignature :: Era era => KeyHash 'Witness (EraCrypto era) -> MultiSig era

pattern RequireMOf :: Era era => Int -> [MultiSig era] -> MultiSig era

evalMultiSig :: Era era => Set (KeyHash 'Witness (EraCrypto era)) -> MultiSig era -> Bool

Script evaluator for native multi-signature scheme. vhks is the set of key hashes that signed the transaction to be validated.

validateMultiSig :: EraTx era => Tx era -> MultiSig era -> Bool

Script validator for native multi-signature scheme.

newtype ScriptHash c

Constructors

ScriptHash (Hash (ADDRHASH c) EraIndependentScript)

nativeMultiSigTag :: ByteString

Magic number "memorialized" in the ValidateScript class under the method: scriptPrefixTag:: Core.Script era -> Bs.ByteString, for the Shelley Era.

eqMultiSigRaw :: MultiSig era -> MultiSig era -> Bool

Check the equality of two underlying types, while ignoring their binary representation, which Eq instance normally does. This is used for testing.

Orphan instances

Crypto c => EraScript (ShelleyEra c)

Instance details

Associated Types

type Script (ShelleyEra c) = (r :: Type)

type NativeScript (ShelleyEra c) = (r :: Type)

Methods

upgradeScript :: Script (PreviousEra (ShelleyEra c)) -> Script (ShelleyEra c)

scriptPrefixTag :: Script (ShelleyEra c) -> ByteString

hashScript :: Script (ShelleyEra c) -> ScriptHash (EraCrypto (ShelleyEra c))

getNativeScript :: Script (ShelleyEra c) -> Maybe (NativeScript (ShelleyEra c))