Cardano.Ledger.Shelley.Transition

Description

Besides capturing all configuration that is necessary to progress to a specific era, this interface also provides fast forward functionality that is used in testing and benchmarking in order to initilize a chain in a particular era without going through the trouble of generating all the history for preceeding eras.

Documentation

class (EraTxOut era, EraGov era, EraGenesis era, ToJSON (TransitionConfig era), FromJSON (TransitionConfig era), Default (StashedAVVMAddresses era)) ⇒ EraTransition era where

Minimal complete definition

mkTransitionConfig, injectIntoTestState, tcPreviousEraConfigL, tcTranslationContextL

Associated Types

data TransitionConfig era ∷ Type

Cumulative configuration that is needed to be able to start in a current era

Methods

mkTransitionConfig

Arguments

∷ TranslationContext era
Translation context necessary for advancing from previous era into the current one. This will usually be the contents of genesis file, if one exists for the current era

→ TransitionConfig (PreviousEra era)
Transition configuration for the previous era.

→ TransitionConfig era

injectIntoTestState

Arguments

∷ TransitionConfig era
Extract data from the given transition configuration and store it in the given state.

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet.

→ NewEpochState era

→ NewEpochState era

tcPreviousEraConfigL ∷ EraTransition (PreviousEra era) ⇒ Lens' (TransitionConfig era) (TransitionConfig (PreviousEra era))

In case when a previous era is available, we should always be able to access TransitionConfig for the previous era, from within the current era's TransitionConfig

tcTranslationContextL ∷ Lens' (TransitionConfig era) (TranslationContext era)

Lens for the TranslationContext for the current era from the TransitionConfig Translation context is a different name for the Genesis type for each era, they are one and the same concept.

tcShelleyGenesisL ∷ Lens' (TransitionConfig era) (ShelleyGenesis (EraCrypto era))

Lens for the ShelleyGenesis from the TransitionConfig. Default implementation looks in the previous era's config

default tcShelleyGenesisL ∷ (EraTransition (PreviousEra era), EraCrypto (PreviousEra era) ~ EraCrypto era) ⇒ Lens' (TransitionConfig era) (ShelleyGenesis (EraCrypto era))

tcInitialPParamsG ∷ SimpleGetter (TransitionConfig era) (PParams era)

Get the initial PParams for the current era from the TransitionConfig. Note that this is only useful for testing and fast forward functionality, because this function assumes no on-chain changes to PParams through PParamsUpdate functionality.

Default implementation will use the PParams from the Previous era and the current TranslationContext to construct PParams for the current era.

Warning - Should only be used in testing and benchmarking

default tcInitialPParamsG ∷ (EraTransition (PreviousEra era), TranslateEra era PParams, TranslationError era PParams ~ Void) ⇒ SimpleGetter (TransitionConfig era) (PParams era)

tcInitialFundsL ∷ (HasCallStack, EraTransition era) ⇒ Lens' (TransitionConfig era) (ListMap (Addr (EraCrypto era)) Coin)

Get the initial funds from the TransitionConfig. This value must be non-empty only during testing and benchmarking, it must never contain anything on a real system.

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet

tcInitialStakingL ∷ (HasCallStack, EraTransition era) ⇒ Lens' (TransitionConfig era) (ShelleyGenesisStaking (EraCrypto era))

Get the initial staking from the TransitionConfig. This value must be non-empty only during testing and benchmarking, it must never contain anything on a real system.

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet

mkShelleyTransitionConfig ∷ ShelleyGenesis c → TransitionConfig (ShelleyEra c)

Constructor for the base Shelley TransitionConfig

createInitialState ∷ ∀ era. (EraTransition era, HasCallStack) ⇒ TransitionConfig era → NewEpochState era

Helper function for constructing the initial state for any era

Warning - Should only be used in testing and benchmarking. Will result in an error when NetworkId is set to Mainnet

This function does not register any initial funds or delegates.

registerInitialFundsThenStaking ∷ EraTransition era ⇒ TransitionConfig era → NewEpochState era → NewEpochState era

toShelleyTransitionConfigPairs ∷ (KeyValue e a, Crypto c) ⇒ TransitionConfig (ShelleyEra c) → [a]

protectMainnet ∷ (HasCallStack, EraTransition era) ⇒ String → TransitionConfig era → (a → Bool) → a → a

protectMainnetLens ∷ (HasCallStack, EraTransition era) ⇒ String → (a → Bool) → Lens' (TransitionConfig era) a → Lens' (TransitionConfig era) a