Cardano.Ledger.Shelley.Translation

Documentation

data FromByronTranslationContext

Required data to translate a Byron ledger into a Shelley ledger.

Constructors

FromByronTranslationContext

Fields

fbtcGenDelegs ∷ !(Map (KeyHash 'Genesis) GenDelegPair)

fbtcProtocolParams ∷ !(PParams ShelleyEra)

fbtcMaxLovelaceSupply ∷ !Word64

emptyFromByronTranslationContext ∷ FromByronTranslationContext

Trivial FromByronTranslationContext value, for use in cases where we do not need to translate from Byron to Shelley.

toFromByronTranslationContext ∷ ShelleyGenesis → FromByronTranslationContext