Cardano.Ledger.Shelley.Translation

Documentation

data FromByronTranslationContext c

Required data to translate a Byron ledger into a Shelley ledger.

Constructors

FromByronTranslationContext

Fields

fbtcGenDelegs ∷ !(Map (KeyHash 'Genesis c) (GenDelegPair c))

fbtcProtocolParams ∷ !(PParams (ShelleyEra c))

fbtcMaxLovelaceSupply ∷ !Word64

emptyFromByronTranslationContext ∷ Crypto c ⇒ FromByronTranslationContext c

Trivial FromByronTranslationContext value, for use in cases where we do not need to translate from Byron to Shelley.

toFromByronTranslationContext ∷ ShelleyGenesis c → FromByronTranslationContext c