Cardano.Ledger.Shelley.Tx

Transaction

data ShelleyTx era where

Bundled Patterns

pattern ShelleyTx ∷ EraTx era ⇒ TxBody era → TxWits era → StrictMaybe (TxAuxData era) → ShelleyTx era

bodyShelleyTxL ∷ EraTx era ⇒ Lens' (ShelleyTx era) (TxBody era)

TxBody setter and getter for ShelleyTx. The setter does update memoized binary representation.

witsShelleyTxL ∷ EraTx era ⇒ Lens' (ShelleyTx era) (TxWits era)

TxWits setter and getter for ShelleyTx. The setter does update memoized binary representation.

auxDataShelleyTxL ∷ EraTx era ⇒ Lens' (ShelleyTx era) (StrictMaybe (TxAuxData era))

TxAuxData setter and getter for ShelleyTx. The setter does update memoized binary representation.

sizeShelleyTxF ∷ Era era ⇒ SimpleGetter (ShelleyTx era) Integer

Size getter for ShelleyTx.

segwitTx ∷ ∀ era. EraTx era ⇒ Annotator (TxBody era) → Annotator (TxWits era) → Maybe (Annotator (TxAuxData era)) → Annotator (ShelleyTx era)

mkBasicShelleyTx ∷ EraTx era ⇒ TxBody era → ShelleyTx era

txwitsScript ∷ EraTx era ⇒ Tx era → Map (ScriptHash (EraCrypto era)) (Script era)

Multi-signature script witness accessor function for Transactions

extractKeyHashWitnessSet ∷ ∀ (r ∷ KeyRole) c. [Credential r c] → Set (KeyHash 'Witness c)

evalNativeMultiSigScript ∷ Era era ⇒ MultiSig era → Set (KeyHash 'Witness (EraCrypto era)) → Bool

Script evaluator for native multi-signature scheme. vhks is the set of key hashes that signed the transaction to be validated.

hashMultiSigScript ∷ ∀ era. (EraScript era, Script era ~ MultiSig (EraCrypto era)) ⇒ MultiSig (EraCrypto era) → ScriptHash (EraCrypto era)

Hashes native multi-signature script.

nativeMultiSigTag ∷ ByteString

Magic number "memorialized" in the ValidateScript class under the method: scriptPrefixTag:: Core.Script era -> Bs.ByteString, for the Shelley Era.

validateNativeMultiSigScript ∷ EraTx era ⇒ MultiSig era → Tx era → Bool

Script validator for native multi-signature scheme.

minfee ∷ EraTx era ⇒ PParams era → Tx era → Coin

Deprecated: In favor of getMinFeeTx

shelleyMinFeeTx ∷ EraTx era ⇒ PParams era → Tx era → Coin

Minimum fee calculation

witsFromTxWitnesses ∷ EraTx era ⇒ Tx era → Set (KeyHash 'Witness (EraCrypto era))

Extract the witness hashes from the Transaction.

Re-exports

newtype ShelleyTxBody era

Constructors

TxBodyConstr (MemoBytes ShelleyTxBodyRaw era)

Bundled Patterns

pattern ShelleyTxBody ∷ EraTxOut era ⇒ Set (TxIn (EraCrypto era)) → StrictSeq (TxOut era) → StrictSeq (DCert (EraCrypto era)) → Withdrawals (EraCrypto era) → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → ShelleyTxBody era
Pattern for use by external users

data ShelleyTxOut era

Constructors

TxOutCompact !(CompactAddr (EraCrypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern ShelleyTxOut ∷ (HasCallStack, Era era, Val (Value era)) ⇒ Addr (EraCrypto era) → Value era → ShelleyTxOut era

data TxIn c

The input of a UTxO.

Constructors

TxIn !(TxId c) !TxIx

newtype TxId c

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

unTxId ∷ SafeHash c EraIndependentTxBody

Orphan instances

Crypto c ⇒ EraTx (ShelleyEra c)

Instance details

Associated Types

type Tx (ShelleyEra c) = (r ∷ Type)

Methods

mkBasicTx ∷ TxBody (ShelleyEra c) → Tx (ShelleyEra c)

bodyTxL ∷ Lens' (Tx (ShelleyEra c)) (TxBody (ShelleyEra c))

witsTxL ∷ Lens' (Tx (ShelleyEra c)) (TxWits (ShelleyEra c))

auxDataTxL ∷ Lens' (Tx (ShelleyEra c)) (StrictMaybe (AuxiliaryData (ShelleyEra c)))

sizeTxF ∷ SimpleGetter (Tx (ShelleyEra c)) Integer

validateScript ∷ PhasedScript 'PhaseOne (ShelleyEra c) → Tx (ShelleyEra c) → Bool

getMinFeeTx ∷ PParams (ShelleyEra c) → Tx (ShelleyEra c) → Coin