Cardano.Ledger.Shelley.TxBody

Documentation

data ShelleyDelegCert c

data Delegation c

Constructors

Delegation

Fields

dDelegator :: !(StakeCredential c)

dDelegatee :: !(KeyHash 'StakePool c)

data GenesisDelegCert c

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot :: !MIRPot

mirRewards :: !(MIRTarget c)

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

data PoolCert c

Constructors

RegPool !(PoolParams c)

RetirePool !(KeyHash 'StakePool c) !EpochNo

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl :: !Url

pmHash :: !ByteString

data PoolParams c

Constructors

PoolParams

Fields

ppId :: !(KeyHash 'StakePool c)

ppVrf :: !(Hash c (VerKeyVRF c))

ppPledge :: !Coin

ppCost :: !Coin

ppMargin :: !UnitInterval

ppRewardAcnt :: !(RewardAcnt c)

ppOwners :: !(Set (KeyHash 'Staking c))

ppRelays :: !(StrictSeq StakePoolRelay)

ppMetadata :: !(StrictMaybe PoolMetadata)

data Ptr

Constructors

Ptr !SlotNo !TxIx !CertIx

data RewardAcnt c

Constructors

RewardAcnt

Fields

getRwdNetwork :: !Network

getRwdCred :: !(Credential 'Staking c)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)

SingleHostName !(StrictMaybe Port) !DnsName

MultiHostName !DnsName

newtype ShelleyTxBody era

Constructors

TxBodyConstr (MemoBytes ShelleyTxBodyRaw era)

Bundled Patterns

pattern ShelleyTxBody :: (EraTxOut era, EncCBOR (TxCert era)) => Set (TxIn (EraCrypto era)) -> StrictSeq (TxOut era) -> StrictSeq (TxCert era) -> Withdrawals (EraCrypto era) -> Coin -> SlotNo -> StrictMaybe (Update era) -> StrictMaybe (AuxiliaryDataHash (EraCrypto era)) -> ShelleyTxBody era
Pattern for use by external users

class (ShelleyEraTxCert era, EraTxBody era) => ShelleyEraTxBody era where

Minimal complete definition

ttlTxBodyL, updateTxBodyL, getTotalDepositsTxBody

Methods

ttlTxBodyL :: ExactEra ShelleyEra era => Lens' (TxBody era) SlotNo

updateTxBodyL :: ProtVerAtMost era 8 => Lens' (TxBody era) (StrictMaybe (Update era))

updateTxBodyG :: SimpleGetter (TxBody era) (StrictMaybe (Update era))

Deprecated: Use updateTxBodyL instead

default updateTxBodyG :: ProtVerAtMost era 8 => SimpleGetter (TxBody era) (StrictMaybe (Update era))

getTotalDepositsTxBody :: PParams era -> CertState era -> TxBody era -> Coin

data ShelleyTxBodyRaw era

Constructors

ShelleyTxBodyRaw

Fields

stbrInputs :: !(Set (TxIn (EraCrypto era)))

stbrOutputs :: !(StrictSeq (TxOut era))

stbrCerts :: !(StrictSeq (TxCert era))

stbrWithdrawals :: !(Withdrawals (EraCrypto era))

stbrTxFee :: !Coin

stbrTTL :: !SlotNo

stbrUpdate :: !(StrictMaybe (Update era))

stbrMDHash :: !(StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

data EraIndependentTxBody

type family TxOut era = (r :: Type) | r -> era

data ShelleyTxOut era

Constructors

TxOutCompact !(CompactAddr (EraCrypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern ShelleyTxOut :: (HasCallStack, Era era, Val (Value era)) => Addr (EraCrypto era) -> Value era -> ShelleyTxOut era

data Url

newtype Withdrawals c

Constructors

Withdrawals

Fields

unWithdrawals :: Map (RewardAcnt c) Coin

type Wdrl c = Withdrawals c

Deprecated: In favor of Withdrawals

pattern Wdrl :: Map (RewardAcnt c) Coin -> Withdrawals c

Deprecated: In favor of Withdrawals

witKeyHash :: WitVKey kr c -> KeyHash 'Witness c

Deprecated: In favor of witVKeyHash

wvkBytes :: WitVKey kr c -> ByteString

Deprecated: In favor of witVKeyBytes

data SizeOfPoolOwners

Constructors

SizeOfPoolOwners

data SizeOfPoolRelays

Constructors

SizeOfPoolRelays

Helpers

addrEitherShelleyTxOutL :: Lens' (ShelleyTxOut era) (Either (Addr (EraCrypto era)) (CompactAddr (EraCrypto era)))

valueEitherShelleyTxOutL :: Val (Value era) => Lens' (ShelleyTxOut era) (Either (Value era) (CompactForm (Value era)))

totalTxDepositsShelley :: ShelleyEraTxBody era => PParams era -> CertState era -> TxBody era -> Coin

Calculates the total amount of deposits needed for all pool registration and stake delegation certificates to be valid.

Orphan instances

Crypto c => EraTxBody (ShelleyEra c)

Instance details

Associated Types

type TxBody (ShelleyEra c) = (r :: Type)

type TxBodyUpgradeError (ShelleyEra c)

Methods

mkBasicTxBody :: TxBody (ShelleyEra c)

inputsTxBodyL :: Lens' (TxBody (ShelleyEra c)) (Set (TxIn (EraCrypto (ShelleyEra c))))

outputsTxBodyL :: Lens' (TxBody (ShelleyEra c)) (StrictSeq (TxOut (ShelleyEra c)))

feeTxBodyL :: Lens' (TxBody (ShelleyEra c)) Coin

withdrawalsTxBodyL :: Lens' (TxBody (ShelleyEra c)) (Withdrawals (EraCrypto (ShelleyEra c)))

auxDataHashTxBodyL :: Lens' (TxBody (ShelleyEra c)) (StrictMaybe (AuxiliaryDataHash (EraCrypto (ShelleyEra c))))

spendableInputsTxBodyF :: SimpleGetter (TxBody (ShelleyEra c)) (Set (TxIn (EraCrypto (ShelleyEra c))))

allInputsTxBodyF :: SimpleGetter (TxBody (ShelleyEra c)) (Set (TxIn (EraCrypto (ShelleyEra c))))

certsTxBodyL :: Lens' (TxBody (ShelleyEra c)) (StrictSeq (TxCert (ShelleyEra c)))

upgradeTxBody :: TxBody (PreviousEra (ShelleyEra c)) -> Either (TxBodyUpgradeError (ShelleyEra c)) (TxBody (ShelleyEra c))

Crypto c => ShelleyEraTxBody (ShelleyEra c)

Instance details

Methods

ttlTxBodyL :: Lens' (TxBody (ShelleyEra c)) SlotNo

updateTxBodyL :: Lens' (TxBody (ShelleyEra c)) (StrictMaybe (Update (ShelleyEra c)))

updateTxBodyG :: SimpleGetter (TxBody (ShelleyEra c)) (StrictMaybe (Update (ShelleyEra c)))

getTotalDepositsTxBody :: PParams (ShelleyEra c) -> CertState (ShelleyEra c) -> TxBody (ShelleyEra c) -> Coin