Cardano.Ledger.Shelley.TxBody

Documentation

data DCert c

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert c)

DCertPool !(PoolCert c)

DCertGenesis !(ConstitutionalDelegCert c)

DCertMir !(MIRCert c)

data DelegCert c

Constructors

RegKey !(StakeCredential c)
A stake key registration certificate.

DeRegKey !(StakeCredential c)
A stake key deregistration certificate.

Delegate !(Delegation c)
A stake delegation certificate.

data Delegation c

The delegation of one stake key to another.

Constructors

Delegation

Fields

dDelegator ∷ !(StakeCredential c)

dDelegatee ∷ !(KeyHash 'StakePool c)

data ConstitutionalDelegCert c

Genesis key delegation certificate

Constructors

ConstitutionalDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ MIRPot

mirRewards ∷ MIRTarget c

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR (Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR Coin

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl ∷ !Url

pmHash ∷ !ByteString

data PoolParams c

A stake pool.

Constructors

PoolParams

Fields

ppId ∷ !(KeyHash 'StakePool c)

ppVrf ∷ !(Hash c (VerKeyVRF c))

ppPledge ∷ !Coin

ppCost ∷ !Coin

ppMargin ∷ !UnitInterval

ppRewardAcnt ∷ !(RewardAcnt c)

ppOwners ∷ !(Set (KeyHash 'Staking c))

ppRelays ∷ !(StrictSeq StakePoolRelay)

ppMetadata ∷ !(StrictMaybe PoolMetadata)

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo !TxIx !CertIx

data RewardAcnt c

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork ∷ !Network

getRwdCred ∷ !(Credential 'Staking c)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

newtype ShelleyTxBody era

Constructors

TxBodyConstr (MemoBytes ShelleyTxBodyRaw era)

Bundled Patterns

pattern ShelleyTxBody ∷ EraTxOut era ⇒ Set (TxIn (EraCrypto era)) → StrictSeq (TxOut era) → StrictSeq (DCert (EraCrypto era)) → Withdrawals (EraCrypto era) → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → ShelleyTxBody era
Pattern for use by external users

class EraTxBody era ⇒ ShelleyEraTxBody era where

Minimal complete definition

ttlTxBodyL, updateTxBodyL, certsTxBodyL

Methods

ttlTxBodyL ∷ ExactEra ShelleyEra era ⇒ Lens' (TxBody era) SlotNo

updateTxBodyL ∷ ProtVerAtMost era 8 ⇒ Lens' (TxBody era) (StrictMaybe (Update era))

updateTxBodyG ∷ SimpleGetter (TxBody era) (StrictMaybe (Update era))

default updateTxBodyG ∷ ProtVerAtMost era 8 ⇒ SimpleGetter (TxBody era) (StrictMaybe (Update era))

certsTxBodyL ∷ ProtVerAtMost era 8 ⇒ Lens' (TxBody era) (StrictSeq (DCert (EraCrypto era)))

certsTxBodyG ∷ SimpleGetter (TxBody era) (StrictSeq (DCert (EraCrypto era)))

default certsTxBodyG ∷ ProtVerAtMost era 8 ⇒ SimpleGetter (TxBody era) (StrictSeq (DCert (EraCrypto era)))

data ShelleyTxBodyRaw era

Constructors

ShelleyTxBodyRaw

Fields

stbrInputs ∷ !(Set (TxIn (EraCrypto era)))

stbrOutputs ∷ !(StrictSeq (TxOut era))

stbrCerts ∷ !(StrictSeq (DCert (EraCrypto era)))

stbrWithdrawals ∷ !(Withdrawals (EraCrypto era))

stbrTxFee ∷ !Coin

stbrTTL ∷ !SlotNo

stbrUpdate ∷ !(StrictMaybe (Update era))

stbrMDHash ∷ !(StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

data EraIndependentTxBody

type family TxOut era = (r ∷ Type) | r → era

The output of a UTxO for a particular era

data ShelleyTxOut era

Constructors

TxOutCompact !(CompactAddr (EraCrypto era)) !(CompactForm (Value era))

Bundled Patterns

pattern ShelleyTxOut ∷ (HasCallStack, Era era, Val (Value era)) ⇒ Addr (EraCrypto era) → Value era → ShelleyTxOut era

data Url

newtype Withdrawals c

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map (RewardAcnt c) Coin

type Wdrl c = Withdrawals c

Deprecated: In favor of Withdrawals

pattern Wdrl ∷ Map (RewardAcnt c) Coin → Withdrawals c

Deprecated: In favor of Withdrawals

module Cardano.Ledger.Keys.WitVKey

witKeyHash ∷ WitVKey kr c → KeyHash 'Witness c

Deprecated: In favor of witVKeyHash

wvkBytes ∷ WitVKey kr c → ByteString

Deprecated: In favor of witVKeyBytes

data SizeOfPoolOwners

The size of the ppOwners Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolOwners

data SizeOfPoolRelays

The size of the ppRelays Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolRelays

Helpers

addrEitherShelleyTxOutL ∷ Lens' (ShelleyTxOut era) (Either (Addr (EraCrypto era)) (CompactAddr (EraCrypto era)))

valueEitherShelleyTxOutL ∷ Val (Value era) ⇒ Lens' (ShelleyTxOut era) (Either (Value era) (CompactForm (Value era)))

Orphan instances

Crypto c ⇒ EraTxBody (ShelleyEra c)

Instance details

Associated Types

type TxBody (ShelleyEra c) = (r ∷ Type)

Methods

mkBasicTxBody ∷ TxBody (ShelleyEra c)

inputsTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) (Set (TxIn (EraCrypto (ShelleyEra c))))

outputsTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) (StrictSeq (TxOut (ShelleyEra c)))

feeTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) (Withdrawals (EraCrypto (ShelleyEra c)))

auxDataHashTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) (StrictMaybe (AuxiliaryDataHash (EraCrypto (ShelleyEra c))))

allInputsTxBodyF ∷ SimpleGetter (TxBody (ShelleyEra c)) (Set (TxIn (EraCrypto (ShelleyEra c))))

Crypto c ⇒ ShelleyEraTxBody (ShelleyEra c)

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) SlotNo

updateTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) (StrictMaybe (Update (ShelleyEra c)))

updateTxBodyG ∷ SimpleGetter (TxBody (ShelleyEra c)) (StrictMaybe (Update (ShelleyEra c)))

certsTxBodyL ∷ Lens' (TxBody (ShelleyEra c)) (StrictSeq (DCert (EraCrypto (ShelleyEra c))))

certsTxBodyG ∷ SimpleGetter (TxBody (ShelleyEra c)) (StrictSeq (DCert (EraCrypto (ShelleyEra c))))