Cardano.Ledger.Shelley.TxCert

Documentation

class EraTxCert era ⇒ ShelleyEraTxCert era where

Methods

mkRegTxCert ∷ StakeCredential (EraCrypto era) → TxCert era

getRegTxCert ∷ TxCert era → Maybe (StakeCredential (EraCrypto era))

mkUnRegTxCert ∷ StakeCredential (EraCrypto era) → TxCert era

getUnRegTxCert ∷ TxCert era → Maybe (StakeCredential (EraCrypto era))

mkDelegStakeTxCert ∷ StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → TxCert era

getDelegStakeTxCert ∷ TxCert era → Maybe (StakeCredential (EraCrypto era), KeyHash 'StakePool (EraCrypto era))

mkGenesisDelegTxCert ∷ ProtVerAtMost era 8 ⇒ GenesisDelegCert (EraCrypto era) → TxCert era

getGenesisDelegTxCert ∷ ProtVerAtMost era 8 ⇒ TxCert era → Maybe (GenesisDelegCert (EraCrypto era))

mkMirTxCert ∷ ProtVerAtMost era 8 ⇒ MIRCert (EraCrypto era) → TxCert era

getMirTxCert ∷ ProtVerAtMost era 8 ⇒ TxCert era → Maybe (MIRCert (EraCrypto era))

pattern MirTxCert ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ MIRCert (EraCrypto era) → TxCert era

pattern GenesisDelegTxCert ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ KeyHash 'Genesis (EraCrypto era) → KeyHash 'GenesisDelegate (EraCrypto era) → Hash (EraCrypto era) (VerKeyVRF (EraCrypto era)) → TxCert era

pattern RegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential (EraCrypto era) → TxCert era

pattern UnRegTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential (EraCrypto era) → TxCert era

pattern DelegStakeTxCert ∷ ShelleyEraTxCert era ⇒ StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → TxCert era

data ShelleyDelegCert c

Constructors

ShelleyRegCert !(StakeCredential c)
A stake credential registration certificate.

ShelleyUnRegCert !(StakeCredential c)
A stake credential deregistration certificate.

ShelleyDelegCert !(StakeCredential c) !(KeyHash 'StakePool c)
A stake delegation certificate.

Bundled Patterns

pattern RegKey ∷ StakeCredential c → ShelleyDelegCert c

Deprecated: In favor of ShelleyRegCert

pattern DeRegKey ∷ StakeCredential c → ShelleyDelegCert c

Deprecated: In favor of ShelleyUnRegCert

pattern Delegate ∷ Delegation c → ShelleyDelegCert c

Deprecated: In favor of ShelleyDelegCert

getVKeyWitnessShelleyTxCert ∷ ShelleyTxCert era → Maybe (KeyHash 'Witness (EraCrypto era))

getScriptWitnessShelleyTxCert ∷ ShelleyTxCert era → Maybe (ScriptHash (EraCrypto era))

delegCWitness ∷ ShelleyDelegCert c → Credential 'Staking c

Deprecated: This was a partial function, logic rewritten in a safer way

Determine the certificate author

data ShelleyTxCert era

A heavyweight certificate.

Constructors

ShelleyTxCertDelegCert !(ShelleyDelegCert (EraCrypto era))

ShelleyTxCertPool !(PoolCert (EraCrypto era))

ShelleyTxCertGenesisDeleg !(GenesisDelegCert (EraCrypto era))

ShelleyTxCertMir !(MIRCert (EraCrypto era))

upgradeShelleyTxCert ∷ EraCrypto era1 ~ EraCrypto era2 ⇒ ShelleyTxCert era1 → ShelleyTxCert era2

GenesisDelegCert

data GenesisDelegCert c

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

genesisCWitness ∷ GenesisDelegCert c → KeyHash 'Genesis c

genesisKeyHashWitness ∷ GenesisDelegCert c → KeyHash 'Witness c

MIRCert

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ !MIRPot

mirRewards ∷ !(MIRTarget c)

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

isRegKey ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Deprecated: Use isRegStakeTxCert instead

Check for ShelleyRegCert constructor

isDeRegKey ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Deprecated: Use isUnRegStakeTxCert instead

Check for ShelleyUnRegCert constructor

isDelegation ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Check for ShelleyDelegCert constructor

isRegPool ∷ EraTxCert era ⇒ TxCert era → Bool

Check for RegPool constructor

isRetirePool ∷ EraTxCert era ⇒ TxCert era → Bool

Check for RetirePool constructor

isGenesisDelegation ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ TxCert era → Bool

Check for GenesisDelegate constructor

isInstantaneousRewards ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ TxCert era → Bool

isReservesMIRCert ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ TxCert era → Bool

isTreasuryMIRCert ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ TxCert era → Bool

requiresVKeyWitness ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ TxCert era → Bool

Deprecated: In favor of getVKeyWitnessTxCert

Returns True for delegation certificates that require at least one witness, and False otherwise. It is mainly used to ensure that calling a variant of cwitness is safe.

Note: This will not compile for Conway, because it is incorrect for Conway, use getVKeyWitnessTxCert instead.

Serialization helpers

shelleyTxCertDelegDecoder ∷ ShelleyEraTxCert era ⇒ Word → Decoder s (Int, TxCert era)

poolTxCertDecoder ∷ EraTxCert era ⇒ Word → Decoder s (Int, TxCert era)

encodeShelleyDelegCert ∷ Crypto c ⇒ ShelleyDelegCert c → Encoding

encodePoolCert ∷ Crypto c ⇒ PoolCert c → Encoding

encodeGenesisDelegCert ∷ Crypto c ⇒ GenesisDelegCert c → Encoding

Deposits and Refunds

shelleyTotalDepositsTxCerts

Arguments

∷ (EraPParams era, Foldable f, EraTxCert era)

⇒ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ f (TxCert era)

→ Coin

Determine the total deposit amount needed from a TxBody. The block may (legitimately) contain multiple registration certificates for the same pool, where the first will be treated as a registration and any subsequent ones as re-registration. As such, we must only take a deposit for the first such registration. It is even possible for a single transaction to have multiple pool registration for the same pool, so as we process pool registrations, we must keep track of those that are already registered, so we do not add a Deposit for the same pool twice.

Note that this is not an issue for key registrations since subsequent registration certificates would be invalid.

shelleyTotalRefundsTxCerts

Arguments

∷ (EraPParams era, Foldable f, EraTxCert era)

⇒ PParams era

→ (StakeCredential (EraCrypto era) → Maybe Coin)
Function that can lookup current deposit, in case when the stake key is registered.

→ f (TxCert era)

→ Coin

Compute the key deregistration refunds in a transaction

Re-exports

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Ord (TxCert era), Eq (TxCert era)) ⇒ EraTxCert era where

Associated Types

type TxCert era = (r ∷ Type) | r → era

type TxCertUpgradeError era

type TxCertUpgradeError era = Void

Methods

upgradeTxCert ∷ TxCert (PreviousEra era) → Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert ∷ TxCert era → Maybe (KeyHash 'Witness (EraCrypto era))

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert ∷ TxCert era → Maybe (ScriptHash (EraCrypto era))

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert ∷ PoolParams (EraCrypto era) → TxCert era

getRegPoolTxCert ∷ TxCert era → Maybe (PoolParams (EraCrypto era))

mkRetirePoolTxCert ∷ KeyHash 'StakePool (EraCrypto era) → EpochNo → TxCert era

getRetirePoolTxCert ∷ TxCert era → Maybe (KeyHash 'StakePool (EraCrypto era), EpochNo)

lookupRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether stake pool is registered or not

→ f (TxCert era)

→ Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (Credential 'Staking (EraCrypto era) → Maybe Coin)
Lookup current deposit for Staking credential if one is registered

→ (Credential 'DRepRole (EraCrypto era) → Maybe Coin)
Lookup current deposit for DRep credential if one is registered

→ f (TxCert era)

→ Coin

Compute the total refunds from a list of certificates.

pattern RegPoolTxCert ∷ EraTxCert era ⇒ PoolParams (EraCrypto era) → TxCert era

pattern RetirePoolTxCert ∷ EraTxCert era ⇒ KeyHash 'StakePool (EraCrypto era) → EpochNo → TxCert era

data Delegation c

The delegation of one stake key to another.

Constructors

Delegation

Fields

dDelegator ∷ !(StakeCredential c)

dDelegatee ∷ !(KeyHash 'StakePool c)

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

poolCWitness ∷ PoolCert c → Credential 'StakePool c

isRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake registering certificate

isUnRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake un-registering certificate

Orphan instances

Crypto c ⇒ EraTxCert (ShelleyEra c)

Instance details

Associated Types

type TxCert (ShelleyEra c) = (r ∷ Type)

type TxCertUpgradeError (ShelleyEra c)

Methods

upgradeTxCert ∷ TxCert (PreviousEra (ShelleyEra c)) → Either (TxCertUpgradeError (ShelleyEra c)) (TxCert (ShelleyEra c))

getVKeyWitnessTxCert ∷ TxCert (ShelleyEra c) → Maybe (KeyHash 'Witness (EraCrypto (ShelleyEra c)))

getScriptWitnessTxCert ∷ TxCert (ShelleyEra c) → Maybe (ScriptHash (EraCrypto (ShelleyEra c)))

mkRegPoolTxCert ∷ PoolParams (EraCrypto (ShelleyEra c)) → TxCert (ShelleyEra c)

getRegPoolTxCert ∷ TxCert (ShelleyEra c) → Maybe (PoolParams (EraCrypto (ShelleyEra c)))

mkRetirePoolTxCert ∷ KeyHash 'StakePool (EraCrypto (ShelleyEra c)) → EpochNo → TxCert (ShelleyEra c)

getRetirePoolTxCert ∷ TxCert (ShelleyEra c) → Maybe (KeyHash 'StakePool (EraCrypto (ShelleyEra c)), EpochNo)

lookupRegStakeTxCert ∷ TxCert (ShelleyEra c) → Maybe (Credential 'Staking (EraCrypto (ShelleyEra c)))

lookupUnRegStakeTxCert ∷ TxCert (ShelleyEra c) → Maybe (Credential 'Staking (EraCrypto (ShelleyEra c)))

getTotalDepositsTxCerts ∷ Foldable f ⇒ PParams (ShelleyEra c) → (KeyHash 'StakePool (EraCrypto (ShelleyEra c)) → Bool) → f (TxCert (ShelleyEra c)) → Coin

getTotalRefundsTxCerts ∷ Foldable f ⇒ PParams (ShelleyEra c) → (Credential 'Staking (EraCrypto (ShelleyEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (ShelleyEra c)) → Maybe Coin) → f (TxCert (ShelleyEra c)) → Coin