Cardano.Ledger.Shelley.TxCert

Documentation

class EraTxCert era ⇒ ShelleyEraTxCert era where

Methods

mkTxCertMir ∷ ProtVerAtMost era 8 ⇒ MIRCert (EraCrypto era) → TxCert era

getTxCertMir ∷ TxCert era → Maybe (MIRCert (EraCrypto era))

mkShelleyTxCertDeleg ∷ ShelleyDelegCert (EraCrypto era) → TxCert era

getShelleyTxCertDeleg ∷ TxCert era → Maybe (ShelleyDelegCert (EraCrypto era))

pattern TxCertMir ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ MIRCert (EraCrypto era) → TxCert era

pattern ShelleyTxCertDeleg ∷ ShelleyEraTxCert era ⇒ ShelleyDelegCert (EraCrypto era) → TxCert era

data ShelleyDelegCert c

Constructors

ShelleyRegCert !(StakeCredential c)
A stake credential registration certificate.

ShelleyUnRegCert !(StakeCredential c)
A stake credential deregistration certificate.

ShelleyDelegCert !(StakeCredential c) !(KeyHash 'StakePool c)
A stake delegation certificate.

Bundled Patterns

pattern RegKey ∷ StakeCredential c → ShelleyDelegCert c

Deprecated: In favor of ShelleyRegCert

pattern DeRegKey ∷ StakeCredential c → ShelleyDelegCert c

Deprecated: In favor of ShelleyUnRegCert

pattern Delegate ∷ Delegation c → ShelleyDelegCert c

Deprecated: In favor of ShelleyDelegCert

delegCWitness ∷ ShelleyDelegCert c → Credential 'Staking c

Determine the certificate author

data ShelleyTxCert era

A heavyweight certificate.

Constructors

ShelleyTxCertDelegCert !(ShelleyDelegCert (EraCrypto era))

ShelleyTxCertPool !(PoolCert (EraCrypto era))

ShelleyTxCertGenesis !(ConstitutionalDelegCert (EraCrypto era))

ShelleyTxCertMir !(MIRCert (EraCrypto era))

data MIRCert c

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ !MIRPot

mirRewards ∷ !(MIRTarget c)

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data MIRTarget c

MIRTarget specifies if funds from either the reserves or the treasury are to be handed out to a collection of reward accounts or instead transfered to the opposite pot.

Constructors

StakeAddressesMIR !(Map (Credential 'Staking c) DeltaCoin)

SendToOppositePotMIR !Coin

isRegKey ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Check for ShelleyRegCert constructor

isDeRegKey ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Check for ShelleyUnRegCert constructor

isDelegation ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Check for ShelleyDelegCert constructor

isRegPool ∷ EraTxCert era ⇒ TxCert era → Bool

Check for RegPool constructor

isRetirePool ∷ EraTxCert era ⇒ TxCert era → Bool

Check for RetirePool constructor

isGenesisDelegation ∷ EraTxCert era ⇒ TxCert era → Bool

Check for GenesisDelegate constructor

isInstantaneousRewards ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

isReservesMIRCert ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

isTreasuryMIRCert ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

requiresVKeyWitness ∷ ShelleyEraTxCert era ⇒ TxCert era → Bool

Returns True for delegation certificates that require at least one witness, and False otherwise. It is mainly used to ensure that calling a variant of cwitness is safe.

Serialization helpers

shelleyTxCertDelegDecoder ∷ ShelleyEraTxCert era ⇒ Word → Decoder s (Int, TxCert era)

commonTxCertDecoder ∷ EraTxCert era ⇒ Word → Decoder s (Int, TxCert era)

encodeShelleyDelegCert ∷ Crypto c ⇒ ShelleyDelegCert c → Encoding

encodePoolCert ∷ Crypto c ⇒ PoolCert c → Encoding

encodeConstitutionalCert ∷ Crypto c ⇒ ConstitutionalDelegCert c → Encoding

Re-exports

class (Era era, DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Eq (TxCert era)) ⇒ EraTxCert era where

Associated Types

type TxCert era = (r ∷ Type) | r → era

Methods

mkTxCertPool ∷ PoolCert (EraCrypto era) → TxCert era

getTxCertPool ∷ TxCert era → Maybe (PoolCert (EraCrypto era))

mkTxCertGenesis ∷ ConstitutionalDelegCert (EraCrypto era) → TxCert era

getTxCertGenesis ∷ TxCert era → Maybe (ConstitutionalDelegCert (EraCrypto era))

pattern TxCertPool ∷ EraTxCert era ⇒ PoolCert (EraCrypto era) → TxCert era

pattern TxCertGenesis ∷ EraTxCert era ⇒ ConstitutionalDelegCert (EraCrypto era) → TxCert era

data Delegation c

The delegation of one stake key to another.

Constructors

Delegation

Fields

dDelegator ∷ !(StakeCredential c)

dDelegatee ∷ !(KeyHash 'StakePool c)

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

poolCWitness ∷ PoolCert c → Credential 'StakePool c

data ConstitutionalDelegCert c

Constitutional key delegation certificate

Constructors

ConstitutionalDelegCert !(KeyHash 'Genesis c) !(KeyHash 'GenesisDelegate c) !(Hash c (VerKeyVRF c))

genesisCWitness ∷ ConstitutionalDelegCert c → KeyHash 'Genesis c

Orphan instances

Crypto c ⇒ EraTxCert (ShelleyEra c)

Instance details

Associated Types

type TxCert (ShelleyEra c) = (r ∷ Type)

Methods

mkTxCertPool ∷ PoolCert (EraCrypto (ShelleyEra c)) → TxCert (ShelleyEra c)

getTxCertPool ∷ TxCert (ShelleyEra c) → Maybe (PoolCert (EraCrypto (ShelleyEra c)))

mkTxCertGenesis ∷ ConstitutionalDelegCert (EraCrypto (ShelleyEra c)) → TxCert (ShelleyEra c)

getTxCertGenesis ∷ TxCert (ShelleyEra c) → Maybe (ConstitutionalDelegCert (EraCrypto (ShelleyEra c)))