Cardano.Ledger.Shelley.UTxO

Documentation

class EraTx era ⇒ EraUTxO era where

Associated Types

type ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

Methods

getConsumedValue

Arguments

∷ PParams era

→ (Credential 'Staking (EraCrypto era) → Maybe Coin)
Function that can lookup current delegation deposits

→ (Credential 'DRepRole (EraCrypto era) → Maybe Coin)
Function that can lookup current drep deposits

→ UTxO era

→ TxBody era

→ Value era

Calculate all the value that is being consumed by the transaction.

getProducedValue

Arguments

∷ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

getScriptsProvided

Arguments

∷ UTxO era
For some era it is necessary to look into the UTxO to find all of the available scripts for the transaction

→ Tx era

→ ScriptsProvided era

Initial eras will look into witness set to find all of the available scripts, but starting with Babbage we can look for available scripts in the UTxO using reference inputs.

getScriptsNeeded ∷ UTxO era → TxBody era → ScriptsNeeded era

Produce all the information required for figuring out which scripts are required for the transaction to be valid, once those scripts are evaluated

getScriptsHashesNeeded ∷ ScriptsNeeded era → Set (ScriptHash (EraCrypto era))

Extract the set of all script hashes that are needed for script validation.

getWitsVKeyNeeded ∷ CertState era → UTxO era → TxBody era → Set (KeyHash 'Witness (EraCrypto era))

Extract all of the KeyHash witnesses that are required for validating the transaction

getMinFeeTxUtxo ∷ PParams era → Tx era → UTxO era → Coin

Minimum fee computation, excluding witnesses and including ref scripts size

newtype ShelleyScriptsNeeded era

Constructors

ShelleyScriptsNeeded (Set (ScriptHash (EraCrypto era)))

scriptsNeeded ∷ ∀ era. EraTx era ⇒ UTxO era → Tx era → Set (ScriptHash (EraCrypto era))

Deprecated: In favor of getScriptsNeeded

Computes the set of script hashes required to unlock the transaction inputs and the withdrawals.

getShelleyScriptsNeeded ∷ EraTxBody era ⇒ UTxO era → TxBody era → ShelleyScriptsNeeded era

scriptCred ∷ Credential kr c → Maybe (ScriptHash c)

Deprecated: In favor of credScriptHash

scriptStakeCred ∷ ShelleyEraTxCert era ⇒ TxCert era → Maybe (ScriptHash (EraCrypto era))

Deprecated: In favor of getScriptWitnessTxCert

getConsumedCoin ∷ EraTxBody era ⇒ PParams era → (Credential 'Staking (EraCrypto era) → Maybe Coin) → UTxO era → TxBody era → Coin

Compute the lovelace which are destroyed by the transaction. This implementation is suitable for Shelley and Allegra only.

shelleyProducedValue

Arguments

∷ EraTxBody era

⇒ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

consumed ∷ EraUTxO era ⇒ PParams era → CertState era → UTxO era → TxBody era → Value era

For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

produced ∷ EraUTxO era ⇒ PParams era → CertState era → TxBody era → Value era

Compute the lovelace which are created by the transaction For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

txup ∷ (EraTx era, ShelleyEraTxBody era) ⇒ Tx era → Maybe (Update era)

Deprecated: In favor of updateTxBodyL

getShelleyMinFeeTxUtxo ∷ EraTx era ⇒ PParams era → Tx era → Coin

getShelleyWitsVKeyNeeded ∷ ∀ era. (EraTx era, ShelleyEraTxBody era) ⇒ CertState era → UTxO era → TxBody era → Set (KeyHash 'Witness (EraCrypto era))

getShelleyWitsVKeyNeededNoGov ∷ ∀ era. EraTx era ⇒ UTxO era → TxBody era → Set (KeyHash 'Witness (EraCrypto era))

Extract witnesses from UTxO and TxBody. Does not enforce witnesses for governance related Keys, i.e. GenDelegs

module Cardano.Ledger.UTxO

Orphan instances

Crypto c ⇒ EraUTxO (ShelleyEra c)

Instance details

Associated Types

type ScriptsNeeded (ShelleyEra c) = (r ∷ Type)

Methods

getConsumedValue ∷ PParams (ShelleyEra c) → (Credential 'Staking (EraCrypto (ShelleyEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (ShelleyEra c)) → Maybe Coin) → UTxO (ShelleyEra c) → TxBody (ShelleyEra c) → Value (ShelleyEra c)

getProducedValue ∷ PParams (ShelleyEra c) → (KeyHash 'StakePool (EraCrypto (ShelleyEra c)) → Bool) → TxBody (ShelleyEra c) → Value (ShelleyEra c)

getScriptsProvided ∷ UTxO (ShelleyEra c) → Tx (ShelleyEra c) → ScriptsProvided (ShelleyEra c)

getScriptsNeeded ∷ UTxO (ShelleyEra c) → TxBody (ShelleyEra c) → ScriptsNeeded (ShelleyEra c)

getScriptsHashesNeeded ∷ ScriptsNeeded (ShelleyEra c) → Set (ScriptHash (EraCrypto (ShelleyEra c)))

getWitsVKeyNeeded ∷ CertState (ShelleyEra c) → UTxO (ShelleyEra c) → TxBody (ShelleyEra c) → Set (KeyHash 'Witness (EraCrypto (ShelleyEra c)))

getMinFeeTxUtxo ∷ PParams (ShelleyEra c) → Tx (ShelleyEra c) → UTxO (ShelleyEra c) → Coin