Cardano.Ledger.Shelley.UTxO

Documentation

class EraTx era => EraUTxO era where

Associated Types

type ScriptsNeeded era = (r :: Type) | r -> era

Methods

getConsumedValue :: PParams era -> (Credential 'Staking (EraCrypto era) -> Maybe Coin) -> (Credential 'DRepRole (EraCrypto era) -> Maybe Coin) -> UTxO era -> TxBody era -> Value era

getProducedValue :: PParams era -> (KeyHash 'StakePool (EraCrypto era) -> Bool) -> TxBody era -> Value era

getScriptsProvided :: UTxO era -> Tx era -> ScriptsProvided era

getScriptsNeeded :: UTxO era -> TxBody era -> ScriptsNeeded era

getScriptsHashesNeeded :: ScriptsNeeded era -> Set (ScriptHash (EraCrypto era))

newtype ShelleyScriptsNeeded era

Constructors

ShelleyScriptsNeeded (Set (ScriptHash (EraCrypto era)))

scriptsNeeded :: forall era. (EraTx era, ShelleyEraTxBody era) => UTxO era -> Tx era -> Set (ScriptHash (EraCrypto era))

Deprecated: In favor of getScriptsNeeded

Computes the set of script hashes required to unlock the transaction inputs and the withdrawals.

getShelleyScriptsNeeded :: ShelleyEraTxBody era => UTxO era -> TxBody era -> ShelleyScriptsNeeded era

scriptCred :: Credential kr c -> Maybe (ScriptHash c)

Deprecated: In favor of credScriptHash

scriptStakeCred :: ShelleyEraTxCert era => TxCert era -> Maybe (ScriptHash (EraCrypto era))

Deprecated: In favor of getScriptWitnessTxCert

getConsumedCoin :: ShelleyEraTxBody era => PParams era -> (Credential 'Staking (EraCrypto era) -> Maybe Coin) -> UTxO era -> TxBody era -> Coin

Compute the lovelace which are destroyed by the transaction

shelleyProducedValue

Arguments

:: ShelleyEraTxBody era

=> PParams era

-> (KeyHash 'StakePool (EraCrypto era) -> Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

-> TxBody era

-> Value era

consumed :: EraUTxO era => PParams era -> CertState era -> UTxO era -> TxBody era -> Value era

For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

produced :: EraUTxO era => PParams era -> CertState era -> TxBody era -> Value era

Compute the lovelace which are created by the transaction For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.

txup :: (EraTx era, ShelleyEraTxBody era, ProtVerAtMost era 8) => Tx era -> Maybe (Update era)

Orphan instances

Crypto c => EraUTxO (ShelleyEra c)

Instance details

Associated Types

type ScriptsNeeded (ShelleyEra c) = (r :: Type)

Methods

getConsumedValue :: PParams (ShelleyEra c) -> (Credential 'Staking (EraCrypto (ShelleyEra c)) -> Maybe Coin) -> (Credential 'DRepRole (EraCrypto (ShelleyEra c)) -> Maybe Coin) -> UTxO (ShelleyEra c) -> TxBody (ShelleyEra c) -> Value (ShelleyEra c)

getProducedValue :: PParams (ShelleyEra c) -> (KeyHash 'StakePool (EraCrypto (ShelleyEra c)) -> Bool) -> TxBody (ShelleyEra c) -> Value (ShelleyEra c)

getScriptsProvided :: UTxO (ShelleyEra c) -> Tx (ShelleyEra c) -> ScriptsProvided (ShelleyEra c)

getScriptsNeeded :: UTxO (ShelleyEra c) -> TxBody (ShelleyEra c) -> ScriptsNeeded (ShelleyEra c)

getScriptsHashesNeeded :: ScriptsNeeded (ShelleyEra c) -> Set (ScriptHash (EraCrypto (ShelleyEra c)))