Cardano.Ledger.Shelley.UTxO

Documentation

newtype ShelleyScriptsNeeded era

Constructors

ShelleyScriptsNeeded (Set (ScriptHash (EraCrypto era)))

scriptsNeeded ∷ ∀ era. (EraTx era, ShelleyEraTxBody era) ⇒ UTxO era → Tx era → Set (ScriptHash (EraCrypto era))

Deprecated: In favor of getScriptsNeeded

Computes the set of script hashes required to unlock the transaction inputs and the withdrawals.

getShelleyScriptsNeeded ∷ ShelleyEraTxBody era ⇒ UTxO era → TxBody era → ShelleyScriptsNeeded era

scriptCred ∷ Credential kr c → Maybe (ScriptHash c)

scriptStakeCred ∷ DCert c → Maybe (ScriptHash c)

getConsumedCoin ∷ ShelleyEraTxBody era ⇒ PParams era → (StakeCredential (EraCrypto era) → Maybe Coin) → UTxO era → TxBody era → Coin

Compute the lovelace which are destroyed by the transaction

getProducedValue

Arguments

∷ ShelleyEraTxBody era

⇒ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

consumed ∷ EraUTxO era ⇒ PParams era → CertState era → UTxO era → TxBody era → Value era

produced ∷ ShelleyEraTxBody era ⇒ PParams era → CertState era → TxBody era → Value era

Compute the lovelace which are created by the transaction

txup ∷ (EraTx era, ShelleyEraTxBody era, ProtVerAtMost era 8) ⇒ Tx era → Maybe (Update era)

module Cardano.Ledger.UTxO

Orphan instances

Crypto c ⇒ EraUTxO (ShelleyEra c)

Instance details

Associated Types

type ScriptsNeeded (ShelleyEra c) = (r ∷ Type)

Methods

getConsumedValue ∷ PParams (ShelleyEra c) → (StakeCredential (EraCrypto (ShelleyEra c)) → Maybe Coin) → UTxO (ShelleyEra c) → TxBody (ShelleyEra c) → Value (ShelleyEra c)

getScriptsNeeded ∷ UTxO (ShelleyEra c) → TxBody (ShelleyEra c) → ScriptsNeeded (ShelleyEra c)

getScriptsHashesNeeded ∷ ScriptsNeeded (ShelleyEra c) → Set (ScriptHash (EraCrypto (ShelleyEra c)))