Test.Cardano.Ledger.Shelley.ImpTest

Documentation

type ImpTestM era = ImpM (LedgerSpec era)

data LedgerSpec era

class EraTest era ⇒ EraSpecificSpec era where

Minimal complete definition

Nothing

Methods

eraSpecificSpec ∷ SpecWith (ImpInit (LedgerSpec era))

data SomeSTSEvent era

Constructors

∀ (rule ∷ Symbol).(Typeable (Event (EraRule rule era)), Eq (Event (EraRule rule era)), ToExpr (Event (EraRule rule era))) ⇒ SomeSTSEvent (Event (EraRule rule era))

data ImpTestState era

data ImpTestEnv era

Constructors

ImpTestEnv

Fields

iteFixup ∷ Tx era → ImpTestM era (Tx era)

itePostSubmitTxHook ∷ ∀ t. Globals → TRC (EraRule "LEDGER" era) → Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (State (EraRule "LEDGER" era), [Event (EraRule "LEDGER" era)]) → ImpM t ()

itePostEpochBoundaryHook ∷ ∀ t. Globals → TRC (EraRule "NEWEPOCH" era) → State (EraRule "NEWEPOCH" era) → ImpM t ()

data ImpException

Stores extra information about the failure of the unit test

Constructors

ImpException

Fields

ieAnnotation ∷ [Doc AnsiStyle]

Description of the IO action that caused the failure

ieThrownException ∷ SomeException

Exception that caused the test to fail

class (ShelleyEraTest era, STS (EraRule "BBODY" era), BaseM (EraRule "BBODY" era) ~ ShelleyBase, Environment (EraRule "BBODY" era) ~ BbodyEnv era, State (EraRule "BBODY" era) ~ ShelleyBbodyState era, Signal (EraRule "BBODY" era) ~ Block BHeaderView era, ToExpr (Event (EraRule "BBODY" era)), State (EraRule "LEDGERS" era) ~ LedgerState era, STS (EraRule "LEDGER" era), BaseM (EraRule "LEDGER" era) ~ ShelleyBase, Signal (EraRule "LEDGER" era) ~ Tx era, State (EraRule "LEDGER" era) ~ LedgerState era, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, Eq (PredicateFailure (EraRule "LEDGER" era)), Show (PredicateFailure (EraRule "LEDGER" era)), ToExpr (PredicateFailure (EraRule "LEDGER" era)), NFData (PredicateFailure (EraRule "LEDGER" era)), EncCBOR (PredicateFailure (EraRule "LEDGER" era)), DecCBOR (PredicateFailure (EraRule "LEDGER" era)), EraRuleEvent "LEDGER" era ~ Event (EraRule "LEDGER" era), Eq (EraRuleEvent "LEDGER" era), ToExpr (EraRuleEvent "LEDGER" era), NFData (EraRuleEvent "LEDGER" era), Typeable (EraRuleEvent "LEDGER" era), STS (EraRule "TICK" era), BaseM (EraRule "TICK" era) ~ ShelleyBase, Signal (EraRule "TICK" era) ~ SlotNo, State (EraRule "TICK" era) ~ NewEpochState era, Environment (EraRule "TICK" era) ~ (), NFData (PredicateFailure (EraRule "TICK" era)), EraRuleEvent "TICK" era ~ Event (EraRule "TICK" era), Eq (EraRuleEvent "TICK" era), ToExpr (EraRuleEvent "TICK" era), NFData (EraRuleEvent "TICK" era), Typeable (EraRuleEvent "TICK" era), ToExpr (PredicateFailure (EraRule "UTXOW" era)), Environment (EraRule "NEWEPOCH" era) ~ (), State (EraRule "NEWEPOCH" era) ~ NewEpochState era, Signal (EraRule "NEWEPOCH" era) ~ EpochNo, EraRuleFailure "BBODY" era ~ PredicateFailure (EraRule "BBODY" era), EraRuleFailure "LEDGER" era ~ PredicateFailure (EraRule "LEDGER" era), InjectRuleFailure "LEDGER" ShelleyDelegPredFailure era, InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era, InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure era, InjectRuleFailure "LEDGER" ShelleyPoolPredFailure era) ⇒ ShelleyEraImp era where

Minimal complete definition

impSatisfyNativeScript, modifyImpInitProtVer, fixupTx, expectTxSuccess, genRegTxCert, genUnRegTxCert

Methods

initGenesis ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m) ⇒ m (Genesis era)

default initGenesis ∷ (Monad m, Genesis era ~ NoGenesis era) ⇒ m (Genesis era)

initNewEpochState ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m) ⇒ m (NewEpochState era)

default initNewEpochState ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m, ShelleyEraImp (PreviousEra era), TranslateEra era NewEpochState, TranslationError era NewEpochState ~ Void, TranslationContext era ~ Genesis era) ⇒ m (NewEpochState era)

initImpTestState ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m) ⇒ m (ImpTestState era)

impSatisfyNativeScript

Arguments

∷ Set (KeyHash 'Witness)
Set of Witnesses that have already been satisfied

→ TxBody era
The transaction body that the script will be applied to

→ NativeScript era

→ ImpTestM era (Maybe (Map (KeyHash 'Witness) (KeyPair 'Witness)))

Try to find a sufficient number of KeyPairs that would satisfy a native script. Whenever script can't be satisfied, Nothing is returned

modifyPParams ∷ (PParams era → PParams era) → ImpTestM era ()

This modifer should change not only the current PParams, but also the future PParams. If the future PParams are not updated, then they will overwrite the mofication of the current PParams at the next epoch.

modifyImpInitProtVer ∷ ShelleyEraImp era ⇒ Version → SpecWith (ImpInit (LedgerSpec era)) → SpecWith (ImpInit (LedgerSpec era))

fixupTx ∷ HasCallStack ⇒ Tx era → ImpTestM era (Tx era)

expectTxSuccess ∷ HasCallStack ⇒ Tx era → ImpTestM era ()

genRegTxCert ∷ Credential 'Staking → ImpTestM era (TxCert era)

genUnRegTxCert ∷ Credential 'Staking → ImpTestM era (TxCert era)

data PlutusArgs

data ScriptTestContext

impWitsVKeyNeeded ∷ EraUTxO era ⇒ TxBody era → ImpTestM era (Set BootstrapAddress, Set (KeyHash 'Witness))

Figure out all the Byron Addresses that need witnesses as well as all of the KeyHashes for Shelley Key witnesses that are required.

modifyPrevPParams ∷ EraGov era ⇒ (PParams era → PParams era) → ImpTestM era ()

Modify the previous PParams in the current state with the given function. For current and future PParams, use modifyPParams

passEpoch ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ ImpTestM era ()

Runs the TICK rule until the next epoch is reached

passNEpochs ∷ ∀ era. ShelleyEraImp era ⇒ Natural → ImpTestM era ()

Runs the TICK rule until the n epochs are passed

passNEpochsChecking ∷ ∀ era. ShelleyEraImp era ⇒ Natural → ImpTestM era () → ImpTestM era ()

Runs the TICK rule until the n epochs are passed, running the checks each time.

passTick ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ ImpTestM era ()

Runs the TICK rule once

freshKeyAddr ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadGen m) ⇒ m (KeyHash 'Payment, Addr)

Generate a random Addr that uses a KeyHash, add the corresponding KeyPair to the known keys in the Imp state, and return the KeyHash as well as the Addr.

freshKeyAddr_ ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadGen m) ⇒ m Addr

Generate a random Addr that uses a KeyHash, and add the corresponding KeyPair to the known keys in the Imp state.

freshKeyHash ∷ ∀ r s g m. (HasKeyPairs s, MonadState s m, HasStatefulGen g m) ⇒ m (KeyHash r)

Generates a fresh KeyHash and stores the corresponding KeyPair in the ImpTestState. If you also need the KeyPair consider using freshKeyPair for generation or getKeyPair to look up the KeyPair corresponding to the KeyHash

freshKeyPair ∷ ∀ r s g m. (HasKeyPairs s, MonadState s m, HasStatefulGen g m) ⇒ m (KeyHash r, KeyPair r)

Generate a random KeyPair and add it to the known keys in the Imp state

getKeyPair ∷ (HasCallStack, HasKeyPairs s, MonadState s m) ⇒ KeyHash r → m (KeyPair r)

Looks up the KeyPair corresponding to the KeyHash. The KeyHash must be created with freshKeyHash for this to work.

freshByronKeyHash ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m) ⇒ m (KeyHash r)

Generates a fresh KeyHash and stores the corresponding ByronKeyPair in the ImpTestState. If you also need the ByronKeyPair consider using freshByronKeyPair for generation or getByronKeyPair to look up the ByronKeyPair corresponding to the KeyHash

freshBootstapAddress ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m) ⇒ m BootstrapAddress

getByronKeyPair ∷ (HasCallStack, HasKeyPairs s, MonadState s m) ⇒ BootstrapAddress → m ByronKeyPair

Looks up the keypair corresponding to the BootstrapAddress. The BootstrapAddress must be created with freshBootstrapAddess for this to work.

freshSafeHash ∷ ImpTestM era (SafeHash a)

Creates a fresh SafeHash

freshKeyHashVRF ∷ ImpTestM era (VRFVerKeyHash (r ∷ KeyRoleVRF))

submitTx ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era (Tx era)

submitTx_ ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era ()

submitTxAnn ∷ (HasCallStack, ShelleyEraImp era) ⇒ String → Tx era → ImpTestM era (Tx era)

submitTxAnn_ ∷ (HasCallStack, ShelleyEraImp era) ⇒ String → Tx era → ImpTestM era ()

submitFailingTx ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

Submit a transaction that is expected to be rejected with the given predicate failures. The inputs and outputs are automatically balanced.

submitFailingTxM ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → (Tx era → ImpTestM era (NonEmpty (PredicateFailure (EraRule "LEDGER" era)))) → ImpTestM era ()

Submit a transaction that is expected to be rejected, and compute the expected predicate failures from the fixed-up tx using the supplied action. The inputs and outputs are automatically balanced.

trySubmitTx ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era)), Tx era) (Tx era))

impShelleyExpectTxSuccess ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era ()

modifyNES ∷ (NewEpochState era → NewEpochState era) → ImpTestM era ()

Modify the current new epoch state with a function

getProtVer ∷ EraGov era ⇒ ImpTestM era ProtVer

getsNES ∷ SimpleGetter (NewEpochState era) a → ImpTestM era a

Get a value from the current new epoch state using the lens

getUTxO ∷ ImpTestM era (UTxO era)

impAddNativeScript ∷ ∀ era. EraScript era ⇒ NativeScript era → ImpTestM era ScriptHash

impAnn ∷ NFData a ⇒ String → ImpM t a → ImpM t a

Annotation for when failure happens. All the logging done within annotation will be discarded if there no failures within the annotation.

impAnnDoc ∷ NFData a ⇒ Doc AnsiStyle → ImpM t a → ImpM t a

impLogToExpr ∷ (HasCallStack, ToExpr a) ⇒ ImpTestM era a → ImpTestM era a

Adds the result of an action to the log, which is only shown if the test fails

runImpRule ∷ ∀ rule era. (HasCallStack, KnownSymbol rule, STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase, NFData (State (EraRule rule era)), NFData (Event (EraRule rule era)), ToExpr (Event (EraRule rule era)), Eq (Event (EraRule rule era)), Typeable (Event (EraRule rule era))) ⇒ Environment (EraRule rule era) → State (EraRule rule era) → Signal (EraRule rule era) → ImpTestM era (State (EraRule rule era))

tryRunImpRule ∷ ∀ rule era. (STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase) ⇒ Environment (EraRule rule era) → State (EraRule rule era) → Signal (EraRule rule era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule rule era))) (State (EraRule rule era), [Event (EraRule rule era)]))

tryRunImpRuleNoAssertions ∷ ∀ rule era. (STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase) ⇒ Environment (EraRule rule era) → State (EraRule rule era) → Signal (EraRule rule era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule rule era))) (State (EraRule rule era), [Event (EraRule rule era)]))

delegateStake ∷ (ShelleyEraImp era, ShelleyEraTxCert era) ⇒ Credential 'Staking → KeyHash 'StakePool → ImpTestM era ()

registerRewardAccount ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ ImpTestM era RewardAccount

registerStakeCredential ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ Credential 'Staking → ImpTestM era RewardAccount

getRewardAccountFor ∷ Credential 'Staking → ImpTestM era RewardAccount

getReward ∷ (HasCallStack, EraCertState era) ⇒ Credential 'Staking → ImpTestM era Coin

Deprecated: In favor of getBalance

lookupReward ∷ EraCertState era ⇒ Credential 'Staking → ImpTestM era (Maybe Coin)

Deprecated: In favor of lookupBalance

freshPoolParams ∷ ShelleyEraImp era ⇒ KeyHash 'StakePool → RewardAccount → ImpTestM era PoolParams

registerPool ∷ ShelleyEraImp era ⇒ KeyHash 'StakePool → ImpTestM era ()

registerPoolWithRewardAccount ∷ ShelleyEraImp era ⇒ KeyHash 'StakePool → RewardAccount → ImpTestM era ()

registerAndRetirePoolToMakeReward ∷ ShelleyEraImp era ⇒ Credential 'Staking → ImpTestM era ()

getBalance ∷ (HasCallStack, EraCertState era) ⇒ Credential 'Staking → ImpTestM era Coin

lookupBalance ∷ EraCertState era ⇒ Credential 'Staking → ImpTestM era (Maybe Coin)

getAccountBalance ∷ (HasCallStack, EraCertState era) ⇒ RewardAccount → ImpTestM era Coin

lookupAccountBalance ∷ (HasCallStack, EraCertState era) ⇒ RewardAccount → ImpTestM era (Maybe Coin)

getRewardAccountAmount ∷ (HasCallStack, EraCertState era) ⇒ RewardAccount → ImpTestM era Coin

Deprecated: In favor of getAccountBalance

shelleyFixupTx ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Tx era)

getImpRootTxOut ∷ ImpTestM era (TxIn, TxOut era)

sendValueTo ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr → Value era → ImpTestM era TxIn

sendValueTo_ ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr → Value era → ImpTestM era ()

sendCoinTo ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr → Coin → ImpTestM era TxIn

sendCoinTo_ ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr → Coin → ImpTestM era ()

expectUTxOContent ∷ (HasCallStack, ToExpr (TxOut era)) ⇒ UTxO era → [(TxIn, Maybe (TxOut era) → Bool)] → ImpTestM era ()

expectRegisteredRewardAddress ∷ (HasCallStack, EraCertState era) ⇒ RewardAccount → ImpTestM era ()

expectNotRegisteredRewardAddress ∷ (HasCallStack, EraCertState era) ⇒ RewardAccount → ImpTestM era ()

expectTreasury ∷ HasCallStack ⇒ Coin → ImpTestM era ()

disableTreasuryExpansion ∷ ShelleyEraImp era ⇒ ImpTestM era ()

updateAddrTxWits ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era (Tx era)

Adds TxWits that will satisfy all of the required key witnesses

addNativeScriptTxWits ∷ ShelleyEraImp era ⇒ Tx era → ImpTestM era (Tx era)

Modifies transaction by adding necessary scripts

addRootTxIn ∷ ShelleyEraImp era ⇒ Tx era → ImpTestM era (Tx era)

This fixup step ensures that there are enough funds in the transaction.

fixupTxOuts ∷ (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Tx era)

fixupFees ∷ (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Tx era)

fixupAuxDataHash ∷ (EraTx era, Applicative m) ⇒ Tx era → m (Tx era)

Adds an auxiliary data hash if auxiliary data present, while the hash of it is not.

impLookupNativeScript ∷ ScriptHash → ImpTestM era (Maybe (NativeScript era))

impGetUTxO ∷ ShelleyEraImp era ⇒ TxIn → ImpTestM era (TxOut era)

defaultInitNewEpochState ∷ ∀ era g s m. (MonadState s m, HasKeyPairs s, HasStatefulGen g m, MonadFail m, ShelleyEraImp era, ShelleyEraImp (PreviousEra era), TranslateEra era NewEpochState, TranslationError era NewEpochState ~ Void, TranslationContext era ~ Genesis era) ⇒ (NewEpochState (PreviousEra era) → NewEpochState (PreviousEra era)) → m (NewEpochState era)

defaultInitImpTestState ∷ ∀ era s g m. (EraGov era, EraTxOut era, HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m) ⇒ NewEpochState era → m (ImpTestState era)

impEraStartEpochNo ∷ ∀ era. Era era ⇒ EpochNo

For debugging purposes we start the era at the epoch number that matches the starting protocol version for the era times a 100

impSetSeed ∷ Int → ImpM t ()

Override the QuickCheck generator using a fixed seed.

shelleyModifyImpInitProtVer ∷ ∀ era. ShelleyEraImp era ⇒ Version → SpecWith (ImpInit (LedgerSpec era)) → SpecWith (ImpInit (LedgerSpec era))

modifyImpInitPostSubmitTxHook ∷ ∀ era. (∀ t. Globals → TRC (EraRule "LEDGER" era) → Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (State (EraRule "LEDGER" era), [Event (EraRule "LEDGER" era)]) → ImpM t ()) → SpecWith (ImpInit (LedgerSpec era)) → SpecWith (ImpInit (LedgerSpec era))

disableImpInitPostSubmitTxHook ∷ SpecWith (ImpInit (LedgerSpec era)) → SpecWith (ImpInit (LedgerSpec era))

modifyImpInitPostEpochBoundaryHook ∷ ∀ era. (∀ t. Globals → TRC (EraRule "NEWEPOCH" era) → State (EraRule "NEWEPOCH" era) → ImpM t ()) → SpecWith (ImpInit (LedgerSpec era)) → SpecWith (ImpInit (LedgerSpec era))

disableImpInitPostEpochBoundaryHook ∷ SpecWith (ImpInit (LedgerSpec era)) → SpecWith (ImpInit (LedgerSpec era))

disableInConformanceIt ∷ ShelleyEraImp era ⇒ String → ImpTestM era () → SpecWith (ImpInit (LedgerSpec era))

minorFollow ∷ ProtVer → ProtVer

A legal ProtVer that differs in the minor Version

majorFollow ∷ ProtVer → ProtVer

A legal ProtVer that moves to the next major Version

cantFollow ∷ ProtVer → ProtVer

An illegal ProtVer that skips 3 minor versions

whenMajorVersion ∷ ∀ (v ∷ Natural) era. (EraGov era, KnownNat v, MinVersion <= v, v <= MaxVersion) ⇒ ImpTestM era () → ImpTestM era ()

whenMajorVersionAtLeast ∷ ∀ (v ∷ Natural) era. (EraGov era, KnownNat v, MinVersion <= v, v <= MaxVersion) ⇒ ImpTestM era () → ImpTestM era ()

whenMajorVersionAtMost ∷ ∀ (v ∷ Natural) era. (EraGov era, KnownNat v, MinVersion <= v, v <= MaxVersion) ⇒ ImpTestM era () → ImpTestM era ()

unlessMajorVersion ∷ ∀ (v ∷ Natural) era. (EraGov era, KnownNat v, MinVersion <= v, v <= MaxVersion) ⇒ ImpTestM era () → ImpTestM era ()

getsPParams ∷ EraGov era ⇒ Lens' (PParams era) a → ImpTestM era a

withEachEraVersion ∷ ∀ era. ShelleyEraImp era ⇒ SpecWith (ImpInit (LedgerSpec era)) → Spec

impSatisfyMNativeScripts

Arguments

∷ ShelleyEraImp era

⇒ Set (KeyHash 'Witness)

→ TxBody era
Set of Witnesses that have already been satisfied

→ Int
The transaction body that the scripts will be applied to

→ StrictSeq (NativeScript era)
Number of scripts to satisfy

→ ImpTestM era (Maybe (Map (KeyHash 'Witness) (KeyPair 'Witness)))
List of scripts that can be satisfied

impSatisfySignature ∷ KeyHash 'Witness → Set (KeyHash 'Witness) → ImpTestM era (Maybe (Map (KeyHash 'Witness) (KeyPair 'Witness)))

shelleyGenRegTxCert ∷ ∀ era. ShelleyEraTxCert era ⇒ Credential 'Staking → ImpTestM era (TxCert era)

shelleyGenUnRegTxCert ∷ ∀ era. ShelleyEraTxCert era ⇒ Credential 'Staking → ImpTestM era (TxCert era)

Logging

data Doc ann

The abstract data type Doc ann represents pretty documents that have been annotated with data of type ann.

More specifically, a value of type Doc represents a non-empty set of possible layouts of a document. The layout functions select one of these possibilities, taking into account things like the width of the output document.

The annotation is an arbitrary piece of data associated with (part of) a document. Annotations may be used by the rendering backends in order to display output differently, such as

color information (e.g. when rendering to the terminal)

mouseover text (e.g. when rendering to rich HTML)

whether to show something or not (to allow simple or detailed versions)

The simplest way to display a Doc is via the Show class.

>>> putStrLn (show (vsep ["hello", "world"]))
hello
world

data AnsiStyle

Render the annotated document in a certain style. Styles not set in the annotation will use the style of the surrounding document, or the terminal’s default if none has been set yet.

style = color Green <> bold
styledDoc = annotate style "hello world"

logDoc ∷ HasCallStack ⇒ Doc AnsiStyle → ImpM t ()

Adds a Doc to the log, which is only shown if the test fails

logText ∷ HasCallStack ⇒ Text → ImpM t ()

Adds a Text to the log, which is only shown if the test fails

logString ∷ HasCallStack ⇒ String → ImpM t ()

Adds a String to the log, which is only shown if the test fails

logToExpr ∷ (HasCallStack, ToExpr a) ⇒ a → ImpM t ()

Adds a ToExpr to the log, which is only shown if the test fails

logInstantStake ∷ ToExpr (InstantStake era) ⇒ HasCallStack ⇒ ImpTestM era ()

Logs the current stake distribution

logFeeMismatch ∷ (EraGov era, EraUTxO era, HasCallStack) ⇒ Tx era → ImpTestM era ()

Combinators

withCustomFixup ∷ ((Tx era → ImpTestM era (Tx era)) → Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Compose given function with the configured fixup

withFixup ∷ (Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Replace all fixup with the given function

withNoFixup ∷ ImpTestM era a → ImpTestM era a

Performs the action without running the fix-up function on any transactions

withPostFixup ∷ (Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Apply given fixup function after the configured fixup

withPreFixup ∷ (Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Apply given fixup function before the configured fixup

impNESL ∷ Lens' (ImpTestState era) (NewEpochState era)

impGlobalsL ∷ Lens' (ImpTestState era) Globals

impLastTickG ∷ SimpleGetter (ImpTestState era) SlotNo

impKeyPairsG ∷ SimpleGetter (ImpTestState era) (Map (KeyHash 'Witness) (KeyPair 'Witness))

impNativeScriptsG ∷ SimpleGetter (ImpTestState era) (Map ScriptHash (NativeScript era))

produceScript ∷ (ShelleyEraImp era, HasCallStack) ⇒ ScriptHash → ImpTestM era TxIn

advanceToPointOfNoReturn ∷ ImpTestM era ()

simulateThenRestore ∷ ImpTestM era a → ImpTestM era a

Runs a simulation action and then restores the ledger state to what it was before the simulation started. This is useful for testing or running actions whose effects on the ledger state should not persist. The return value of the simulation is preserved, but any changes to the internal state (e.g., the UTxO set, protocol parameters, etc.) are discarded and replaced with the original snapshot.

ImpSpec re-exports

data ImpM t a

data ImpInit t