Test.Cardano.Ledger.Shelley.ImpTest

Documentation

data ImpTestM era a

data SomeSTSEvent era

Constructors

∀ (rule ∷ Symbol).(Typeable (Event (EraRule rule era)), Eq (Event (EraRule rule era)), ToExpr (Event (EraRule rule era))) ⇒ SomeSTSEvent (Event (EraRule rule era))

runImpTestM ∷ ShelleyEraImp era ⇒ Maybe Int → ImpTestState era → ImpTestM era b → IO (b, ImpTestState era)

runImpTestM_ ∷ ShelleyEraImp era ⇒ Maybe Int → ImpTestState era → ImpTestM era b → IO ()

evalImpTestM ∷ ShelleyEraImp era ⇒ Maybe Int → ImpTestState era → ImpTestM era b → IO b

execImpTestM ∷ ShelleyEraImp era ⇒ Maybe Int → ImpTestState era → ImpTestM era b → IO (ImpTestState era)

runImpTestGenM ∷ ShelleyEraImp era ⇒ ImpTestState era → ImpTestM era b → Gen (IO (b, ImpTestState era))

runImpTestGenM_ ∷ ShelleyEraImp era ⇒ ImpTestState era → ImpTestM era b → Gen (IO ())

evalImpTestGenM ∷ ShelleyEraImp era ⇒ ImpTestState era → ImpTestM era b → Gen (IO b)

execImpTestGenM ∷ ShelleyEraImp era ⇒ ImpTestState era → ImpTestM era b → Gen (IO (ImpTestState era))

data ImpTestState era

data ImpTestEnv era

Constructors

ImpTestEnv

Fields

iteState ∷ !(IORef (ImpTestState era))

iteFixup ∷ Tx era → ImpTestM era (Tx era)

iteQuickCheckSize ∷ !Int

iteCborRoundTripFailures ∷ !Bool

Expect failures in CBOR round trip serialization tests for predicate failures

data ImpException

Stores extra information about the failure of the unit test

Constructors

ImpException

Fields

ieAnnotation ∷ [Doc AnsiStyle]

Description of the IO action that caused the failure

ieThrownException ∷ SomeException

Exception that caused the test to fail

class (EraGov era, EraUTxO era, EraTxOut era, EraPParams era, ShelleyEraTxCert era, ShelleyEraScript era, ToExpr (Tx era), NFData (Tx era), ToExpr (TxBody era), ToExpr (TxOut era), ToExpr (Value era), ToExpr (PParams era), ToExpr (PParamsHKD Identity era), ToExpr (PParamsHKD StrictMaybe era), Show (NewEpochState era), ToExpr (NewEpochState era), ToExpr (GovState era), Eq (StashedAVVMAddresses era), Show (StashedAVVMAddresses era), ToExpr (StashedAVVMAddresses era), NFData (StashedAVVMAddresses era), Default (StashedAVVMAddresses era), STS (EraRule "BBODY" era), BaseM (EraRule "BBODY" era) ~ ShelleyBase, Environment (EraRule "BBODY" era) ~ BbodyEnv era, State (EraRule "BBODY" era) ~ ShelleyBbodyState era, Signal (EraRule "BBODY" era) ~ Block (BHeaderView (EraCrypto era)) era, State (EraRule "LEDGERS" era) ~ LedgerState era, STS (EraRule "LEDGER" era), BaseM (EraRule "LEDGER" era) ~ ShelleyBase, Signal (EraRule "LEDGER" era) ~ Tx era, State (EraRule "LEDGER" era) ~ LedgerState era, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, Eq (PredicateFailure (EraRule "LEDGER" era)), Show (PredicateFailure (EraRule "LEDGER" era)), ToExpr (PredicateFailure (EraRule "LEDGER" era)), NFData (PredicateFailure (EraRule "LEDGER" era)), EncCBOR (PredicateFailure (EraRule "LEDGER" era)), DecCBOR (PredicateFailure (EraRule "LEDGER" era)), EraRuleEvent "LEDGER" era ~ Event (EraRule "LEDGER" era), Eq (EraRuleEvent "LEDGER" era), ToExpr (EraRuleEvent "LEDGER" era), NFData (EraRuleEvent "LEDGER" era), Typeable (EraRuleEvent "LEDGER" era), STS (EraRule "TICK" era), BaseM (EraRule "TICK" era) ~ ShelleyBase, Signal (EraRule "TICK" era) ~ SlotNo, State (EraRule "TICK" era) ~ NewEpochState era, Environment (EraRule "TICK" era) ~ (), NFData (PredicateFailure (EraRule "TICK" era)), EraRuleEvent "TICK" era ~ Event (EraRule "TICK" era), Eq (EraRuleEvent "TICK" era), ToExpr (EraRuleEvent "TICK" era), NFData (EraRuleEvent "TICK" era), Typeable (EraRuleEvent "TICK" era), ToExpr (PredicateFailure (EraRule "UTXOW" era)), DSIGN (EraCrypto era) ~ Ed25519DSIGN, NFData (VerKeyDSIGN (DSIGN (EraCrypto era))), VRFAlgorithm (VRF (EraCrypto era)), HashAlgorithm (HASH (EraCrypto era)), DSIGNAlgorithm (DSIGN (EraCrypto era)), Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody), ADDRHASH (EraCrypto era) ~ Blake2b_224) ⇒ ShelleyEraImp era where

Minimal complete definition

impSatisfyNativeScript, fixupTx

Methods

initGenesis ∷ (HasKeyPairs s (EraCrypto era), MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (Genesis era)

default initGenesis ∷ (Monad m, Genesis era ~ NoGenesis era) ⇒ m (Genesis era)

initNewEpochState ∷ (HasKeyPairs s (EraCrypto era), MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (NewEpochState era)

default initNewEpochState ∷ (HasKeyPairs s (EraCrypto era), MonadState s m, HasStatefulGen (StateGenM s) m, ShelleyEraImp (PreviousEra era), TranslateEra era NewEpochState, TranslationError era NewEpochState ~ Void, TranslationContext era ~ Genesis era, EraCrypto era ~ EraCrypto (PreviousEra era)) ⇒ m (NewEpochState era)

initImpTestState ∷ (HasKeyPairs s (EraCrypto era), MonadState s m, HasSubState s, SubState s ~ StateGen QCGen, HasStatefulGen (StateGenM s) m) ⇒ m (ImpTestState era)

impSatisfyNativeScript

Arguments

∷ Set (KeyHash 'Witness (EraCrypto era))
Set of Witnesses that have already been satisfied

→ NativeScript era

→ ImpTestM era (Maybe (Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era))))

Try to find a sufficient number of KeyPairs that would satisfy a native script. Whenever script can't be satisfied, Nothing is returned

modifyPParams ∷ (PParams era → PParams era) → ImpTestM era ()

This modifer should change not only the current PParams, but also the future PParams. If the future PParams are not updated, then they will overwrite the mofication of the current PParams at the next epoch.

fixupTx ∷ HasCallStack ⇒ Tx era → ImpTestM era (Tx era)

data PlutusArgs

data ScriptTestContext

impWitsVKeyNeeded ∷ EraUTxO era ⇒ TxBody era → ImpTestM era (Set (BootstrapAddress (EraCrypto era)), Set (KeyHash 'Witness (EraCrypto era)))

Figure out all the Byron Addresses that need witnesses as well as all of the KeyHashes for Shelley Key witnesses that are required.

modifyPrevPParams ∷ EraGov era ⇒ (PParams era → PParams era) → ImpTestM era ()

Modify the previous PParams in the current state with the given function. For current and future PParams, use modifyPParams

passEpoch ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ ImpTestM era ()

Runs the TICK rule until the next epoch is reached

passNEpochs ∷ ∀ era. ShelleyEraImp era ⇒ Natural → ImpTestM era ()

Runs the TICK rule until the n epochs are passed

passNEpochsChecking ∷ ∀ era. ShelleyEraImp era ⇒ Natural → ImpTestM era () → ImpTestM era ()

Runs the TICK rule until the n epochs are passed, running the checks each time.

passTick ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ ImpTestM era ()

Runs the TICK rule once

freshKeyAddr ∷ (HasKeyPairs s c, MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (KeyHash r c, Addr c)

Generate a random Addr that uses a KeyHash, add the corresponding KeyPair to the known keys in the Imp state, and return the KeyHash as well as the Addr.

freshKeyAddr_ ∷ (HasKeyPairs s c, MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (Addr c)

Generate a random Addr that uses a KeyHash, and add the corresponding KeyPair to the known keys in the Imp state.

freshKeyHash ∷ (HasKeyPairs s c, MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (KeyHash r c)

Generates a fresh KeyHash and stores the corresponding KeyPair in the ImpTestState. If you also need the KeyPair consider using freshKeyPair for generation or lookupKeyPair to look up the KeyPair corresponding to the KeyHash

freshKeyPair ∷ (HasKeyPairs s c, MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (KeyHash r c, KeyPair r c)

Generate a random KeyPair and add it to the known keys in the Imp state

lookupKeyPair ∷ (HasCallStack, HasKeyPairs s c, MonadState s m) ⇒ KeyHash r c → m (KeyPair r c)

Looks up the KeyPair corresponding to the KeyHash. The KeyHash must be created with freshKeyHash for this to work.

freshByronKeyHash ∷ (HasKeyPairs s c, MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (KeyHash r c)

Generates a fresh KeyHash and stores the corresponding ByronKeyPair in the ImpTestState. If you also need the ByronKeyPair consider using freshByronKeyPair for generation or lookupByronKeyPair to look up the ByronKeyPair corresponding to the KeyHash

freshBootstapAddress ∷ (HasKeyPairs s c, MonadState s m, HasStatefulGen (StateGenM s) m) ⇒ m (BootstrapAddress c)

lookupByronKeyPair ∷ (HasCallStack, HasKeyPairs s c, MonadState s m) ⇒ BootstrapAddress c → m ByronKeyPair

Looks up the keypair corresponding to the BootstrapAddress. The BootstrapAddress must be created with freshBootstrapAddess for this to work.

freshSafeHash ∷ Era era ⇒ ImpTestM era (SafeHash (EraCrypto era) a)

Creates a fresh SafeHash

freshKeyHashVRF ∷ Era era ⇒ ImpTestM era (Hash (HASH (EraCrypto era)) (VerKeyVRF (EraCrypto era)))

submitTx ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era (Tx era)

submitTx_ ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era ()

submitTxAnn ∷ (HasCallStack, ShelleyEraImp era) ⇒ String → Tx era → ImpTestM era (Tx era)

submitTxAnn_ ∷ (HasCallStack, ShelleyEraImp era) ⇒ String → Tx era → ImpTestM era ()

submitFailingTx ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

Submit a transaction that is expected to be rejected with the given predicate failures. The inputs and outputs are automatically balanced.

submitFailingTxM ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → (Tx era → ImpTestM era (NonEmpty (PredicateFailure (EraRule "LEDGER" era)))) → ImpTestM era ()

Submit a transaction that is expected to be rejected, and compute the expected predicate failures from the fixed-up tx using the supplied action. The inputs and outputs are automatically balanced.

trySubmitTx ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era)), Tx era) (Tx era))

modifyNES ∷ (NewEpochState era → NewEpochState era) → ImpTestM era ()

Modify the current new epoch state with a function

getProtVer ∷ EraGov era ⇒ ImpTestM era ProtVer

getsNES ∷ SimpleGetter (NewEpochState era) a → ImpTestM era a

Get a value from the current new epoch state using the lens

getUTxO ∷ ImpTestM era (UTxO era)

impAddNativeScript ∷ ∀ era. EraScript era ⇒ NativeScript era → ImpTestM era (ScriptHash (EraCrypto era))

impAnn ∷ NFData a ⇒ String → ImpTestM era a → ImpTestM era a

Annotation for when failure happens. All the logging done within annotation will be discarded if there no failures within the annotation.

impAnnDoc ∷ NFData a ⇒ Doc AnsiStyle → ImpTestM era a → ImpTestM era a

impLogToExpr ∷ (HasCallStack, ToExpr a) ⇒ ImpTestM era a → ImpTestM era a

Adds the result of an action to the log, which is only shown if the test fails

runImpRule ∷ ∀ rule era. (HasCallStack, KnownSymbol rule, STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase, NFData (State (EraRule rule era)), NFData (Event (EraRule rule era)), ToExpr (Event (EraRule rule era)), Eq (Event (EraRule rule era)), Typeable (Event (EraRule rule era))) ⇒ Environment (EraRule rule era) → State (EraRule rule era) → Signal (EraRule rule era) → ImpTestM era (State (EraRule rule era))

tryRunImpRule ∷ ∀ rule era. (STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase) ⇒ Environment (EraRule rule era) → State (EraRule rule era) → Signal (EraRule rule era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule rule era))) (State (EraRule rule era), [Event (EraRule rule era)]))

delegateStake ∷ ShelleyEraImp era ⇒ Credential 'Staking (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → ImpTestM era ()

registerRewardAccount ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ ImpTestM era (RewardAccount (EraCrypto era))

registerStakeCredential ∷ ∀ era. (HasCallStack, ShelleyEraImp era) ⇒ Credential 'Staking (EraCrypto era) → ImpTestM era (RewardAccount (EraCrypto era))

getRewardAccountFor ∷ Credential 'Staking (EraCrypto era) → ImpTestM era (RewardAccount (EraCrypto era))

lookupReward ∷ HasCallStack ⇒ Credential 'Staking (EraCrypto era) → ImpTestM era Coin

poolParams ∷ ShelleyEraImp era ⇒ KeyHash 'StakePool (EraCrypto era) → RewardAccount (EraCrypto era) → ImpTestM era (PoolParams (EraCrypto era))

registerPool ∷ ShelleyEraImp era ⇒ KeyHash 'StakePool (EraCrypto era) → ImpTestM era ()

registerPoolWithRewardAccount ∷ ShelleyEraImp era ⇒ KeyHash 'StakePool (EraCrypto era) → RewardAccount (EraCrypto era) → ImpTestM era ()

registerAndRetirePoolToMakeReward ∷ ShelleyEraImp era ⇒ Credential 'Staking (EraCrypto era) → ImpTestM era ()

getRewardAccountAmount ∷ RewardAccount (EraCrypto era) → ImpTestM era Coin

withImpState ∷ ShelleyEraImp era ⇒ SpecWith (ImpTestState era) → Spec

withImpStateModified ∷ ∀ era. ShelleyEraImp era ⇒ (ImpTestState era → ImpTestState era) → SpecWith (ImpTestState era) → Spec

shelleyFixupTx ∷ ∀ era. (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Tx era)

lookupImpRootTxOut ∷ ImpTestM era (TxIn (EraCrypto era), TxOut era)

sendValueTo ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr (EraCrypto era) → Value era → ImpTestM era (TxIn (EraCrypto era))

sendCoinTo ∷ (ShelleyEraImp era, HasCallStack) ⇒ Addr (EraCrypto era) → Coin → ImpTestM era (TxIn (EraCrypto era))

expectRegisteredRewardAddress ∷ RewardAccount (EraCrypto era) → ImpTestM era ()

expectNotRegisteredRewardAddress ∷ RewardAccount (EraCrypto era) → ImpTestM era ()

expectTreasury ∷ HasCallStack ⇒ Coin → ImpTestM era ()

updateAddrTxWits ∷ (HasCallStack, ShelleyEraImp era) ⇒ Tx era → ImpTestM era (Tx era)

Adds TxWits that will satisfy all of the required key witnesses

addNativeScriptTxWits ∷ ShelleyEraImp era ⇒ Tx era → ImpTestM era (Tx era)

Modifies transaction by adding necessary scripts

addRootTxIn ∷ ShelleyEraImp era ⇒ Tx era → ImpTestM era (Tx era)

This fixup step ensures that there are enough funds in the transaction.

fixupTxOuts ∷ ShelleyEraImp era ⇒ Tx era → ImpTestM era (Tx era)

fixupFees ∷ (ShelleyEraImp era, HasCallStack) ⇒ Tx era → ImpTestM era (Tx era)

fixupAuxDataHash ∷ (EraTx era, Applicative m) ⇒ Tx era → m (Tx era)

Adds an auxiliary data hash if auxiliary data present, while the hash of it is not.

impGetNativeScript ∷ ScriptHash (EraCrypto era) → ImpTestM era (Maybe (NativeScript era))

impLookupUTxO ∷ ShelleyEraImp era ⇒ TxIn (EraCrypto era) → ImpTestM era (TxOut era)

defaultInitNewEpochState ∷ ∀ era s m. (MonadState s m, HasKeyPairs s (EraCrypto era), HasStatefulGen (StateGenM s) m, ShelleyEraImp era, ShelleyEraImp (PreviousEra era), TranslateEra era NewEpochState, TranslationError era NewEpochState ~ Void, TranslationContext era ~ Genesis era, EraCrypto era ~ EraCrypto (PreviousEra era)) ⇒ (NewEpochState (PreviousEra era) → NewEpochState (PreviousEra era)) → m (NewEpochState era)

defaultInitImpTestState ∷ ∀ era s m. (EraGov era, EraTxOut era, DSIGN (EraCrypto era) ~ Ed25519DSIGN, ADDRHASH (EraCrypto era) ~ Blake2b_224, HasKeyPairs s (EraCrypto era), MonadState s m, HasStatefulGen (StateGenM s) m, HasSubState s, SubState s ~ StateGen QCGen) ⇒ NewEpochState era → m (ImpTestState era)

impEraStartEpochNo ∷ ∀ era. Era era ⇒ EpochNo

For debugging purposes we start the era at the epoch number that matches the starting protocol version for the era times a 100

impSetSeed ∷ Int → ImpTestM era ()

Override the QuickCheck generator using a fixed seed.

Logging

data Doc ann

The abstract data type Doc ann represents pretty documents that have been annotated with data of type ann.

More specifically, a value of type Doc represents a non-empty set of possible layouts of a document. The layout functions select one of these possibilities, taking into account things like the width of the output document.

The annotation is an arbitrary piece of data associated with (part of) a document. Annotations may be used by the rendering backends in order to display output differently, such as

color information (e.g. when rendering to the terminal)

mouseover text (e.g. when rendering to rich HTML)

whether to show something or not (to allow simple or detailed versions)

The simplest way to display a Doc is via the Show class.

>>> putStrLn (show (vsep ["hello", "world"]))
hello
world

data AnsiStyle

Render the annotated document in a certain style. Styles not set in the annotation will use the style of the surrounding document, or the terminal’s default if none has been set yet.

style = color Green <> bold
styledDoc = annotate style "hello world"

logDoc ∷ HasCallStack ⇒ Doc AnsiStyle → ImpTestM era ()

Adds a Doc to the log, which is only shown if the test fails

logString ∷ HasCallStack ⇒ String → ImpTestM era ()

Adds a String to the log, which is only shown if the test fails

logToExpr ∷ (HasCallStack, ToExpr a) ⇒ a → ImpTestM era ()

Adds a ToExpr to the log, which is only shown if the test fails

logStakeDistr ∷ HasCallStack ⇒ ImpTestM era ()

Logs the current stake distribution

logFeeMismatch ∷ (EraGov era, EraUTxO era, HasCallStack) ⇒ Tx era → ImpTestM era ()

Combinators

withCustomFixup ∷ ((Tx era → ImpTestM era (Tx era)) → Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Compose given function with the configured fixup

withFixup ∷ (Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Replace all fixup with the given function

withNoFixup ∷ ImpTestM era a → ImpTestM era a

Performs the action without running the fix-up function on any transactions

withPostFixup ∷ (Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Apply given fixup function after the configured fixup

withPreFixup ∷ (Tx era → ImpTestM era (Tx era)) → ImpTestM era a → ImpTestM era a

Apply given fixup function before the configured fixup

withCborRoundTripFailures ∷ ImpTestM era a → ImpTestM era a

impNESL ∷ Lens' (ImpTestState era) (NewEpochState era)

impGlobalsL ∷ Lens' (ImpTestState era) Globals

impLastTickG ∷ SimpleGetter (ImpTestState era) SlotNo

impKeyPairsG ∷ SimpleGetter (ImpTestState era) (Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era)))

impNativeScriptsG ∷ SimpleGetter (ImpTestState era) (Map (ScriptHash (EraCrypto era)) (NativeScript era))

produceScript ∷ ShelleyEraImp era ⇒ ScriptHash (EraCrypto era) → ImpTestM era (TxIn (EraCrypto era))

advanceToPointOfNoReturn ∷ ImpTestM era ()